% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{add_opacity}
\alias{add_opacity}
\title{Add Opacity to Hexadecimal Colors}
\usage{
add_opacity(hexcol, opacity)
}
\arguments{
\item{hexcol}{A character vector of hexadecimal color values.}

\item{opacity}{A numeric value specifying the opacity to be added, ranging from 0 (transparent) to 1 (opaque).}
}
\value{
A character vector of the same length as `hexcol`, containing the RGBA color values with added opacity.
}
\description{
This function adds opacity to a vector of hexadecimal color values.
}
\details{
This function takes a character vector of hexadecimal color values `hexcol` and 
a numeric value `opacity` as input and adds the specified opacity to each color value. 
The function uses the `toRGB` function from the `plotly` package to convert the hexadecimal 
color values to RGBA format, then modifies the alpha channel according to the specified `opacity` value. 
The resulting RGBA color values are returned as a character vector.
}
\examples{
\dontrun{
# example code
hexcol <- c("#FF0000", "#00FF00", "#0000FF")
add_opacity(hexcol, 0.5)
}
}
