% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{ai.cmap_create_infographic}
\alias{ai.cmap_create_infographic}
\title{Create an infographic from a drug connectivity report.}
\usage{
ai.cmap_create_infographic(
  report,
  model,
  filename,
  aspectratio = c("4:3", "16:9", "3:4")[2],
  add.fallback = TRUE
)
}
\description{
Create an infographic from a drug connectivity report.
}
