% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-image.R
\name{ai.create_image}
\alias{ai.create_image}
\title{Generate image with generative AI model. When specified with
multiple models, it will try next model if a model fails.}
\usage{
ai.create_image(
  prompt,
  model = IMAGE_MODELS,
  format = c("file", "base64", "raw")[1],
  filename = "image.png",
  aspect_ratio = "16:9"
)
}
\description{
Generate image with generative AI model. When specified with
multiple models, it will try next model if a model fails.
}
