% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-image.R
\name{ai.create_image_openai}
\alias{ai.create_image_openai}
\title{Generate image with OpenAI's dallE. Note the limitation of the
prompt of about 1000 characters.}
\usage{
ai.create_image_openai(
  prompt,
  model = "gpt-image-2",
  size = 1024,
  aspect_ratio = "3:2",
  format = c("file", "base64", "raw"),
  filename = "image.png",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  base_url = "https://api.openai.com/v1",
  response_format = NULL,
  user = NULL,
  organization = NULL
)
}
\description{
Generate image with OpenAI's dallE. Note the limitation of the
prompt of about 1000 characters.
}
