% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-image.R
\name{ai.create_infographic}
\alias{ai.create_infographic}
\title{Given a report this function creates an infographic by calling
Gemini3. The genAI model is given the report and asked to adhere to
the included or external given diagram (in DOT format).}
\usage{
ai.create_infographic(
  report,
  model,
  diagram = NULL,
  aspectratio = c("4:3", "16:9", "3:4")[2],
  prompt = NULL,
  add.fallback = FALSE,
  format = c("file", "image")[1],
  filename = "infographic.png"
)
}
\description{
Given a report this function creates an infographic by calling
Gemini3. The genAI model is given the report and asked to adhere to
the included or external given diagram (in DOT format).
}
