% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-summary.R, R/pgx-drugs.R
\name{ai.summarize_drug_connectivity}
\alias{ai.summarize_drug_connectivity}
\title{Summarize drug connectivity results}
\usage{
ai.summarize_drug_connectivity(
  pgx,
  ct,
  model,
  drugs = NULL,
  db = 1,
  ntop = 10,
  ntop2 = 50
)

ai.summarize_drug_connectivity(
  pgx,
  ct,
  model,
  drugs = NULL,
  db = 1,
  ntop = 10,
  ntop2 = 50
)
}
\description{
Summarize drug connectivity results

Summarize drug connectivity results
}
