% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R, R/pgx-ensembl.R
\name{allSpecies.ANNOTHUB}
\alias{allSpecies.ANNOTHUB}
\title{Return all species that are supported by the ANNOTHUB annotation
engine.}
\usage{
allSpecies.ANNOTHUB()

allSpecies.ANNOTHUB()
}
\value{
character vector of species names

character vector of species names
}
\description{
Return all species that are supported by the ANNOTHUB annotation
engine.

Return all species that are supported by the ANNOTHUB annotation
engine.
}
