% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{allSpecies.ORTHOGENE}
\alias{allSpecies.ORTHOGENE}
\title{Return all species that are supported by the ORTHOGENE annotation
engine.}
\usage{
allSpecies.ORTHOGENE()
}
\value{
character vector of species names
}
\description{
Return all species that are supported by the ORTHOGENE annotation
engine.
}
