% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{annotate_phospho_residue}
\alias{annotate_phospho_residue}
\title{Annotate phosphosite with residue symbol. Feature names must be of
form 'uniprot_position'. NOTE!!! Annotation is currently done here
in feature name but it would be 'better' to add phosphosite
modification type in the pgx$genes general annotation table.}
\usage{
annotate_phospho_residue(features, detect.only = FALSE)
}
\description{
Annotate phosphosite with residue symbol. Feature names must be of
form 'uniprot_position'. NOTE!!! Annotation is currently done here
in feature name but it would be 'better' to add phosphosite
modification type in the pgx$genes general annotation table.
}
