% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{annotate_phosphotype}
\alias{annotate_phosphotype}
\title{Annotate phosphosite with residue symbol. Feature names must be of
form 'uniprot_position'.}
\usage{
annotate_phosphotype(features, organism = NA, detect.only = FALSE)
}
\description{
Annotate phosphosite with residue symbol. Feature names must be of
form 'uniprot_position'.
}
