% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{AnnotationDbi_select_2pass}
\alias{AnnotationDbi_select_2pass}
\title{Two-pass implemenation of AnnotationDbi::select that checks for
original probe names and then matches clean names.}
\usage{
AnnotationDbi_select_2pass(orgdb, keys, columns, keytype, verbose = 1)
}
\description{
Two-pass implemenation of AnnotationDbi::select that checks for
original probe names and then matches clean names.
}
