% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{averageByGroup}
\alias{averageByGroup}
\title{Calculate group means}
\usage{
averageByGroup(mat, group, FUN = mean)
}
\arguments{
\item{mat}{Numeric matrix with columns as samples.}

\item{group}{Grouping vector or factor.}

\item{FUN}{Function for aggregation. Default is mean.}
}
\value{
Matrix with group means.
}
\description{
Calculates the column means within groups defined by a grouping variable.
}
\details{
This function calculates the column means of \code{X} within groups
 defined by \code{y}. It calculates the mean for each column within each
group. The output is a matrix with rows corresponding to groups and columns
corresponding to samples.
}
\examples{
\dontrun{
mat <- matrix(rnorm(100), ncol = 10)
groups <- gl(2, 5)
means <- averageByGroup(mat, groups)
}

}
