% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-rankcor.r
\name{avgRCLR_sparse}
\alias{avgRCLR_sparse}
\title{Fast version of average ranked-CLR for sparse matrices.}
\usage{
avgRCLR_sparse(X, matG)
}
\description{
Fast version of average ranked-CLR for sparse matrices.
}
