% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{betaToM}
\alias{betaToM}
\title{betaToM}
\usage{
betaToM(beta, offset = 1e-06)
}
\arguments{
\item{X}{Matrix of Beta values. Probes in rows; samples in columns.}
}
\value{
Matrix of M values in log2-scale
}
\description{
Convert Beta to M values; 450K+850K Meth Array.
}
