% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pheno.R
\name{binarizeNumericalColumns}
\alias{binarizeNumericalColumns}
\title{Binarize numerical columns in phenotype matrix}
\usage{
binarizeNumericalColumns(S)
}
\arguments{
\item{S}{Data frame containing phenotype data}
}
\value{
Discretized phenotype data frame
}
\description{
Binarize numerical columns in phenotype matrix
}
\details{
This function binarizes numerical columns of phenotype
  data frame \code{df} into categorical variables.  Numerical
  variables are converted into "high" and "low" values based on the
  median.  Categorical variables are kept as factors.

The number of categories is controlled by \code{min.ncat} and \code{max.ncat} parameters.
Duplicate factor levels can be removed by setting \code{remove.dup = TRUE}.
}
