% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{cached.csv}
\alias{cached.csv}
\title{Cache a CSV file locally or on S3}
\usage{
cached.csv(file, FUN = read.csv, force = FALSE, ...)
}
\arguments{
\item{file}{Character string specifying the path to the CSV file to cache}

\item{FUN}{Function to use for reading the CSV file. Default is read.csv.}

\item{...}{Other arguments passed to FUN.}

\item{bucket}{Character string specifying the S3 bucket name if caching on S3. Default is NULL for local caching.}
}
\value{
A data frame containing the contents of the cached CSV file.
}
\description{
Cache a CSV file locally or on S3
}
\details{
This function checks if a cached copy of the CSV file exists locally in /tmp or on S3. 
If not, it reads the file using FUN(), caches it locally or on S3, and returns the contents.
On subsequent calls it returns the cached copy instead of re-reading the file.
}
\examples{
\dontrun{
# Cache locally
df <- cached.csv("data.csv") 

# Cache on S3  
df <- cached.csv("s3://mybucket/data.csv", bucket = "mybucket")
}
}
