% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{calculateCompoundSignificance}
\alias{calculateCompoundSignificance}
\title{Calculate compound significance scores per gene}
\usage{
calculateCompoundSignificance(
  wgcna,
  collapse = TRUE,
  sort.by = "score1",
  digits = 4
)
}
\arguments{
\item{wgcna}{WGCNA result object with stats.}
}
\value{
Data frame of compound significance scores.
}
\description{
Calculate compound significance scores per gene
}
