% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{cbind_sparse_matrix}
\alias{cbind_sparse_matrix}
\title{Merge Sparse Matrix}
\usage{
cbind_sparse_matrix(m1, m2)
}
\arguments{
\item{m1}{The first sparse matrix.}

\item{m2}{The second sparse matrix.}
}
\value{
A merged sparse matrix with combined columns and shared
  features on rows
}
\description{
Merges two sparse matrices by combining column-wise (aka cbind) but
ensuring the same gene order. Union of features (gene/proteins) are
on rows (i.e. rownames are features names). Commonly needed for
merging multiple pgx$GMT matrices (gene on rows).
}
