% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{CCLE_RNA_CANCERTYPE}
\alias{CCLE_RNA_CANCERTYPE}
\title{A dataset used by `compute_deconvolution`}
\format{
## `CCLE_RNA_CANCERTYPE`
}
\source{
unknown
}
\usage{
CCLE_RNA_CANCERTYPE
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
