% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{check_duplicate_cols.DEPRECATED}
\alias{check_duplicate_cols.DEPRECATED}
\title{Input Checks}
\usage{
check_duplicate_cols.DEPRECATED(data)
}
\arguments{
\item{data}{A data frame or matrix.}
}
\value{
Logical indicating output of the checks.
}
\description{
Checks if a data frame or matrix contains duplicate column names.
}
\details{
This function takes a data frame or matrix \code{data} as input and checks if it contains any duplicate column names.
It compares the column names against each other to look for duplicates.

The output is a logical value indicating whether any duplicate names were found.
\code{TRUE} means duplicate names were detected, \code{FALSE} means no duplicates.

This can be used to validate data before further analysis, to ensure no columns are duplicated.
}
\examples{
\dontrun{
data <- data.frame(A = 1:3, B = 4:6, A = 7:9)
check_duplicate_cols(data)
# Returns TRUE

data <- data.frame(A = 1:3, B = 4:6, C = 7:9)
check_duplicate_cols(data)
# Returns FALSE
}
}
