% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R, R/pgx-ensembl.R
\name{check_probetype.ORTHOGENE}
\alias{check_probetype.ORTHOGENE}
\title{Check if probes of organism are automatically recognized by
ORTHOGENE annotation engine.}
\usage{
check_probetype.ORTHOGENE(organism, probes)

check_probetype.ORTHOGENE(organism, probes)
}
\value{
TRUE  if probes are recognized

FALSE if probes are not recognized

TRUE  if probes are recognized

FALSE if probes are not recognized
}
\description{
Check if probes of organism are automatically recognized by
ORTHOGENE annotation engine.

Check if probes of organism are automatically recognized by
ORTHOGENE annotation engine.
}
