% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{check_species_probetype}
\alias{check_species_probetype}
\title{Automatically detects species by trying to detect probetype from
list of test_species. Warning. bit slow.}
\usage{
check_species_probetype(
  probes,
  test_species = c("Human", "Mouse", "Rat"),
  datatype = NULL,
  annot.cols = NULL
)
}
\description{
Automatically detects species by trying to detect probetype from
list of test_species. Warning. bit slow.
}
