% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct-diagnostics.R
\name{checkConfounders}
\alias{checkConfounders}
\title{Check for confounders}
\usage{
checkConfounders(pheno, model.par, max.rho = 0.3)
}
\arguments{
\item{pheno}{Dataframe containing sample metadata/covariates.}

\item{model.par}{Vector of column names of covariates of interest \code{pheno}.}

\item{max.rho}{Maximum allowed correlation.}
}
\value{
List of confounding and not-confounding factors. Correlation matrix rho.
}
\description{
Performs a correlation analysis on a phenotype matrix to detect possible confounding factors.
}
\details{
This function performs a correlation analysis on the factor with respect to the factor
of interest. Correlation is compute on each level of all factors and a factor is deemed confounding
if any level has a correlation larger than max.rho. For example, such factors should not be used
for batch correction as they may remove 'genuine' signal correlated with the factor of interest.
}
