% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{checkProbes}
\alias{checkProbes}
\title{Check if probes are valid for organism}
\usage{
checkProbes(organism, probes, ah = NULL)
}
\value{
TRUE    if probes match any probetype of organism

FALSE   if probes do not any probetype of organism
}
\description{
Check if probes are valid for organism
}
