% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{clean_gmt}
\alias{clean_gmt}
\title{Clean GMT}
\usage{
clean_gmt(gmt.all, gmt.db)
}
\arguments{
\item{gmt.all}{A list representing the GMT object.}

\item{gmt.db}{A character vector representing the GMT database.}
}
\value{
The cleaned GMT object.
}
\description{
Cleans the GMT (Gene Matrix Transposed) object by modifying its names and removing duplicates.
}
