% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R, R/pgx-ensembl.R
\name{clean_probe_names}
\alias{clean_probe_names}
\title{Cleanup probe names from postfixes or version numbers}
\usage{
clean_probe_names(probes)

clean_probe_names(probes)
}
\description{
Cleanup probe names from postfixes or version numbers

Cleanup probe names from postfixes or version numbers
}
