% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R, R/pgx-ensembl.R
\name{cleanupAnnotation}
\alias{cleanupAnnotation}
\title{Cleanup annotation}
\usage{
cleanupAnnotation(genes)

cleanupAnnotation(genes)
}
\description{
Cleanup annotation

Cleanup annotation
}
