% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{clustermap}
\alias{clustermap}
\title{Title}
\usage{
clustermap(
  x,
  nc = 6,
  nr = 6,
  na = 4,
  q = 0.8,
  p = 2,
  method = "multidist",
  nca = 24,
  col.annot = NULL,
  row.annot = NULL,
  plot = TRUE,
  na.fill = FALSE,
  nrlab = nrow(x),
  nclab = ncol(x),
  labrow = rownames(x),
  labcol = colnames(x),
  ...
)
}
\arguments{
\item{x}{value}

\item{nc}{value}

\item{nr}{value}

\item{na}{value}

\item{q}{value}

\item{p}{value}

\item{method}{value}

\item{nca}{value}

\item{col.annot}{value}

\item{row.annot}{value}

\item{plot}{value}

\item{na.fill}{value}

\item{nrlab}{value}

\item{nclab}{value}

\item{labrow}{value}

\item{labcol}{value}

\item{...}{value}
}
\description{
Title
}
