% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{cmap.summarize_results}
\alias{cmap.summarize_results}
\title{Summarize drug connectivity results}
\usage{
cmap.summarize_results(
  pgx,
  ct,
  model,
  drugs = NULL,
  db = 1,
  ntop = 10,
  ntop2 = 50
)
}
\description{
Summarize drug connectivity results
}
