% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{collapseTraitMatrix}
\alias{collapseTraitMatrix}
\title{Collapses an expanded (binarized) trait matrix to its original
categorical phenotype with levels. Colnames must be "pheno1=A",
"pheno1=B" etc.}
\usage{
collapseTraitMatrix(Y)
}
\description{
Collapses an expanded (binarized) trait matrix to its original
categorical phenotype with levels. Colnames must be "pheno1=A",
"pheno1=B" etc.
}
