% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{colSignedRanks}
\alias{colSignedRanks}
\title{Calculate columnwise signed ranks for matrices. Rank on absolute
value and add sign. Values are rescaled to [-1,1].}
\usage{
colSignedRanks(x)
}
\description{
Calculate columnwise signed ranks for matrices. Rank on absolute
value and add sign. Values are rescaled to [-1,1].
}
