% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-extra.R
\name{compute_cellcycle_gender}
\alias{compute_cellcycle_gender}
\title{Compute Cell Cycle and Gender Inference}
\usage{
compute_cellcycle_gender(pgx, rna.counts = pgx.countScaleMatrix(pgx))
}
\arguments{
\item{pgx}{An object containing the input data for analysis.}

\item{rna.counts}{A matrix or data frame of RNA expression counts.
Defaults to the counts in the input object.}
}
\value{
An updated object with cell cycle and gender inference results.
}
\description{
This function performs cell cycle phase inference and gender estimation based on the input RNA expression data.
The cell cycle phase inference is performed using the Seurat package.
}
\examples{
\dontrun{
pgx <- playbase::GEIGER_PGX
deconv <- compute_cellcycle_gender(pgx)
}
}
