% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_CV}
\alias{compute_CV}
\title{Compute coefficient of variation (CV, %)
This function computes coefficient of variation (CV) of each row (feature) of data matrix
It does so in the linear space.}
\usage{
compute_CV(counts, na.rm = TRUE)
}
\arguments{
\item{counts}{A numeric matrix}
}
\value{
CV (%) for each row (feature) of the input matrix.
}
\description{
Compute coefficient of variation (CV, %)
This function computes coefficient of variation (CV) of each row (feature) of data matrix
It does so in the linear space.
}
