% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-extra.R
\name{compute_drugSensitivityEnrichment}
\alias{compute_drugSensitivityEnrichment}
\title{Title}
\usage{
compute_drugSensitivityEnrichment(ngs, cmap.dir)
}
\arguments{
\item{cmap.dir}{}
}
\description{
Title
}
