% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-extra.R
\name{compute_omicsGraphs}
\alias{compute_omicsGraphs}
\title{Compute Omics Graphs}
\usage{
compute_omicsGraphs(ngs)
}
\arguments{
\item{ngs}{An object containing the input data for analysis.}
}
\value{
An updated object with omics graphs and path scores.
}
\description{
This function computes omics graphs and path scores based on the input data.
It creates the omics graph using the input object and calculates path scores using the omics graph.
It also computes a reduced graph and path scores based on the reduced graph.
}
