% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_core.R
\name{compute_pgx}
\alias{compute_pgx}
\title{Compute the pgx object}
\usage{
compute_pgx(
  pgx,
  max.genes = 19999,
  max.genesets = 9999,
  gx.methods = c("ttest.welch", "trend.limma", "edger.qlf"),
  gset.methods = c("fisher", "gsva", "fgsea"),
  do.cluster = TRUE,
  use.design = TRUE,
  prune.samples = FALSE,
  extra.methods = c("meta.go", "deconv", "infer", "drugs", "wordcloud"),
  lib.dir = "../lib",
  progress = NULL
)
}
\arguments{
\item{pgx}{datatype}

\item{max.genes}{datatype}

\item{max.genesets}{datatype}

\item{gx.methods}{datatype}

\item{gset.methods}{datatype}

\item{do.cluster}{datatype}

\item{use.design}{datatype}

\item{prune.samples}{datatype}

\item{extra.methods}{datatype}

\item{lib.dir}{datatype}

\item{progress}{datatype}
}
\value{
a list (representing a pgx object)
}
\description{
This function works on a pgx object that has already been created from
counts/samples/contrasts data using the `pgx_create()` function.
}
\details{
It performs the following functions:
 - runs all standard statistical methods on genes
 - runs all standard statistical methods on genesets
 - computes any extra statistical methods as specified
}
\examples{
counts <- read_counts(example_file("counts.csv"))
samples <- read_samples(example_file("samples.csv"))
contrasts <- read_contrasts(example_file("contrasts.csv"))
## takes too long
# mypgx <- create_pgx(counts, samples, contrasts)
# mypgx <- compute_pgx(mypgx)
}
