% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genes.R
\name{compute_testGenes}
\alias{compute_testGenes}
\title{Title}
\usage{
compute_testGenes(
  pgx,
  contr.matrix,
  max.features = 1000,
  test.methods = c("trend.limma", "deseq2.wald", "edger.qlf"),
  use.design = TRUE,
  prune.samples = FALSE,
  remove.outputs = TRUE
)
}
\arguments{
\item{remove.outputs}{}
}
\description{
Title
}
