% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{compute_testGenesets}
\alias{compute_testGenesets}
\title{Compute Test Genesets}
\usage{
compute_testGenesets(
  pgx,
  max.features = 1000,
  custom.geneset = list(gmt = NULL, info = NULL),
  test.methods = c("gsva", "camera", "fgsea"),
  remove.outputs = TRUE
)
}
\arguments{
\item{pgx}{A data object representing multi-omics data.}

\item{max.features}{Maximum number of features to consider.}

\item{custom.geneset}{Custom geneset object.}

\item{test.methods}{Methods to use for testing.}

\item{remove.outputs}{Logical indicating whether to remove large outputs.}
}
\value{
The updated \code{pgx} object with computed test genesets.
}
\description{
Computes test genesets for the given multi-omics data.
}
