% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genes.R
\name{compute_testGenesMultiOmics}
\alias{compute_testGenesMultiOmics}
\title{Title}
\usage{
compute_testGenesMultiOmics(
  pgx,
  contr.matrix,
  max.features = 1000,
  test.methods = c("trend.limma", "deseq2.wald", "edger.qlf"),
  use.design = TRUE,
  prune.samples = FALSE,
  remove.outputs = TRUE
)
}
\arguments{
\item{pgx}{value}

\item{contr.matrix}{value}

\item{max.features}{value}

\item{test.methods}{value}

\item{use.design}{value}

\item{prune.samples}{value}

\item{remove.outputs}{value}
}
\description{
Title
}
