% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genes.R
\name{compute_testGenesSingleOmics}
\alias{compute_testGenesSingleOmics}
\title{Title}
\usage{
compute_testGenesSingleOmics(
  pgx,
  contr.matrix,
  max.features = 1000,
  filter.low = TRUE,
  remove.outputs = TRUE,
  use.design = TRUE,
  prune.samples = FALSE,
  test.methods = c("trend.limma", "deseq2.wald", "edger.qlf")
)
}
\arguments{
\item{test.methods}{}
}
\description{
Title
}
