% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{contrastAsLabels}
\alias{contrastAsLabels}
\title{Title}
\usage{
contrastAsLabels(contr.matrix, as.factor = FALSE)
}
\arguments{
\item{contr.matrix}{value}

\item{as.factor}{value}
}
\description{
Title
}
