% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{contrasts.addTimeInteraction}
\alias{contrasts.addTimeInteraction}
\title{Expand contrast matrix with time-interaction terms}
\usage{
contrasts.addTimeInteraction(contrasts, samples)
}
\arguments{
\item{contrasts}{Matrix of contrasts}

\item{samples}{Sample information dataframe}
}
\value{
Contrast matrix
}
\description{
Autodetect 'time' variable in sample dataframe and expand contrasts matrix with interaction term.
}
\details{
This function takes the contrast matrix and samples dataframe as input.
It checks if 'time' variable is in samples. It then checks for valid contrasts
in the contrast matrix ie. contrasts suitable for time-interaction testing.
If any suited contrast is detected, the contrast matrix is expanded.
These contrasts will be automatically detected later for DGE testing
where time-interaction (with or without spline) will be performed.
}
