% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{contrasts.convertToLabelMatrix}
\alias{contrasts.convertToLabelMatrix}
\title{Convert old-style contrast matrix to sample-wise labeled matrix}
\usage{
contrasts.convertToLabelMatrix(contrasts, samples)
}
\arguments{
\item{contrasts}{Matrix of contrasts}

\item{samples}{Sample information dataframe}
}
\value{
Contrast matrix
}
\description{
Convert old-style contrast to new-style contrast of sample-wise labels.
}
\details{
Converts old-style contrast matrix to sample-wise labeled contrast matrix

This function takes a matrix of sample labels as input, where the column
names indicate the sample groups being compared (e.g. "Group1_vs_Group2"). It parses
the column names to extract the two groups, then constructs a labels contrast matrix by
assigning the proper condition name to samples belonging to each group.

The resulting contrast matrix has rows corresponding to samples, and columns
corresponding to the label matrix column names. This encodes the contrasts between
each pair of groups.
}
