% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{CONTRASTS}
\alias{CONTRASTS}
\title{Example contrasts data from unknown GEO dataset}
\format{
## `CONTRASTS`
data.frame with rows and columns as contrasts, values are -1, 0 or 1.
}
\source{
unknown
}
\usage{
CONTRASTS
}
\description{
Example contrasts data from unknown GEO dataset
}
\keyword{datasets}
