% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-check.R
\name{contrasts_conversion}
\alias{contrasts_conversion}
\title{Convert contrasts for OPG}
\usage{
contrasts_conversion(SAMPLES, CONTRASTS)
}
\arguments{
\item{CONTRASTS}{data.frame. The data frame corresponding to the input file as in playbase::CONTRASTS}

\item{SAMPLE}{data.frame. The data frame corresponding to the input file as in playbase::SAMPLES}
}
\value{
converted contrast df
}
\description{
Convert contrasts for OPG
}
