% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{convert.gmt}
\alias{convert.gmt}
\title{Convert long vector of genes and pathways (eg. cols from data.frame) to gmt format}
\usage{
convert.gmt(genes, gs_name)
}
\arguments{
\item{genes}{a long vector of genes}

\item{gs_name}{a long vector of gene sets/pathway names matching the genes vector length}

\item{source}{value}
}
\description{
Convert long vector of genes and pathways (eg. cols from data.frame) to gmt format
}
