% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{cor.pvalue}
\alias{cor.pvalue}
\title{P-value for Pearson's Correlation Coefficient}
\usage{
cor.pvalue(x, n)
}
\arguments{
\item{x}{A numeric value representing Pearson's correlation coefficient.}

\item{n}{A numeric value representing the sample size.}
}
\value{
A numeric value representing the p-value for Pearson's correlation coefficient.
}
\description{
This function calculates the p-value for Pearson's correlation coefficient.
}
\details{
The p-value is calculated using the standard normal distribution,
where the test statistic is calculated as `z = x / sqrt((1 - x^2) / (n - 2))`.
The p-value represents the probability of observing a correlation coefficient as
extreme or more extreme than `x` given that the null hypothesis (no correlation) is true.
}
\examples{
\dontrun{
# example code
cor.pvalue(0.8, 100)
}
}
