% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{corclust}
\alias{corclust}
\title{Correlation-based Clustering}
\usage{
corclust(x)
}
\arguments{
\item{x}{A numeric matrix or data frame containing the data to be clustered.}
}
\value{
An object of class `hclust` representing the hierarchical clustering 
of the input data.
}
\description{
This function performs hierarchical clustering on a 
dataset using a correlation-based distance metric.
}
\details{
Correlation-based Clustering

The function first calculates the pairwise correlation between 
the columns of the input data using the `cor` function from the `stats` 
package. The correlation matrix is then transformed into a distance matrix by 
subtracting it from 1. The resulting distance matrix is used as input to the 
`hclust` function from the `fastcluster` package, which performs hierarchical 
clustering using Ward's method.
}
\examples{
\dontrun{
# Generate some example data
set.seed(123)
x <- matrix(rnorm(1000), ncol = 10)

# Perform correlation-based clustering
hc <- corclust(x)

# Plot the resulting dendrogram
plot(hc)
}

}
