% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-fisher.r
\name{corpora.fastFET}
\alias{corpora.fastFET}
\title{Wrapper superfast version of Fisher Exact Test from 'corpora' R
package. This is the fastest implementation currently
available. Uses phyper inside.}
\usage{
corpora.fastFET(
  a,
  b,
  c,
  d,
  alternative = c("two.sided", "less", "greater")[3],
  log.p = FALSE
)
}
\description{
setAn ¬setA
       setB  a     b | a+b
      ¬setB  c     d | c+d
         ------------|-----
            a+c   b+d| a+b+c+d
}
