% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{correctMarchSeptemberGenes}
\alias{correctMarchSeptemberGenes}
\title{Correct gene symbols with September/March suffix}
\usage{
correctMarchSeptemberGenes(gg)
}
\arguments{
\item{gg}{A character vector of gene symbols.}
}
\value{
The character vector \code{gg} with any September/March suffixes converted to full month names.
}
\description{
Corrects gene symbols that use "Sep" or "Mar" suffix by converting them to full month names.
}
\details{
This function checks if any gene symbols in \code{gg} end with "-Sep", "-SEP", "-Mar", or "-MAR", 
which represent September or March. If found, it converts these to full month names by replacing
"-Sep" with "-September" and "-Mar" with "-March".

It first replaces any "-SEP" or "-MAR" with "-Sep" and "-Mar" respectively. 
Then it uses \code{\link[plyr]{mapvalues}} to match any "-Sep" or "-Mar" suffixes to the 
corresponding "-September" or "-March" and replaces them.
}
