% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{COUNTS}
\alias{COUNTS}
\title{Example counts data from unknown GEO dataset}
\format{
## `COUNTS`
data.frame with genes as rows and samples as columns
}
\source{
unknown
}
\usage{
COUNTS
}
\description{
Example counts data from unknown GEO dataset
}
\keyword{datasets}
