% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{darkmode}
\alias{darkmode}
\title{Dark Mode Theme}
\usage{
darkmode(p, dim = 2)
}
\arguments{
\item{p}{A plotly plot object}

\item{dim}{The number of dimensions for the plot (2 or 3)}
}
\value{
The modified plotly plot object with a dark theme.
}
\description{
Changes the theme of a plotly plot to a dark mode style.
}
\details{
This function modifies a plotly plot to use a dark theme style.
It changes the colors of the background, axes, text, etc to darker colors.

The \code{p} argument is a plotly plot object.
The \code{dim} argument specifies whether it is a 2D or 3D plot.

It sets the plot and paper background colors to dark greys.
The font and axis colors are changed to light greys.
For 3D plots the z-axis is also styled.
}
