% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{detect_decimal}
\alias{detect_decimal}
\title{Detect delimiter of text file from first 10 lines. Assumes there is
a header and rownames column.}
\usage{
detect_decimal(file)
}
\description{
Detect delimiter of text file from first 10 lines. Assumes there is
a header and rownames column.
}
