% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{detect_delim}
\alias{detect_delim}
\title{Detect delimiter of text file from header (or first line)}
\usage{
detect_delim(file, delims = c(",", "\\t", " ", "|", ":", ";"))
}
\description{
Detect delimiter of text file from header (or first line)
}
