% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{detect_probe_DEPRECATED}
\alias{detect_probe_DEPRECATED}
\title{Detect probe type from probe set}
\usage{
detect_probe_DEPRECATED(probes, organism)
}
\arguments{
\item{probes}{Character vector of gene/transcript identifiers.}

\item{organism}{Organism name (e.g. "Human", "Mouse", "Rat").}
}
\value{
The probe type with the most matches to the input identifiers.
}
\description{
Detects the most likely probe type (ENSEMBL, SYMBOL, etc) 
for a set of gene/transcript identifiers. This function has been deprecated
and is mantained for fallback purposes.
}
\details{
Checks the input identifiers against known gene/transcript 
identifiers from an organism-specific annotation package to determine the
most likely probe type. Useful for identifying the type of identifiers
in a gene expression matrix when not explicitly provided.
}
\examples{
probes <- probes <- c("NM_001081979", "NM_001081980", "NM_001081981", "NM_001081982", "NM_001081983")
organism <- "Mouse"
type <- detect_probe_DEPRECATED(probes, organism)

}
