% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{detect_probetype.ANNOTHUB}
\alias{detect_probetype.ANNOTHUB}
\title{Detect probe type from probe set}
\usage{
detect_probetype.ANNOTHUB(
  organism,
  probes,
  orgdb = NULL,
  nprobe = 1000,
  use.ah = NULL,
  datatype = NULL,
  verbose = TRUE
)
}
\description{
Detect probe type from probe set
}
