% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{detect_probetype.GPROFILER}
\alias{detect_probetype.GPROFILER}
\title{Detect/validate features with gprofiler}
\usage{
detect_probetype.GPROFILER(
  organism,
  probes,
  nprobe = 1000,
  datatype = NULL,
  verbose = TRUE
)
}
\description{
Detect/validate features with gprofiler
}
