% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{detect_probetype}
\alias{detect_probetype}
\title{Get human ortholog from given symbols of organism by using
  gprofiler2 package. This package needs internet connection.}
\usage{
detect_probetype(
  organism,
  probes,
  orgdb = NULL,
  nprobe = 1000,
  use.ah = NULL,
  datatype = NULL,
  verbose = TRUE
)
}
\description{
Get human ortholog from given symbols of organism by using
  gprofiler2 package. This package needs internet connection.

Detect probe type from probe set
}
