% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{detect_species}
\alias{detect_species}
\title{ensembl_species <- useDataset(dataset = dataset_hsa$dataset, mart = ensembl)}
\usage{
detect_species(x)
}
\description{
ensembl_species <- useDataset(dataset = dataset_hsa$dataset, mart = ensembl)
}
