% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{detect_species_probetype}
\alias{detect_species_probetype}
\title{Automatically detects species by trying to detect probetype from
list of test_species. Warning. bit slow.}
\usage{
detect_species_probetype(probes, test_species = c("human", "mouse", "rat"))
}
\description{
Automatically detects species by trying to detect probetype from
list of test_species. Warning. bit slow.
}
