% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{DICE_SIGNATURE1000}
\alias{DICE_SIGNATURE1000}
\title{A dataset used by `compute_deconvolution`}
\format{
## `DICE_SIGNATURE1000`
}
\source{
unknown
}
\usage{
DICE_SIGNATURE1000
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
