% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report.R
\name{.ai_report_wanted}
\alias{.ai_report_wanted}
\title{Should this module be (re)generated?}
\usage{
.ai_report_wanted(pgx, module, ai)
}
\description{
Should this module be (re)generated?
}
\keyword{internal}
