% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{.append_gmt_to_matrix}
\alias{.append_gmt_to_matrix}
\title{Internal use: append gmt (list) to a sparse gene set matrix}
\usage{
.append_gmt_to_matrix(gmt, G, all_genes, minsize, maxsize)
}
\description{
Internal use: append gmt (list) to a sparse gene set matrix
}
