% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{.check_probetype.GPROFILER}
\alias{.check_probetype.GPROFILER}
\title{Check if probes of organism are automatically recognized by
ORTHOGENE annotation engine.}
\usage{
.check_probetype.GPROFILER(organism, probes, min.map = 0.2)
}
\value{
TRUE  if probes are recognized

FALSE if probes are not recognized
}
\description{
Check if probes of organism are automatically recognized by
ORTHOGENE annotation engine.
}
