% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{.convert_orthologs}
\alias{.convert_orthologs}
\title{Internal helper function.}
\usage{
.convert_orthologs(
  genes,
  species,
  methods = c("homologene", "gprofiler", "babelgene", "gprofiler2", "uppercase"),
  target_species = "Human",
  verbose = 1
)
}
\description{
Internal helper function.
}
