% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{.getHumanOrtholog.GPROFILER}
\alias{.getHumanOrtholog.GPROFILER}
\title{Get human ortholog from given symbols of organism by using
  gprofiler2 package. This package needs internet connection.}
\usage{
.getHumanOrtholog.GPROFILER(query, organism, target = "hsapiens", verbose = 1)
}
\description{
Get human ortholog from given symbols of organism by using
  gprofiler2 package. This package needs internet connection.
}
