% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{.ngs.fitConstrastsWithDESEQ2.nodesign}
\alias{.ngs.fitConstrastsWithDESEQ2.nodesign}
\title{Title}
\usage{
.ngs.fitConstrastsWithDESEQ2.nodesign(
  counts,
  contr.matrix,
  test = "Wald",
  prune.samples = FALSE,
  conform.output = FALSE,
  X = NULL
)
}
\arguments{
\item{X}{}
}
\description{
Title
}
