% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{.ngs.fitContrastsWithEDGER.nodesign.pruned}
\alias{.ngs.fitContrastsWithEDGER.nodesign.pruned}
\title{Title}
\usage{
.ngs.fitContrastsWithEDGER.nodesign.pruned(
  counts,
  contr.matrix,
  group = NULL,
  method = c("qlf", "lrt"),
  X = NULL,
  conform.output = FALSE,
  robust = TRUE,
  plot = TRUE
)
}
\arguments{
\item{counts}{value}

\item{contr.matrix}{value}

\item{group}{value}

\item{method}{value}

\item{X}{value}

\item{conform.output}{value}

\item{robust}{value}

\item{plot}{value}
}
\description{
Title
}
