% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{.ngs.fitContrastsWithEDGER.nodesign}
\alias{.ngs.fitContrastsWithEDGER.nodesign}
\title{Title}
\usage{
.ngs.fitContrastsWithEDGER.nodesign(
  dge,
  contr.matrix,
  method = c("qlf", "lrt"),
  X = NULL,
  conform.output = FALSE,
  robust = TRUE,
  plot = TRUE
)
}
\arguments{
\item{plot}{}
}
\description{
Title
}
