% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_utils.R
\name{.probe_to_symbol}
\alias{.probe_to_symbol}
\title{Convert probe names to symbol names}
\usage{
.probe_to_symbol(probes, type = NULL, org = "human", keep.na = FALSE)
}
\arguments{
\item{probes}{unknown.}

\item{type}{unknown.}

\item{org}{unknown.}

\item{keep.na}{unknown.}
}
\value{
unknown.
}
\description{
More explanation is needed for this function
}
\examples{
x <- 1
}
