% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{.query_orthologs_bridged}
\alias{.query_orthologs_bridged}
\title{Bridged lookup of genes/features to human orthologs. Uses an
intermediate bridge organism for mapping to human ortholog. This
can be useful for poorly annotated species.}
\usage{
.query_orthologs_bridged(
  genes,
  species,
  bridge_species,
  target_species = "hsapiens",
  verbose = 1
)
}
\description{
Bridged lookup of genes/features to human orthologs. Uses an
intermediate bridge organism for mapping to human ortholog. This
can be useful for poorly annotated species.
}
