% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_read.R
\name{.read_as_matrix}
\alias{.read_as_matrix}
\title{Read a file as a matrix}
\usage{
.read_as_matrix(file)
}
\arguments{
\item{file}{string. filepath to data}
}
\value{
matrix
}
\description{
This function is used to read counts/samples/contrasts data from file.
}
\examples{
x <- 1
}
