% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{double_center_scale_fast}
\alias{double_center_scale_fast}
\title{Double Center and Scale a Matrix}
\usage{
double_center_scale_fast(X)
}
\arguments{
\item{X}{A numeric matrix to be centered and scaled}
}
\value{
A transposed matrix that has been double centered and scaled
}
\description{
This function performs double centering and scaling on a matrix by:
1. Subtracting row means
2. Subtracting column means
3. Scaling by column standard deviations
}
