% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report-drugs.R
\name{drugs_assemble_prompt}
\alias{drugs_assemble_prompt}
\title{Assemble drug-connectivity static-report prompt for a single drug DB.}
\usage{
drugs_assemble_prompt(pgx, method, ai)
}
\description{
Assemble drug-connectivity static-report prompt for a single drug DB.
}
\keyword{internal}
