% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{duplicated.dgCMatrix}
\alias{duplicated.dgCMatrix}
\title{Test if rows or columns of sparse matrix are duplicated}
\usage{
\method{duplicated}{dgCMatrix}(dgCMat, MARGIN)
}
\description{
Test if rows or columns of sparse matrix are duplicated
}
