% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-getgeo.R
\name{eset.getPhenoData}
\alias{eset.getPhenoData}
\alias{eset.parsePhenoFromTitle}
\title{Extract phenotype data field from ExpressionSet}
\usage{
eset.getPhenoData(eset, field)

eset.parsePhenoFromTitle(title, split = NULL)
}
\arguments{
\item{eset}{ExpressionSet object}

\item{field}{Character string specifying phenotype data field name}
}
\value{
Vector of values for the specified field
}
\description{
Extract phenotype data field from ExpressionSet
}
\details{
Extracts a specific phenotype data field from an ExpressionSet object.
The \code{field} parameter specifies the phenotype data column name to extract.
The phenotype data pData slot is extracted from the ExpressionSet; the specified
field is returned as a vector.
}
\section{Functions}{
\itemize{
\item \code{eset.parsePhenoFromTitle()}: Phenotype data from the title of an
ExpressionSet object by splitting the title on a specified delimiter and guessing column names.

}}
