% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct-diagnostics.R
\name{estimateBatchCorrectionVectors}
\alias{estimateBatchCorrectionVectors}
\title{Estimate batch correction vectors from corrected cX and uncorrected
matrix X.}
\usage{
estimateBatchCorrectionVectors(cX, X, k = NULL, threshold = 0.8)
}
\description{
Estimate batch correction vectors from corrected cX and uncorrected
matrix X.
}
