% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\name{example_file}
\alias{example_file}
\title{Get path to omp example dataset(s)}
\usage{
example_file(file = NULL)
}
\arguments{
\item{file}{string. Name of file. If `NULL`, the example files will
be listed.}
}
\description{
`playbase` comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access. This function was
taken from tidyverse/readr.
}
\examples{
example_file()
example_file("counts.csv")
}
