% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{EXAMPLE_GMT}
\alias{EXAMPLE_GMT}
\title{Example GMT (geneset matrix transpose) of genes targeted by microRNA}
\format{
## `GMT`
list of genes (targeted by a microRNA), and the list name is the microRNA
}
\source{
C3: regulatory target gene sets, MIR: microRNA targets from https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp
}
\usage{
EXAMPLE_GMT
}
\description{
Example GMT (geneset matrix transpose) of genes targeted by microRNA
}
\keyword{datasets}
