% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{expandAnnotationMatrixSAVE}
\alias{expandAnnotationMatrixSAVE}
\title{Expand Annotation Matrix}
\usage{
expandAnnotationMatrixSAVE(A)
}
\arguments{
\item{A}{A data frame representing the annotation matrix to be expanded.}
}
\value{
A data frame representing the expanded annotation matrix.
}
\description{
This function expands an annotation matrix by 
converting categorical variables into dummy variables and ranking numerical variables.
}
\details{
The function takes an annotation matrix `A` as input, 
where each column represents a variable and each row represents an 
observation. For each column, if the variable is numerical, the function 
ranks the values and stores them in a new column. If the variable is 
categorical, the function creates dummy variables for each level of the c
ategorical variable and stores them in new columns.

The resulting expanded annotation matrix is returned as a data frame, 
where each row represents an observation and each column represents a 
variable or a level of a categorical variable.
}
