% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{expmat2contrast}
\alias{expmat2contrast}
\title{Convert expression matrix to contrast matrix}
\usage{
expmat2contrast(exp.matrix)
}
\arguments{
\item{exp.matrix}{Expression matrix with samples in columns}
}
\value{
A list containing:
\itemize{
\item{contr.matrix}{Contrast matrix}  
\item{group}{Factor of sample groups}
\item{exp.matrix}{Original expression matrix}
}
}
\description{
Converts an expression matrix into a contrast matrix encoding comparisons between sample groups.
}
\details{
This function takes an expression matrix and generates a contrast matrix from it.

It creates a group factor by pasting column names. The number of unique groups is used to name contrast columns as "groupN".

The contrast matrix contains a row for each unique group, with 0/1/-1 values encoding the sample membership.

The original group factor and expression matrix are also returned.
}
