% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{extend_metabolite_sets}
\alias{extend_metabolite_sets}
\title{Extend metabolite sets in M by adding neighbour-metabolites that
are connected (neighbours) according to given PPI. Only
metabolite-metabolite neighbours are considered.}
\usage{
extend_metabolite_sets(
  M,
  ppi,
  add = TRUE,
  postfix = "(extended)",
  maxcost = 0.33
)
}
\description{
Extend metabolite sets in M by adding neighbour-metabolites that
are connected (neighbours) according to given PPI. Only
metabolite-metabolite neighbours are considered.
}
