% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{extend_metabolite_sets2}
\alias{extend_metabolite_sets2}
\title{Extend metabolite sets in M by adding neighbouring-metabolites that
are connected (neighbours) with members in metabolite set according
to given PPI. Any metabolite-metabolite or protein-metabolite
neighbours are considered.}
\usage{
extend_metabolite_sets2(
  M,
  ppi,
  add = TRUE,
  postfix = "(extended)",
  maxcost = 0.33
)
}
\description{
Extend metabolite sets in M by adding neighbouring-metabolites that
are connected (neighbours) with members in metabolite set according
to given PPI. Any metabolite-metabolite or protein-metabolite
neighbours are considered.
}
