% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{fastTOMsimilarity}
\alias{fastTOMsimilarity}
\title{Faster implementation of TOM computation using low-rank SVD
approximation.}
\usage{
fastTOMsimilarity(A, tomtype = "signed", lowrank = 20)
}
\description{
Faster implementation of TOM computation using low-rank SVD
approximation.
}
\seealso{
WGCNAplus::fastTOMsimilarity
}
