% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{filterFamily}
\alias{filterFamily}
\title{Filter Genes by Family}
\usage{
filterFamily(genes, family, ngs)
}
\arguments{
\item{genes}{A data frame of genes to be filtered, with row names representing probe names.}

\item{family}{A character string specifying the name of the gene family to filter by.}

\item{ngs}{A list containing gene families, where `ngs$families` is a 
named list of character vectors representing gene families.}
}
\value{
A character vector representing the row names of the rows in the 
`genes` data frame that match the specified gene family.
}
\description{
This function filters a data frame of genes by a specified gene family.
}
\details{
The function takes a data frame of genes `genes`, a gene family 
name `family`, and a list of gene families `ngs` as input. 
The function searches for rows in the `genes` data frame where the probe 
name, short probe name, or gene name matches a gene in the specified gene family. 
The row names of the matching rows are returned.
}
