% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-proteomics.R
\name{fit.weibull2}
\alias{fit.weibull2}
\title{Fit a bivariate Weibull distribution}
\usage{
fit.weibull2(x, y)
}
\arguments{
\item{x}{A numeric vector for the x variable}

\item{y}{A numeric vector for the y variable}
}
\value{
A list with the estimated Weibull parameters
}
\description{
Fits a bivariate Weibull distribution to two input variables.
}
\details{
Fit a Weibull distribution to bivariate data

This function takes two numeric vectors \code{x} and \code{y} and fits a bivariate Weibull distribution.
The Weibull distribution has shape and scale parameters for each variable.

Maximum likelihood estimation is used to estimate the 4 Weibull parameters from the input data.
The estimated parameters are returned as a named list.
}
