% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{fixContrastMatrix}
\alias{fixContrastMatrix}
\title{Fix a contrast matrix by checking and conforming column entries
with the contrast name 'c:a_vs_b' so that all column entries are
strictly in {a,b}. Checks ignoring case. For multi-class column,
'c:a_vs_b1.b2' will replace any b1, b2 or b1.b2 to class 'b1.b2'.}
\usage{
fixContrastMatrix(contrasts)
}
\description{
Fix a contrast matrix by checking and conforming column entries
with the contrast name 'c:a_vs_b' so that all column entries are
strictly in {a,b}. Checks ignoring case. For multi-class column,
'c:a_vs_b1.b2' will replace any b1, b2 or b1.b2 to class 'b1.b2'.
}
