% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEIGER_PGX}
\alias{GEIGER_PGX}
\title{Example pgx object}
\format{
## `GEIGER_PGX`
A pgx object
\describe{
  \item{name}{PGX name}
  \item{description}{PGX description}
  \item{date}{PGX data created}
  ...
}
}
\source{
Geiger2016
}
\usage{
GEIGER_PGX
}
\description{
This is an example pgx object already created. It comes from
Geiger2016-arginine.
}
\keyword{datasets}
