% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-colors.R
\name{generate_colors}
\alias{generate_colors}
\title{Function to generate color palettes}
\usage{
generate_colors(hex, discrete = FALSE)
}
\arguments{
\item{hex}{Hex code for main color as string}

\item{discrete}{Reorder for categorical data to place main color first}
}
\value{
A set of three colors
}
\description{
Function to generate color palettes
}
\examples{
generate_colors("#2EC09C")
generate_colors("#BE34EF")
}
