% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_call_stack}
\alias{get_call_stack}
\title{Get Call Stack}
\usage{
get_call_stack(n = 3, n0 = 2)
}
\arguments{
\item{n}{Number of calls to include in call stack.}

\item{n0}{Number of calls to exclude from the call stack.}
}
\value{
Character vector of call stack.
}
\description{
Returns the call stack up to a specified depth.
}
\examples{
\dontrun{
f3 <- function(x) {
  print("Show whole stack")
  message(get_call_stack())
  print("Remove message call")
  message(get_call_stack(1))
  x <- x + 1
  x
}

f2 <- function(x) {
  x <- f3(x)
  x
}

f1 <- function(x) {
  x
  x <- f2(x)
}
 f0 <- function(x) {
  x
  x <- f1(x)
}
f0(2)
}
}
