% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mini_example_data}
\alias{get_mini_example_data}
\title{Get mini example dataset}
\usage{
get_mini_example_data(item = "all")
}
\arguments{
\item{item}{Select which elements to return c("counts", "samples", "contrats", "all")
by default it returns a list with all the data elements to build a pgx object.}
}
\value{
List containing:
\itemize{
  \item samples - data.frame with sample info
  \item counts - matrix of gene counts
  \item contrast - data.frame with contrasts
}
}
\description{
Returns a mini example dataset with sample information,
gene expression counts, and contrasts.
}
\examples{
d <- get_mini_example_data()
}
