% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-timeseries.R
\name{get_timevars}
\alias{get_timevars}
\title{Get time variable patterns}
\usage{
get_timevars()
}
\value{
A character string containing pipe-separated time-related variable patterns
}
\description{
Returns a regex pattern string containing common time-related variable names
used for detecting time variables in sample metadata.
}
\details{
Get time-related variable patterns
}
\examples{
get_timevars()
# Returns "second|minute|hour|day|week|month|year|time|age"

}
