% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getCustomAnnotation2}
\alias{getCustomAnnotation2}
\title{Like getCustomAnnotation() but allows custom column names mapping
to feature, symbol and title. Also simplified
implementation. Should eventually replace getCustomAnnotation().}
\usage{
getCustomAnnotation2(
  probes,
  custom_annot,
  feature.col = "feature",
  symbol.col = "symbol",
  gene_title.col = "gene_title",
  ortholog.col = "ortholog",
  extra.columns = TRUE
)
}
\description{
Like getCustomAnnotation() but allows custom column names mapping
to feature, symbol and title. Also simplified
implementation. Should eventually replace getCustomAnnotation().
}
