% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{getDatasetFromSample}
\alias{getDatasetFromSample}
\title{Extract dataset name from sample identifiers}
\usage{
getDatasetFromSample(samples)
}
\arguments{
\item{samples}{Character vector of sample identifiers ("dataset::sample")}
}
\value{
Character vector of dataset names
}
\description{
Extract dataset name from sample identifiers
}
