% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getGeneAnnotation.ANNOTHUB}
\alias{getGeneAnnotation.ANNOTHUB}
\title{Get gene annotation data using AnnotationHub}
\usage{
getGeneAnnotation.ANNOTHUB(
  organism,
  probes,
  use.ah = NULL,
  probe_type = NULL,
  second.pass = TRUE,
  bridge = NULL,
  bridge_species = NULL,
  verbose = TRUE
)
}
\arguments{
\item{organism}{Organism name, e.g. "hsapiens_gene_ensembl".}

\item{probes}{Character vector of gene/transcript identifiers to retrieve annotation for.}

\item{probe_type}{Character specifying the type of input identifiers. If NULL,
it will be automatically detected. Options are "ensembl_gene_id", "ensembl_transcript_id", etc.}

\item{verbose}{Logical indicating whether to print status messages.}
}
\value{
Data frame with gene annotation data for the input identifiers. Columns are:
\itemize{
  \item \code{feature}: The probe identifier.
  \item \code{sybmol}: Human readable gene name.
  \item \code{human_homolog}: Gene symbol for human. Only present if working with non-human dataset.
  \item \code{gene_title}: Gene description
  \item \code{gene_biotype}: Gene biotype
  \item \code{chr}: Chromosome
  \item \code{pos}: Transcript start position
  \item \code{tx_len}: Transcript length
  \item \code{map}: Chromosome band
  \item \code{gene_name}: equivalent to the rownames. Kept for back compatibility
}
}
\description{
Retrieves gene annotation information from AnnotationHub for a set of input
gene/transcript identifiers.
}
\details{
This function queries AnnotHub to retrieve key gene annotation data for
a set of input gene/transcript identifiers. It can detect the identifier
type automatically if not provided.
}
\examples{
\dontrun{
probes <- c("ENSG00000142192", "ENST00000288602")
result <- getGeneAnnotation(organism, probes)
head(result)
}
}
