% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-meta.r
\name{getGeneSetTables}
\alias{getGeneSetTables}
\title{Get Gene Set Tables}
\usage{
getGeneSetTables(path)
}
\arguments{
\item{path}{The path to the directory containing the gene set tables.}
}
\value{
A list containing the gene set tables and associated metadata.
}
\description{
This function retrieves gene set tables from a specified directory path and organizes the data into a list.
}
