% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-graph.R
\name{getGOgraph}
\alias{getGOgraph}
\title{Create a Gene Ontology Graph}
\usage{
getGOgraph()
}
\value{
An igraph graph object representing the Gene Ontology hierarchy.
}
\description{
This function creates a graph representation of the Gene Ontology (GO) hierarchy, using data from the GO.db package.
}
\details{
The function first extracts the GO terms and their associated information from the GOTERM table in the GO.db package.
Then, it removes any duplicate terms and sets the row names of the resulting data frame to the GO IDs.

Next, the function extracts the parent-child relationships for each of the three GO
domains (Biological Process, Molecular Function, and Cellular Component) from the
corresponding tables in the GO.db package. It combines these relationships into a single data.frame
and creates a graph object using the igraph package.
}
