% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{getHSGeneInfo}
\alias{getHSGeneInfo}
\title{Get human gene information from annotation packages}
\usage{
getHSGeneInfo(eg, as.link = TRUE)
}
\arguments{
\item{eg}{Character vector of gene symbols}

\item{as.link}{Logical indicating whether to return symbols as hyperlinks. Default is TRUE.}
}
\value{
Named character vector with gene information.
}
\description{
Get human gene information from annotation packages
}
\details{
This function retrieves basic gene information for human genes from Bioconductor annotation packages.
It gets the symbol, name, map location, OMIM IDs, KEGG IDs and GO terms for each input gene symbol.

If as.link is TRUE, the gene symbols are returned as hyperlinks to GeneCards and OMIM IDs are returned as hyperlinks.
}
\examples{
\dontrun{
genes <- c("TP53", "BRCA1", "ABCD1")
info <- getHSGeneInfo(genes)

}
}
