% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getHumanOrtholog.biomart}
\alias{getHumanOrtholog.biomart}
\title{Get human ortholog from given symbols of organism by using
  biomart package. This package needs internet connection.
This is an alternative to orthogene::map_genes failure.
Unfortunately, biomart is *much* less reliable than orthogene (often down).}
\usage{
getHumanOrtholog.biomart(organism, symbols, verbose = 1)
}
\description{
Get human ortholog from given symbols of organism by using
  biomart package. This package needs internet connection.
This is an alternative to orthogene::map_genes failure.
Unfortunately, biomart is *much* less reliable than orthogene (often down).
}
