% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getHumanOrtholog}
\alias{getHumanOrtholog}
\title{Get human ortholog from given symbols of organism by using
  orthogene package. This package needs internet connection.}
\usage{
getHumanOrtholog(
  organism,
  symbols,
  ortho.methods = c("homologene", "gprofiler", "babelgene", "gprofiler2"),
  verbose = 1
)
}
\description{
Get human ortholog from given symbols of organism by using
  orthogene package. This package needs internet connection.
}
