% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getHumanOrtholog.SAVE}
\alias{getHumanOrtholog.SAVE}
\title{Get human ortholog from given symbols of organism by using
  orthogene package. This package needs internet connection.}
\usage{
getHumanOrtholog.SAVE(organism, symbols)
}
\description{
Get human ortholog from given symbols of organism by using
  orthogene package. This package needs internet connection.
}
