% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{getLevels}
\alias{getLevels}
\title{Get Levels of Group Variables}
\usage{
getLevels(Y)
}
\arguments{
\item{Y}{A data frame containing the data to be analyzed.}
}
\value{
A character vector representing the levels of the group variables in the input data frame.
}
\description{
This function retrieves the levels of group variables in a data frame.
}
\details{
The function takes a data frame `Y` as input and searches for
columns that represent group variables. Group variables are identified as
columns that have a name that does not contain "title", "name", "sample", or
"patient" and have a majority of non-unique values.
Numeric columns are excluded from the search.
The levels of the identified group variables are then extracted and returned as a character vector.
}
