% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{getMetaboliteInfo}
\alias{getMetaboliteInfo}
\title{Given a ChEBI id this provides information about the metabolite
using internal annotation databases METABOLITE_METADATA and
METABOLITE_ID. Also provides links to external metabolite
databases.}
\usage{
getMetaboliteInfo(organism = "Human", id)
}
\description{
Given a ChEBI id this provides information about the metabolite
using internal annotation databases METABOLITE_METADATA and
METABOLITE_ID. Also provides links to external metabolite
databases.
}
