% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getMultiProbeAnnotation}
\alias{getMultiProbeAnnotation}
\title{Convert multi-omics probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.}
\usage{
getMultiProbeAnnotation(organism, probes)
}
\description{
Convert multi-omics probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.
}
