% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getMultiSpeciesProbeAnnotation}
\alias{getMultiSpeciesProbeAnnotation}
\title{Annotate multi-species probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.}
\usage{
getMultiSpeciesProbeAnnotation(
  probes,
  organisms,
  probetype,
  datatype = "rna-seq"
)
}
\description{
Annotate multi-species probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.
}
