% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-info.R
\name{getOrgGeneInfo}
\alias{getOrgGeneInfo}
\title{Retrieve gene information from different databases (AnnotHub,
orgdb) and wrap hyperlinks.}
\usage{
getOrgGeneInfo(organism, gene, feature, ortholog, datatype, as.link = TRUE)
}
\description{
Retrieve gene information from different databases (AnnotHub,
orgdb) and wrap hyperlinks.
}
