% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getOrtholog}
\alias{getOrtholog}
\title{Get ortholog from given symbols of organism by
  aggregation of multiple methods. This package needs internet
  connection.}
\usage{
getOrtholog(
  symbols,
  organism,
  target_species,
  ortho.methods = c("homologene", "gprofiler", "babelgene", "gprofiler2", "uppercase"),
  verbose = 1
)
}
\description{
Get ortholog from given symbols of organism by
  aggregation of multiple methods. This package needs internet
  connection.
}
