% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{getPhenotypeValuesTileDB}
\alias{getPhenotypeValuesTileDB}
\title{Get unique values for a phenotype column}
\usage{
getPhenotypeValuesTileDB(tiledb_path, phenotype, datasets = NULL)
}
\arguments{
\item{tiledb_path}{Path to the TileDB database}

\item{phenotype}{Name of the phenotype column}

\item{datasets}{Optional dataset names to filter by}
}
\value{
Character vector of unique values for the phenotype
}
\description{
Get unique values for a phenotype column
}
