% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{getPlaydataGMT}
\alias{getPlaydataGMT}
\title{Combined GSETxGENE/MSETxMETABOLITE geneset matrix from playdata}
\usage{
getPlaydataGMT()
}
\description{
Combined GSETxGENE/MSETxMETABOLITE geneset matrix from playdata
}
\seealso{
WGCNAplus::getPlaydataGMT
}
