% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{getPrior}
\alias{getPrior}
\title{Get prior value for normalization for non-gx data.
For other data types (proteomics, metabolomics, lipidomics) we use min positive.}
\usage{
getPrior(counts)
}
\arguments{
\item{counts}{Multiomics data matrix: counts. Features in rows and samples in columns.}
}
\value{
prior value
}
\description{
Get prior
}
\details{
Get prior. Conventionally 1 for logCPM (used for RNAseq)
For other data types (proteomics, metabolomics, lipidomics) we use min positive.
}
