% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getProbeAnnotation}
\alias{getProbeAnnotation}
\title{Get probetype annotation for organism and datatype. For multi-omics
probe names must be prefixed with data type.}
\usage{
getProbeAnnotation(
  organism,
  probes,
  datatype,
  meth_type = NULL,
  probetype = "",
  ortholog_species = "hsapiens",
  annot_table = NULL
)
}
\description{
Get probetype annotation for organism and datatype. For multi-omics
probe names must be prefixed with data type.
}
