% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{getSampleShortName}
\alias{getSampleShortName}
\title{Extract short sample name from sample identifiers}
\usage{
getSampleShortName(samples)
}
\arguments{
\item{samples}{Character vector of sample identifiers ("dataset::sample")}
}
\value{
Character vector of short sample names
}
\description{
Extract short sample name from sample identifiers
}
