% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{getSIGDB.DIR}
\alias{getSIGDB.DIR}
\title{Get the SIGDB Directory}
\usage{
getSIGDB.DIR()
}
\value{
A character vector containing the path to the SIGDB directory.
}
\description{
This function returns the path to the SIGDB directory, which is used to store data related to the SIGDB database.
}
\details{
The function first checks if the `SIGDB.DIR` variable exists in the global environment.
If it does not exist, the function checks if the `FILESX` variable exists in the global environment.
If `FILESX` exists, the function constructs the path to the SIGDB directory by concatenating `FILESX` and a subdirectory named "sigdb".
The resulting path is returned as a character vector of length 2, containing both `FILESX` and the full path to the SIGDB directory.
If `SIGDB.DIR` already exists in the global environment, its value is returned directly.
}
