% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R, R/pgx-ensembl.R
\name{getSpeciesTable}
\alias{getSpeciesTable}
\title{Get species table in AnnotationHub/OrgDB}
\usage{
getSpeciesTable(ah = NULL)

getSpeciesTable(ah = NULL)
}
\description{
Get species table in AnnotationHub/OrgDB

Get species table in AnnotationHub/OrgDB
}
