% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{gmt.map2symbol}
\alias{gmt.map2symbol}
\title{Maps a gmt list to symbol using annotation table. We go via
sparsematrix because it is much faster than list filtering.}
\usage{
gmt.map2symbol(gmt, annot, target = "symbol")
}
\description{
Maps a gmt list to symbol using annotation table. We go via
sparsematrix because it is much faster than list filtering.
}
