% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gmt2mat_old}
\alias{gmt2mat_old}
\title{Convert GMT file to binary matrix}
\usage{
gmt2mat_old(
  gmt,
  max.genes = -1,
  ntop = -1,
  sparse = TRUE,
  bg = NULL,
  use.multicore = TRUE
)
}
\arguments{
\item{gmt}{A list representing the GMT file, where each element is a character
vector representing a gene set.}

\item{max.genes}{The maximum number of genes to include in the binary matrix.
Defaults to -1, which includes all genes.}

\item{ntop}{The number of top genes to consider for each gene set. Defaults to -1,
which includes all genes.}

\item{sparse}{A logical value indicating whether to create a sparse matrix
(from the 'Matrix' package) or a dense matrix. Defaults to `TRUE`.}

\item{bg}{A character vector representing the background set of genes. Defaults to `NULL`,
which considers all unique genes from the gene sets.}

\item{use.multicore}{A logical value indicating whether to use parallel processing
(via the 'parallel' package) for faster execution. Defaults to `TRUE`.}
}
\value{
A binary matrix representing the presence or absence of genes in each gene set.
        Rows correspond to genes, and columns correspond to gene sets.
}
\description{
This function converts a GMT file (Gene Matrix Transposed) to a binary matrix,
where rows represent genes and columns represent gene sets. The binary matrix
indicates the presence or absence of genes in each gene set.
}
