% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{go.merge_duplicates}
\alias{go.merge_duplicates}
\title{Merge duplicated GO sets in a gmt collection by merging the
terms. This is often needed after merging to GO collections if
retrieved by different methods.}
\usage{
go.merge_duplicates(gmt)
}
\description{
Merge duplicated GO sets in a gmt collection by merging the
terms. This is often needed after merging to GO collections if
retrieved by different methods.
}
