% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-graph.r
\name{graph.cut_crossings}
\alias{graph.cut_crossings}
\title{Cut graph crossings}
\usage{
graph.cut_crossings(g, idx, max.wt = 9999)
}
\arguments{
\item{g}{An igraph graph object}

\item{idx}{Vector of node membership indices}

\item{max.wt}{Maximum edge weight to cut. Default 9999.}
}
\value{
An igraph graph with cross-cluster edges removed
}
\description{
Cuts inter-cluster edges in a graph based on node membership
}
\details{
Cut graph crossings

This function takes an igraph graph object \code{g} and a vector of node membership indices \code{idx}.
It identifies edges that connect nodes of different clusters based on \code{idx}.
Any edges with weight less than \code{max.wt} that link across clusters are removed from the graph.
The resulting pruned graph is returned.

By default only edges with very high weight (9999) are kept. This effectively cuts crossings between clusters.
The \code{max.wt} parameter can be lowered to allow more inter-cluster edges.
}
