% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.barplot}
\alias{gsea.barplot}
\title{GSEA Barplot}
\usage{
gsea.barplot(
  scores,
  names = NULL,
  xlab = "score",
  xlim = NULL,
  cex.text = 1,
  main = "enrichment",
  n = 16
)
}
\arguments{
\item{scores}{A numeric vector of GSEA scores.}

\item{names}{A character vector of names for the scores.}

\item{xlab}{The label for the x-axis.}

\item{xlim}{The limits for the x-axis.}

\item{cex.text}{The size of the text labels.}

\item{main}{The main title of the plot.}

\item{n}{The number of scores to display in the barplot.}
}
\value{
barplot
}
\description{
Generates a barplot visualization of GSEA scores.
}
