% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.enplot}
\alias{gsea.enplot}
\title{GSEA Enrichment Plot}
\usage{
gsea.enplot(
  rnk,
  gset,
  names = NULL,
  main = NULL,
  decreasing = TRUE,
  cex = 1,
  cex.main = 0.95,
  len.main = 40,
  lab.line = c(0.8, 2),
  cex.lab = 0.85,
  main.line = 0.3,
  xlab = "Rank in ordered dataset",
  res = 1200,
  ylab = "Rank metric"
)
}
\arguments{
\item{rnk}{A numeric vector representing the ranked list.}

\item{gset}{A character vector specifying the gene set of interest.}

\item{names}{A character vector of names for the ranked list.}

\item{main}{The main title of the plot.}

\item{decreasing}{Logical indicating whether the ranked list should be sorted in decreasing order.}

\item{cex}{The size of the text labels.}

\item{cex.main}{The size of the main title text.}

\item{len.main}{The maximum number of characters for the main title before line breaks are applied.}

\item{lab.line}{The position of the axis labels.}

\item{cex.lab}{The size of the axis labels.}

\item{main.line}{The position of the main title.}

\item{xlab}{The label for the x-axis.}

\item{res}{The resolution of the plot.}

\item{ylab}{The label for the y-axis.}
}
\value{
plot
}
\description{
Generates an enrichment plot for GSEA analysis.
}
