% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.enplot.UPDN}
\alias{gsea.enplot.UPDN}
\title{Title}
\usage{
gsea.enplot.UPDN(
  rnk,
  gset.up,
  gset.dn,
  names = NULL,
  main = NULL,
  decreasing = TRUE,
  cex.main = 0.9,
  len.main = 40,
  sum.trace = TRUE,
  res = 1200,
  xlab = "Rank in ordered dataset",
  ylab = "Ranked list metric"
)
}
\arguments{
\item{ylab}{}
}
\description{
Title
}
