% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{gsea.enplotly}
\alias{gsea.enplotly}
\title{GSEA Enrichment Plot}
\usage{
gsea.enplotly(
  fc,
  gset,
  cex = 1,
  main = NULL,
  xlab = NULL,
  ticklen = 0.25,
  ylab = NULL,
  yth = 1,
  tooltips = NULL,
  cex.text = 1,
  cex.title = 1.4,
  cex.axis = 1,
  cbar.width = 32
)
}
\arguments{
\item{fc}{A numeric vector of gene-level stats (e.g. log2 fold changes) ordered by decreasing statistic.}

\item{gset}{A character vector of gene ids in the gene set.}

\item{cex}{The magnification to be used for points and text.}

\item{main}{The plot title.}

\item{xlab}{The x-axis label.}

\item{ticklen}{Length of the tick marks as a fraction of the height of a line of text.}

\item{ylab}{The y-axis label.}

\item{yth}{The length of the y-axis tick marks.}

\item{tooltips}{An optional vector of tooltips to display for each point.}

\item{cex.text}{The magnification to be used for point labels.}

\item{cex.title}{The magnification to be used for the title.}

\item{cbar.width}{The width of the color bar.}
}
\value{
A ggplot object containing the enrichment plot.
}
\description{
Creates an enrichment plot visualization for gene set enrichment analysis.
}
\details{
This function takes a ranked list of gene-level stats and a gene set, and produces an enrichment plot visualization using ggplot2.
It computes a running enrichment score and plots it against the ranked list. The goal is to visualize if the gene set is enriched at either end of the ranked list.
}
