% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.enplotPDF}
\alias{gsea.enplotPDF}
\title{Title}
\usage{
gsea.enplotPDF(gsea.dir, gs, output.pdf)
}
\arguments{
\item{gsea.dir}{value}

\item{gs}{value}

\item{output.pdf}{value}
}
\description{
Title
}
