% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.fitAllContrasts}
\alias{gsea.fitAllContrasts}
\title{Title}
\usage{
gsea.fitAllContrasts(
  X,
  gmt,
  design,
  contr.matrix,
  output_dir,
  set.min = 15,
  set.max = 500,
  fdr = 0.25,
  skip.done = TRUE
)
}
\arguments{
\item{skip.done}{}
}
\description{
Title
}
