% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.ftplot}
\alias{gsea.ftplot}
\title{gsea.ftplot}
\usage{
gsea.ftplot(
  x,
  rft = "var",
  cft = "var",
  rmax = nrow(x),
  cmax = ncol(x),
  main = "",
  cex.main = 1.5,
  xnames = NULL,
  ynames = NULL,
  cft.lab = "col metric",
  rft.lab = "row metric",
  cexRow = 0.65,
  cexCol = 0.65,
  scale = "rows",
  rsort = "hclust",
  csort = "hclust",
  p = 1,
  mar = c(8, 4, 4, 2),
  layout.reset = TRUE,
  sort.decreasing = FALSE,
  col = gplots::bluered(64),
  legend = NULL,
  cex.legend = 1,
  layout.widths = c(0.15, 0.15, 1),
  layout.heights = c(0.15, 0.15, 1)
)
}
\arguments{
\item{x}{The input data matrix.}

\item{rft}{The metric to use for row ordering. If character, the metric function will be applied to each row of the matrix \code{x}.}

\item{cft}{The metric to use for column ordering. If character, the metric function will be applied to each column of the matrix \code{x}.}

\item{rmax}{The maximum number of rows to include in the plot.}

\item{cmax}{The maximum number of columns to include in the plot.}

\item{main}{The title of the plot.}

\item{cex.main}{The size of the main title.}

\item{xnames}{The names to use for the rows of the matrix.}

\item{ynames}{The names to use for the columns of the matrix.}

\item{cft.lab}{The label for the column metric.}

\item{rft.lab}{The label for the row metric.}

\item{cexRow}{The size of the row names.}

\item{cexCol}{The size of the column names.}

\item{scale}{The scaling method for the heatmap. Options are "rows" (scale rows) and "cols" (scale columns).}

\item{rsort}{The method for ordering the rows. Options are "hclust" (hierarchical clustering), "metric" (order by metric value), and "none" (no ordering).}

\item{csort}{The method for ordering the columns. Options are "hclust" (hierarchical clustering), "metric" (order by metric value), and "none" (no ordering).}

\item{p}{The power to raise the heatmap values to.}

\item{mar}{The margin sizes for the plot.}

\item{layout.reset}{Whether to reset the layout after plotting.}

\item{sort.decreasing}{Whether to sort the metric values in decreasing order.}

\item{col}{The color palette for the heatmap.}

\item{legend}{The legend labels for the heatmap colors.}

\item{cex.legend}{The size of the legend labels.}

\item{layout.widths}{The widths of the plot layout.}

\item{layout.heights}{The heights of the plot layout.}
}
\value{
None
}
\description{
A function to generate a GSEA (Gene Set Enrichment Analysis) plot.
}
