% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.quick_report}
\alias{gsea.quick_report}
\title{gsea.quick_report}
\usage{
gsea.quick_report(output.dir, pattern = NULL)
}
\arguments{
\item{output.dir}{The directory containing the metaGSEA results.}

\item{pattern}{A pattern to match specific files in the directory (optional).}
}
\value{
A data frame with the summary of metaGSEA results.
}
\description{
A function to generate a quick report of metaGSEA results in a specified folder.
}
