% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gseaLeadingEdgeHeatmap}
\alias{gseaLeadingEdgeHeatmap}
\title{Leading Edge Heatmap}
\usage{
gseaLeadingEdgeHeatmap(
  gsea,
  maxrow = 60,
  maxcol = 60,
  gsets = NULL,
  render = "gx.heatmap",
  info.text = TRUE
)
}
\arguments{
\item{gsea}{An object containing GSEA results.}

\item{maxrow}{Maximum number of rows to display in the heatmap.}

\item{maxcol}{Maximum number of columns to display in the heatmap.}

\item{gsets}{A character vector specifying the gene sets to include in the analysis.}

\item{render}{The rendering method for the heatmap. Options are "gx.heatmap", "d3heatmap", or "heatmaply".}

\item{info.text}{Logical indicating whether to display additional information text.}
}
\value{
NULL if the leading edge matrix is too small (less than or equal to 1 row or 1 column),
otherwise, it returns the generated heatmap visualization.
}
\description{
Generates a heatmap visualization of the leading edge analysis results.
}
