% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gseaPlotEnplots}
\alias{gseaPlotEnplots}
\title{Plot GSEA Enrichment Plots}
\usage{
gseaPlotEnplots(gsea, gsets = NULL, ncol = 5)
}
\arguments{
\item{gsea}{The GSEA analysis results obtained from \code{\link{getGseaOutput}}.}

\item{gsets}{Optional character vector specifying the gene sets to plot.
If NULL (default), all available gene sets will be plotted.}

\item{ncol}{The number of columns in the grid arrangement of plots.
Defaults to 5.}
}
\value{
None
}
\description{
Plot the GSEA enrichment plots from the GSEA analysis results.
}
