% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-rankcor.r
\name{gset.averageFC}
\alias{gset.averageFC}
\title{Compute geneset expression as the average foldchange of genes in
the geneset. Requires FoldChange matrix F and (sparse) geneset
matrix matG.}
\usage{
gset.averageFC(F, matG, use.rank = FALSE)
}
\description{
Compute geneset expression as the average foldchange of genes in
the geneset. Requires FoldChange matrix F and (sparse) geneset
matrix matG.
}
