% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-fisher.r
\name{gset.fastFET}
\alias{gset.fastFET}
\title{Calculate fast Fisher exact test.}
\usage{
gset.fastFET(genes, G, bg, report.genes = FALSE)
}
\arguments{
\item{genes}{Vector of significant genes}

\item{G}{Sparse matrix containing gene sets}

\item{bg}{Vector of genes as background}

\item{report.genes}{Logical to report gene set genes in output}
}
\description{
Calculate fast Fisher exact test.
}
