% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-fisher.r
\name{gset.fisher2}
\alias{gset.fisher2}
\title{Title}
\usage{
gset.fisher2(
  genes.up,
  genes.dn,
  genesets,
  background = NULL,
  fdr = 0.05,
  mc = TRUE,
  sort.by = "zratio",
  nmin = 3,
  verbose = 1,
  min.genes = 15,
  max.genes = 500,
  method = "fast.fisher",
  check.background = TRUE,
  common.genes = TRUE
)
}
\arguments{
\item{genes.up}{value}

\item{genes.dn}{value}

\item{genesets}{value}

\item{background}{value}

\item{fdr}{value}

\item{mc}{value}

\item{sort.by}{value}

\item{nmin}{value}

\item{verbose}{value}

\item{min.genes}{value}

\item{max.genes}{value}

\item{method}{value}

\item{check.background}{value}

\item{common.genes}{value}
}
\description{
Title
}
