% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-meta.r
\name{gset.fitContrastsWithAllMethods}
\alias{gset.fitContrastsWithAllMethods}
\title{Title}
\usage{
gset.fitContrastsWithAllMethods(
  gmt,
  X,
  Y,
  G,
  design,
  contr.matrix,
  methods,
  mc.threads = 1,
  mc.cores = NULL,
  batch.correct = TRUE
)
}
\arguments{
\item{batch.correct}{}
}
\description{
Title
}
