% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-rankcor.r
\name{gset.gsva}
\alias{gset.gsva}
\title{Single sample genesets expression scores. See Foroutan 2018. This
seems to be a nice and efficient algorithm.}
\usage{
gset.gsva(X, geneSets, method = "gsva")
}
\description{
Single sample genesets expression scores. See Foroutan 2018. This
seems to be a nice and efficient algorithm.
}
