% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-rankcor.r
\name{gset.rankcor}
\alias{gset.rankcor}
\title{Title}
\usage{
gset.rankcor(rnk, gset, compute.p = FALSE, use.rank = TRUE)
}
\arguments{
\item{rnk}{value}

\item{gset}{value}

\item{compute.p}{value}

\item{use.rank}{value}
}
\description{
Title
}
