% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-rankcor.r
\name{gset.singscore}
\alias{gset.singscore}
\title{Single sample genesets expression scores. See Foroutan 2018. This
seems to be a nice and efficient algorithm.}
\usage{
gset.singscore(X, geneSets, center = FALSE, return.score = TRUE)
}
\description{
Single sample genesets expression scores. See Foroutan 2018. This
seems to be a nice and efficient algorithm.
}
