% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-rankcor.r
\name{gset.ttest}
\alias{gset.ttest}
\title{Statistical testing of differentially enrichment on foldchange fc
using one-sample t-test or z-test.}
\usage{
gset.ttest(
  fc,
  G,
  sort.by = "pvalue",
  nprior = 0,
  method = c("ttest", "ztest")[1]
)
}
\arguments{
\item{fc}{Vector of logFC values}

\item{G}{Sparse matrix of gene sets. Non-zero entry indicates
gene/feature is part of gene sets. Features on rows, gene sets on
columns.}
}
\description{
This function performs statistical testing for differential
enrichment using plaid
}
