% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase_data.R
\docType{data}
\name{GSET_SPARSEG_XL}
\alias{GSET_SPARSEG_XL}
\title{Some kind of sparse matrix dataset?}
\format{
## `GSET_SPARSEG_XL`
A dgCMatrix / Matrix object
}
\source{
unknown
}
\usage{
GSET_SPARSEG_XL
}
\description{
No idea what this data is but it is used in `test_genesets`.
}
\keyword{datasets}
