% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{GTEX_RNA_TISSUE_TPM}
\alias{GTEX_RNA_TISSUE_TPM}
\title{A dataset used by `compute_deconvolution`}
\format{
## `GTEX_RNA_TISSUE_TPM`
}
\source{
unknown
}
\usage{
GTEX_RNA_TISSUE_TPM
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
