% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-plot.r
\name{gx.b3plot}
\alias{gx.b3plot}
\title{Title}
\usage{
gx.b3plot(
  x,
  y,
  first = NULL,
  width = 1,
  bar = TRUE,
  bee = TRUE,
  sig.stars = FALSE,
  ymax = NULL,
  bee.cex = 0.3,
  max.stars = 5,
  srt = NULL,
  xoff = 0,
  names.cex = 1,
  names = TRUE,
  max.points = 100,
  col = "grey80",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
