% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-plot.r
\name{gx.barplot}
\alias{gx.barplot}
\title{Title}
\usage{
gx.barplot(
  x,
  main = "",
  cex.main = 1.2,
  cex.names = 0.85,
  cex.legend = 0.9,
  srt = 0,
  xlab = "",
  ylab = "",
  group = NULL,
  group.names = NULL,
  bar.names = NULL,
  voff = 2,
  legend = TRUE
)
}
\arguments{
\item{legend}{}
}
\description{
Title
}
