% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{gx.heatmap}
\alias{gx.heatmap}
\title{Title}
\usage{
gx.heatmap(
  gx,
  values = NULL,
  clust.method = "ward.D2",
  dist.method = "pearson",
  col.dist.method = "euclidean",
  plot.method = "heatmap.2",
  col = colorRampPalette(c("royalblue3", "grey90", "indianred3"))(64),
  softmax = FALSE,
  scale = "row",
  verbose = 1,
  symm = FALSE,
  col.annot = NULL,
  row.annot = NULL,
  annot.ht = 1,
  nmax = 1000,
  cmax = NULL,
  show_colnames = TRUE,
  indent.names = FALSE,
  ...
)
}
\arguments{
\item{gx}{value}

\item{values}{value}

\item{clust.method}{value}

\item{dist.method}{value}

\item{col.dist.method}{value}

\item{plot.method}{value}

\item{col}{value}

\item{softmax}{value}

\item{scale}{value}

\item{verbose}{value}

\item{symm}{value}

\item{col.annot}{value}

\item{row.annot}{value}

\item{annot.ht}{value}

\item{nmax}{value}

\item{cmax}{value}

\item{show_colnames}{value}

\item{indent.names}{value}

\item{...}{value}
}
\description{
Title
}
