% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-plot.r
\name{gx.hist}
\alias{gx.hist}
\title{Histogram of gene expression values}
\usage{
gx.hist(gx, main = "", ylim = NULL)
}
\arguments{
\item{gx}{Gene expression matrix with genes in rows and samples in columns.}

\item{main}{Title for the plot.}

\item{ylim}{Limits for the y-axis.}
}
\value{
A histogram plot is generated, no value is returned.
}
\description{
Generate a histogram of gene expression values.
}
\details{
This function generates a histogram of the gene expression values in \code{gx}.
It first creates a histogram of all values using \code{\link[graphics]{hist}}.
It then overlays density histograms for each sample, with colors corresponding to column numbers.
}
\examples{
\dontrun{
gx <- matrix(rnorm(100 * 10), 100, 10)
gx.hist(gx)
}
}
