% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{gx.markermap}
\alias{gx.markermap}
\title{Title}
\usage{
gx.markermap(X, splitx, n = 5, ...)
}
\arguments{
\item{X}{value}

\item{splitx}{value}

\item{n}{value}

\item{...}{value}
}
\description{
Title
}
