% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-combat.r
\name{gx.nearestReferenceCorrection.DEPRECATED}
\alias{gx.nearestReferenceCorrection.DEPRECATED}
\title{Title}
\usage{
gx.nearestReferenceCorrection.DEPRECATED(x, y, ref, k = 3, dist.method = "cor")
}
\arguments{
\item{x}{value}

\item{y}{value}

\item{ref}{value}

\item{k}{value}

\item{dist.method}{value}
}
\description{
Title
}
