% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-combat.r
\name{gx.nnmcorrect2.NOTWORKING}
\alias{gx.nnmcorrect2.NOTWORKING}
\title{Title}
\usage{
gx.nnmcorrect2.NOTWORKING(
  X,
  y,
  pairs = NULL,
  use.design = TRUE,
  center.x = TRUE,
  center.m = TRUE,
  dist.method = "cor"
)
}
\arguments{
\item{X}{value}

\item{y}{value}

\item{pairs}{value}

\item{use.design}{value}

\item{center.x}{value}

\item{center.m}{value}

\item{dist.method}{value}
}
\description{
Title
}
