% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-limma.r
\name{gx.test.groups}
\alias{gx.test.groups}
\title{Title}
\usage{
gx.test.groups(
  sig,
  class.label,
  fdr = 0.2,
  test.method = c("wilcox", "limma", "ttest", "fisher"),
  running.name = NULL,
  output.to.file = TRUE
)
}
\arguments{
\item{output.to.file}{}
}
\description{
Title
}
