% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-volcano.r
\name{gx.volcanoPlot.XY}
\alias{gx.volcanoPlot.XY}
\title{Title}
\usage{
gx.volcanoPlot.XY(
  x,
  pv,
  gene,
  ma_plot = FALSE,
  ma = NULL,
  p.sig = 0.05,
  lfc = 1,
  render = "scatterD3",
  n = 1000,
  highlight = NULL,
  main = "",
  cex = 1,
  lab.cex = 1,
  nlab = 10,
  xlim = NULL,
  ylim = NULL,
  use.fdr = FALSE,
  xlab = "effect size",
  ylab = "significance (-log10p)",
  use.rpkm = FALSE,
  maxchar = 40,
  hi.col = "#1e60bb",
  cex.main = 1.2,
  axes = TRUE,
  cex.axis = 1
)
}
\arguments{
\item{x}{value}

\item{pv}{value}

\item{gene}{value}

\item{ma_plot}{value}

\item{ma}{value}

\item{p.sig}{value}

\item{lfc}{value}

\item{render}{value}

\item{n}{value}

\item{highlight}{value}

\item{main}{value}

\item{cex}{value}

\item{lab.cex}{value}

\item{nlab}{value}

\item{xlim}{value}

\item{ylim}{value}

\item{use.fdr}{value}

\item{xlab}{value}

\item{ylab}{value}

\item{use.rpkm}{value}

\item{maxchar}{value}

\item{hi.col}{value}

\item{cex.main}{value}

\item{axes}{value}

\item{cex.axis}{value}
}
\description{
Title
}
