% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{h5exists}
\alias{h5exists}
\title{Check if an Object Exists in an HDF5 File}
\usage{
h5exists(h5.file, obj)
}
\arguments{
\item{h5.file}{A character string specifying the path to the HDF5 file.}

\item{obj}{A character string specifying the name of the object to check for existence in the HDF5 file.}
}
\value{
A logical value indicating whether the specified object exists in the HDF5 file (TRUE) or not (FALSE).
}
\description{
This function checks if a specified object exists in an HDF5 file.
}
\details{
The function uses the `h5ls` function from the rhdf5 package to list all objects in the specified HDF5 file.
It then constructs the full path for each object by concatenating the group and name columns of the resulting data frame, separated by a forward slash.
The function checks if the specified object name is present in this list of full paths, ignoring any leading forward slashes.
}
