% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-graph.R
\name{hclustGraph}
\alias{hclustGraph}
\title{Hierarchical clustering of graph nodes}
\usage{
hclustGraph(g, k = NULL, mc.cores = 2)
}
\arguments{
\item{g}{An igraph object representing the graph}

\item{k}{Number of clusters. If NULL, optimal number is found.}

\item{mc.cores}{Number of cores for parallel processing.}
}
\value{
Matrix with hierarchical clustering membership at each level.
}
\description{
Performs hierarchical clustering of nodes in a graph using the Louvain algorithm.
}
\details{
Hierarchical Clustering of Graph


This function takes an igraph graph object \code{g} and performs hierarchical clustering
of the nodes using the Louvain community detection algorithm.

It iteratively clusters the nodes, splitting communities into sub-communities to find
the optimal number of clusters. The number of clusters \code{k} can be specified,
otherwise the optimal number is found automatically.

The clustering is done in parallel using \code{mc.cores} cores if available.
The output is a matrix with the hierarchical clustering membership at each level.
}
