% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{heatmap.3}
\alias{heatmap.3}
\title{Title}
\usage{
heatmap.3(
  x,
  Rowv = TRUE,
  Colv = if (symm) "Rowv" else TRUE,
  distfun = dist,
  hclustfun = hclust,
  dendrogram = c("both", "row", "column", "none"),
  symm = FALSE,
  scale = c("none", "row", "column"),
  na.rm = TRUE,
  revC = identical(Colv, "Rowv"),
  add.expr,
  breaks,
  symbreaks = max(x < 0, na.rm = TRUE) || scale != "none",
  col = "heat.colors",
  colsep,
  rowsep,
  sepcolor = "white",
  sepwidth = c(0.05, 0.05),
  cellnote,
  notecex = 1,
  notecol = "cyan",
  na.color = par("bg"),
  trace = c("none", "column", "row", "both"),
  tracecol = "cyan",
  hline = median(breaks),
  vline = median(breaks),
  linecol = tracecol,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  side.height.fraction = 0.3,
  cexRow = 0.2 + 1/log10(nr),
  cexCol = 0.2 + 1/log10(nc),
  labRow = NULL,
  labCol = NULL,
  key = TRUE,
  keysize = 1.5,
  density.info = c("none", "histogram", "density"),
  denscol = tracecol,
  symkey = max(x < 0, na.rm = TRUE) || symbreaks,
  densadj = 0.25,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  lmat = NULL,
  lhei = NULL,
  lwid = NULL,
  NumColSideColors = 1,
  NumRowSideColors = 1,
  KeyValueName = "Value",
  ...
)
}
\arguments{
\item{x}{value}

\item{Rowv}{value}

\item{Colv}{value}

\item{distfun}{value}

\item{hclustfun}{value}

\item{dendrogram}{value}

\item{symm}{value}

\item{scale}{value}

\item{na.rm}{value}

\item{revC}{value}

\item{add.expr}{value}

\item{breaks}{value}

\item{symbreaks}{value}

\item{col}{value}

\item{colsep}{value}

\item{rowsep}{value}

\item{sepcolor}{value}

\item{sepwidth}{value}

\item{cellnote}{value}

\item{notecex}{value}

\item{notecol}{value}

\item{na.color}{value}

\item{trace}{value}

\item{tracecol}{value}

\item{hline}{value}

\item{vline}{value}

\item{linecol}{value}

\item{margins}{value}

\item{ColSideColors}{value}

\item{RowSideColors}{value}

\item{side.height.fraction}{value}

\item{cexRow}{value}

\item{cexCol}{value}

\item{labRow}{value}

\item{labCol}{value}

\item{key}{value}

\item{keysize}{value}

\item{density.info}{value}

\item{denscol}{value}

\item{symkey}{value}

\item{densadj}{value}

\item{main}{value}

\item{xlab}{value}

\item{ylab}{value}

\item{lmat}{value}

\item{lhei}{value}

\item{lwid}{value}

\item{NumColSideColors}{value}

\item{NumRowSideColors}{value}

\item{KeyValueName}{value}

\item{...}{value}
}
\description{
Title
}
