% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{heatmapWithAnnot}
\alias{heatmapWithAnnot}
\title{heatmapWithAnnot}
\usage{
heatmapWithAnnot(
  F,
  anno.type = c("boxplot", "barplot"),
  bar.height = NULL,
  map.height = NULL,
  row_fontsize = 9,
  column_fontsize = 9,
  inset = c(-0.025, -0.1),
  mar = c(0, 0, 0, 0),
  legend = TRUE,
  ...
)
}
\arguments{
\item{F}{Numeric data matrix to visualize as a heatmap}

\item{anno.type}{Type of annotation to show - "boxplot" or "barplot"}

\item{bar.height}{Height of the annotation barplot in cm}

\item{map.height}{Height of the heatmap in cm}

\item{row_fontsize}{Fontsize for row labels}

\item{column_fontsize}{Fontsize for column labels}

\item{inset}{Inset margins for annotation}

\item{mar}{Heatmap margins}

\item{legend}{Show legend for annotation colors}

\item{...}{Other arguments passed to heatmap functions}
}
\value{
A heatmap grob object
}
\description{
Creates a heatmap visualization with side annotation bars.
}
\details{
This function generates a heatmap from the input data matrix F.
Side annotation bars can be added, either as boxplots or barplots.
The annotation height, heatmap height, font sizes, margins, and other
heatmap parameters can be customized.
}
