% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{HPA_RNA_CELLINE}
\alias{HPA_RNA_CELLINE}
\title{A dataset used by `compute_deconvolution`}
\format{
## `HPA_RNA_CELLINE`
}
\source{
unknown
}
\usage{
HPA_RNA_CELLINE
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
