% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{human2mouse}
\alias{human2mouse}
\alias{mouse2human}
\title{Convert human gene symbols to mouse}
\usage{
human2mouse(x)

mouse2human(x)
}
\arguments{
\item{x}{Character vector of human gene symbols}
}
\value{
Character vector of converted mouse gene symbols
}
\description{
Convert human gene symbols to mouse
}
\section{Functions}{
\itemize{
\item \code{mouse2human()}: Convert human to mouse gene symbols using HomoloGene database

}}
