% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{IMMUNOSTATES_MATRIX}
\alias{IMMUNOSTATES_MATRIX}
\title{A dataset used by `compute_deconvolution`}
\format{
## `IMMUNOSTATES_MATRIX`
}
\source{
unknown
}
\usage{
IMMUNOSTATES_MATRIX
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
