% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-impute.R
\name{imputeMissing.mox}
\alias{imputeMissing.mox}
\title{Impute missing values separately for each omics type in multi-omics data.}
\usage{
imputeMissing.mox(
  X,
  method = c("LLS", "bpca", "msImpute", "SVD", "SVD2", "NMF", "RF", "knn", "QRILC",
    "MLE", "MinDet", "MinProb", "min", "zero", "nbavg", "rowmeans", "Perseus")[1:3],
  rf.ntree = 100,
  nv = 5,
  keep.limits = FALSE,
  infinite.na = TRUE,
  plot = FALSE
)
}
\arguments{
\item{X}{A multiomics matrix containing the input data.}
}
\value{
Imputed multi-omics matrix
}
\description{
Generic function to impute missing (NA) value. Input is a matrix.

It uses the imputeMissing function.
}
