% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-impute.R
\name{imputeMissing}
\alias{imputeMissing}
\title{Impute missing values}
\usage{
imputeMissing(
  X,
  method = c("LLS", "bpca", "msImpute", "SVD", "SVD2", "NMF", "RF", "knn", "QRILC",
    "MLE", "MinDet", "MinProb", "min", "zero", "nbavg", "rowmeans")[1:3],
  rf.ntree = 100,
  nv = 5,
  keep.limits = FALSE,
  plot = FALSE
)
}
\arguments{
\item{X}{A matrix containing the input data}

\item{plot}{Boolean for generating histogram plots}

\item{methods}{Methods for imputation. Default is c("LLS", "bpca", "msImpute")}
}
\value{
An updated matrix containing imputed values
}
\description{
Generic function to impute missing (NA) value. Input is a matrix.
}
