% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-methyl.R
\name{infer_sex_methyl}
\alias{infer_sex_methyl}
\title{infer_sex_methyl}
\usage{
infer_sex_methyl(data, genes = NULL, meth_type = "450K array")
}
\arguments{
\item{genes}{Annotation matrix: pgx$genes. if NULL, automatically retrieved.}

\item{meth_type}{Array type. '450K array' or 'EPIC array'}

\item{X}{Matrix of methyation values. Can be M or Beta. Beta values will be used.}
}
\value{
list of 3 elements: named vector of sex, chrX and chrY median beta values.
}
\description{
Infer biological sex from array methylomics data
}
