% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_message}
\alias{info_message}
\title{Print Info Message}
\usage{
info_message(msg)
}
\arguments{
\item{msg}{Message to print.}
}
\description{
Print a formatted info message.
}
\examples{
info_message("Hello")

}
