% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infomessage}
\alias{infomessage}
\title{Print Info Message}
\usage{
infomessage(..., n = 3, n0 = 2)
}
\arguments{
\item{n}{Number of calls to include in call stack.}

\item{n0}{Number of calls to exclude from the call stack.}

\item{msg}{Message to print.}
}
\value{
Prints a formatted message.
}
\description{
Print a formatted info message. By default it includes the last 3 calls.
}
\examples{
\dontrun{
info_message("Hello")
f <- function(x) {
  playbase::infomessage("test")
  x <- x + 1
  x
}
f1 <- function(x) {
  playbase::infomessage("test1")
  x <- f(x)
}
f2 <- function(x) {
  playbase::infomessage("test2")
  x <- f1(x)
}

f3 <- function(x) {
  playbase::infomessage("test3")
  x <- f2(x)
}

f4 <- function(x) {
  playbase::infomessage("test3")
  x <- f3(x)
}

f(1)
f4(1)
}
}
