% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{infoTileDB}
\alias{infoTileDB}
\title{Get TileDB database info}
\usage{
infoTileDB(tiledb_path)
}
\arguments{
\item{tiledb_path}{Path to the TileDB database}
}
\value{
List with database metadata (invisibly)
}
\description{
Get TileDB database info
}
