% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{is.Date}
\alias{is.Date}
\title{Check if a vector contains date values}
\usage{
is.Date(x)
}
\arguments{
\item{x}{A vector to check if it contains date values}
}
\value{
Logical. TRUE if the vector contains valid date values, FALSE otherwise.
}
\description{
Check if a vector contains date values
}
\details{
This function checks if a vector contains valid date values that can be converted to Date class.
It first replaces any "NA" values with actual NA values.
It then tries converting the vector to Date class using various date formats.
If none of the dates throw an error, it returns TRUE indicating the vector contains valid dates.
}
\examples{
\dontrun{
x <- c("2020-01-01", "01/15/2021", "foo")
is.Date(x) # TRUE
}
}
