% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{is.multiomics}
\alias{is.multiomics}
\title{Test if probes are multi-omics. All probe names should have a
'colon' (:) and prefixes are alpha-numeric without any spaces or
other characters. This is to avoid colons in unprefixed names (like
sometimes in metabolomics).}
\usage{
is.multiomics(probes)
}
\description{
Test if probes are multi-omics. All probe names should have a
'colon' (:) and prefixes are alpha-numeric without any spaces or
other characters. This is to avoid colons in unprefixed names (like
sometimes in metabolomics).
}
