% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{is.POSvsNEG}
\alias{is.POSvsNEG}
\title{Check if two groups represent a positive vs negative comparison}
\usage{
is.POSvsNEG(pgx)
}
\arguments{
\item{pgx}{pgx object containing the data}
}
\value{
Logical indicating if the two groups represent a positive vs negative comparison
}
\description{
Check if two groups represent a positive vs negative comparison
}
\details{
This function examines two groups from the pgx object to determine if they represent a positive vs negative comparison.
It checks the mean expression of signature genes in each group to see if one group has overall higher expression.  
It also looks for keywords like "neg", "untr", "wt" etc in the group names.
Based on these checks it returns a logical indicating if the groups represent a positive vs negative comparison.
}
