% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{is.num}
\alias{is.num}
\alias{isanumber}
\title{Check if a vector contains numeric data}
\usage{
is.num(y)

isanumber(x)
}
\arguments{
\item{y}{A vector to check if it contains numeric data}

\item{x}{A vector to check if it contains numeric data}
}
\value{
Logical. TRUE if the vector contains mostly numeric data, FALSE otherwise.
}
\description{
Check if a vector contains numeric data
}
\details{
This function checks if a vector contains mostly numeric data.
It first replaces any "NA" values with actual NA values.
It then removes any empty values and converts the vector to numeric.
If more than 50% of the resulting numeric vector is not NA, it returns TRUE.
}
\section{Functions}{
\itemize{
\item \code{is.num()}: The function tests if the input \code{x} is of numeric type.

}}
\examples{
\dontrun{
x <- c("1", "2", "foo", "3")
isanumber(x) # FALSE

x <- c("1", "2", NA, "3")
isanumber(x) # TRUE
}
}
