% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{justGSEA}
\alias{justGSEA}
\title{Run GSEA analysis using pre-ranked or phenotype comparison mode}
\usage{
justGSEA(
  X,
  gmt,
  Y = NULL,
  fdr = 1,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  rpt.label = "analysis",
  output.dir = NULL,
  xmx = 10,
  scoring = "weighted",
  sort.y = "ref.last",
  ref.type = c("0", "nt", "ref", "dmso", "wt", "dsmo"),
  clean.files = TRUE
)
}
\arguments{
\item{X}{The gene expression matrix or pre-ranked gene list.}

\item{gmt}{The GMT file or gene set collection.}

\item{Y}{The class labels or binary matrix indicating the phenotypes for phenotype comparison mode.}

\item{fdr}{The false discovery rate (FDR) threshold for selecting significant gene sets. Default is 1.}

\item{set.min}{The minimum gene set size. Default is 15.}

\item{set.max}{The maximum gene set size. Default is 500.}

\item{topgs}{The number of top gene sets to report. Default is 100.}

\item{nperm}{The number of permutations for estimating the statistical significance. Default is 1000.}

\item{rpt.label}{The label for the analysis report. Default is "analysis".}

\item{output.dir}{The output directory for saving the results. Default is NULL (results are not saved).}

\item{xmx}{The maximum memory allocation for the GSEA program in gigabytes (GB). Default is 10.}

\item{scoring}{The scoring scheme for ranking genes. Default is "weighted".}

\item{sort.y}{The sorting order of phenotypes for phenotype comparison mode. Default is "ref.last".}

\item{ref.type}{The reference type for phenotype comparison mode. Default is "0".}

\item{clean.files}{Logical value indicating whether to clean intermediate files. Default is TRUE.}
}
\value{
A list containing the results of the GSEA analysis, including q-values and NES (Normalized Enrichment Scores).
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) using either pre-ranked mode or phenotype comparison mode.
In pre-ranked mode, the function performs GSEA on a pre-ranked gene list. In phenotype comparison mode,
the function compares different phenotypes to identify enriched gene sets.
}
