% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{knnMedianFilter}
\alias{knnMedianFilter}
\title{Apply k-Nearest Neighbors Median Filter}
\usage{
knnMedianFilter(x, pos, k = 10)
}
\arguments{
\item{x}{A numeric vector of values to be filtered.}

\item{pos}{A matrix of positions for each element in `x`.}

\item{k}{An optional numeric value specifying the number of nearest neighbors to use for filtering.
The default value is 10.}
}
\value{
A numeric vector of the same length as `x`, containing the filtered values.
}
\description{
This function applies a k-nearest neighbors median filter to a vector of values.
}
\details{
This function takes a numeric vector `x` and a matrix of positions `pos` as input and applies a k-nearest neighbors median filter to `x`. 
The number of nearest neighbors used for filtering is specified by the `k` parameter. 
For each element in `x`, the function finds its `k` nearest neighbors in `pos` and calculates the median of their values in `x`. 
The resulting filtered values are returned as a numeric vector of the same length as `x`.
}
