% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{labels2rainbow}
\alias{labels2rainbow}
\title{Map module colors to rainbow palette}
\usage{
labels2rainbow(net)
}
\arguments{
\item{net}{A WGCNA network object.}
}
\value{
A named character vector mapping the original module colors to rainbow colors.
}
\description{
Maps the module colors from a WGCNA network to a rainbow palette.
}
\details{
This function takes a WGCNA network object and maps the module colors 
to a rainbow palette based on the modules' hierarchical clustering order.

It extracts the hierarchical clustering dendrogram order, gets the number
of unique module colors, ranks the module color means based on the 
dendrogram order, and assigns rainbow colors accordingly.

This allows easier visualization and interpretation of modules in the
standard rainbow palette order.
}
