% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{lasagna.multisolve}
\alias{lasagna.multisolve}
\title{Solves lasagna graph iteratively for multiple contrasts. The
solution is the RMS graph.}
\usage{
lasagna.multisolve(
  obj,
  min_rho = 0.2,
  traits = NULL,
  max_edges = 100,
  value.type = "rho",
  fc.weights = TRUE,
  prune = TRUE,
  sp.weight = FALSE
)
}
\description{
Solves lasagna graph iteratively for multiple contrasts. The
solution is the RMS graph.
}
