% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{length_encode}
\alias{length_encode}
\title{Encode values to length scales}
\usage{
length_encode(x, r = 0.1, a = 0.25)
}
\arguments{
\item{x}{A numeric vector to encode.}

\item{r}{The resolution of the length scale. Default is 0.1.}

\item{a}{The scale parameter for length scale. Default is 0.25.}
}
\value{
A sparse matrix encoding the input values to lengths.
}
\description{
Encodes a numeric vector to a length scale based on the logarithm of the values.
}
\details{
This function takes a numeric vector \code{x} and encodes the values to a length scale.
It first takes the log of the values in \code{x}, excluding any NA values.
It then constructs a set of length bins based on the min and max log values, with bin width determined by the \code{r} parameter.
Values in \code{x} are assigned to bins based on their log values.
The \code{a} parameter controls the scale of the length encoding.

The output is a sparse matrix with rows corresponding to the input \code{x} and columns for the length bins.
}
\examples{
\dontrun{
x <- rpois(100)
len <- length_encode(x)  
}
}
