% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{listDatasetsTileDB}
\alias{listDatasetsTileDB}
\title{List datasets in TileDB database}
\usage{
listDatasetsTileDB(tiledb_path)
}
\arguments{
\item{tiledb_path}{Path to the TileDB database}
}
\value{
Character vector of dataset names (sorted alphabetically)
}
\description{
Returns all unique dataset names stored in the TileDB database.
}
