% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{listGenesTileDB}
\alias{listGenesTileDB}
\title{List genes in TileDB database}
\usage{
listGenesTileDB(tiledb_path)
}
\arguments{
\item{tiledb_path}{Path to the TileDB database}
}
\value{
Character vector of gene names
}
\description{
List genes in TileDB database
}
