% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{listPhenotypesTileDB}
\alias{listPhenotypesTileDB}
\title{List phenotype columns in TileDB database}
\usage{
listPhenotypesTileDB(tiledb_path)
}
\arguments{
\item{tiledb_path}{Path to the TileDB database}
}
\value{
Character vector of phenotype column names
}
\description{
List phenotype columns in TileDB database
}
