% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{listSamplesTileDB}
\alias{listSamplesTileDB}
\title{List samples in TileDB database}
\usage{
listSamplesTileDB(tiledb_path)
}
\arguments{
\item{tiledb_path}{Path to the TileDB database}
}
\value{
Character vector of sample names (format: "filename::samplename")
}
\description{
List samples in TileDB database
}
