% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{LM22}
\alias{LM22}
\title{A dataset used by `compute_deconvolution`}
\format{
## `LM22`
}
\source{
unknown
}
\usage{
LM22
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
