% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{log1s}
\alias{log1s}
\title{Scaled log2(x/s+1) function. A generalization of log2(x+1) where we
allow a scaling factor s. This is also related to log2(x+p) with
custom prior p. However log1s maintains a minimum value of log1s=0
for all values of s.}
\usage{
log1s(counts, q = 0.2)
}
\arguments{
\item{counts}{A numeric matrix of count data, where rows represent
features and columns represent samples.}

\item{q}{numeric value corresponding to quantile to be used for
scaling. Default q=0.20.}
}
\value{
A numeric matrix of log-normalized count data.
}
\description{
This function normalizes count data using the scaled
  log1p method.
}
\examples{
\dontrun{
counts <- matrix(rnbinom(100, mu = 10, size = 1), ncol = 10)
norm_counts <- log1s(counts, q=0.01)
}
}
