% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{logMaxSumNorm}
\alias{logMaxSumNorm}
\title{Normalization for TMT and Silac data: logMaxSumNorm}
\usage{
logMaxSumNorm(counts, toLog = TRUE, prior = 0)
}
\arguments{
\item{counts}{Numeric matrix of raw intensities: genes in rows, samples in columns.}

\item{prior}{Pseudocount to add prior to log transform. Default is 1.}
}
\value{
Normalized matrix of log2-transformed values.
}
\description{
This function normalizes TMT and Silac data using max sum of intensities
}
\details{
Most used norm. for TMT/Silac data: make median or sum of each sample equal to max median or sum.
}
\examples{
\dontrun{
counts <- matrix(rnbinom(100 * 10, mu = 100, size = 1), 100, 10)
norm_counts <- logMaxSumNorm(counts)
}
}
