% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{make_unique_ws}
\alias{make_unique_ws}
\title{add empty white spaces after name to make unique.}
\usage{
make_unique_ws(s, sep = "", iter = 10 * length(s))
}
\description{
add empty white spaces after name to make unique.
}
