% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{makeAcronym}
\alias{makeAcronym}
\title{Make Acronym from String}
\usage{
makeAcronym(x)
}
\arguments{
\item{x}{A character string from which to create an acronym.}
}
\value{
A character string representing the acronym created from the input string.
}
\description{
This function creates an acronym from a character string.
}
\details{
The function takes a character string `x` as input and splits it into words using `_`, `-`, and space as delimiters.
If the resulting list of words contains only one word, the first two characters of the word are returned.
Otherwise, the first character of each word is extracted, capitalized, and concatenated to form the acronym.
}
\examples{
\dontrun{
# example code
makeAcronym("United Nations")
makeAcronym("World Health Organization")
}
}
