% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{makeContrastFromSamples}
\alias{makeContrastFromSamples}
\title{Create a labeled contrast matrix from samples dataframe.}
\usage{
makeContrastFromSamples(samples, ref = NULL)
}
\description{
Create a labeled contrast matrix from samples dataframe.
}
