% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{makeContrastsFromPairs}
\alias{makeContrastsFromPairs}
\title{Title}
\usage{
makeContrastsFromPairs(
  main.group,
  ref.group,
  groups = NULL,
  comparisons = NULL
)
}
\arguments{
\item{comparisons}{}
}
\description{
Title
}
