% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{makeDirectContrasts}
\alias{makeDirectContrasts}
\title{Title}
\usage{
makeDirectContrasts(Y, ref, na.rm = TRUE)
}
\arguments{
\item{Y}{value}

\item{ref}{value}

\item{na.rm}{value}
}
\description{
Title
}
