% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{makeDirectContrasts000}
\alias{makeDirectContrasts000}
\title{Title}
\usage{
makeDirectContrasts000(Y, ref, na.rm = TRUE, warn = FALSE)
}
\arguments{
\item{Y}{value}

\item{ref}{value}

\item{na.rm}{value}

\item{warn}{value}
}
\description{
Title
}
