% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{map2symbol}
\alias{map2symbol}
\title{Map any feature identifier to other feature column. Warning does
not retain original match order or length. Default mapping to
symbol. Searches all columns of annot for matching.}
\usage{
map2symbol(
  annot,
  genes,
  target = "symbol",
  column = NULL,
  ignore.case = TRUE,
  na.rm = TRUE
)
}
\description{
Map any feature identifier to other feature column. Warning does
not retain original match order or length. Default mapping to
symbol. Searches all columns of annot for matching.
}
