% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-util.r
\name{mat2hugo}
\alias{mat2hugo}
\title{Matrix rownames to HUGO gene symbols}
\usage{
mat2hugo(x)
}
\arguments{
\item{x}{A numeric matrix with rownames to convert.}
}
\value{
The input matrix with rownames converted to HUGO symbols.
}
\description{
Converts the rownames of a matrix to official HUGO gene symbols.
}
\details{
This function orders the rows by standard deviation, maps the
rownames to HUGO symbols using the playbase::alias2hugo function,
removes duplicates, and returns the resulting matrix with HUGO symbols
as rownames.
}
\examples{
\dontrun{
TODO
}

}
