% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{match.dataframe}
\alias{match.dataframe}
\title{Return a matched dataframe by matching id to any of its columns of
df. Returned dataframe has length(id) rows, and ncol(df) columns.}
\usage{
match.dataframe(id, df, parallel = TRUE)
}
\description{
Return a matched dataframe by matching id to any of its columns of
df. Returned dataframe has length(id) rows, and ncol(df) columns.
}
