% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{match_probe_names}
\alias{match_probe_names}
\title{Match dirty probe names to clean key names}
\usage{
match_probe_names(probes, orgdb, probe_type = NULL)
}
\description{
Match dirty probe names to clean key names
}
