% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{match_probes}
\alias{match_probes}
\title{Map any feature identifier to other feature column. Retain original
match order. Default mapping to rownames.}
\usage{
match_probes(
  annot,
  genes,
  column = NULL,
  ignore.case = FALSE,
  target = "rownames"
)
}
\description{
Map any feature identifier to other feature column. Retain original
match order. Default mapping to rownames.
}
