% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{matGroupMeans}
\alias{matGroupMeans}
\title{Calculate group-wise summary statistics from a matrix}
\usage{
matGroupMeans(X, group, FUN = rowMeans, dir = 1)
}
\arguments{
\item{X}{Numeric matrix with samples in columns.}

\item{group}{Factor indicating group membership for each column of X.}

\item{FUN}{Function to calculate row-wise statistics. Default is rowMeans.}

\item{dir}{Direction to apply FUN. 1=row-wise, 2=column-wise.}
}
\value{
Matrix of summary statistics, with rows corresponding to groups.
}
\description{
Calculate group-wise summary statistics from a matrix
}
\details{
This function calculates a row-wise or column-wise summary statistic
  for each group defined in the \code{group} vector.

  It applies the function specified by \code{FUN} (default is rowMeans) to summarize
  either rows or columns of \code{X}, depending on the \code{dir} argument.

  The result is a matrix with rows corresponding to the unique groups, and columns
  equal to the number of rows/columns summarized.
}
\examples{
\dontrun{
X <- matrix(rnorm(100 * 20), 100, 20)
groups <- sample(letters[1:5], 20, replace = TRUE)
means <- matGroupMeans(X, groups)
}
}
