% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{matrix_combine_p}
\alias{matrix_combine_p}
\title{Matrix version for combining p-values using fisher or stouffer
method. Much faster than doing metap::sumlog() and metap::sumz()}
\usage{
matrix_combine_p(plist, method = "fisher")
}
\description{
Matrix version for combining p-values using fisher or stouffer
method. Much faster than doing metap::sumlog() and metap::sumz()
}
