% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{matrix_onesample_ttest}
\alias{matrix_onesample_ttest}
\title{Fast one sample t-test for matrix object F (e.g. foldchanges) and
grouping matrix G (e.g. gene sets).}
\usage{
matrix_onesample_ttest(F, G)
}
\description{
Fast one sample t-test for matrix object F (e.g. foldchanges) and
grouping matrix G (e.g. gene sets).
}
