% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{matrix_twosample_ttest}
\alias{matrix_twosample_ttest}
\title{Fast one sample t-test for matrix object X (e.g. expression) and
grouping matrix G (e.g. gene sets).}
\usage{
matrix_twosample_ttest(X, G)
}
\description{
Fast one sample t-test for matrix object X (e.g. expression) and
grouping matrix G (e.g. gene sets).
}
