% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{maxSumNormalization.logX}
\alias{maxSumNormalization.logX}
\title{Total sum normalization. Warning this is different than
maxSumNormalization because is takes the sum of logX rather than X.}
\usage{
maxSumNormalization.logX(X)
}
\description{
Total sum normalization. Warning this is different than
maxSumNormalization because is takes the sum of logX rather than X.
}
