% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{merge_annot_table}
\alias{merge_annot_table}
\title{Merges any missing annotation in df with non-missing annotation of
df2.}
\usage{
merge_annot_table(df, df2, priority = 1)
}
\description{
Merges any missing annotation in df with non-missing annotation of
df2.
}
