% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{merge_sparse_matrix}
\alias{merge_sparse_matrix}
\title{Default merge by columns (cbind) with shared features on
rows. Features are union of input matrices.}
\usage{
merge_sparse_matrix(m1, m2, margin = NULL, verbose = 1)
}
\description{
Default merge by columns (cbind) with shared features on
rows. Features are union of input matrices.
}
