% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mgsea.compute_enrichment}
\alias{mgsea.compute_enrichment}
\title{Compute multi-enrichment for contrasts}
\usage{
mgsea.compute_enrichment(F, G, filter = NULL, ntop = 1000)
}
\description{
Compute multi-enrichment for contrasts
}
