% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.compute_clusters}
\alias{mofa.compute_clusters}
\title{Compute clusters for MOFA factors.}
\usage{
mofa.compute_clusters(xx, matF, along = "samples")
}
\description{
Compute clusters for MOFA factors.
}
