% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.compute_enrichment}
\alias{mofa.compute_enrichment}
\title{Compute enrichment for MOFA factors.}
\usage{
mofa.compute_enrichment(
  ww,
  ww.types = NULL,
  filter = NULL,
  G = NULL,
  ntop = NULL
)
}
\description{
Compute enrichment for MOFA factors.
}
