% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.compute_geneset_mofa}
\alias{mofa.compute_geneset_mofa}
\title{Compute enrichment for MOFA factors.}
\usage{
mofa.compute_geneset_mofa(
  mofa,
  kernel = "mofa",
  GMT = NULL,
  numfactors = 20,
  factorname = "Module"
)
}
\description{
Compute enrichment for MOFA factors.
}
