% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.create_report}
\alias{mofa.create_report}
\title{Create report object.}
\usage{
mofa.create_report(
  mofa,
  llm_model,
  graph = NULL,
  annot = NULL,
  ntop = 100,
  psig = 0.05,
  do.diagram = TRUE,
  img_model = NULL,
  userprompt = "",
  format = "markdown",
  verbose = 1,
  progress = NULL
)
}
\description{
Create report object.
}
