% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.imputeMissing}
\alias{mofa.imputeMissing}
\title{Impute missing values for a multiomics expression matrix
X. Features must be prefixed with datatype.}
\usage{
mofa.imputeMissing(X, method = "SVD2")
}
\description{
Impute missing values for a multiomics expression matrix
X. Features must be prefixed with datatype.
}
