% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.intNMF}
\alias{mofa.intNMF}
\title{Integrated NMF.}
\usage{
mofa.intNMF(
  datasets,
  k = NULL,
  method = "RcppML",
  opt.method = "RcppML",
  maxiter = 200,
  separate.sign = FALSE,
  shift = TRUE,
  scale = TRUE,
  exponentiate = FALSE
)
}
\description{
Integrated NMF.
}
