% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{mofa.log1s}
\alias{mofa.log1s}
\title{Normalize a multi-omics count matrix with log1s per datatype.}
\usage{
mofa.log1s(counts, q = 0.2)
}
\description{
Normalize a multi-omics count matrix with log1s per datatype.
}
