% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.merge_data}
\alias{mofa.merge_data}
\title{This merges list of multi-omics data to a single matrix. Columns
must match. Merge data by row after adding prefix.}
\usage{
mofa.merge_data(xx)
}
\description{
This merges list of multi-omics data to a single matrix. Columns
must match. Merge data by row after adding prefix.
}
