% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.normalizeExpression}
\alias{mofa.normalizeExpression}
\title{Normalize matrix for a multiomics expression matrix X. Features
must be prefixed with datatype.}
\usage{
mofa.normalizeExpression(X, method1 = "maxMedian", method2 = "none")
}
\description{
See also: normalizeMultiOmics()
}
