% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.plot_biplot}
\alias{mofa.plot_biplot}
\title{Plots from results from compute.mofa()}
\usage{
mofa.plot_biplot(r, pheno, nfeat = 5, comp = c(1, 2), arrow = TRUE, cex = 1)
}
\description{
Plots from results from compute.mofa()
}
