% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.plot_covariate_correlation}
\alias{mofa.plot_covariate_correlation}
\title{Covariate x Factor correlation plot. Shows how factors are
correlated with response covariates.}
\usage{
mofa.plot_covariate_correlation(
  mofa,
  collapse = FALSE,
  mar = c(5, 10),
  nmax = 60,
  ...
)
}
\description{
Covariate x Factor correlation plot. Shows how factors are
correlated with response covariates.
}
