% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report-mofa.R
\name{mofa_assemble_prompt}
\alias{mofa_assemble_prompt}
\title{Assemble the MOFA report prompt.}
\usage{
mofa_assemble_prompt(slice, pgx, ai)
}
\description{
Assemble the MOFA report prompt.
}
\keyword{internal}
