% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report-mofa.R
\name{mofa_build_methods}
\alias{mofa_build_methods}
\title{Build deterministic methods section for MOFA report.}
\usage{
mofa_build_methods(mofa, pgx)
}
\description{
Build deterministic methods section for MOFA report.
}
