% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report-mofa.R
\name{mofa_rank_factors}
\alias{mofa_rank_factors}
\title{Rank MOFA factors for reporting.}
\usage{
mofa_rank_factors(factor_summary)
}
\arguments{
\item{factor_summary}{MOFA factor summary table.}
}
\value{
Character vector of factor names.
}
\description{
Rank MOFA factors for reporting.
}
