% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{mToBeta}
\alias{mToBeta}
\title{mToBeta}
\usage{
mToBeta(m)
}
\arguments{
\item{X}{Matrix of M values. Probes in rows; samples in columns.}
}
\value{
Matrix of Beta values.
}
\description{
Convert  M to Beta values; 450K+850K (EPIC) Meth Array.
}
