% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{multi.dist}
\alias{multi.dist}
\title{Title}
\usage{
multi.dist(x, p = 4, method = c("pearson", "euclidean", "manhattan"))
}
\arguments{
\item{x}{value}

\item{p}{value}

\item{method}{value}
}
\description{
Title
}
