% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{multimatch}
\alias{multimatch}
\title{matches id to rows on any columns of dataframe df. Generalization
of match().}
\usage{
multimatch(id, df, parallel = TRUE)
}
\description{
matches id to rows on any columns of dataframe df. Generalization
of match().
}
