% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.annotateLipids.000}
\alias{mx.annotateLipids.000}
\title{Annotate lipids using RefMet and rgoslin. One-pass method. Only
annotate with given ID.}
\usage{
mx.annotateLipids.000(
  name,
  db = c("rgoslin", "refmet", "ramp"),
  add_id = TRUE,
  harmonize = TRUE,
  verbose = 1
)
}
\description{
Annotate lipids using RefMet and rgoslin. One-pass method. Only
annotate with given ID.
}
