% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.annotateLipids}
\alias{mx.annotateLipids}
\title{Two-pass annotation of lipids using RefMet and rgoslin. This is a
two-pass approach where missing annotations are retried using
synonyms IDs of the original ID. This can improve annotation
coverage.}
\usage{
mx.annotateLipids(
  name,
  db = c("rgoslin", "refmet", "ramp"),
  harmonize = TRUE,
  two.pass = TRUE
)
}
\description{
Two-pass annotation of lipids using RefMet and rgoslin. This is a
two-pass approach where missing annotations are retried using
synonyms IDs of the original ID. This can improve annotation
coverage.
}
