% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.check_mapping}
\alias{mx.check_mapping}
\title{Check metabolite mapping in databases. For each probe return
matched database or NA if not found.}
\usage{
mx.check_mapping(
  probes,
  all.db = c("playdata", "annothub", "refmet"),
  check.first = TRUE
)
}
\description{
Check metabolite mapping in databases. For each probe return
matched database or NA if not found.
}
