% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.convert_probe}
\alias{mx.convert_probe}
\title{Convert IDs to CHEBI using internal playdata annotation table. Can
handle different target columns.}
\usage{
mx.convert_probe(probes, probe_type = NULL, target_id = "ID")
}
\description{
Convert IDs to CHEBI using internal playdata annotation table. Can
handle different target columns.
}
