% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.create_metabolite_sets}
\alias{mx.create_metabolite_sets}
\title{This adds some 'gene sets' based on the species class of the lipids
according to the annotation table.}
\usage{
mx.create_metabolite_sets(annot, gmin = 0, metmin = 5, as_matrix = FALSE)
}
\description{
This adds some 'gene sets' based on the species class of the lipids
according to the annotation table.
}
