% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.get_metabolite_mapping}
\alias{mx.get_metabolite_mapping}
\title{This uses different databases to create a ID mapping table for
given 'id' to other ID types (ChEBI, HMDB, LIPIDMAPS, KEGG,
REFMET).}
\usage{
mx.get_metabolite_mapping(
  met,
  method = c("refmet", "playdata", "annhub"),
  verbose = 1
)
}
\description{
This uses different databases to create a ID mapping table for
given 'id' to other ID types (ChEBI, HMDB, LIPIDMAPS, KEGG,
REFMET).
}
