% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.harmonizeMainclassNames}
\alias{mx.harmonizeMainclassNames}
\title{Translates rgoslin lipid classes to RefMet lipid classes}
\usage{
mx.harmonizeMainclassNames(name)
}
\description{
Translates rgoslin lipid classes to RefMet lipid classes
}
