% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.harmonizeSubclassNames}
\alias{mx.harmonizeSubclassNames}
\title{Translates rgoslin lipid classes to RefMet lipid classes}
\usage{
mx.harmonizeSubclassNames(name)
}
\description{
Translates rgoslin lipid classes to RefMet lipid classes
}
