% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.harmonizeSuperclassNames}
\alias{mx.harmonizeSuperclassNames}
\title{Translates rgoslin lipid classes to RefMet lipid classes}
\usage{
mx.harmonizeSuperclassNames(name)
}
\description{
Translates rgoslin lipid classes to RefMet lipid classes
}
