% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{myplot_ly}
\alias{myplot_ly}
\title{Dark Mode for Plotly Plots}
\usage{
myplot_ly(..., theme = "default")
}
\arguments{
\item{p}{A plotly plot object.}

\item{dim}{The number of dimensions of the plot (2 or 3).}
}
\value{
A modified plotly plot object with the dark mode theme applied.
}
\description{
This function applies a dark mode theme to a plotly plot object.
}
\details{
This function modifies the layout of the input plotly 
plot object to apply a dark mode theme. It changes the background 
color of the plot and paper to a dark color, and sets the font and 
axis colors to a light gray color.
}
