% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-cook.r
\name{ngs.cookForDESEQ2}
\alias{ngs.cookForDESEQ2}
\title{Title}
\usage{
ngs.cookForDESEQ2(
  counts,
  samples,
  genes,
  remove.batch = TRUE,
  test = "Wald",
  prior.cpm = 0,
  filter = TRUE
)
}
\arguments{
\item{filter}{}
}
\description{
Title
}
