% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitConstrastsWithDESEQ2}
\alias{ngs.fitConstrastsWithDESEQ2}
\title{Title}
\usage{
ngs.fitConstrastsWithDESEQ2(
  counts,
  group,
  contr.matrix,
  design,
  X = NULL,
  genes = NULL,
  test = "Wald",
  prune.samples = FALSE,
  conform.output = FALSE
)
}
\arguments{
\item{counts}{value}

\item{group}{value}

\item{contr.matrix}{value}

\item{design}{value}

\item{X}{value}

\item{genes}{value}

\item{test}{value}

\item{prune.samples}{value}

\item{conform.output}{value}
}
\description{
Title
}
