% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithAllMethods}
\alias{ngs.fitContrastsWithAllMethods}
\title{Title}
\usage{
ngs.fitContrastsWithAllMethods(
  counts,
  X = NULL,
  samples,
  design,
  contr.matrix,
  genes = NULL,
  prior.cpm = 1,
  prune.samples = FALSE,
  conform.output = TRUE,
  do.filter = TRUE,
  cpm.scale = 1e+06,
  remove.batch = TRUE,
  methods = ALL.GENETEST.METHODS,
  correct.AveExpr = TRUE,
  custom = NULL,
  custom.name = NULL
)
}
\arguments{
\item{custom.name}{}
}
\description{
Title
}
