% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithDESEQ2.regress.covs}
\alias{ngs.fitContrastsWithDESEQ2.regress.covs}
\title{ngs.fitContrastsWithDESEQ2.regress.covs. See ngs.fitContrastsWithDESEQ2().}
\usage{
ngs.fitContrastsWithDESEQ2.regress.covs(
  counts,
  X,
  y,
  covariate,
  covariate.name = "covariate",
  test = "Wald",
  fitType = "mean"
)
}
\arguments{
\item{counts}{counts matrix. Features in rows; samples in columns.}

\item{X}{log2-transformed (and normalized) data matrix. Features in rows; samples in columns.}

\item{y}{phenotype vector (main contrast)}

\item{covariate}{covariate vector}

\item{covariate.name}{covariate name.}
}
\value{
Table with p-values of main contrast adjusted for the covariate.
}
\description{
Regress out covariate from main contrast using DESeq2.
}
