% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithEDGER}
\alias{ngs.fitContrastsWithEDGER}
\title{Title}
\usage{
ngs.fitContrastsWithEDGER(
  counts,
  group,
  contr.matrix,
  design,
  method = c("qlf", "lrt"),
  prune.samples = FALSE,
  X = NULL,
  conform.output = FALSE,
  robust = TRUE,
  plot = TRUE
)
}
\arguments{
\item{plot}{}
}
\description{
Title
}
