% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithEDGER.regress.covs}
\alias{ngs.fitContrastsWithEDGER.regress.covs}
\title{ngs.fitContrastsWithEDGER.regress.covs. See ngs.fitContrastsWithEDGER().}
\usage{
ngs.fitContrastsWithEDGER.regress.covs(
  counts,
  X,
  group,
  y,
  covariate,
  covariate.name = "covariate",
  method,
  robust = TRUE,
  calc.tmm = TRUE
)
}
\arguments{
\item{counts}{counts matrix. Features in rows; samples in columns.}

\item{X}{log2-transformed (and normalized) data matrix. Features in rows; samples in columns.}

\item{y}{phenotype vector (main contrast)}

\item{covariate}{covariate vector}

\item{covariate.name}{covariate name.}

\item{method}{"qlf" or "lrt"}
}
\value{
Table with p-values of main contrast adjusted for the covariate.
}
\description{
Regress out covariate from main contrast using EdgeR.
}
