% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithLIMMA}
\alias{ngs.fitContrastsWithLIMMA}
\title{Title}
\usage{
ngs.fitContrastsWithLIMMA(
  X,
  contr.matrix,
  design,
  method = c("voom", "limma"),
  trend = TRUE,
  robust = TRUE,
  prune.samples = FALSE,
  conform.output = FALSE,
  plot = FALSE
)
}
\arguments{
\item{X}{value}

\item{contr.matrix}{value}

\item{design}{value}

\item{method}{value}

\item{trend}{value}

\item{robust}{value}

\item{prune.samples}{value}

\item{conform.output}{value}

\item{plot}{value}
}
\description{
Title
}
