% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithLIMMA.regress.covs}
\alias{ngs.fitContrastsWithLIMMA.regress.covs}
\title{ngs.fitContrastsWithLIMMA.regress.covs. See ngs.fitContrastsWithLIMMA().}
\usage{
ngs.fitContrastsWithLIMMA.regress.covs(
  X,
  y,
  covariate,
  covariate.name = "covariate",
  trend = TRUE,
  robust = TRUE
)
}
\arguments{
\item{X}{log2-transformed data matrix. Features in rows; samples in columns.}

\item{y}{phenotype vector (main contrast)}

\item{covariate}{covariate vector}
}
\value{
Table with p-values of main contrast adjusted for the covariate.
}
\description{
Regress out covariate from main contrast using limma.
}
