% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithTTEST}
\alias{ngs.fitContrastsWithTTEST}
\title{Title}
\usage{
ngs.fitContrastsWithTTEST(
  X,
  contr.matrix,
  design,
  method = "welch",
  conform.output = 0
)
}
\arguments{
\item{conform.output}{}
}
\description{
Title
}
