% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{ngs.getGeneAnnotation_DEPRECATED}
\alias{ngs.getGeneAnnotation_DEPRECATED}
\title{Get gene annotation data}
\usage{
ngs.getGeneAnnotation_DEPRECATED(probes, probe_type, organism)
}
\arguments{
\item{probes}{Character vector of gene/transcript identifiers.}

\item{probe_type}{Type of identifiers provided in probes (e.g. "ENSEMBL").}

\item{organism}{Organism name (e.g. "Human", "Mouse", "Rat").}
}
\value{
Data frame containing annotation data for the provided identifiers.
}
\description{
Retrieves gene annotation data from an organism-specific
annotation package. This function has been deprecated
and is mantained for fallback purposes.
}
\details{
Queries an annotation package for the specified organism 
using the provided identifiers and probe type. Returns a cleaned and 
reformatted data frame with feature identifiers as row names.
}
\examples{
probes <- head(rownames(counts), 100)
probe_type <- "ENSEMBL"
organism <- "Human"
annotation <- ngs.getGeneAnnotation_DEPRECATED(probes, probe_type, organism)

}
