% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-salmon.r
\name{ngs.rawSalmon}
\alias{ngs.rawSalmon}
\title{Import Salmon quantification results}
\usage{
ngs.rawSalmon(sampleTable, gencode, txi = NULL)
}
\arguments{
\item{sampleTable}{A data frame with sample info and paths to Salmon sf files.}

\item{gencode}{Annotations for transcripts (e.g. from Ensembl).}

\item{txi}{Pre-computed Salmon tximport object. If NULL will be calculated.}
}
\value{
A DGEList containing imported counts, sample data, and gene annotations.
}
\description{
Imports Salmon transcript-level quantification results into an NGS object.
}
\details{
This function takes a data frame \code{sampleTable} containing sample information and paths to Salmon sf files generated by Salmon quantification. 
It runs \code{\link{tximport}} to import the Salmon results and aggregate to gene-level counts.

Gene annotations are added from the provided \code{gencode} data frame based on Ensembl IDs.
The output is an \code{\link[edgeR]{DGEList}} object containing the imported counts and sample data ready for differential analysis.
 
If a \code{txi} object is already provided, it will be used instead of re-importing the Salmon results.
}
