% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-salmon.r
\name{ngs.tximportSalmon}
\alias{ngs.tximportSalmon}
\title{Import Salmon Abundance Estimates}
\usage{
ngs.tximportSalmon(
  sf.files,
  count.type = "lengthScaledTPM",
  organism = "Hsapiens",
  txOut = FALSE
)
}
\arguments{
\item{sf.files}{A named character vector of Salmon SF files.}

\item{count.type}{The type of counts to extract, either "TPM", "NumReads" or "lengthScaledTPM".}

\item{organism}{The organism name, either "Hsapiens" or "Mmusculus".}

\item{txOut}{Whether to output transcript-level estimates in addition to gene-level.}
}
\value{
A DGEList object containing the abundance estimates, ready for use with edgeR.
}
\description{
This function imports Salmon transcript abundance estimates from Salmon SF files
and converts them to a DGEList object for use with edgeR.
}
\details{
This function takes a named character vector of Salmon SF files, extracts the abundance
estimates, converts them to a DGEList object with integer counts scaled to library size.
It requires the EnsDb and org packages for the corresponding organism.
}
