% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-impute.R
\name{nmfImpute}
\alias{nmfImpute}
\title{Impute missing values with non-negative matrix factorization}
\usage{
nmfImpute(x, k = 5)
}
\description{
Imputes missing values in matrix non-negative matrix factorization
}
