% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-impute.R
\name{nmfImpute2}
\alias{nmfImpute2}
\title{Impute missing values with iterative NMF}
\usage{
nmfImpute2(x, k = 5, niter = 10, init = 0.05)
}
\description{
Imputes missing values in matrix with iterative
   non-negative matrix factorization
}
