% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct-superbatch.R
\name{nnmCorrect2}
\alias{nnmCorrect2}
\title{Nearest neighbor matching batch correction}
\usage{
nnmCorrect2(
  X,
  y,
  r = 0.35,
  center.x = TRUE,
  center.m = TRUE,
  scale.x = FALSE,
  center.y = TRUE,
  mode = "sym",
  knn = 1,
  sdtop = 2000,
  return.B = FALSE,
  use.design = TRUE,
  delete.kin = TRUE
)
}
\arguments{
\item{X}{Numeric matrix of gene expression values (genes in rows, samples in columns).}

\item{y}{Factor vector indicating batch for each sample.}

\item{center.x}{Logical for whether to center gene expression by row means.}

\item{center.m}{Logical for whether to center expression by batch means.}

\item{sdtop}{Number of top variable genes to use for correlation.}

\item{dist.method}{Distance metric to use for matching ('cor' or 'euclidean').}
}
\value{
List containing:
\itemize{
  \item X - Batch corrected gene expression matrix
  \item pairings - Matrix of sample pairings used for correction
}
}
\description{
Correct for batch effects in a gene expression matrix using nearest neighbor matching.
}
\details{
This function performs batch correction using the following steps:
\enumerate{
  \item Compute distance matrix between all samples
  \item Find nearest neighbor matches between batches
  \item Construct full paired dataset with matches
  \item Apply limma batch correction to paired data
  \item Average paired samples back to original samples
}
}
\examples{
# TODO

}
\seealso{
\code{\link[limma]{removeBatchEffect}} for the batch correction method used.
}
