% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{normalize_matrix_by_row}
\alias{normalize_matrix_by_row}
\title{Normalize geneset matrix by row}
\usage{
normalize_matrix_by_row(G)
}
\arguments{
\item{g}{a sparse matrix (eg. genesets x genes)}
}
\description{
Normalize geneset matrix by row
}
