% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{normalize_multirank}
\alias{normalize_multirank}
\title{Normalize a rank vector for each datatype to [-1;1]. This is useful
for doing integrated GSEA on multi-omics data.}
\usage{
normalize_multirank(rnk)
}
\description{
Normalize a rank vector for each datatype to [-1;1]. This is useful
for doing integrated GSEA on multi-omics data.
}
