% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{normalize_rows}
\alias{normalize_rows}
\title{Normalize Matrix by Row}
\usage{
normalize_rows(G)
}
\arguments{
\item{G}{The matrix to be normalized.}
}
\value{
The normalized matrix.
}
\description{
Normalizes a matrix by dividing each row by the sum of its elements.
}
