% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{normalizeData}
\alias{normalizeData}
\title{All-in-one normalization function.}
\usage{
normalizeData(
  pgx,
  do.impute = TRUE,
  do.regress = TRUE,
  impute.method = "SVD2",
  scale.method = "cpm"
)
}
\description{
All-in-one normalization function.
}
