% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{normalizeMethylation}
\alias{normalizeMethylation}
\title{normalizeMethylation}
\usage{
normalizeMethylation(X, method = "BMIQ", meth_type = "450K array", nfit = 2000)
}
\arguments{
\item{X}{Matrix of Beta or M-values. We use Beta. Probes in rows; samples in columns.}

\item{method}{Normalization method(s) to use. At the moment BMIQ or quantile. To expand.}

\item{nfit}{Number of probes type I and II for the fitting. In most cases, 5000 or 10000 is ok.}
}
\value{
Normalized Beta values matrix.
}
\description{
Normalizes a matrix of 450K and EPIC Methylation data.
}
