% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{normalizeMultiOmics}
\alias{normalizeMultiOmics}
\title{Normalize multiomics counts}
\usage{
normalizeMultiOmics(
  X,
  prior = NULL,
  gx.method = "CPM",
  non.gx.method = "maxMedian"
)
}
\arguments{
\item{prior}{If NULL, computes it.}

\item{counts}{Multiomics data matrix: counts. Features in rows and samples in columns.}
}
\value{
Normalized multi-omics data matrix
}
\description{
Normalize each data type independently in a multiomics data matrix (counts)
}
\details{
This function normalizes each datatype independently.
Default normalization methods are logCPM for gx and maxMedian for non-gx data.
}
