% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{normalizeRLE}
\alias{normalizeRLE}
\title{Normalize count data using RLE method}
\usage{
normalizeRLE(counts, log = FALSE, use = "deseq2")
}
\arguments{
\item{counts}{A numeric matrix of count data, where rows represent features and columns represent samples.}

\item{log}{Logical, whether to return log-transformed normalized counts.}

\item{use}{Character, the package to use for normalization, either "edger" or "deseq2".}
}
\value{
A numeric matrix of normalized count data.
}
\description{
This function normalizes count data using the Relative Log Expression (RLE) method.
It supports normalization using either the edgeR or DESeq2 package.
}
\examples{
\dontrun{
counts <- matrix(rnbinom(100, mu=10, size=1), ncol=10) 
norm_counts <- normalizeRLE(counts)
}
}
