% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-colors.R
\name{omics_colors}
\alias{omics_colors}
\title{Function to extract omics colors as hex codes}
\usage{
omics_colors(...)
}
\arguments{
\item{...}{Character names of colors}
}
\description{
Function to extract omics colors as hex codes
}
\examples{
omics_colors()
omics_colors("brand_blue")

}
