% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-colors.R
\name{omics_pal_c}
\alias{omics_pal_c}
\title{Return function to interpolate a continuous omics color palette}
\usage{
omics_pal_c(palette = "brand_blue", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omics_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to [grDevices::colorRampPalette()]}
}
\description{
Return function to interpolate a continuous omics color palette
}
\examples{
omics_pal_c()(10)
omics_pal_c("red")(3)
omics_pal_c("bLUE_rED", reverse = TRUE)(5)

}
