% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-hiveplot.R
\name{omx.makeHivePlotData_}
\alias{omx.makeHivePlotData_}
\title{Create data for hive plot visualization}
\usage{
omx.makeHivePlotData_(
  res,
  rho.min = 0.15,
  cxi = 0.11,
  use.alpha = TRUE,
  ntop = 1000,
  ew = 4,
  nv = 10,
  dx.dir = c(-1, 1),
  rx = 2.2
)
}
\arguments{
\item{res}{Data frame of correlation statistics between features and phenotype.}

\item{rho.min}{Minimum correlation threshold.}

\item{cxi}{Node circle size parameter.}

\item{use.alpha}{Use transparency based on correlation strength.}

\item{ntop}{Number of top features to include.}

\item{ew}{Edge width parameter.}

\item{nv}{Number of vertices for node polygons.}

\item{dx.dir}{X axis direction.}

\item{rx}{Hive plot x-radius.}
}
\value{
A list with two data frames 'nodes' and 'edges' containing node and edge properties
for generating a hive plot.
}
\description{
Generates data frames required to visualize associations between omics datasets
using a hive plot.
}
\details{
This function takes a data frame of correlation statistics between features in
gene expression, copy number, methylation datasets vs a phenotype. It filters,
orders, and processes the data into formats needed for generating a hive plot using the
function.

The main outputs are two nested data frames defining the nodes and edges for the plot.
Visual parameters like node size, color, transparency, etc. are set based on the input
correlation data.
}
