% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-getgeo.R
\name{parse_geo_series_matrix}
\alias{parse_geo_series_matrix}
\title{Parse GEO series matrix file}
\usage{
parse_geo_series_matrix(
  SERIES_FILE,
  PLATFORM_FILE,
  GENE_COLUMN = "GENE_SYMBOL",
  EXPRESSION_OUTPUT_FILE = "expression.csv",
  ANNOTATION_OUTPUT_FILE = "annotation.csv",
  write.file = TRUE
)
}
\arguments{
\item{SERIES_FILE}{Character string. Path to GEO series matrix file.}

\item{PLATFORM_FILE}{Character string. Path to GEO platform file.}

\item{GENE_COLUMN}{Character string. Name of column containing gene symbols. Default "GENE_SYMBOL".}

\item{EXPRESSION_OUTPUT_FILE}{Character string. Name of expression matrix output file. Default "expression.csv".}

\item{ANNOTATION_OUTPUT_FILE}{Character string. Name of annotation output file. Default "annotation.csv".}

\item{write.file}{Logical. Whether to write output files. Default TRUE.}
}
\value{
List containing expression matrix and annotation data frame if write.file is FALSE.
}
\description{
Parse GEO series matrix file
}
\details{
Parses a GEO series matrix file and corresponding platform file to extract expression data and annotations.
The gene symbols, expression values, and annotations are extracted into R objects.
The expression matrix and annotation data frame can be optionally written to CSV files.
}
