% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report-pathways.R
\name{pathways_assemble_prompt}
\alias{pathways_assemble_prompt}
\title{Assemble the pathway-enrichment report prompt.}
\usage{
pathways_assemble_prompt(slice, pgx, ai)
}
\description{
Assemble the pathway-enrichment report prompt.
}
\keyword{internal}
