% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{pcsf.cut_and_relayout}
\alias{pcsf.cut_and_relayout}
\title{Cut graph into components using Louvain clustering and calculate
merged layout.}
\usage{
pcsf.cut_and_relayout(
  net,
  ncomp = -1,
  cluster.method = c("louvain", "leiden")[1],
  leiden.resolution = 0.1,
  layout = c("kk", "tree", "circle", "graphopt")[1],
  as_grid = TRUE
)
}
\description{
Cut graph into components using Louvain clustering and calculate
merged layout.
}
