% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-graph.R
\name{pgx._addSourceSink}
\alias{pgx._addSourceSink}
\title{Add source and sink nodes}
\usage{
pgx._addSourceSink(gr)
}
\arguments{
\item{gr}{An igraph pathway graph object}
}
\value{
An igraph object with source and sink nodes added
}
\description{
Adds source and sink nodes to a pathway graph to enable path scoring.
}
\details{
Add source and sink nodes to pathway graph


This function takes an igraph pathway graph object \code{gr} and adds
source and sink nodes to it.

The source and sink nodes are named "SOURCE" and "SINK" respectively.
They are assigned layout positions above and below the existing nodes.
The fold changes and scaled data for the source and sink are set to 1s.

Edges are added from the source to the top layer nodes, and from the bottom layer nodes to the sink.
All edge weights are set to 1.
}
