% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{pgx._computeNumSig}
\alias{pgx._computeNumSig}
\title{Compute number of significant genes}
\usage{
pgx._computeNumSig(ngs, X, contrast = NULL, fc = 0, qv = 0.05)
}
\arguments{
\item{ngs}{Object containing NGS analysis data}

\item{X}{Expression matrix}

\item{contrast}{Contrast to test}

\item{fc}{Fold change cutoff}

\item{qv}{FDR cutoff}
}
\value{
Named vector containing number of significant genes for each parameter combination
}
\description{
Counts number of significant genes for given parameters.
}
\details{
This function fits the specified contrast and counts the number of genes
significant at the given FDR and fold change cutoffs.

It requires an NGS analysis object, expression matrix, and contrast. Batch
correction is performed using parameters defined in the NGS object. The contrast is
then fit using limma-voom and number of significant genes counted.

Fold change and FDR cutoffs can also be specified. The number of significant genes
passing these thresholds is returned for each parameter combination.
}
