% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{pgx._runComputeNumSig}
\alias{pgx._runComputeNumSig}
\title{Compute Number of Significant Genes}
\usage{
pgx._runComputeNumSig(
  ngs,
  parcomb,
  contrast,
  resample = -1,
  normalization = NORMALIZATION.METHODS,
  show.progress = 1
)
}
\arguments{
\item{contrast}{Character vector of contrasts to test}

\item{resample}{Resampling fraction}

\item{normalization}{Character vector of normalization methods to try}

\item{show.progress}{Show progress}

\item{obj}{A pgx object}

\item{batch}{A character vector of batch parameters}

\item{niter}{Number of iterations for resampling}
}
\value{
Returns a data frame with the number of significant genes for each parameter/method combination.
}
\description{
Counts number of significant genes for each combination of batch correction 
  parameters and normalization methods. Used to optimize batch correction.
}
\details{
This function systematically goes through different combinations of batch correction
  parameters and normalization methods to correct the data in obj. For each combination, it performs 
  differential expression analysis for the specified contrasts, and counts the number of significant
  genes. This can be used to optimize batch correction by finding the combination that maximizes 
  the number of significant genes.
  
  The batch parameters, normalization methods, number of iterations and resampling fraction can be
  specified. Progress is shown if show.progress is TRUE.
}
