% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{pgx.addEnrichmentSignaturesH5}
\alias{pgx.addEnrichmentSignaturesH5}
\title{Add Enrichment Signatures to HDF5 File}
\usage{
pgx.addEnrichmentSignaturesH5(
  h5.file,
  X = NULL,
  mc.cores = 0,
  methods = c("gsea", "rankcor")[2]
)
}
\arguments{
\item{h5.file}{The HDF5 file path to write enrichment results to.}

\item{X}{The gene expression matrix. If NULL, matrix is read from HDF5 file.}

\item{mc.cores}{Number of cores for parallel processing. Default 0.}

\item{methods}{Enrichment methods to run. Default c("gsea", "rankcor")[2].}
}
\value{
No value is returned. Enrichment signatures are written to the HDF5 file.
}
\description{
Adds gene set enrichment results as signatures to an HDF5 file.
}
\details{
This function runs gene set enrichment analysis on a gene expression matrix X,
using the specified methods. It writes the enrichment results as signatures to
the specified HDF5 file under the "enrichment" group.

Parallel processing can be enabled by setting mc.cores to use multiple CPU cores.
By default, it runs only the "rankcor" method. Alternative methods like "gsea" can
be specified in the methods parameter.
}
