% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.barplot.PLOTLY}
\alias{pgx.barplot.PLOTLY}
\title{Bar plot using plotly}
\usage{
pgx.barplot.PLOTLY(
  data,
  x = NULL,
  y = NULL,
  title = NULL,
  color = "#3181de",
  fillcolor = "#A6CEE3",
  linecolor = "#3181de",
  titlecolor = "#1f77b4",
  hoverinfo = "y",
  hoverformat = ".2f",
  yaxistitle = FALSE,
  xaxistitle = FALSE,
  xlen = NULL,
  yrange = NULL,
  barmode = "relative",
  font_family = "Lato",
  margin = list(l = 0, r = 0, b = 0, t = 0),
  grouped = TRUE,
  annotations = NULL,
  pct.NA = NULL
)
}
\arguments{
\item{data}{Data frame to plot}

\item{x}{Column in data to use for x-axis. Default NULL.}

\item{y}{Column in data to use for y-axis. Default NULL.}

\item{title}{Plot title. Default NULL.}

\item{color}{Bar color. Default "#3181de".}

\item{fillcolor}{Bar fill color. Default "#2fb5e3".}

\item{linecolor}{Bar border color. Default "#3181de".}

\item{titlecolor}{Title text color. Default "#1f77b4".}

\item{hoverinfo}{Determines hover label info. Default "y".}

\item{hoverformat}{Format for hover labels. Default ".2f".}

\item{yaxistitle}{Show y-axis title? Default FALSE.}

\item{xaxistitle}{Show x-axis title? Default FALSE.}

\item{xlen}{Length of x-axis. Default NULL for automatic.}

\item{yrange}{Limits for y-axis. Default NULL for automatic.}

\item{font_family}{Font family for text. Default "Lato".}

\item{margin}{Plot margins. Default c(0,0,0,0).}

\item{grouped}{Calculate mean and SD across groups? Default TRUE.}

\item{annotations}{Additional plot annotations. Default NULL.}
}
\value{
A plotly bar plot object
}
\description{
Bar plot using plotly
}
