% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.boxplot.PLOTLY}
\alias{pgx.boxplot.PLOTLY}
\title{Box plot using plotly}
\usage{
pgx.boxplot.PLOTLY(
  data,
  x = NULL,
  y = NULL,
  title = NULL,
  color = "#3181de",
  fillcolor = "#2fb5e3",
  linecolor = "#3181de",
  hoverinfo = "y",
  hoverformat = ".2f",
  yaxistitle = FALSE,
  xaxistitle = FALSE,
  font_family = "Lato",
  margin = list(l = 10, r = 10, b = 10, t = 10)
)
}
\arguments{
\item{data}{Data frame to plot.}

\item{x}{Column in data to use for x-axis grouping. Default NULL.}

\item{y}{Column in data to use for y-axis values. Default NULL.}

\item{title}{Plot title text. Default NULL.}

\item{color}{Box fill color. Default "#3181de".}

\item{fillcolor}{Box inside fill color. Default "#2fb5e3".}

\item{linecolor}{Box border color. Default "#3181de".}

\item{hoverinfo}{Determines hover label info. Default "y".}

\item{hoverformat}{Format for hover labels. Default ".2f".}

\item{yaxistitle}{Show y-axis title? Logical. Default FALSE.}

\item{xaxistitle}{Show x-axis title? Logical. Default FALSE.}

\item{font_family}{Font family for text. Default "Lato".}

\item{margin}{Plot margins. Default c(10,10,10,10).}
}
\value{
A plotly boxplot object.
}
\description{
Box plot using plotly
}
