% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{pgx.checkPGX}
\alias{pgx.checkPGX}
\title{Check input files for pgx.computePGX}
\usage{
pgx.checkPGX(df, type = c("SAMPLES", "COUNTS", "EXPRESSION", "CONTRASTS")[0])
}
\arguments{
\item{df}{a data.frame correcponding to the input file as described in Omics Playground documentation}

\item{type}{one of "counts", "samples", "contrasts"}
}
\description{
Check input files for pgx.computePGX
}
