% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-check.R
\name{pgx.checkPGX_all}
\alias{pgx.checkPGX_all}
\title{Check all input files for pgx.computePGX}
\usage{
pgx.checkPGX_all(SAMPLES = NULL, COUNTS = NULL, CONTRASTS = NULL)
}
\arguments{
\item{COUNTS}{data.frame. The data frame corresponding to the input file as in playbase::COUNTS}

\item{CONTRASTS}{data.frame. The data frame corresponding to the input file as in playbase::CONTRASTS}

\item{SAMPLE}{data.frame. The data frame corresponding to the input file as in playbase::SAMPLES}
}
\value{
a list with FIVE elements: SAMPLES, COUNTS and CONTRASTS that are the cleaned version of the
 input data frames, `checks` which contains the status of the checks, and
 `PASS` which contains the overall status of the check.
}
\description{
Check all input files for pgx.computePGX
}
