% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterBigMatrix}
\alias{pgx.clusterBigMatrix}
\title{Title}
\usage{
pgx.clusterBigMatrix(
  X,
  methods = c("pca", "tsne", "umap"),
  dims = c(2, 3),
  perplexity = 30,
  reduce.sd = 1000,
  reduce.pca = 50,
  center.features = TRUE,
  scale.features = FALSE,
  find.clusters = FALSE,
  svd.gamma = 1,
  umap.pkg = "uwot"
)
}
\arguments{
\item{umap.pkg}{}
}
\description{
Title
}
