% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterMatrix}
\alias{pgx.clusterMatrix}
\title{Title}
\usage{
pgx.clusterMatrix(
  X,
  perplexity = 30,
  dims = c(2, 3),
  ntop = 1000,
  npca = 50,
  prefix = "c",
  row.center = TRUE,
  row.scale = FALSE,
  find.clusters = TRUE,
  kclust = 1,
  clust.detect = c("louvain", "hclust"),
  method = c("tsne", "umap", "pca")
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
