% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterSamples2}
\alias{pgx.clusterSamples2}
\title{Title}
\usage{
pgx.clusterSamples2(
  pgx,
  methods = c("pca", "tsne", "umap"),
  dims = c(2, 3),
  reduce.sd = 1000,
  reduce.pca = 50,
  perplexity = 30,
  center.rows = TRUE,
  scale.rows = FALSE,
  X = NULL,
  umap.pkg = "uwot",
  replace.orig = TRUE
)
}
\arguments{
\item{pgx}{value}

\item{methods}{value}

\item{dims}{value}

\item{reduce.sd}{value}

\item{reduce.pca}{value}

\item{perplexity}{value}

\item{center.rows}{value}

\item{scale.rows}{value}

\item{X}{value}

\item{umap.pkg}{value}

\item{replace.orig}{value}
}
\description{
Title
}
