% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-predict.R
\name{pgx.compute_importance}
\alias{pgx.compute_importance}
\title{Main compute driver function. Compute variable importance for one
specific contrast.}
\usage{
pgx.compute_importance(
  pgx,
  pheno,
  level = "genes",
  filter_features = NULL,
  select_features = NULL,
  select_samples = NULL,
  nfeatures = 60,
  multiomics = NULL,
  do.survival = FALSE
)
}
\description{
Main compute driver function. Compute variable importance for one
specific contrast.
}
