% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{pgx.compute_mofa}
\alias{pgx.compute_mofa}
\title{Main function to call MOFA analysis on pgx object.}
\usage{
pgx.compute_mofa(
  pgx,
  kernel = "MOFA",
  numfactors = 10,
  ntop = 2000,
  gset.ntop = 2000,
  add_gsets = FALSE
)
}
\description{
Main function to call MOFA analysis on pgx object.
}
