% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{pgx.computeComboEnrichment}
\alias{pgx.computeComboEnrichment}
\title{Compute enrichment for drug combinations}
\usage{
pgx.computeComboEnrichment(
  obj,
  X,
  xdrugs,
  ntop = 10,
  nsample = 20,
  nprune = 250,
  contrasts = NULL,
  res.mono = NULL
)
}
\arguments{
\item{obj}{PGX object containing gene expression data}

\item{X}{Gene expression matrix}

\item{xdrugs}{Character vector of drug names}

\item{ntop}{Number of top drugs to combine per sample}

\item{nsample}{Number of samples to sample combinations}

\item{nprune}{Number of genes to use for GSEA}

\item{contrasts}{Contrasts to compute enrichment}

\item{res.mono}{Single drug enrichment results to use}
}
\value{
List with enrichment results for each drug combination
}
\description{
Computes enrichment of drug combinations using single drug enrichment results.
}
\details{
This function takes single drug enrichment results and computes enrichment for 
combinations by combining the individual drug profiles. It samples combinations from the 
top drugs per sample then runs GSEA to test for enrichment.

The output is a list containing the enrichment results for each sampled drug combination.
This allows searching for synergistic combinations with higher enrichment than individual drugs.
}
