% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{pgx.computeGeneSetExpression}
\alias{pgx.computeGeneSetExpression}
\title{Compute gene set expression}
\usage{
pgx.computeGeneSetExpression(
  X,
  gmt,
  method = NULL,
  min.size = 10,
  center = TRUE
)
}
\arguments{
\item{X}{Numeric gene expression matrix with genes in rows and samples in columns.}

\item{gmt}{List of gene sets, each containing gene names.}

\item{method}{Character vector of methods to use for computing gene set expression. Options are "gsva", "ssgsea", "spearman", "average". Default uses all.}

\item{min.size}{Minimum gene set size required. Sets with fewer genes are removed. Default 10.}

\item{center}{Logical indicating whether gene expression should be centered first. Default is TRUE.}
}
\value{
A list with gene set expression matrices, named by method.
}
\description{
This function calculates gene set expression scores from an input gene expression matrix using different methods like GSVA, ssGSEA, etc.
}
\details{
This function takes a gene expression matrix X and gene sets defined in gmt. It calculates gene set expression scores using the specified methods.

By default it computes GSVA, ssGSEA, spearman correlation and mean expression scores. X is centered by gene means if center=TRUE. Small gene sets are filtered out.

The output is a list containing the gene set expression matrices from each method.
}
