% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correlation.R
\name{pgx.computeGlassoAroundGene}
\alias{pgx.computeGlassoAroundGene}
\title{Compute partial correlations around a gene}
\usage{
pgx.computeGlassoAroundGene(X, gene, nmax = 100)
}
\arguments{
\item{X}{Expression matrix, genes in rows and samples in columns.}

\item{gene}{Gene name or index for target gene.}

\item{nmax}{Maximum number of genes to use.}
}
\value{
List containing:
\itemize{
  \item cor: Correlation matrix
  \item pcor: Partial correlation matrix
  \item wi: Inverse covariance matrix
  \item sampleSize: Number of samples used
}
}
\description{
Estimates partial correlations between a target gene and other
genes using glasso.
}
\details{
This function calculates the correlation between a target gene and
the top nmax correlated genes.

It uses the glasso algorithm to estimate a sparse inverse covariance
matrix and derive partial correlations.

The correlation matrix, partial correlation matrix, inverse covariance
matrix (precision matrix), and number of samples used are returned.
}
