% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correlation.R
\name{pgx.computePartialCorrelationAroundGene}
\alias{pgx.computePartialCorrelationAroundGene}
\title{Computes partial correlation between a given gene and all other genes in a data matrix X.}
\usage{
pgx.computePartialCorrelationAroundGene(
  X,
  gene,
  method = PCOR.METHODS,
  nmax = 100,
  fast = FALSE
)
}
\arguments{
\item{X}{Numeric gene expression matrix with genes as rows.}

\item{gene}{Gene name to compute partial correlation around.}

\item{method}{Partial correlation method(s) to use. Default is PCOR.METHODS.}

\item{nmax}{Maximum number of genes to return correlations for. Default 100.}

\item{fast}{Use fast approximation methods if TRUE. Default FALSE.}
}
\value{
A named list with components:
\itemize{
  \item rho - matrix of partial correlations
  \item p.value - matrix of p-values
}
}
\description{
Computes partial correlation between a given gene and all other genes in a data matrix X.
}
\details{
This function calculates the partial correlation between a specified gene and 
all other genes in the expression matrix X.
It returns a list containing the partial correlation matrix and p-values.

The \code{method} argument specifies which partial correlation algorithm to use. 
The default \code{PCOR.METHODS} uses both Pearson and Spearman correlation.

The number of top genes to return correlations for can be controlled with \code{nmax}. 
Setting \code{fast=TRUE} will use faster approximations for large matrices.
}
