% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-graph.R
\name{pgx.computePathscores}
\alias{pgx.computePathscores}
\title{Compute path scores}
\usage{
pgx.computePathscores(graph, strict.pos = TRUE)
}
\arguments{
\item{graph}{An igraph object representing the pathway graph}

\item{strict.pos}{Logical indicating whether to enforce strictly positive fold changes. Default is TRUE.}
}
\value{
A matrix of path scores for each node and sample
}
\description{
Computes path scores for a graph based on node fold changes.
}
\details{
Compute path scores for a graph

This function takes an igraph pathway graph object and calculates path scores for each node.

It first adds source and sink nodes if they do not exist. 
It then calculates edge weights based on the fold changes, enforcing strictly positive values if strict.pos=TRUE.

Path scores are computed as the path integral from source to sink going through each node. 
The path score indicates how well connected a node is to the observed fold changes.
}
