% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{pgx.computePCSF}
\alias{pgx.computePCSF}
\title{Compute PCSF solution from pgx object}
\usage{
pgx.computePCSF(
  pgx,
  contrast,
  datatypes = NULL,
  as_prize = "rho",
  use_rank = FALSE,
  ntop = 250,
  ncomp = 3,
  beta = 1,
  rm.negedge = TRUE,
  highcor = 0.8,
  dir = "both",
  level = "gene",
  ppi = c("STRING", "GRAPHITE")
)
}
\description{
Compute PCSF solution from pgx object for given contrast.
}
