% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{pgx.computePCSF_gset}
\alias{pgx.computePCSF_gset}
\title{Compute PCSF solution from pgx object for GENESET level}
\usage{
pgx.computePCSF_gset(
  pgx,
  contrast,
  gmt.rho = 0.8,
  gset.filter = NULL,
  highcor = 0.9,
  ntop = 250,
  ncomp = 3,
  beta = 1,
  rm.negedge = TRUE,
  dir = "both"
)
}
\description{
Compute PCSF solution from pgx object for GENESET level
}
