% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{pgx.computePCSF_multiomics}
\alias{pgx.computePCSF_multiomics}
\title{Compute PCSF solution from pgx object for MULTI-OMICS
  (datatype prefixed) data.}
\usage{
pgx.computePCSF_multiomics(
  pgx,
  contrast,
  datatypes = NULL,
  ntop = 250,
  ncomp = 3,
  beta = 1,
  rm.negedge = TRUE,
  highcor = 0.8,
  dir = "both",
  ppi = c("STRING", "GRAPHITE"),
  as.name = NULL
)
}
\description{
Compute PCSF solution from pgx object for MULTI-OMICS
  (datatype prefixed) data.
}
