% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct-diagnostics.R
\name{pgx.computeTechnicalEffects}
\alias{pgx.computeTechnicalEffects}
\title{Estimate technical effects variation}
\usage{
pgx.computeTechnicalEffects(X, is.count = FALSE, nmin = 3, nv = 1)
}
\arguments{
\item{X}{Gene expression matrix, with genes in rows and samples in columns}

\item{is.count}{Logical indicating if X contains counts (TRUE) or log-expression values (FALSE)}
}
\value{
List containing:
\itemize{
 \item{pct.mito}{Percent mitochondrial genes}
 \item{pct.ribo}{Percent ribosomal genes}
 \item{biological}{Biological coefficient of variation}
}
}
\description{
Estimates the biological variation and fraction of mitochondrial and ribosomal genes from a gene expression matrix.
}
\details{
This function calculates the biological variation (BCV) for each gene as the coefficient of variation of expression across samples.

It also calculates the percentage of mitochondrial and ribosomal genes based on gene symbols.

If the input matrix X contains counts, it will be transformed to log2-CPM.
If X contains log-expression values, it will be shifted to start at the 1% quantile.
}
\examples{
\dontrun{
data(sample.ExpressionSet)
results <- pgx.computeTechnicalEffects(sample.ExpressionSet)
head(results$biological)
}
}
