% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{pgx.correlateSignatureH5}
\alias{pgx.correlateSignatureH5}
\title{Correlate SignatureH5}
\usage{
pgx.correlateSignatureH5(fc, h5.file, nsig = 100, ntop = 200, nperm = 10000)
}
\arguments{
\item{fc:}{fold change matrix}

\item{h5.file:}{HDF5 file of reference expression signatures}

\item{nsig:}{number of significant genes}

\item{ntop:}{number of top signatures (in abs(rho)) to report}

\item{nperm:}{number of permuations for fGSEA}
}
\description{
pgx.correlateSignatureH5 computes correlation and gene set enrichment between a
signature and datasets in an HDF5 file using on-disk chunked computations
}
