% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{pgx.countNormalization.beta}
\alias{pgx.countNormalization.beta}
\title{Normalize expression data}
\usage{
pgx.countNormalization.beta(X, method = "CPM", ref = NULL, prior = 1)
}
\arguments{
\item{method}{Normalization method(s) to use.}

\item{ref}{Reference for referenceNormalization}

\item{prior}{Offset added in X prior to log2}

\item{x}{Matrix of log2-transformed data, with genes in rows and samples in columns.}
}
\value{
Normalized, log2-transformed matrix.
}
\description{
Normalizes a matrix of RNA-seq/proteomics data using different methods. Uses log2 data.
}
\details{
This function normalizes a log2-transformed data matrix using different normalization methods:
}
