% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-normalize.R
\name{pgx.countNormalization}
\alias{pgx.countNormalization}
\title{Normalize count data}
\usage{
pgx.countNormalization(x, methods)
}
\arguments{
\item{x}{Matrix of count data, with genes in rows and samples in columns.}

\item{methods}{Normalization method(s) to use. Options are "none", "scale", "CPM", "TMM", "RLE".}

\item{keep.zero}{Logical indicating whether to retain zero counts. Default is TRUE.}
}
\value{
The normalized count matrix.
}
\description{
Normalizes a matrix of RNA-seq count data using different methods.
}
\details{
This function normalizes a matrix of RNA-seq count data using different normalization methods:

- "none": No normalization
- "scale": Scale normalization by column means
- "CPM": Counts per million
- "TMM": Trimmed mean of M-values (edgeR)
- "RLE": Relative log expression (DESeq2)

Zero counts can be retained or set to NA after normalization based on the keep.zero parameter.
}
\examples{
\dontrun{
counts <- matrix(rnbinom(10000, mu = 10, size = 1), 100, 100)
normalized <- pgx.countNormalization(counts, c("TMM", "RLE"))
}
}
