% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{pgx.createComboDrugAnnot}
\alias{pgx.createComboDrugAnnot}
\title{Create annotation for drug combinations}
\usage{
pgx.createComboDrugAnnot(combo, annot0)
}
\arguments{
\item{combo}{Character vector of drug combinations (separated by '+')}

\item{annot0}{Data frame with annotation for individual drugs}
}
\value{
Data frame with annotation for the drug combinations
}
\description{
Creates annotation for drug combinations by combining annotation
  from individual drugs.
}
\details{
Create annotation for drug combinations

This function takes a vector of drug combinations and an annotation
  data frame for individual drugs. It splits the combinations into individual
  drugs, looks up their annotation, and combines it into annotation for the combos.

  For example, if combo is 'DrugA + DrugB', it will look up mode of action (moa)
  and targets for DrugA and DrugB, and combine them into moa and targets for
  the combination. NA values are removed.
}
