% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{pgx.createCreedsSigDB}
\alias{pgx.createCreedsSigDB}
\title{Create CREEDS signature database}
\usage{
pgx.createCreedsSigDB(gmt.files, h5.file, update.only = FALSE)
}
\arguments{
\item{gmt.files}{Character vector of paths to CREEDS GMT files containing gene signatures.}

\item{h5.file}{Path to HDF5 file to write signature database to.}

\item{update.only}{Logical indicating whether to only update existing signatures. Default is FALSE.}
}
\value{
NULL. The CREEDS signature database is written to the specified HDF5 file.
}
\description{
Reads gene signatures from CREEDS GMT files and writes to an HDF5 database.
}
\details{
This function takes a set of CREEDS GMT files containing up/down regulated gene signatures.
It extracts the signatures and writes them to an HDF5 file in a standard format that can be used for
gene set enrichment analysis.

Each GMT file contains multiple up and down regulated signatures for a perturbation experiment.
The signatures are named using the perturbation and direction, e.g. "DrugX-up".

If update.only=FALSE, any existing signatures in the HDF5 file will be overwritten.
If TRUE, only signatures not already present will be added.
}
