% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-graph.R
\name{pgx.createOmicsGraph}
\alias{pgx.createOmicsGraph}
\alias{pgx.reduceOmicsGraph}
\alias{pgx.createVipGeneLayer}
\title{Create an omics graph}
\usage{
pgx.createOmicsGraph(ngs, do.intersect = TRUE)

pgx.reduceOmicsGraph(ngs)

pgx.createVipGeneLayer(gr, genes, z = 0, reconnect = 40)
}
\arguments{
\item{ngs}{An NGS analysis object containing gene expression data}

\item{do.intersect}{Logical indicating whether to intersect genesets. Default is TRUE.}

\item{gr:}{igraph object}

\item{genes:}{character vector of VIP genes}

\item{z:}{z position for new VIP layer}

\item{reconnect:}{minimum edges to keep when removing shortcuts}
}
\value{
An igraph object representing the omics graph.
}
\description{
Constructs an omics graph by connecting genes to gene sets and merging nodes.
}
\details{
Create an omics graph from an NGS analysis

This function takes an NGS analysis object and constructs a graph where genes are connected to the gene sets they belong to.
It first creates a bipartite graph connecting genes and gene sets. The graph is then simplified by merging genes into clusters.

If do.intersect is TRUE, gene sets are intersected when more than one gene set contains the same genes. This reduces redundancy.

The graph layout positions nodes along the z-axis based on the type, with genes lowest and gene sets highest. Edges are added from
a source node to genes and from gene sets to a sink node.

The graph contains node data on expression, log2 fold changes between conditions, and metadata. It can be used for pathway analysis
and visualizations.
}
\section{Functions}{
\itemize{
\item \code{pgx.reduceOmicsGraph()}: reduces a full omics graph into a simplified graph by clustering nodes into groups.

\item \code{pgx.createVipGeneLayer()}: Creates a separate layer in a graph for VIP genes by assigning them a new z position and removing shortcut links.

}}
