% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{pgx.createSignatureDatabaseH5.fromMatrix}
\alias{pgx.createSignatureDatabaseH5.fromMatrix}
\title{Create signature database from matrix}
\usage{
pgx.createSignatureDatabaseH5.fromMatrix(h5.file, X, update.only = FALSE)
}
\arguments{
\item{h5.file}{Path to HDF5 file to write signature database to}

\item{X}{Numeric matrix of gene expression data}

\item{update.only}{Logical indicating whether to update existing signatures only. Default is FALSE.}
}
\value{
NULL. The signature database is written to the HDF5 file.
}
\description{
Reads a gene expression matrix and saves it to an HDF5 database file. Also extracts per-dataset top up/down regulated signatures.
}
\details{
Create signature database from matrix

This function takes a gene expression matrix \code{X} and writes it to an HDF5 file at \code{h5.file} in a standard format.

It also calculates the top 100 up and down regulated genes (signatures) for each dataset in \code{X}. These signatures are saved in the HDF5 file.

If \code{update.only=TRUE}, existing signatures in the HDF5 file will not be overwritten. Only new signatures from \code{X} will be added.
}
