% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{pgx.createSignatureDatabaseH5}
\alias{pgx.createSignatureDatabaseH5}
\title{Create signature database}
\usage{
pgx.createSignatureDatabaseH5(h5.file, pgx.files, update.only = FALSE)
}
\arguments{
\item{h5.file}{Path to the H5 file to store the signature database.}

\item{pgx.files}{Character vector of paths to PGX files to read fold change matrices from.}

\item{update.only}{Logical indicating if only new PGX files should be added. Default is FALSE.}
}
\value{
The combined signature database matrix is saved to the specified H5 file.
}
\description{
Reads fold change matrices from multiple PGX files and combines them into a single signature database stored in an H5 file.
}
\details{
Create a signature database from PGX files

This function reads fold change matrices from multiple PGX experiment result files (.pgx).
It extracts the fold change values, normalizes the gene names, and combines the matrices into a single signature database.
This combined matrix is then stored in an H5 file for later use.
}
