% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.detect_timevar}
\alias{pgx.detect_timevar}
\title{Detect time variables in expression matrix}
\usage{
pgx.detect_timevar(Y)
}
\arguments{
\item{Y}{Expression matrix with samples in columns}
}
\value{
A character vector of detected time variables
}
\description{
Detects column names in an expression matrix \code{Y} that represent time variables.
}
\details{
This function checks the column names of the expression matrix \code{Y} for common time variable names like
"time", "hour", "hr", or "day". It returns a character vector of any matched column names that represent time points.

This is useful for identifying potential time variables that could be used for time course analyses.
}
