% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-singlecell.R
\name{pgx.downsample}
\alias{pgx.downsample}
\title{Random downsampling by group targeting equal sizes per group.}
\usage{
pgx.downsample(counts, meta, group, target_n = 20)
}
\description{
Random downsampling by group targeting equal sizes per group.
}
