% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-feature.R
\name{pgx.featureSetScore.byContrast}
\alias{pgx.featureSetScore.byContrast}
\title{Compute featureset score using limma on contrasts}
\usage{
pgx.featureSetScore.byContrast(pgx)
}
\description{
Compute featureset score using limma on contrasts
}
