% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-feature.R
\name{pgx.featureSetScore.Ftest}
\alias{pgx.featureSetScore.Ftest}
\title{Compute featureset score using F-test on phenotypes.}
\usage{
pgx.featureSetScore.Ftest(pgx)
}
\description{
Compute featureset score using F-test on phenotypes.
}
