% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-feature.R
\name{pgx.featureSetScore}
\alias{pgx.featureSetScore}
\title{Create feature sets (like genesets but collections of features)
based on gene families and custom class columns in gene annation.}
\usage{
pgx.featureSetScore(pgx)
}
\description{
Create feature sets (like genesets but collections of features)
based on gene families and custom class columns in gene annation.
}
