% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.FindClusters}
\alias{pgx.FindClusters}
\title{Title}
\usage{
pgx.FindClusters(
  X,
  method = c("kmeans", "hclust", "louvain", "meta"),
  top.sd = 1000,
  npca = 50
)
}
\arguments{
\item{npca}{}
}
\description{
Title
}
