% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.findLouvainClusters.SNN}
\alias{pgx.findLouvainClusters.SNN}
\title{Title}
\usage{
pgx.findLouvainClusters.SNN(
  X,
  prefix = "c",
  level = 1,
  gamma = 1,
  small.zero = 0.01
)
}
\arguments{
\item{X}{value}

\item{prefix}{value}

\item{level}{value}

\item{gamma}{value}

\item{small.zero}{value}
}
\description{
Title
}
