% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{pgx.gene_table}
\alias{pgx.gene_table}
\title{Retrieve gene annotation table}
\usage{
pgx.gene_table(pgx, organism)
}
\arguments{
\item{pgx}{PGX object with a counts table.}

\item{organism}{Char. Organism name. For more info see \code{\link{playbase::SPECIES_TABLE}}.}
}
\value{
Updated PGX object with gene annotation table
}
\description{
Retrieves a gene annotation table for the given organism
from Ensembl using biomaRt. Adds the table to the PGX object.
}
\details{
Queries the Ensembl database to get a gene annotation table
containing external gene IDs mapped to Ensembl IDs. Handles retries in case
of temporary Ensembl API errors.
}
\examples{
\dontrun{
pgx <- list()
pgx$counts <- matrix(rnorm(4), nrow = 2)
rownames(pgx$counts) <- c("ENSG00000142192", "ENSG00000288602")
pgx <- pgx.gene_table(pgx, "Human")
}
}
