% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pheno.R
\name{pgx.getCategoricalPhenotypes}
\alias{pgx.getCategoricalPhenotypes}
\title{Get Categorical Phenotypes}
\usage{
pgx.getCategoricalPhenotypes(
  df,
  min.ncat = 2,
  max.ncat = 20,
  remove.dup = FALSE,
  remove.dot = TRUE
)
}
\arguments{
\item{df}{Data frame containing phenotype data.}

\item{min.ncat}{Minimum number of categories required. Default 2.}

\item{max.ncat}{Maximum number of categories allowed. Default 20.}

\item{remove.dup}{Logical indicating whether to remove duplicate factor
levels. Default FALSE.}
}
\value{
A character vector of categorical phenotype column names
}
\description{
Identifies categorical phenotype columns in a data frame.
}
\details{
This function examines a phenotype data frame \code{df} and identifies
columns that represent categorical variables.
It first excludes any columns that appear to contain sample IDs.
It then selects columns that are factors, or character/numeric with
between \code{min.ncat} and \code{max.ncat} unique values.

The function returns a character vector of column names representing putative categorical phenotypes.
Setting \code{remove.dup = TRUE} will remove any duplicate factor levels prior to counting number of categories.
}
