% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{pgx.getDrugMOATable}
\alias{pgx.getDrugMOATable}
\title{Returns MOA enrichment matrix from pgx object}
\usage{
pgx.getDrugMOATable(
  pgx,
  contrast,
  db = 1,
  drugs = NULL,
  type = c("drugClass", "targetGene")[1]
)
}
\description{
Returns MOA enrichment matrix from pgx object
}
