% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-feature.R
\name{pgx.getFeatureSets}
\alias{pgx.getFeatureSets}
\title{Create feature sets (like genesets but collections of
features/rownames) based on gene families and custom class columns
in gene annation. This allows to create sets where probes have no
symbol (like in metabolomics).}
\usage{
pgx.getFeatureSets(pgx, min.size = 1)
}
\description{
Create feature sets (like genesets but collections of
features/rownames) based on gene families and custom class columns
in gene annation. This allows to create sets where probes have no
symbol (like in metabolomics).
}
