% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{pgx.getGeneFamilies}
\alias{pgx.getGeneFamilies}
\alias{pgx.getGeneSetCollections}
\title{Get Gene Set Collections}
\usage{
pgx.getGeneFamilies(genes, min.size = 10, max.size = 500)

pgx.getGeneSetCollections(gsets, min.size = 10, max.size = 500)
}
\arguments{
\item{min.size}{The minimum gene set size to include. Default 10.}

\item{max.size}{The maximum gene set size to include. Default 500.}

\item{gsets}{A list of gene sets}
}
\value{
A list containing the extracted gene set collections.
}
\description{
Extracts gene set collections from a list of gene sets based on size filters.
}
\details{
This function takes a list of gene sets and extracts collections
of gene sets that fall within a specified size range, defined by min.size
and max.size parameters. It removes any gene sets that are smaller than
`min.size` or larger than `max.size`. This allows filtering gene sets into size-based
collections for different types of analyses.
}
\section{Functions}{
\itemize{
\item \code{pgx.getGeneFamilies()}: Get the gene families.

}}
