% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-api.R
\name{pgx.getMetaFoldChangeMatrix}
\alias{pgx.getMetaFoldChangeMatrix}
\title{Title}
\usage{
pgx.getMetaFoldChangeMatrix(pgx, what = "meta", level = "gene")
}
\arguments{
\item{pgx}{value}

\item{what}{value}

\item{level}{value}
}
\description{
Title
}
