% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{pgx.getMOAmatrix}
\alias{pgx.getMOAmatrix}
\title{Returns MOA enrichment matrix from pgx object}
\usage{
pgx.getMOAmatrix(pgx, db = 1, type = c("drugClass", "targetGene")[1])
}
\description{
Returns MOA enrichment matrix from pgx object
}
