% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{pgx.getOrganism}
\alias{pgx.getOrganism}
\title{Determine organism from count matrix}
\usage{
pgx.getOrganism(pgx, capitalise = FALSE)
}
\arguments{
\item{pgx}{A pgx object with the pgx$organism information and the pgx$counts slot for the
guessing approach.}

\item{capitalise}{logical: by default FALSE. Parameter to capitalise the first letter of the 
specie if mouse or human.}
}
\value{
Character string the organism.
}
\description{
Determines if count data is from human or mouse based on gene identifiers.
}
\details{
This function retreives the pgx$organism slot. If it is not found, then it examines 
the gene identifiers in the row names of a count matrix to determine if the data is from human 
or mouse (main organism supported in the old playbase version). It checks if the identifiers 
match common patterns found in mouse genes, like "rik", "loc", "orf". If more than 20% match
 these mouse patterns, it assigns the organism as "mouse". Otherwise it assigns "human".

The function calculates the fraction of row names that do NOT match the mouse gene patterns.
If this fraction is >0.8, it assigns "human". This relies on the assumption that human data
will tend to have more uppercase ENSEMBL identifiers.
}
