% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-getgeo.R
\name{pgx.getSymbolFromFeatureData}
\alias{pgx.getSymbolFromFeatureData}
\title{Extract gene symbols from GEO feature data}
\usage{
pgx.getSymbolFromFeatureData(fdata)
}
\arguments{
\item{fdata}{The featureData table from a GEOquery GEO dataset object.}
}
\value{
A character vector of gene symbols, or NULL if symbols could not be extracted.
}
\description{
Extracts official gene symbols from feature datatable of a GEO dataset downloaded with GEOquery.
It first looks for a column containing gene symbols by matching against the org.Hs.egSYMBOL database.
If no direct symbol column is found, it looks for an ENTREZ identifier and maps to symbols using org.Hs.egSYMBOL.
Then it looks at REFSEQ identifiers. Then it looks at ENSEMBLE IDs. If no approach works, it returns NULL.
}
