% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-api.R
\name{pgx.getTopGeneSets}
\alias{pgx.getTopGeneSets}
\title{Title}
\usage{
pgx.getTopGeneSets(pgx, n = 10, ng = 100, dir = 0, sym = FALSE, filt = NULL)
}
\arguments{
\item{pgx}{value}

\item{n}{value}

\item{ng}{value}

\item{dir}{value}

\item{sym}{value}

\item{filt}{value}
}
\description{
Title
}
