% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-deconv.R
\name{pgx.inferCellCyclePhase}
\alias{pgx.inferCellCyclePhase}
\title{Infer Cell Cycle Phase}
\usage{
pgx.inferCellCyclePhase(counts)
}
\arguments{
\item{counts}{Gene expression matrix with genes in rows and cells in columns.}
}
\value{
Seurat object containing cell cycle scores and phase assignments.
}
\description{
Infers cell cycle phase based on expression of known marker genes.
}
\details{
This function takes a gene expression matrix and calculates cell cycle 
scores and assigns cell cycle phases using the Seurat package. It uses curated lists 
of S phase and G2/M phase marker genes from Tirosh et al. 2015. The gene expression 
matrix is normalized then scored based on the phase marker genes. The output Seurat
object contains the scores and assigned phases for each cell.
}
