% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-deconv.R
\name{pgx.inferCellType}
\alias{pgx.inferCellType}
\title{Infer cell types from gene expression}
\usage{
pgx.inferCellType(
  counts,
  low.th = 0.01,
  add.unknown = FALSE,
  min.prob = 0.2,
  min.count = 3,
  scalex = FALSE,
  normalize.mat = TRUE,
  method = "NNLM",
  collapse = "max",
  markers = NULL
)
}
\arguments{
\item{counts}{Numeric gene expression matrix with genes in rows.}

\item{low.th}{Filtering threshold to remove lowly expressed genes.}

\item{add.unknown}{Add "unknown" cell type if no match.}

\item{min.prob}{Minimum probability threshold for calling a cell type.}

\item{min.count}{Minimum count threshold for including a gene.}

\item{scalex}{Scale the expression values before inference.}

\item{normalize.mat}{Normalize the expression matrix before inference.}

\item{method}{Inference method to use (NNLM, LDA, ...).}

\item{collapse}{Method for collapsing predictions from multiple markers.}

\item{markers}{Cell type marker genes, if provided.}
}
\value{
Matrix of predicted cell type probabilities for each sample.
}
\description{
Infer cell types based on known marker genes and their expression levels.
}
\details{
This function identifies cell types based on the expression of known marker genes.
It filters the expression matrix, matches to the marker genes, and runs inference using NNLM or LDA.
The output is a matrix of predicted probabilities for each sample and cell type.
Parameters allow customizing the marker genes, expression filters, and inference method.
}
