% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-deconv.R
\name{pgx.inferCellTypeLM22}
\alias{pgx.inferCellTypeLM22}
\title{Infer cell types using LM22 signature}
\usage{
pgx.inferCellTypeLM22(
  counts,
  low.th = 0.01,
  add.unknown = FALSE,
  normalize.mat = TRUE,
  method = "NNLM",
  min.count = 3,
  celltype0 = NULL,
  min.prob = 0.2
)
}
\arguments{
\item{counts}{A counts matrix with genes as rows and samples as columns.}

\item{min.count}{The minimum count threshold. Default is 3.}

\item{markers}{A data frame with positive and negative cell type markers.}
}
\value{
A list with logical vectors for each cell type indicating marker presence.
}
\description{
This function infers the proportion of cell types in a bulk 
RNA-seq sample using the LM22 signature.
}
\details{
This function checks if the canonical cell type markers from the
LM22 signature are expressed above a minimum threshold in the input counts 
matrix. It returns a list with logical vectors indicating which cell types
are detected based on their marker expression.
}
