% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cna.R
\name{pgx.inferCNV}
\alias{pgx.inferCNV}
\title{Title}
\usage{
pgx.inferCNV(ngs, refgroup = NULL, progress = NULL)
}
\arguments{
\item{ngs}{value}

\item{refgroup}{value}

\item{progress}{value}
}
\description{
Title
}
