% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgx.load}
\alias{pgx.load}
\title{Load an R Object from a File}
\usage{
pgx.load(file, verbose = 0)
}
\arguments{
\item{file}{A character string specifying the path to the file containing the R object to be loaded.}

\item{verbose}{An integer specifying the level of verbosity for printing messages during the loading process (default is 0).}
}
\value{
The function does not return a value, but makes the loaded R object available in the local environment.
}
\description{
This function loads an R object from a file and makes it available in the local environment.
}
\details{
The function uses the `load` function from base R to load the R object from the specified file.
The loaded object is then made available in the local environment using the `local` function.
}
