% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.makeStratifiedContrasts}
\alias{pgx.makeStratifiedContrasts}
\title{Generate contrasts within strata}
\usage{
pgx.makeStratifiedContrasts(Y, strata, ref)
}
\arguments{
\item{Y}{Factor vector indicating groups}

\item{strata}{Factor vector indicating strata}

\item{ref}{Reference group to contrast against}
}
\value{
Contrast matrix
}
\description{
Generates contrasts between groups within strata.
}
\details{
This function takes a group factor Y and strata factor, and generates contrasts between
the groups within each stratum.

First, dummy variables are created for Y and the reference group is subtracted to create
contrasts. Then these are crossed with the strata dummy variables to encode comparisons  
within each stratum.

The result is a contrast matrix with columns encoding the group comparisons within
each stratum.
}
