% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.makeStratifiedContrastsDF}
\alias{pgx.makeStratifiedContrastsDF}
\title{Title}
\usage{
pgx.makeStratifiedContrastsDF(data, vars, strata, ref)
}
\arguments{
\item{data}{value}

\item{vars}{value}

\item{strata}{value}

\item{ref}{value}
}
\description{
Title
}
