% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-deconv.R
\name{pgx.multipleDeconvolution}
\alias{pgx.multipleDeconvolution}
\title{Run multiple deconvolution methods}
\usage{
pgx.multipleDeconvolution(
  counts,
  refmat,
  methods = c("I-NNLS", "CIBERSORT", "DCQ", "DeconRNAseq", "EPIC", "NNLM", "cor",
    "SingleR")
)
}
\arguments{
\item{counts}{Numeric gene expression matrix with genes in rows and samples in columns.}

\item{refmat}{List of reference expression matrices or a single matrix.}

\item{methods}{Character vector of deconvolution methods to run.}
}
\value{
List containing results from each deconvolution method.
}
\description{
Runs multiple deconvolution methods using different reference profiles.
}
\details{
This function runs deconvolution across multiple methods like CIBERSORT, DeconRNASeq, EPIC, etc.
It takes a gene expression matrix and one or more reference matrices as input. The reference matrix/matrices
represent cell type-specific expression profiles.

Deconvolution is performed for each reference separately using the specified methods. The results
from each method are returned as nested lists, with the outer list split by reference profile name.
}
