% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cna.R
\name{pgx.plotCNAHeatmap}
\alias{pgx.plotCNAHeatmap}
\title{Plot CNA heatmap}
\usage{
pgx.plotCNAHeatmap(
  ngs,
  res,
  annot = NA,
  pca.filter = -1,
  lwd = 1,
  downsample = 10,
  order.by = "clust",
  clip = 0,
  lab.cex = 0.6
)
}
\arguments{
\item{ngs}{An NGS object containing copy number data}

\item{res}{CNA segmentation results from pgx.segmentCN()}

\item{annot}{Data frame with sample annotations}

\item{pca.filter}{Filter samples by PCA clustering (-1 to disable)}

\item{lwd}{Line width for chromosome lines}

\item{downsample}{Downsample CNA matrix for plotting (integer factor)}

\item{order.by}{Order samples by "clust" or "annot"}

\item{clip}{Clip copy number ratio values (0 to disable)}

\item{lab.cex}{Label cex size}
}
\value{
None. Plot is produced as a side-effect.
}
\description{
Generates a heatmap to visualize copy number alterations across samples.
}
\details{
Plot copy number alteration heatmap


This function takes CNA segmentation results from pgx.segmentCN() and 
plots a heatmap of the copy number ratios. Samples can be ordered by 
annotations or clustering. The full CNA matrix can be downsampled for 
easier visualization. Chromosome lines and sample annotations are added.
}
