% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{pgx.plotDrugConnectivity}
\alias{pgx.plotDrugConnectivity}
\title{Plot drug connectivity as enrichment plots}
\usage{
pgx.plotDrugConnectivity(
  pgx,
  contrast,
  db = "L1000/activity",
  drugs = NULL,
  nplots = 25
)
}
\description{
Plot drug connectivity as enrichment plots
}
