% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{pgx.plotDrugMOA}
\alias{pgx.plotDrugMOA}
\title{Plot MOA enrichment as barplot}
\usage{
pgx.plotDrugMOA(
  pgx,
  contrast,
  db = "L1000/activity",
  type = c("drugClass", "targetGene")[1],
  ntop = ntop
)
}
\description{
Plot MOA enrichment as barplot
}
