% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-graph.R
\name{pgx.plotDualProjection}
\alias{pgx.plotDualProjection}
\title{Create Omics Graph}
\usage{
pgx.plotDualProjection(
  gr,
  gene = NULL,
  geneset = NULL,
  cex = 1,
  fx = NULL,
  main = NULL,
  plot = TRUE
)
}
\arguments{
\item{gr}{An igraph object containing the gene-geneset network}

\item{gene}{Character name of gene to highlight}

\item{geneset}{Character name of gene set to highlight}

\item{cex}{Size of node labels}

\item{fx}{PGX results matrix to color nodes by}

\item{main}{Plot title}

\item{plot}{Logical indicating whether to plot the graph}
}
\value{
An igraph object containing the omics graph.
}
\description{
Generates an igraph object representing associations between genes, gene sets and clinical outcomes from PGX results.
}
\details{
Create an omics graph from PGX results


This function takes a gene-geneset igraph object and PGX results and generates an omics graph.
Nodes represent genes and gene sets. Edge widths indicate the association strength between them.

Nodes can be colored by PGX results to highlight genes/sets associated with outcomes.
Specific genes and sets can also be highlighted by specifying their names.

The graph can be plotted directly using various layouts, or the igraph object can be returned.
}
