% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.plotExpression}
\alias{pgx.plotExpression}
\title{Plot expression for a gene}
\usage{
pgx.plotExpression(
  pgx,
  probe,
  comp,
  logscale = TRUE,
  level = "gene",
  grouped = FALSE,
  srt = NULL,
  cex = 1,
  collapse.others = TRUE,
  showothers = TRUE,
  max.points = 200,
  group.names = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  names = TRUE,
  plotly.annotations = NULL,
  plotly.margin = NULL,
  plotlib = "base"
)
}
\arguments{
\item{pgx}{PGX object with expression data}

\item{probe}{Gene name or ID to plot}

\item{comp}{Sample grouping variable}

\item{logscale}{Log2 transform expression values}

\item{level}{Data level to extract ("gene", "exon", etc)}

\item{grouped}{Group samples}

\item{srt}{Label rotation angle}

\item{cex}{Point size}

\item{collapse.others}{Collapse non-highlighted samples}

\item{showothers}{Show non-highlighted samples}

\item{max.points}{Maximum number of points to plot}

\item{group.names}{Group names for legend}

\item{main}{Plot title}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{names}{Show sample names}

\item{plotly.annotations}{Plotly annotations}

\item{plotly.margin}{Plotly margins}

\item{plotlib}{Plotting library}
}
\value{
A gene expression plot
}
\description{
Generate a gene expression plot from a PGX object.
}
\details{
This function extracts the expression for a given \code{probe}
from the \code{pgx} object at the specified \code{level}.
It generates a dot plot with samples grouped on the x-axis.
Use \code{comp} to specify grouping.

Options to collapse non-highlighted samples, log-transform values,
show sample names, set axis labels, etc.
}
