% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{pgx.plotMitoRibo}
\alias{pgx.plotMitoRibo}
\title{Plot mitochondrial and ribosomal counts}
\usage{
pgx.plotMitoRibo(counts, percentage = TRUE)
}
\arguments{
\item{counts}{A numeric matrix of count data, where rows represent features and columns represent samples.}

\item{percentage}{Logical, whether to plot the counts as percentages of total counts.}
}
\value{
A bar plot of mitochondrial and ribosomal counts.
}
\description{
This function plots the mitochondrial and ribosomal counts in a given count matrix.
}
