% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.plotOmicsNetwork}
\alias{pgx.plotOmicsNetwork}
\title{Plot omics network from PGX object}
\usage{
pgx.plotOmicsNetwork(
  pgx,
  gene = NULL,
  reduced = NULL,
  levels = c("gene", "geneset"),
  contrast = NULL,
  layout = NULL,
  colorcluster = FALSE,
  hilight = NULL
)
}
\arguments{
\item{pgx}{PGX object containing network}

\item{gene}{Gene name to highlight}

\item{reduced}{Use reduced network}

\item{levels}{Data levels to include}

\item{contrast}{Contrast to color by}

\item{layout}{Network layout algorithm}

\item{colorcluster}{Color by module}

\item{hilight}{Additional genes to highlight}
}
\value{
A network plot object
}
\description{
Visualize the omics network from a PGX analysis
}
\details{
This function extracts the omics network stored in a PGX object.
It generates a network graph, coloring and sizing nodes by the 
specified \code{contrast}. 

The \code{reduced} network can be plotted for a simplified view.
\code{levels} determines which data levels are included.

\code{gene} and \code{hilight} highlight specific genes. 
\code{colorcluster} colors nodes by module.
\code{layout} specifies the network layout algorithm.
}
