% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correlation.R
\name{pgx.plotPartialCorrelationAroundGene}
\alias{pgx.plotPartialCorrelationAroundGene}
\title{Plot Partial Correlation Network Around a Gene}
\usage{
pgx.plotPartialCorrelationAroundGene(
  res,
  gene,
  rho.min = 0.8,
  pcor.min = 0,
  nsize = -1,
  main = "",
  what = c("cor", "pcor", "graph"),
  edge.width = 10,
  layout = "fr"
)
}
\arguments{
\item{res}{The correlation analysis result list returned by \code{\link{pgx.pcor}}.}

\item{gene}{The gene name to extract partial correlations around.}

\item{rho.min}{Minimum absolute partial correlation threshold.}

\item{pcor.min}{Minimum absolute meta-partial correlation threshold.}

\item{nsize}{Node size parameter.}

\item{main}{Plot title.}

\item{what}{Type of plot, "cor", "pcor" or "graph".}

\item{edge.width}{Edge width for graph.}

\item{layout}{Layout algorithm for graph plot.}
}
\value{
Either a correlogram or graph plot centered on the given gene is produced.
}
\description{
Visualize a partial correlation network for a given gene using the correlation results from pgx.pcor().
}
\details{
This function extracts the partial correlation matrix around the specified 
\code{gene} from the \code{res} object.
It then optionally thresholds the correlations, and visualizes them as a 
correlogram or network graph, with the central gene in the middle.

For the graph, the \code{rho.min} and \code{pcor.min} arguments filter edges to 
plot. Node size is proportional to the number of connections. The graph can be 
drawn using different layout algorithms like \code{"fr"} or \code{"circle"}.
}
