% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correlation.R
\name{pgx.plotPartialCorrelationGraph}
\alias{pgx.plotPartialCorrelationGraph}
\title{Compute partial correlation around a gene}
\usage{
pgx.plotPartialCorrelationGraph(
  res,
  gene,
  rho.min = 0.1,
  nsize = -1,
  main = "",
  vsize = 10,
  edge.width = 10,
  label.cex = 0.8,
  radius = -1,
  plot = TRUE,
  layout = "fr"
)
}
\arguments{
\item{gene}{Character string specifying the gene name to compute partial correlation around.}

\item{rho.min}{Numeric threshold for including genes in partial correlation graph.}

\item{nsize}{Integer specifying number of top correlated genes to include.}

\item{main}{Character string for plot main title.}

\item{edge.width}{Numeric line width for graph edges.}

\item{layout}{Layout algorithm for graph. Options are "fr", "kk", "lgl", etc.}

\item{pgx}{pgx object}

\item{pcor.min}{Numeric threshold for including genes in partial correlation graph.}

\item{what}{Character vector specifying which plots to generate. Options are 
"cor", "pcor", "graph".}
}
\value{
List with components:
\itemize{
 \item rho: List of partial correlation matrices 
 \item meta.pcor: Matrix of mean partial correlations
 \item timings: Matrix of timings
}
}
\description{
Computes partial correlation between a given gene and all other genes.
}
\details{
Compute partial correlation around a gene

This function computes the partial correlation between a given 
gene and all other genes in the dataset. 
It returns plots of the top correlated genes, a correlation graph, and 
optionally a matrix of all gene-gene correlations.
}
