% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.plotPhenotypeMatrix}
\alias{pgx.plotPhenotypeMatrix}
\title{Visualize phenotype matrix as a heatmap}
\usage{
pgx.plotPhenotypeMatrix(annot)
}
\arguments{
\item{annot}{Phenotype matrix with samples as rows and variables as columns.}
}
\value{
A heatmap visualization of the phenotype matrix.
}
\description{
Generates a heatmap to visualize a phenotype matrix.
}
\details{
This function takes a phenotype matrix \code{annot} with samples as rows and phenotype
variables as columns. It converts the columns to factors and visualizes the matrix
as a heatmap, with samples in rows and variables in columns.

Useful for exploring sample metadata.
}
