% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.plotPhenotypeMatrix0}
\alias{pgx.plotPhenotypeMatrix0}
\title{Visualize phenotype matrix as a heatmap}
\usage{
pgx.plotPhenotypeMatrix0(annot, annot.ht = 5, cluster.samples = TRUE)
}
\arguments{
\item{annot}{Phenotype matrix with samples as rows and variables as columns.}

\item{annot.ht}{Height of annotation labels in mm. Default is 5.}

\item{cluster.samples}{Logical, cluster samples before plotting. Default is TRUE.}
}
\value{
A heatmap visualization of the phenotype matrix.
}
\description{
Generates a heatmap to visualize a phenotype matrix.
}
\details{
This function takes a phenotype matrix \code{annot} with samples as rows and phenotype
variables as columns. It converts the columns to factors and visualizes the matrix
as a heatmap, with samples in rows and variables in columns.

Samples are clustered before plotting if \code{cluster.samples=TRUE}.
The annotation label height is set by \code{annot.ht}.

Usef
}
