% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.plotSampleClustering}
\alias{pgx.plotSampleClustering}
\title{Plot Sample Clustering}
\usage{
pgx.plotSampleClustering(
  x,
  dim = 2,
  method = c("tsne", "umap", "pca"),
  ntop = 1000,
  ...
)
}
\arguments{
\item{x}{A numeric matrix or data frame containing the data to be clustered.}

\item{dim}{The number of dimensions to use for the dimensionality reduction.}

\item{method}{A character vector specifying the dimensionality reduction method to use.
Possible values are "tsne", "umap", and "pca".}

\item{ntop}{The number of top features to use for the dimensionality reduction.}

\item{...}{Additional arguments passed to the `plot` function.}
}
\value{
The function returns invisibly.
}
\description{
This function creates a plot of the sample clustering of a 
data matrix using one of several available dimensionality reduction methods.
}
\details{
This function first calls the `pgx.clusterMatrix` function to perform 
dimensionality reduction on the input data using the specified method. The 
resulting low-dimensional representation of the data is then plotted using the 
`plot` function from the base R graphics system.
}
