% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wordcloud.R
\name{pgx.plotWordCloud}
\alias{pgx.plotWordCloud}
\title{Plot wordcloud graphic from pgx}
\usage{
pgx.plotWordCloud(pgx, contrast)
}
\arguments{
\item{pgx}{A NGStest object containing gene set results.}

\item{progress}{A progress object to track progress.}

\item{pg.unit}{Increment value for progress tracking.}
}
\value{
A wordcloud plot
}
\description{
Plot wordcloud graphic from pgx
}
