% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-purifyExpression.R
\name{pgx.purifyExpression}
\alias{pgx.purifyExpression}
\title{Purify tumor expression profiles}
\usage{
pgx.purifyExpression(
  tumorX,
  normalX,
  method = c("nnlm", "nnmf", "isopurer", "demixt", "undo")
)
}
\arguments{
\item{method}{Character vector of deconvolution methods to apply.}

\item{expr}{Numeric gene expression matrix, rows are genes, columns are samples.}

\item{normal}{Numeric gene expression matrix from normal samples.}
}
\value{
List containing purified expression profiles and estimated tumor purity.
}
\description{
Estimates and removes normal cell contamination from tumor expression profiles.
}
\details{
This function takes a tumor expression matrix and an expression matrix from normal samples.
It applies computational deconvolution methods like ISOpure, DeMixT, etc. to estimate normal contamination.

The normal profile is removed from each tumor sample to generate a purified tumor expression profile.
Results from all methods are returned as a list, along with numeric vectors of estimated tumor purity.
}
