% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgx.readOptions}
\alias{pgx.readOptions}
\title{Read PGX Options}
\usage{
pgx.readOptions(file = "./OPTIONS")
}
\arguments{
\item{file}{The path to the PGX options file. Default is "./OPTIONS".}
}
\value{
A named list containing the PGX options.
}
\description{
Reads PGX analysis options from a file.
}
\details{
This function reads a simple text file containing PGX analysis options,
one option per line in the format:

\code{option=value}

Options include parameters like:

\code{fdr=0.05} - FDR threshold
\code{logfc=1} - Log fold-change threshold

The options file allows saving a set of parameters for easily re-running
an analysis with the same settings.
}
