% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{pgx.removeBatchEffect}
\alias{pgx.removeBatchEffect}
\title{Normalize count data}
\usage{
pgx.removeBatchEffect(
  X,
  batch,
  model.vars = NULL,
  method = c("ComBat", "BMC", "limma", "MNN", "fastMNN")
)
}
\arguments{
\item{method}{Normalization method to use. Options are "TMM", "RLE", "upperquartile", or "none".}

\item{counts}{Matrix of count data, with genes in rows and samples in columns.}

\item{log}{Logical, if TRUE apply log2 transformation after normalization.}

\item{prior}{Count offset to add prior to normalization. Default 1.}
}
\value{
The normalized count matrix.
}
\description{
Normalizes a matrix of RNA-seq count data using different methods.
}
\details{
This function normalizes a matrix of RNA-seq count data using different normalization methods:

- "TMM": Trimmed Mean of M-values method from edgeR
- "RLE": Relative Log Expression method
- "upperquartile": Upper quartile normalization
- "none": No normalization

Counts are offset by prior value before normalization. A log2 transformation can also be applied after normalization.
}
\examples{
\dontrun{
counts <- matrix(rnbinom(10000, mu = 10, size = 1), 100, 100) 
normalized <- pgx.countNormalization(counts, method="TMM", log=TRUE)
}
}
