% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{pgx.removePC}
\alias{pgx.removePC}
\title{Remove principal components from gene expression data}
\usage{
pgx.removePC(X, nv)
}
\arguments{
\item{X}{Gene expression matrix, genes in rows, samples in columns.}

\item{nv}{Number of principal components to remove.}
}
\value{
The gene expression matrix \code{X} with the top \code{nv} principal components removed.
}
\description{
Removes the top principal components from a gene expression matrix.
}
\details{
Remove principal components from gene expression data


This function calculates the top \code{nv} principal components of the gene expression matrix \code{X} using \code{irlba::irlba}. 
It then removes these principal components from \code{X} using \code{limma::removeBatchEffect}.

The goal is to remove technical noise or batch effects captured by the top principal components.
}
