% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.SankeyFromMatrixList.PLOTLY}
\alias{pgx.SankeyFromMatrixList.PLOTLY}
\title{Create a sankey diagram from a list of matrices}
\usage{
pgx.SankeyFromMatrixList.PLOTLY(matlist, contrast = NULL)
}
\arguments{
\item{matlist}{A list of matrices to compare}

\item{contrast}{Optional contrast name}
}
\value{
A plotly sankey diagram object
}
\description{
Generates a sankey diagram comparing matrices in a list
}
\details{
This function takes a list of matrices as input matlist. 
It standardizes each matrix, calculates cross-tables between matrices, 
and generates a sankey diagram showing the relationships.

The contrast parameter can optionally be used to weight the edges.

Useful for visualizing relationships and flows between data matrices.
}
