% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.SankeyFromPhenotypes.PLOTLY}
\alias{pgx.SankeyFromPhenotypes.PLOTLY}
\alias{pgx.SankeyFromPhenotypes.GGPLOT}
\title{Create a sankey diagram from phenotypes}
\usage{
pgx.SankeyFromPhenotypes.PLOTLY(
  pgx,
  phenotypes,
  mat = NULL,
  fill = NULL,
  nmin = 1,
  title = ""
)

pgx.SankeyFromPhenotypes.GGPLOT(
  pgx,
  phenotypes,
  mat = NULL,
  fill = NULL,
  sort = FALSE,
  nmin = 1,
  title = ""
)
}
\arguments{
\item{pgx}{PGX object with results}

\item{phenotypes}{Data frame of sample phenotypes}

\item{mat}{The results matrix to use, default is pgx$MAT}

\item{fill}{Whether to color edges by matrix values}

\item{nmin}{Minimum sample size for phenotype levels}

\item{title}{Plot title}
}
\value{
A plotly sankey diagram object
}
\description{
Generates a sankey diagram from sample phenotypes
}
\details{
This function takes a PGX object and sample phenotype data frame.
It aggregates the results matrix mat by phenotypes, filters low count levels,
calculates a cross-table, and visualizes the relationships between phenotypes
as a sankey diagram.

Useful for exploring sample relationships and phenotype flows.
}
\section{Functions}{
\itemize{
\item \code{pgx.SankeyFromPhenotypes.GGPLOT()}: pgx.SankeyFromPhenotypes.GGPLOT (ggplot2 version)

}}
