% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgx.save}
\alias{pgx.save}
\title{Save PGX Object}
\usage{
pgx.save(pgx, file, update.date = TRUE, light = TRUE, system = FALSE)
}
\arguments{
\item{pgx}{PGX object to save}

\item{file}{File path to save PGX object to}

\item{update.date}{Logical indicating whether to update date field. Default TRUE.}

\item{light}{Logical indicating whether to save a light version without some large fields. Default TRUE.}

\item{system}{Logical indicating whether to keep system-level objects like omicsnet. Default FALSE.}
}
\value{
NULL. The PGX object is saved to the specified file path.
}
\description{
Saves a PGX object to an R binary file.
}
\details{
Save PGX object to file

This function saves a PGX object to a .rda or .RData file.
By default it saves a light version without some large objects like gene set matrices.
It also optionally updates the date field and removes system-level objects like the omicsnet.
The PGX object can later be loaded back into R using \code{pgx.load()}.
}
