% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.scatterPlot}
\alias{pgx.scatterPlot}
\title{Scatter plot for PGX object}
\usage{
pgx.scatterPlot(
  pgx,
  pheno = NULL,
  gene = NULL,
  geneset = NULL,
  contrast = NULL,
  level = "gene",
  method = "tsne",
  pos = NULL,
  plotlib = "base",
  ...
)
}
\arguments{
\item{pgx}{PGX object with results}

\item{pheno}{Phenotype data.frame or vector}

\item{gene}{Gene name to color by expression}

\item{contrast}{Contrast name to color by correlation}

\item{level}{Data level to use ("gene", "geneset")}

\item{pos}{Positions for points (if level="geneset")}

\item{plotlib}{Plotting library ("base", "plotly")}

\item{...}{Other arguments passed to plotting functions}
}
\value{
A scatter plot grob or plotly object
}
\description{
Generates a scatter plot from a PGX object.
}
\details{
Samples or genes can be plotted, colored by phenotype,
gene expression, or correlation with a contrast. For gene sets,
positions need to be provided. Interactive plotly plots or
static ggplot2 plots can be generated.

Useful for exploring sample relationships and associations.
}
