% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.scatterPlotXY.D3}
\alias{pgx.scatterPlotXY.D3}
\title{Scatter Plot XY using scatterD3}
\usage{
pgx.scatterPlotXY.D3(
  pos,
  var = NULL,
  type = NULL,
  col = NULL,
  na.color = "#AAAAAA55",
  cex = 1,
  cex.lab = 0.8,
  cex.title = 1.2,
  cex.clust = 1.5,
  cex.legend = 1,
  zoom = 1,
  legend = TRUE,
  bty = "n",
  hilight = NULL,
  zlim = NULL,
  zlog = FALSE,
  softmax = FALSE,
  xlab = NULL,
  ylab = NULL,
  hilight2 = hilight,
  hilight.lwd = 0,
  hilight2.lwd = 0,
  zsym = TRUE,
  xlim = NULL,
  ylim = NULL,
  opacity = 1,
  label.clusters = FALSE,
  labels = NULL,
  legend.ysp = 0.85,
  legend.pos = "bottomright",
  tooltip = NULL,
  theme = NULL,
  set.par = TRUE,
  title = NULL,
  barscale = 0.8
)
}
\arguments{
\item{pos}{A data frame or matrix containing the x and y coordinates of the points to be plotted.}

\item{var}{An optional numeric vector specifying the value of a third variable to be used for coloring the points.}

\item{type}{Not used.}

\item{col}{Not used.}

\item{cex}{A numeric value specifying the size of the points.}

\item{cex.lab}{Not used.}

\item{cex.title}{Not used.}

\item{cex.clust}{Not used.}

\item{cex.legend}{Not used.}

\item{zoom}{Not used.}

\item{legend}{Not used.}

\item{bty}{Not used.}

\item{hilight}{Not used.}

\item{zlim}{Not used.}

\item{zlog}{Not used.}

\item{softmax}{Not used.}

\item{xlab}{A character string specifying the label for the x-axis.}

\item{ylab}{A character string specifying the label for the y-axis.}

\item{hilight2}{Not used.}

\item{opacity}{Not used.}

\item{label.clusters}{Not used.}

\item{labels}{Not used.}

\item{legend.ysp}{Not used.}

\item{legend.pos}{Not used.}

\item{tooltip}{Not used.}

\item{theme}{Not used.}

\item{set.par}{Not used.}

\item{title}{Not used.}

\item{barscale}{Not used.}
}
\description{
This function creates a scatter plot of two variables
using the scatterD3 library for interactive scatter plots.
}
