% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.scatterPlotXY.PLOTLY}
\alias{pgx.scatterPlotXY.PLOTLY}
\title{Interactive scatter plot using plotly}
\usage{
pgx.scatterPlotXY.PLOTLY(
  pos,
  var = NULL,
  type = NULL,
  col = NULL,
  cex = NULL,
  cex.lab = 0.8,
  cex.title = 1.2,
  cex.clust = 1.5,
  cex.legend = 1,
  cex.axis = 1,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  axis = TRUE,
  zoom = 1,
  legend = TRUE,
  bty = "n",
  hilight = NULL,
  hilight2 = hilight,
  hilight.col = NULL,
  hilight.cex = NULL,
  hilight.lwd = 0.8,
  zlim = NULL,
  zlog = FALSE,
  zsym = FALSE,
  softmax = FALSE,
  opacity = 1,
  bgcolor = NULL,
  box = TRUE,
  label.clusters = FALSE,
  labels = NULL,
  label.type = NULL,
  tooltip = NULL,
  theme = NULL,
  set.par = TRUE,
  title = "",
  title.y = 1,
  gridcolor = NULL,
  source = NULL,
  key = NULL,
  displayModeBar = FALSE
)
}
\arguments{
\item{pos}{Data frame containing x and y coordinates}

\item{var}{Grouping variable}

\item{type}{Variable type, either "continuous" or "factor"}

\item{col}{Colors}

\item{cex}{Point expansion factor}

\item{cex.lab}{Label size factor}

\item{cex.title}{Title size factor}

\item{cex.clust}{Cluster label size factor}

\item{cex.legend}{Legend text size factor}

\item{cex.axis}{Axis text size factor}

\item{xlab, ylab}{Axis labels}

\item{xlim, ylim}{Axis limits}

\item{axis}{Show axis? Logical}

\item{zoom}{Zoom factor}

\item{legend}{Show legend? Logical}

\item{bty}{Legend box type}

\item{hilight}{Indices of points to highlight}

\item{hilight2}{Secondary highlight indices}

\item{hilight.col}{Color for highlighted points}

\item{hilight.cex}{Size for highlighted points}

\item{hilight.lwd}{Line width for highlighted points}

\item{zlim}{Color scale limits}

\item{zlog}{Log scale z axis? Logical}

\item{zsym}{Symmetrize z axis? Logical}

\item{softmax}{Apply softmax to z axis? Logical}

\item{opacity}{Opacity factor, between 0 and 1}

\item{bgcolor}{Background color}

\item{box}{Draw box around plot? Logical}

\item{label.clusters}{Logical, label clusters?}

\item{labels}{Point labels}

\item{label.type}{Label type, "text" or "box"}

\item{tooltip}{Tooltip text}

\item{theme}{Theme parameters}

\item{set.par}{Reset par? Logical}

\item{title}{Plot title}

\item{title.y}{Title position}

\item{gridcolor}{Grid line color}

\item{source}{Character label for plot source}

\item{key}{Color key specification}

\item{displayModeBar}{Show plotly modebar? Logical}
}
\value{
A plotly interactive scatterplot object
}
\description{
Interactive scatter plot using plotly
}
