% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.scatterPlotXY}
\alias{pgx.scatterPlotXY}
\title{Scatter Plot XY}
\usage{
pgx.scatterPlotXY(..., plotlib = "base")
}
\arguments{
\item{...}{Additional arguments passed to the selected plotting function.}

\item{plotlib}{A character string specifying the plotting library to use.
Possible values are "base", "plotly", "ggplot", and "scatterD3".}
}
\value{
A plot object created by the selected plotting library.
}
\description{
This function creates a scatter plot of two variables 
using one of several available plotting libraries.
}
\details{
Depending on the value of the `plotlib` parameter, this function 
calls one of several internal functions to create the scatter plot using the 
specified plotting library. The available plotting libraries are:
- "base": The base R graphics system.
- "plotly": The plotly library for interactive plots.
- "ggplot": The ggplot2 library for static plots based on the Grammar of Graphics.
- "scatterD3": The scatterD3 library for interactive scatter plots using D3.js.
}
