% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-deconv.R
\name{pgx.scoreCellCycle}
\alias{pgx.scoreCellCycle}
\title{Score cell cycle phase from gene expression}
\usage{
pgx.scoreCellCycle(counts)
}
\arguments{
\item{expr}{Gene expression matrix with genes in rows and cells in columns.}
}
\value{
DataFrame with cell cycle phase scores and assignments for each cell.
}
\description{
Scores cell cycle phase for single cells based on expression of known marker genes.
}
\details{
This function takes a gene expression matrix and calculates cell cycle phase
scores and assignments for each cell using curated gene lists. It calculates S and G2M
scores based on phase-specific marker genes from Tirosh et al. 2015. The difference between
S and G2M scores is used to assign cell cycle phases (G1, S or G2M).

The output data frame contains the scores and assigned phases for each cell. This provides
an overview of cell cycle state in the data based on expression of canonical markers.
}
