% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.splitHeatmap}
\alias{pgx.splitHeatmap}
\title{Split heatmap from PGX object}
\usage{
pgx.splitHeatmap(
  ngs,
  splitx = NULL,
  top.mode = "specific",
  annot.pheno = NULL,
  row_annot_width = 0.03,
  scale = "row.center",
  ntop = 50,
  colors = NULL,
  rowcex = rowcex,
  colcex = colcex
)
}
\arguments{
\item{ngs}{PGX object containing expression data}

\item{splitx}{Sample groups to split heatmap columns}

\item{top.mode}{Method for selecting top rows ("specific", "sd", "pca")}

\item{annot.pheno}{Dataframe of sample annotations}

\item{row_annot_width}{Width of row annotation labels}

\item{scale}{Scaling method for expression data ("row.center", "row.zscore", etc)}

\item{ntop}{Number of top rows to include}

\item{colors}{Vector of colors to use for heatmap}

\item{rowcex}{Row text size scaling factor}

\item{colcex}{Column text size scaling factor}
}
\value{
A split heatmap grob object
}
\description{
Generates an interactive split heatmap visualization from a PGX object.
}
\details{
This function takes a PGX object and generates a split heatmap using the ComplexHeatmap package.
The heatmap can be split into groups by a sample metadata variable in \code{splitx}.

The top \code{ntop} most variable rows are selected based on \code{top.mode}, which can be "specific", "sd", or "pca".
Additional sample annotations can be provided in \code{annot.pheno}.

Various graphical parameters like colors, scaling, text sizes can be customized.
The output is a ComplexHeatmap grob object containing the split heatmap plot.
}
