% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.splitHeatmapFromMatrix}
\alias{pgx.splitHeatmapFromMatrix}
\title{Split heatmap from matrix using iHeatmapR}
\usage{
pgx.splitHeatmapFromMatrix(
  X,
  annot = NULL,
  idx = NULL,
  splitx = NULL,
  xtips = NULL,
  ytips = NULL,
  row_clust = TRUE,
  col_clust = TRUE,
  row_annot_width = 0.03,
  scale = "row.center",
  colors = NULL,
  lmar = 60,
  na_text = NULL,
  rowcex = 1,
  colcex = 1,
  show_legend = TRUE,
  zlim = NULL,
  symm = NULL,
  return_x_matrix = FALSE,
  splitx_order = NULL,
  heatmap_colors = NULL
)
}
\arguments{
\item{X}{Numeric data matrix}

\item{annot}{Data frame with row and column annotations}

\item{idx}{Rows to highlight}

\item{splitx}{Columns to split heatmap}

\item{xtips}{Custom column tooltips}

\item{ytips}{Custom row tooltips}

\item{row_clust}{Cluster rows? Default is TRUE.}

\item{col_clust}{Cluster columns within each split group? Default is TRUE.
Set to FALSE to keep columns in the supplied matrix order (e.g. for a
user-defined sample ordering).}

\item{row_annot_width}{Width for row annotations. Default is 0.03.}

\item{scale}{Scaling for data. Default is "row.center".}

\item{colors}{Vector of colors to use for annotation. Default is RColorBrewer Set1.}

\item{lmar}{Label margin parameter. Default is 60.}

\item{rowcex}{Row text size scaling. Default is 1.}

\item{colcex}{Column text size scaling. Default is 1.}

\item{show_legend}{Show color legend? Default is TRUE.}

\item{return_x_matrix}{Return the X matrix of the plot along the plotly object (as a list)}
}
\value{
A plotly split heatmap object
}
\description{
Split heatmap from matrix using iHeatmapR
}
\details{
This function generates an interactive split heatmap from a data matrix.
Rows and/or columns can be split according to the input annotations.
Data is scaled and clustered using hierarchical clustering by default.
Various graphical parameters like colors, labels, text sizes can be adjusted.
}
