% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.stackedBarplot}
\alias{pgx.stackedBarplot}
\title{Stacked barplot}
\usage{
pgx.stackedBarplot(x, showlegend, ylab = NULL, xlab = NULL, horiz = FALSE)
}
\arguments{
\item{x}{Numeric data matrix. Rows are stacked, columns are groups.}

\item{showlegend}{Logical indicating if legend should be displayed.}

\item{ylab}{Character string for y-axis label. Default NULL.}

\item{xlab}{Character string for x-axis label. Default NULL.}

\item{horiz}{Logical indicating if bars should be horizontal. Default FALSE.}
}
\value{
A plotly stacked barplot object.
}
\description{
Stacked barplot
}
\details{
This function takes a numeric data matrix and generates an interactive 
stacked barplot using plotly. Rows of the input matrix are stacked, and columns 
represent groups. Set horiz=TRUE to generate horizontal stacked bars.
}
\examples{
\dontrun{
x <- matrix(rnorm(100), ncol=5)
pgx.stackedBarplot(x)
}
}
