% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct-superbatch.R
\name{pgx.superBatchCorrect}
\alias{pgx.superBatchCorrect}
\title{Supervised batch correction}
\usage{
pgx.superBatchCorrect(
  X,
  pheno,
  model.par,
  partype = NULL,
  batch.par = "*",
  lib.correct = TRUE,
  bio.correct = c("mito", "ribo", "cell_cycle", "gender"),
  sva.correct = TRUE,
  pca.correct = TRUE,
  hc.correct = TRUE,
  mnn.correct = NULL,
  nnm.correct = TRUE,
  bc.methods = NULL,
  max.rho = 0.3,
  max.iter = 10,
  hc.top = 50
)
}
\arguments{
\item{X}{Gene expression matrix, genes in rows, samples in columns.}

\item{pheno}{Dataframe containing sample metadata/covariates. Must match colnames of \code{X}.}

\item{model.par}{Vector of column names of biological covariates in \code{pheno}.}

\item{batch.par}{Vector of column names of batch covariates in \code{pheno}.}

\item{method}{Batch correction method to apply ("combat", "limma", "sva", etc).}
}
\value{
The batch corrected gene expression matrix.
}
\description{
Performs supervised batch correction on a gene
  expression matrix, using known technical factors and biological
  covariates.
}
\details{
This function performs supervised batch correction, using known batch groups and biological covariates.
It constructs a design matrix containing batch groups and covariates.
The \code{method} batch correction is then applied to the expression matrix \code{X}, using this design matrix.

Technical effects are estimated from the data as follows:
\enumerate{
\item Compute average expression within each batch
\item Perform PCA on the batch means
\item Extract PCs explaining at least 80\% variance
}

Biological effects are included using the specified \code{model.par} covariates.
}
