% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-singlecell.R
\name{pgx.supercell}
\alias{pgx.supercell}
\title{SuperCell down sampling. Uniform down samplsing using gamma = 20.}
\usage{
pgx.supercell(
  counts,
  meta,
  group = NULL,
  gamma = 20,
  nvargenes = 1000,
  log.transform = TRUE
)
}
\description{
SuperCell down sampling. Uniform down samplsing using gamma = 20.
}
