% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-singlecell.R
\name{pgx.supercell2}
\alias{pgx.supercell2}
\title{SuperCell downsampling by group targeting equal sizes per group.}
\usage{
pgx.supercell2(counts, meta, group, target_n = 20)
}
\description{
SuperCell downsampling by group targeting equal sizes per group.
}
