% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-predict.R
\name{pgx.survivalVariableImportance}
\alias{pgx.survivalVariableImportance}
\alias{pgx.variableImportance}
\title{Variable importance for survival models}
\usage{
pgx.survivalVariableImportance(
  X,
  time,
  status,
  methods = c("glmnet", "randomforest", "xgboost", "pls")
)

pgx.variableImportance(
  X,
  y,
  methods = c("glmnet", "randomforest", "xgboost", "splsda", "correlation", "ftest",
    CARET.METHODS)[1:6],
  scale = TRUE,
  reduce = 1000,
  resample = 0,
  add.noise = 0
)
}
\arguments{
\item{X}{Numeric matrix of predictor variables, rows are variables}

\item{time}{Vector of event/censoring times}

\item{status}{Vector indicating event (1) or censoring (0)}

\item{methods}{Methods for computing variable importance. Options are "glmnet", "randomforest", "xgboost", "pls".}

\item{y}{Multiclass factor response variable. Contains the class
labels for each sample}
}
\value{
Named list with variable importance scores by method.
}
\description{
Compute variable importance scores for predictors in a survival analysis model using different methods.
}
\details{
This function calculates variable importance scores using a variety of methods suitable for survival analysis.
The input data consists of a predictor matrix \code{X}, a vector of event/censoring times \code{time} and a status indicator vector \code{status}.

The following methods can be selected via the \code{methods} parameter:
\itemize{
\item glmnet: Absolute value of coefficients from elastic net Cox model
\item randomForest: Variable importance from random survival forest
\item xgboost: Importance scores from XGBoost survival model
\item pls: Absolute coefficients from partial least squares Cox model
}

Variable importance scores are returned for each method in a named list.
These scores can be used to select important predictors for survival modeling.
}
\section{Functions}{
\itemize{
\item \code{pgx.variableImportance()}: Calculates variable
importance scores for predictors of a multiclass response using
various methods.

}}
