% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correlation.R
\name{pgx.testPhenoCorrelation}
\alias{pgx.testPhenoCorrelation}
\title{Test correlation of phenotype with expression data}
\usage{
pgx.testPhenoCorrelation(df, plot = TRUE, cex = 1)
}
\arguments{
\item{df}{Data frame containing phenotype data}

\item{plot}{Logical, whether to generate a correlation heatmap plot}

\item{cex}{Text size factor for plot labels}
}
\value{
A matrix of correlation p-values
}
\description{
Test correlation between phenotype variables and gene expression, using
appropriate methods based on data types.
}
\details{
This function takes a data frame \code{df} containing phenotype data, including both
continuous and discrete variables. It calculates correlation p-values between the phenotype
variables and gene expression data in a PGX object, using either Pearson or ANOVA. Discrete
variables are compared between groups using Fisher's exact test. The results are returned as
a matrix of p-values, with rows corresponding to expression profiles. If \code{plot=TRUE},
a correlation plot of the significance (-log10 p-values) is also generated.
}
