% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correlation.R
\name{pgx.testTraitRelationship}
\alias{pgx.testTraitRelationship}
\title{Test relationships between gene expression and traits}
\usage{
pgx.testTraitRelationship(me, df, plot = TRUE, cex = 1)
}
\arguments{
\item{me}{A gene expression matrix}

\item{df}{A data frame containing trait data}

\item{plot}{Logical indicating whether to plot results. Default is TRUE.}

\item{cex}{Size of text in plots. Default is 1.}
}
\value{
A list with correlation and ANOVA results
}
\description{
Tests for relationships between gene expression (me) and trait data (df),
using correlation for continuous variables and ANOVA for discrete variables.
}
\details{
This function takes a gene expression matrix (me) and a data frame (df) containing 
trait data. It separates df into continuous (numeric) and discrete (factor) variables.

For continuous variables, it calculates the correlation between each trait and each gene.
For discrete variables, it performs an ANOVA between each trait and each gene.

The results are returned as matrices of p-values for the correlations and ANOVAs.
If plot = TRUE, scatterplots and boxplots are generated to visualize the results.
}
