% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgx.updateInfoPGX}
\alias{pgx.updateInfoPGX}
\title{Update PGX-table with new pgx object.}
\usage{
pgx.updateInfoPGX(pgxinfo, pgx, remove.old = TRUE)
}
\description{
Update PGX-table with new pgx object.
}
