% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{PGX_CREATE}
\alias{PGX_CREATE}
\title{GENES_TABLE dataset}
\format{
A PGX object
}
\source{
Proteome profiles of activated vs resting human naive T cells at different times (Geiger et al., Cell 2016).
}
\usage{
PGX_CREATE
}
\description{
A PGX object after running pgx.createPGX() on the mini example counts, 
samples and contrast data
}
\keyword{datasets}
