% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_core.R
\name{pgx_options}
\alias{pgx_options}
\title{Alter the pgx options list}
\usage{
pgx_options(
  is.logx = NULL,
  batch.correct = TRUE,
  auto.scale = TRUE,
  filter.genes = TRUE,
  prune.samples = FALSE,
  only.known = TRUE,
  only.hugo = TRUE,
  convert.hugo = TRUE,
  do.cluster = TRUE,
  cluster.contrasts = FALSE,
  do.clustergenes = TRUE,
  only.proteincoding = TRUE
)
}
\arguments{
\item{is.logx}{boolean.}

\item{batch.correct}{boolean.}

\item{auto.scale}{boolean.}

\item{filter.genes}{boolean.}

\item{prune.samples}{boolean.}

\item{only.known}{boolean.}

\item{only.hugo}{boolean.}

\item{convert.hugo}{boolean.}

\item{do.cluster}{boolean.}

\item{cluster.contrasts}{boolean.}

\item{do.clustergenes}{boolean.}

\item{only.proteincoding}{boolean.}
}
\value{
list of options
}
\description{
This function creates the options list which will be used (eventually) to
control how the pgx object is created and computed.
}
\examples{
opt <- pgx_options()
}
