% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgxinfo.addPgx}
\alias{pgxinfo.addPgx}
\title{Update PGX-info file with new pgx object (WIP)}
\usage{
pgxinfo.addPgx(pgx.dir, pgx, file = "datasets-info.csv", update.fc = TRUE)
}
\description{
Update PGX-info file with new pgx object (WIP)
}
