% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pgxinfo.R
\name{pgxinfo.delete}
\alias{pgxinfo.delete}
\title{Delete pgx entries in pgx info objects}
\usage{
pgxinfo.delete(pgx.dir, pgxname, purge.h5 = FALSE)
}
\arguments{
\item{pgx.dir}{The folder containing pgxinfo metadata}

\item{pgxname}{The name of the pgx object for which metadata to delete}

\item{purge.h5}{Logical indicating whether to remove entry in big H5 file}
}
\description{
Removes entries the pgxinfo metadata files from a pgxname
}
\details{
This function takes a pgxname as input and removes all entries in the pgx metadata files.
}
