% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgxinfo.deletePgx}
\alias{pgxinfo.deletePgx}
\title{Delete pgx entry in datasets-info table in files (WIP)}
\usage{
pgxinfo.deletePgx(pgx.dir, pgxname, delete.fc = FALSE)
}
\description{
Delete pgx entry in datasets-info table in files (WIP)
}
