% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgxinfo.updateDatasetFolder}
\alias{pgxinfo.updateDatasetFolder}
\title{Update}
\usage{
pgxinfo.updateDatasetFolder(
  pgx.dir,
  force = FALSE,
  delete.old = FALSE,
  new.pgx = NULL,
  verbose = TRUE
)
}
\description{
Update
}
