% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pgxinfo.R
\name{pgxinfo.write}
\alias{pgxinfo.write}
\title{Write dataset information file (aka PGX info)}
\usage{
pgxinfo.write(pgxinfo, pgx.dir, info.file = "datasets-info.csv")
}
\arguments{
\item{pgxinfo}{Character string specifying the pgxinfo dataframe}

\item{pgx.dir}{Character string specifying the target folder}

\item{info.file}{Character string specifying the filename for the dataset info file. Default is "datasets-info.csv".}
}
\description{
Write the dataset information CSV file with some safety processing
}
\details{
This function writes the dataset info CSV located in a PGX directory. It contains metadata like datatype, organism, sample sizes, etc.

The \code{pgxinfo} argument specifies the PGXinfo dataframe
The \code{pgx.dir} argument specifies the target folder
The \code{info.file} argument specifies the filename for the dataset info file. By default this is "datasets-info.csv".
}
