% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-package.R
\docType{package}
\name{playbase}
\alias{playbase}
\alias{playbase-package}
\title{Playbase package}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
