% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plot_ggbarplot}
\alias{plot_ggbarplot}
\title{Bar plot using ggplot2}
\usage{
plot_ggbarplot(
  mat,
  xlab = "x",
  ylab = "y",
  srt = 0,
  main = NULL,
  las = NULL,
  beside = FALSE,
  legend.pos = c(0.016, 1),
  legend.cex = 1,
  axis.cex = 1,
  label.cex = 1,
  horiz = FALSE,
  cpal = "Blues",
  bar_width = 0.7,
  base_size = 12,
  group.name = "group"
)
}
\arguments{
\item{mat}{Numeric data matrix with samples in rows and groups in columns.}

\item{xlab}{Character label for x-axis. Default is "x".}

\item{ylab}{Character label for y-axis. Default is "y".}

\item{srt}{Rotation angle for x-axis labels. Default is 0.}

\item{main}{Character title for the plot. Default is NULL.}

\item{las}{Style for axis labels. Default is NULL.}

\item{beside}{Logical to draw bars beside each other. Default is FALSE.}

\item{legend.pos}{Numeric vector of legend position. Default is c(0.016, 1).}

\item{legend.cex}{Legend text size. Default is 1.}

\item{bar_width}{Width of bars. Default is 0.7.}

\item{base_size}{Base font size for plot. Default is 12.}

\item{group.name}{Label for legend groups. Default is "group".}

\item{col}{Vector of colors to use for bars. Default is NULL.}
}
\value{
A ggplot bar plot object is returned.
}
\description{
Creates a bar plot visualization using ggplot2.
}
\details{
This function takes a numeric matrix and generates a bar plot
visualization using ggplot2. Bars represent the groups, colored by row names.
Many graphical parameters like labels, colors, text sizes, etc. can be adjusted.
}
