% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plot_gghist}
\alias{plot_gghist}
\title{Plot Histogram with ggplot2}
\usage{
plot_gghist(x)
}
\arguments{
\item{x}{A numeric vector of values to plot.}
}
\value{
A ggplot histogram plot object.
}
\description{
Generate a histogram using ggplot2.
}
\details{
This function takes a numeric vector \code{x} and generates a histogram 
plot using ggplot2. The data is passed to \code{ggplot(data=x)} to generate a histogram
geom automatically. Titles, axis labels and other customizations can be added via 
additional ggplot commands.
}
