% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plot_ggscatterFILL}
\alias{plot_ggscatterFILL}
\title{Scatter Plot with Filled Colors}
\usage{
plot_ggscatterFILL(
  x,
  y = NULL,
  col = NULL,
  shape = NULL,
  main = NULL,
  cex = 1,
  pch = 20,
  legend = TRUE,
  xlab = NULL,
  ylab = NULL,
  legend.ysp = 0.8,
  cex.legend = 1,
  barscale = 0.3,
  opacity = 1,
  gamma = 1
)
}
\arguments{
\item{x}{Numeric vector for x-axis values.}

\item{y}{Numeric vector for y-axis values. Default is NULL.}

\item{col}{Vector of values used to determine point colors. Default is NULL.}

\item{shape}{Vector of values used to determine point shapes. Default is NULL.}

\item{main}{Title for the plot. Default is NULL.}

\item{cex}{Point expansion factor. Default is 1.}

\item{pch}{Point shape. Default is 20.}

\item{legend}{Show legend for colors? Default is TRUE.}

\item{xlab}{Label for x-axis. Default is NULL.}

\item{ylab}{Label for y-axis. Default is NULL.}

\item{legend.ysp}{Vertical space for legend. Default is 0.8.}

\item{cex.legend}{Legend text size. Default is 1.}

\item{barscale}{Color bar size scaling factor. Default is 0.3.}

\item{opacity}{Opacity of points. Default is 1.}

\item{gamma}{Gamma correction factor for colors. Default is 1.}
}
\value{
A ggplot2 scatter plot object is returned.
}
\description{
Creates a scatter plot with points colored and sized based on input vectors.
}
\details{
This function generates a scatter plot from x and y vectors. Points can be colored
and sized based on input col and shape vectors. A color bar legend is shown when col is provided.
Various graphical parameters like main title, axis labels, legend, and text sizes can be adjusted.
}
