% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plot_ggsplom}
\alias{plot_ggsplom}
\title{Scatterplot matrix using ggplot2}
\usage{
plot_ggsplom(F, F2 = NULL, title_cex = 2, no.axes = FALSE, ...)
}
\arguments{
\item{F}{A data frame or matrix.}

\item{F2}{Optional second data frame or matrix to plot on y-axis. Default is NULL.}

\item{title_cex}{Title text size. Default is 2.}

\item{no.axes}{Hide axis labels and tick marks. Default is FALSE.}

\item{...}{Other arguments passed to ggplot2 functions.}
}
\value{
A ggplot scatterplot matrix object
}
\description{
Generate a scatterplot matrix from a data frame or matrix using ggplot2.
}
\details{
Scatterplot matrix using ggplot2

This function takes a data frame or matrix \code{F} and generates a scatterplot matrix.
Each variable in \code{F} is plotted against each other variable. An optional second data frame 
or matrix \code{F2} can be provided for the y-axis variables. Titles and axis labels are added
automatically. \code{title_cex} controls the title size. \code{no.axes} hides axes.
}
