% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plot_ggviolin}
\alias{plot_ggviolin}
\title{Violin plot using ggplot2}
\usage{
plot_ggviolin(
  x,
  y,
  group = NULL,
  main = "",
  ylim = NULL,
  add.dots = TRUE,
  col = "#AAAAAA",
  cex = 1,
  xlab = "",
  ylab = "y",
  srt = 0,
  pdodge = 1.5,
  n.dodge = 1,
  base_size = 13
)
}
\arguments{
\item{x}{Factor variable for x-axis groups.}

\item{y}{Numeric vector of values to plot.}

\item{group}{Optional grouping variable.}

\item{main}{Plot title.}

\item{ylim}{Limits for y-axis. Calculated automatically if NULL.}

\item{add.dots}{Logical to add jittered data points. Default TRUE.}

\item{col}{Violin fill color. Default "#AAAAAA".}

\item{cex}{Point expansion factor. Default 1.}

\item{xlab}{Label for x-axis. Default "".}

\item{ylab}{Label for y-axis. Default "y".}

\item{srt}{Rotation angle for x-axis labels. Default 0.}

\item{pdodge}{Amount to dodge violins when group is used. Default 1.5.}

\item{n.dodge}{Number of levels to dodge. Default 1.}

\item{base_size}{Base font size. Default 13.}
}
\value{
A ggplot2 violin plot object.
}
\description{
Violin plot using ggplot2
}
