% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plot_grid.sharedAxisLabels}
\alias{plot_grid.sharedAxisLabels}
\title{Share axis labels in plot grid}
\usage{
plot_grid.sharedAxisLabels(plotList, nrow)
}
\arguments{
\item{plotList}{A list of ggplot objects to arrange in a grid}

\item{nrow}{Number of rows in the plot grid}
}
\value{
A combined ggplot object with shared axis labels
}
\description{
Arranges multiple plots in a grid sharing common axis labels.
}
\details{
This function takes a list of ggplot objects and arranges them in a grid 
layout with a specified number of rows. It shares common x and y axis labels among
the plots to avoid repetition.

Axis labels are removed from all plots except the bottom row and leftmost column.
The plots are then combined using cowplot::plot_grid() with shared axes.

Useful for creating multi-panel figure grids with clean axis labels.
}
