% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{plot_lasagna}
\alias{plot_lasagna}
\title{Plot lasagna model specified with 3D positions in posx using Grimon.}
\usage{
plot_lasagna(posx, vars = NULL, num_edges = 20)
}
\description{
Plot lasagna model specified with 3D positions in posx using Grimon.
}
