% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plotlyCytoplot}
\alias{plotlyCytoplot}
\title{Interactive cytoPlot using plotly}
\usage{
plotlyCytoplot(
  pgx,
  gene1,
  gene2,
  samples,
  nbinsx,
  nbinsy,
  lab.unit = "(log2CPM)",
  reversescale = TRUE,
  marker.size = 5,
  contour.coloring = "none",
  marker.color = "black",
  showgrid = TRUE
)
}
\arguments{
\item{pgx}{Expression data matrix}

\item{gene1}{First gene for x-axis}

\item{gene2}{Second gene for y-axis}

\item{samples}{Samples to include}

\item{nbinsx}{Number of bins for x-axis}

\item{nbinsy}{Number of bins for y-axis}

\item{lab.unit}{Units for axis labels. Default "(log2CPM)".}

\item{reversescale}{Reverse color scale? Default TRUE.}

\item{marker.size}{Size of scatter markers. Default 5.}

\item{contour.coloring}{Add density shading? "none", "color", or 
"black". Default "none".}

\item{marker.color}{Color of markers. Default "black".}

\item{showgrid}{Show grid lines? Logical. Default TRUE.}
}
\value{
A plotly interactive cytoPlot object
}
\description{
Interactive cytoPlot using plotly
}
\details{
This function generates an interactive cytoPlot using the plotly package.
Expression values for two genes are binned and plotted with one gene on each axis.
Samples to include and graphical parameters like color scale, marker size, etc. can be adjusted.
}
