% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plotlyMA}
\alias{plotlyMA}
\title{Interactive MA plot using plotly}
\usage{
plotlyMA(
  x,
  y,
  names,
  label.names = names,
  group.names = c("group1", "group2"),
  xlab = "average expression (log2.CPM)",
  ylab = "effect size (log2.FC)",
  lfc = 1,
  psig = 0.05,
  showlegend = TRUE,
  highlight = NULL,
  marker.size = 5,
  label = NULL,
  label.cex = 1,
  shape = "circle",
  color_up_down = TRUE,
  colors = c(up = "#f23451", notsig = "#8F8F8F", down = "#3181de"),
  marker.type = "scatter",
  source = "plot1",
  displayModeBar = TRUE
)
}
\arguments{
\item{x}{Numeric vector of average expression values (log2 CPM)}

\item{y}{Numeric vector of effect sizes (log2 fold change)}

\item{names}{Vector of gene or feature names}

\item{group.names}{Character vector of group names for legend.}

\item{xlab}{Label for x-axis. Default "average expression (log2.CPM)".}

\item{ylab}{Label for y-axis. Default "effect size (log2.FC)".}

\item{lfc}{Cutoff for absolute log fold change. Default 1.}

\item{psig}{Cutoff for adjusted p-value. Default 0.05.}

\item{showlegend}{Show legend? Logical. Default TRUE.}

\item{highlight}{Vector of genes to highlight. Default NULL.}

\item{marker.size}{Marker size. Default 5.}

\item{label}{Vector of labels for highlighted genes. Default NULL.}

\item{label.cex}{Text size for labels. Default 1.}

\item{shape}{Shape of points. Default is filled circle.}

\item{color_up_down}{Color up/down regulated features}

\item{marker.type}{Marker type (scatter, line, etc). Default "scatter".}

\item{source}{Character label for plot source. Default "plot1".}

\item{displayModeBar}{Show plotly modebar? Logical. Default TRUE.}
}
\value{
A plotly interactive MA plot object.
}
\description{
Interactive MA plot using plotly
}
