% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plotlyVolcano}
\alias{plotlyVolcano}
\title{Interactive volcano plot using plotly}
\usage{
plotlyVolcano(
  x,
  y,
  names,
  label.names = names,
  group.names = c("group1", "group2"),
  xlab = "effect size (logFC)",
  ylab = "significance (-log10p)",
  lfc = 1,
  psig = 0.05,
  showlegend = TRUE,
  highlight = NULL,
  marker.size = 5,
  label = NULL,
  label.cex = 1,
  shape = "circle",
  max.absy = NULL,
  xlim = NULL,
  ylim = NULL,
  color_up_down = TRUE,
  colors = c(up = "#f23451", notsig = "#8F8F8F", down = "#3181de", notsel = "#cccccc88"),
  title = "Volcano plot",
  marker.type = "scatter",
  displayModeBar = TRUE,
  source = "plot1"
)
}
\arguments{
\item{x}{Numeric vector of effect sizes (log fold changes)}

\item{y}{Numeric vector of p-values (-log10 transformed)}

\item{names}{Vector of gene or feature names}

\item{group.names}{Character vector of group names for legend.}

\item{xlab}{Label for x-axis. Default "effect size (logFC)".}

\item{ylab}{Label for y-axis. Default "significance (-log10p)".}

\item{lfc}{Cutoff for absolute log fold change. Default 1.}

\item{psig}{Cutoff for adjusted p-value. Default 0.05.}

\item{showlegend}{Show legend? Logical. Default TRUE.}

\item{highlight}{Vector of genes to highlight. Default NULL.}

\item{marker.size}{Marker size. Default 5.}

\item{label}{Vector of labels for highlighted genes. Default NULL.}

\item{label.cex}{Text size for labels. Default 1.}

\item{shape}{Shape of points. Default is filled circle.}

\item{xlim}{Numeric length-2 vector setting the x-axis range. Default NULL (auto).}

\item{ylim}{Numeric length-2 vector setting the y-axis range. Default NULL (auto).}

\item{color_up_down}{Color up/down regulated features}

\item{marker.type}{Marker type (scatter, line, etc). Default "scatter".}

\item{displayModeBar}{Show plotly modebar? Logical. Default TRUE.}

\item{source}{Character label for plot source. Default "plot1".}
}
\value{
A plotly interactive volcano plot object.
}
\description{
Interactive volcano plot using plotly
}
