% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plotMultiPartiteGraph2}
\alias{plotMultiPartiteGraph2}
\title{Plot multi-partite graph as parallel plot with base plot.}
\usage{
plotMultiPartiteGraph2(
  graph,
  layers = NULL,
  min.rho = 0.5,
  ntop = 25,
  labels = NULL,
  cex.label = 1,
  vx.cex = 1,
  xpos = NULL,
  xlim = NULL,
  justgraph = FALSE,
  edge.cex = 1,
  edge.alpha = 0.33,
  xdist = 1,
  normalize.edges = FALSE,
  yheight = 2,
  edge.sign = "both",
  edge.type = "both",
  labpos = NULL,
  value.name = NULL,
  strip.prefix = FALSE,
  strip.prefix2 = FALSE,
  prune = FALSE,
  layout = c("parallel", "hive")[1]
)
}
\description{
Plot multi-partite graph as parallel plot with base plot.
}
