% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{plotPCSF.IGRAPH}
\alias{plotPCSF.IGRAPH}
\title{Plot static igraph network from PCSF results object}
\usage{
plotPCSF.IGRAPH(
  net,
  fx = NULL,
  vertex.cex = 1,
  label.fx = fx,
  label.cex = 4,
  edge.cex = 1,
  layoutMatrix = NULL,
  ...
)
}
\arguments{
\item{x}{}
}
\description{
Plot static igraph PCSF graph from result of
  pgx.computePCSF. To be called from plotPCSF.
}
