% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{plotPCSF}
\alias{plotPCSF}
\title{Plot PCSF graph from results object}
\usage{
plotPCSF(
  pcsf,
  sizeby = "prize",
  colorby = "foldchange",
  highlightby = c("centrality", "prize")[1],
  labels = NULL,
  plotlib = c("visnet", "igraph")[1],
  layout = c("layout_with_kk", "hierarchical")[1],
  layoutMatrix = NULL,
  asp = 1,
  physics = TRUE,
  node_cex = 1,
  node_alpha = 0.8,
  node_gamma = 1,
  label_cex = 1,
  label_gamma = 2,
  nlabel = -1,
  border_width = 1.5,
  edge_cex = 1,
  edge_length = NULL,
  cut.clusters = FALSE,
  cut.layout = "kk",
  cut.resolution = 0.05,
  as_grid = TRUE,
  nlargest = -1,
  hilight_single = NULL
)
}
\description{
Plot PCSF graph (visnetwork or igraph) from result of
  pgx.computePCSF
}
