% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plotTimeSeries.groups}
\alias{plotTimeSeries.groups}
\title{Plots time series facetted by module/colors with groups. X is an
expression matrix. Time is a vector of times for each sample.}
\usage{
plotTimeSeries.groups(
  time,
  y,
  group = NULL,
  main = "",
  lwd = 3,
  xlab = "time",
  time.factor = TRUE
)
}
\description{
Plots time series facetted by module/colors with groups. X is an
expression matrix. Time is a vector of times for each sample.
}
