% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plotTimeSeries.modules}
\alias{plotTimeSeries.modules}
\title{Plots time series facetted by module/colors. X is an expression
matrix. Time is a vector of times for each sample.}
\usage{
plotTimeSeries.modules(time, xx, modules, main = "")
}
\description{
Plots time series facetted by module/colors. X is an expression
matrix. Time is a vector of times for each sample.
}
