% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pubmed.R
\name{pmid.buildGraph}
\alias{pmid.buildGraph}
\title{Build gene association graph from PubMed identifiers}
\usage{
pmid.buildGraph(P)
}
\arguments{
\item{P}{Sparse matrix mapping PMIDs to gene symbols}
}
\value{
An igraph undirected graph object representing the gene association network.
}
\description{
Builds an undirected gene association graph from PubMed 
identifiers (PMIDs) mapped to gene symbols.
}
\details{
This function takes as input a sparse matrix \code{P} that maps 
PubMed identifiers (PMIDs) to gene symbols. 
It filters \code{P} to only include PMIDs mapped to 2-10 genes.

It then calculates a co-occurrence matrix \code{M} between genes by multiplying 
\code{P} and its transpose. \code{M[i,j]} indicates the number of PMIDs shared 
between gene i and gene j.

An undirected graph is constructed from \code{M} using the igraph package, 
with genes as nodes and edge weights proportional to their PMID co-occurrences.
}
