% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pubmed.R
\name{pmid.buildMatrix}
\alias{pmid.buildMatrix}
\title{Build PMID annotation matrix}
\usage{
pmid.buildMatrix()
}
\value{
A sparse matrix with rows as PMIDs, columns as gene symbols, and 
entries indicating a mapping between a PMID and symbol.
}
\description{
Builds an annotation matrix mapping PubMed IDs to gene symbols 
based on mappings in org.Hs.eg.db.
}
\details{
This function retrieves PubMed ID to Entrez Gene ID mappings from 
org.Hs.eg.db and converts them to a sparse matrix mapping PMIDs to 
gene symbols. It filters to PMIDs associated with <=10 genes.

It collapses duplicate PMID mappings and builds a sparse matrix 
with rows as PMIDs, columns as gene symbols, and entries indicating 
an association between a PMID and symbol.
}
