% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pubmed.R
\name{pmid.getGeneContext}
\alias{pmid.getGeneContext}
\title{Get gene context from PubMed}
\usage{
pmid.getGeneContext(gene, keyword)
}
\arguments{
\item{gene}{Gene name or symbol.}

\item{keyword}{Optional keyword or phrases to search for context.}
}
\value{
A list containing:
\itemize{
\item rifs: Character vector of PMIDs for articles matching the gene and keyword filter.
\item table: Contingency table used for calculating p-value.  
\item p.value: Fisher's exact test p-value.
\item context: Word context scores.
}
}
\description{
Retrieves PubMed article information related to a gene and keyword context.
}
\details{
This function searches PubMed for articles related to the input \code{gene}, 
optionally filtered by \code{keyword} context phrases.

It first searches PubMed for the gene name and synonyms. It then filters these 
articles based on matches to the \code{keyword} terms.  The PubMed ids (PMIDs) 
of matching articles are returned.

Statistics are calculated to assess enrichment of the \code{keyword} within articles 
matching the \code{gene}. These include a Fisher's exact test p-value, and word 
co-occurrence statistics.
}
