% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pubmed.R
\name{pmid.getPubMedContext}
\alias{pmid.getPubMedContext}
\title{Get PubMed context for a gene}
\usage{
pmid.getPubMedContext(gene, context)
}
\arguments{
\item{gene}{Gene name or symbol to search for.}

\item{context}{Context words or phrases to search for along with gene.}
}
\value{
A list with the PubMed IDs, article titles, and abstract excerpts containing
the gene name and context words.
}
\description{
Retrieves PubMed abstract text containing a gene name and context words.
}
\details{
This function searches PubMed for articles containing the specified gene name 
and context words. It uses the EUtils API to search titles and abstracts in 
PubMed for the gene name and context terms.

It returns a list containing the PubMed IDs, article titles, and abstract 
excerpts (RIF strings) containing both the gene and context.
}
\examples{
\dontrun{
res <- pmid.getPubMedContext("TP53", "DNA damage")
}
}
