% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{pos.compact}
\alias{pos.compact}
\title{Compact Positions}
\usage{
pos.compact(pos, d = 0.01)
}
\arguments{
\item{pos}{A numeric matrix of positions to be compacted.}

\item{d}{An optional numeric value specifying the maximum distance between adjacent positions after compaction.
The default value is 0.01.}
}
\value{
A numeric matrix of the same size as `pos`, containing the compacted positions.
}
\description{
This function compacts a matrix of positions by removing white space.
}
\details{
This function takes a numeric matrix of positions `pos` and an optional numeric value `d` as input and compacts the positions by removing white space.
For each column in `pos`, the function calculates the distances between adjacent positions and replaces any distance greater than `d` with `d`.
The resulting compacted positions are returned as a numeric matrix of the same size as `pos`.
}
