% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-proteomics.R
\name{prot.nmfImpute}
\alias{prot.nmfImpute}
\title{Impute missing values with NMF}
\usage{
prot.nmfImpute(X, groups, k = 10, r = 0.5)
}
\arguments{
\item{X}{Numeric matrix of proteomics data}

\item{groups}{Factor indicating groups for samples}

\item{k}{Number of NMF factors. Default 10.}

\item{r}{Cutoff proportion of missing values to impute within group. Default 0.5.}
}
\value{
Matrix with imputed values
}
\description{
Imputes missing values in a proteomics matrix using non-negative matrix factorization (NMF).
}
\details{
This function takes a numeric matrix \code{X} containing proteomics data, with samples in columns.
It uses \code{groups} to define sample groups.

It first replaces missing values (0s) with NAs.
For each sample group, any rows with more than \code{r} proportion of NAs are set to 0.

NMF with \code{k} factors is then applied to impute the remaining NAs.
The imputed matrix is returned.
}
