% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{psort}
\alias{psort}
\title{Sort Data Frame by P-value Column}
\usage{
psort(x, p.col = NULL)
}
\arguments{
\item{x}{A data frame representing the input data.}

\item{p.col}{An optional character string specifying the name of the column
containing p-values.}
}
\value{
A data frame representing the input data sorted by the p-value column.
}
\description{
This function sorts a data frame by a column containing p-values.
}
\details{
The function takes a data frame `x` as input and searches
for a column containing p-values. If the `p.col` parameter is not provided,
the function searches for a column with a name containing "p.value", "p",
"p-val", or "pval" (case-insensitive). If multiple columns match, the first one is used.

The data frame is then sorted in ascending order by the p-value column and returned.
}
\examples{
\dontrun{
# example code
x <- data.frame(a = c(1, 2, 3), p.value = c(0.05, 0.01, 0.1))
psort(x)
}
}
