% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pubmed.R
\name{pubmedlink}
\alias{pubmedlink}
\title{Create HTML link to PubMed article}
\usage{
pubmedlink(s)
}
\arguments{
\item{s}{PubMed ID (PMID) to link to.}
}
\value{
HTML link to the PubMed article page.
}
\description{
Generates an HTML hyperlink for a given PubMed ID (PMID).
}
\details{
Link to PubMed article

This function takes a PubMed ID as input and returns 
an HTML hyperlink that will open the article page on PubMed website.

The PMID is embedded in the link URL along with link text containing 
the PMID. This allows easy creation of hyperlinks to PubMed articles.
}
\examples{
\dontrun{
link <- pubmedlink("12345678")
cat(link)
}
}
