% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R
\name{purpleGreyYellow}
\alias{purpleGreyYellow}
\title{Purple-grey-yellow color ramp}
\usage{
purpleGreyYellow(n)
}
\description{
Purple-grey-yellow color ramp
}
\seealso{
WGCNAplus::purpleGreyYellow
}
