% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\docType{class}
\name{random_aggregator}
\alias{random_aggregator}
\alias{gaussian_aggregator}
\title{Plot data aggregators}
\format{
An \code{R6::R6Class} object

An \code{R6::R6Class} object
}
\description{
Plot data aggregators

Plot data aggregators
}
\details{
The data aggregators are usefult to downsample data before plotting. This becomes
very helpful when plotting large datasets. The aggregators are used by the 
shinyHugePlot::downsample objects. Currently, the following aggregators are
random aggregator and gaussian aggregator. The random aggregator select the 
data following a uniform probability distribution. The gaussian aggregator
assumes the data is a multivariate gaussian distribution and select the data
from the less dense areas in order to preserve the information.
}
\section{Functions}{
\itemize{
\item \code{gaussian_aggregator}: gaussian aggregator constructor for the downsampler object

}}
\section{Super class}{
\code{\link[shinyHugePlot:aggregator]{shinyHugePlot::aggregator}} -> \code{random_aggregator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-random_aggregator-new}{\code{random_aggregator$new()}}
\item \href{#method-random_aggregator-clone}{\code{random_aggregator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shinyHugePlot" data-topic="aggregator" data-id="aggregate"><a href='../../shinyHugePlot/html/aggregator.html#method-aggregator-aggregate'><code>shinyHugePlot::aggregator$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyHugePlot" data-topic="aggregator" data-id="set_parameters"><a href='../../shinyHugePlot/html/aggregator.html#method-aggregator-set_parameters'><code>shinyHugePlot::aggregator$set_parameters()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_aggregator-new"></a>}}
\if{latex}{\out{\hypertarget{method-random_aggregator-new}{}}}
\subsection{Method \code{new()}}{
Constructor of the Aggregator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{random_aggregator$new(
  ...,
  interleave_gaps,
  coef_gap,
  NA_position,
  accepted_datatype
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interleave_gaps, coef_gap, NA_position, accepted_datatype, ...}}{Arguments pass to the constructor of \code{aggregator} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_aggregator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-random_aggregator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{random_aggregator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[shinyHugePlot:aggregator]{shinyHugePlot::aggregator}} -> \code{gaussian_aggregator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussian_aggregator-new}{\code{gaussian_aggregator$new()}}
\item \href{#method-gaussian_aggregator-clone}{\code{gaussian_aggregator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shinyHugePlot" data-topic="aggregator" data-id="aggregate"><a href='../../shinyHugePlot/html/aggregator.html#method-aggregator-aggregate'><code>shinyHugePlot::aggregator$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shinyHugePlot" data-topic="aggregator" data-id="set_parameters"><a href='../../shinyHugePlot/html/aggregator.html#method-aggregator-set_parameters'><code>shinyHugePlot::aggregator$set_parameters()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussian_aggregator-new"></a>}}
\if{latex}{\out{\hypertarget{method-gaussian_aggregator-new}{}}}
\subsection{Method \code{new()}}{
Constructor of the Aggregator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussian_aggregator$new(
  ...,
  interleave_gaps,
  coef_gap,
  NA_position,
  accepted_datatype
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interleave_gaps, coef_gap, NA_position, accepted_datatype, ...}}{Arguments pass to the constructor of \code{aggregator} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussian_aggregator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussian_aggregator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussian_aggregator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
