% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_uniprots}
\alias{rank_uniprots}
\title{Re-order uniprot IDs (when a feature has multiple uniprot ids)
Rules:
If there are Swiss-Prot entries, Swiss-Prot entries come first, followed by TrEMBL entries.
Among Swiss-Prot entries, if there are canonical entries, canonical entries come first (eg P05067),
followed by isoforms (P05067-2). In absence of canonical entries, convert isoforms to canonical
(eg P05067-2 to P05067).
Among Swiss-Prot entries (after step 2), longer proteins come first.
Among TrEMBL entires, longer proteins come first.}
\usage{
rank_uniprots(feature, lengths = NULL, verbose = FALSE)
}
\arguments{
\item{feature}{feature ids, collapsed with ";".}

\item{lengths}{Character vector of feature lengths, collapsed with ";". Default NULL.}
}
\description{
Re-order uniprot IDs (when a feature has multiple uniprot ids)
Rules:
If there are Swiss-Prot entries, Swiss-Prot entries come first, followed by TrEMBL entries.
Among Swiss-Prot entries, if there are canonical entries, canonical entries come first (eg P05067),
followed by isoforms (P05067-2). In absence of canonical entries, convert isoforms to canonical
(eg P05067-2 to P05067).
Among Swiss-Prot entries (after step 2), longer proteins come first.
Among TrEMBL entires, longer proteins come first.
}
