% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{rbind_sparse_matrix}
\alias{rbind_sparse_matrix}
\title{Merge Sparse Matrix}
\usage{
rbind_sparse_matrix(m1, m2)
}
\arguments{
\item{m1}{The first sparse matrix.}

\item{m2}{The second sparse matrix.}
}
\value{
A merged sparse matrix with combined rows and shared
  features in columns
}
\description{
Merges two sparse matrices by combining row-wise (aka rbind) but
ensuring the same gene order. Union of features (gene/proteins) are
on columns (i.e. column ames are features names). This is commonly
done for merging GSETxGENE and MSETxMETABOLITE matrices.
}
