% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pathways.R
\name{reactomeDiagrams}
\alias{reactomeDiagrams}
\title{Pathway ids for which a native Reactome diagram can be shown.}
\usage{
reactomeDiagrams()
}
\description{
Includes sub-pathways that resolve to an ancestor's diagram.
}
