% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{read_annot}
\alias{read_annot}
\title{Read gene/probe annotation file}
\usage{
read_annot(
  file,
  drop_na_rows = TRUE,
  drop_na_cols = TRUE,
  first = FALSE,
  unique = TRUE,
  paste_char = "_"
)
}
\description{
Read gene/probe annotation file
}
