% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{read_cellranger_output}
\alias{read_cellranger_output}
\title{Read 10X Cell Ranger Software output (version V3 onwards).}
\usage{
read_cellranger_output(file)
}
\arguments{
\item{file}{.tar.gz or .zip compressed directory}
}
\value{
Count gene expression data matrix (sparse dgCMatrix)
}
\description{
Read 10X Cell Ranger Software output (version V3 onwards).
}
