% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_read.R
\name{read_expression}
\alias{read_expression}
\title{Read expression data from file}
\usage{
read_expression(file, convert_names = TRUE)
}
\arguments{
\item{file}{string. path to file}

\item{convert_names}{boolean.}
}
\value{
matrix. the file with the data
}
\description{
Read expression data from file
}
\examples{
counts <- read_expression(playbase::example_file("counts.csv"))
}
