% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{read_Olink_NPX}
\alias{read_Olink_NPX}
\title{Read Olink NPX data and create a counts matrix}
\usage{
read_Olink_NPX(NPX_data)
}
\arguments{
\item{NPX_data}{Path to input Olink NPX data file. Must be standard Olink NPX format as per OlinkAnalyze R package.}
}
\value{
data matrix (features on rows; samples on columns)
}
\description{
Read Olink NPX data and create a counts matrix
}
