% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{read_Olink_samples}
\alias{read_Olink_samples}
\title{Read Olink NPX data and automatically create samples dataframe}
\usage{
read_Olink_samples(NPX_data)
}
\arguments{
\item{NPX_data}{Path to input Olink NPX data file. Must be standard Olink NPX format as per OlinkAnalyze R package.}
}
\value{
dataframe with the metadata
}
\description{
Read Olink NPX data and automatically create samples dataframe
}
