% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_read.R
\name{read_samples}
\alias{read_samples}
\title{Read samples data from file}
\usage{
read_samples(file)
}
\arguments{
\item{file}{string. path to file}
}
\value{
dataframe the file with the data
}
\description{
Read samples data from file
}
\examples{
samples <- read_samples(playbase::playbase_example('samples.csv'))
}
