% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{read_spectronaut_hPTM}
\alias{read_spectronaut_hPTM}
\title{Read Spectronaut output hPTM abundance data file.}
\usage{
read_spectronaut_hPTM(file, use_ptm_norm = TRUE)
}
\arguments{
\item{use_ptm_norm}{Boolean. Use PTM abudances normalized to the global proteome. Default TRUE.}

\item{Path}{to Spectronaut output hPTM abundance file.}
}
\value{
Abundance data matrix + annotation (features on rows; samples on columns)
}
\description{
Read Spectronaut output hPTM abundance data file.
}
