% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-preprocess.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pp.alignCounts}
\alias{pp.batchCorrect}
\alias{pp.convertSpace}
\alias{pp.countScaleMatrix}
\alias{pp.deduplicate}
\alias{pp.filterFeatures}
\alias{pp.impute}
\alias{pp.normalize}
\alias{pp.removeOutliers}
\alias{pp.toCountScale}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{playbase.preprocess}{\code{\link[playbase.preprocess]{pp.alignCounts}}, \code{\link[playbase.preprocess]{pp.batchCorrect}}, \code{\link[playbase.preprocess]{pp.convertSpace}}, \code{\link[playbase.preprocess]{pp.countScaleMatrix}}, \code{\link[playbase.preprocess]{pp.deduplicate}}, \code{\link[playbase.preprocess]{pp.filterFeatures}}, \code{\link[playbase.preprocess]{pp.impute}}, \code{\link[playbase.preprocess]{pp.normalize}}, \code{\link[playbase.preprocess]{pp.removeOutliers}}, \code{\link[playbase.preprocess]{pp.toCountScale}}}
}}

