% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{relevelFactorFirst}
\alias{relevelFactorFirst}
\title{Relevel Factor by First Occurrence}
\usage{
relevelFactorFirst(f)
}
\arguments{
\item{f}{A factor to be releveled.}
}
\value{
A factor with the same levels as the input factor, but with the
levels reordered according to their first appearance in the data.
}
\description{
This function relevels a factor by the order of first
occurrence of its levels.
}
\details{
The function takes a factor `f` as input and relevels it by
the order of first occurrence of its levels in the data.
The new factor is created with the same levels as the original factor,
but with the levels reordered according to their first appearance in the data.
}
\examples{
\dontrun{
# example code
f <- factor(c("b", "a", "c", "b", "a"))
relevelFactorFirst(f)
}
}
