% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{rename_by_humansymbol}
\alias{rename_by_humansymbol}
\title{Collapse object rownames/names to human symbol. Warning this function
does not maintain the original dimensions/length of object.}
\usage{
rename_by_humansymbol(obj, annot)
}
\description{
Collapse object rownames/names to human symbol. Warning this function
does not maintain the original dimensions/length of object.
}
