% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{reverse.AvsB}
\alias{reverse.AvsB}
\title{Reverse Comparison Strings}
\usage{
reverse.AvsB(comp)
}
\arguments{
\item{comp}{A character vector of comparison strings to be reversed.}
}
\value{
A character vector of the same length as `comp`, containing the reversed comparison strings.
}
\description{
This function reverses the order of comparison strings in the format "A_vs_B".
}
\details{
The function takes a character vector of comparison strings `comp` as
input and reverses the order of the strings in the format "A_vs_B" or "A_VS_B".
The function also preserves any prefix before a colon (":") or postfix after an at
symbol ("@") in the input strings.
}
\examples{
\dontrun{
# example code
comp <- c("A_vs_B", "C_VS_D:prefix@postfix")
reverse.AvsB(comp)
}
}
