% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{RNA_TISSUE_MATRIX}
\alias{RNA_TISSUE_MATRIX}
\title{A dataset used by `compute_deconvolution`}
\format{
## `RNA_TISSUE_MATRIX`
}
\source{
unknown
}
\usage{
RNA_TISSUE_MATRIX
}
\description{
No idea what this data is but it is used in `compute_deconvolution`.
}
\keyword{datasets}
