% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{rowFUN}
\alias{rowFUN}
\title{Calculate group-wise row function call FUN from a matrix. Like rowmean
but generalized. NOTE: only with functions from matrixStats.}
\usage{
rowFUN(X, group = rownames(X), FUN = matrixStats::colMeans2, reorder = TRUE)
}
\description{
Calculate group-wise row function call FUN from a matrix. Like rowmean
but generalized. NOTE: only with functions from matrixStats.
}
