% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{rowmean}
\alias{rowmean}
\title{Calculate group-wise row means (like base::rowsum) from a
matrix. Faster than matGroupMeans.}
\usage{
rowmean(X, group = rownames(X), reorder = TRUE)
}
\description{
Calculate group-wise row means (like base::rowsum) from a
matrix. Faster than matGroupMeans.
}
