% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rownames.data.table}
\alias{rownames.data.table}
\title{Get row names for a data.table}
\usage{
rownames.data.table(dt)
}
\arguments{
\item{dt}{A data.table object.}
}
\value{
Character vector of row names
}
\description{
This function aims to extract the rownames of a data.table object as
if it would be a data.frame. The implementation is slow but it helps to maintain back
compatibility with data.frame objects.
}
\examples{
\dontrun{
dt <- data.table(rn = letters[1:5], x = 1:5)
rownames(dt)
}
}
