% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{rowscale}
\alias{rowscale}
\title{Row Scale a Matrix}
\usage{
rowscale(x)
}
\arguments{
\item{x}{A numeric matrix to be scaled.}
}
\value{
A numeric matrix of the same size as `x`, containing the scaled values.
}
\description{
This function scales the rows of a matrix by subtracting the row 
means and dividing by the row standard deviations.
}
\details{
This function takes a numeric matrix `x` as input and scales its rows by 
subtracting the row means and dividing by the row standard deviations. 
The resulting scaled matrix is returned.
}
\examples{
\dontrun{
# example code
x <- matrix(c(1, 2, 3, 4), nrow = 2)
rowscale(x)
}

}
