% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{rowsduplicated.dgCMatrix}
\alias{rowsduplicated.dgCMatrix}
\title{Fast test if rows are duplicated in sparse matrix M using random
vector and fast matrix multiplication.}
\usage{
rowsduplicated.dgCMatrix(M)
}
\description{
Fast test if rows are duplicated in sparse matrix M using random
vector and fast matrix multiplication.
}
