% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-old-report.R
\name{rpt.compile_sections}
\alias{rpt.compile_sections}
\title{Format and render contents as section.}
\usage{
rpt.compile_sections(contents, hlevel = 2, shift = TRUE)
}
\description{
Format and render contents as section.
}
