% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report.R
\name{rpt.create_summary_report}
\alias{rpt.create_summary_report}
\title{Format and render contents as section.}
\usage{
rpt.create_summary_report(
  pgx,
  llm,
  ntop = 100,
  format = "short",
  add_prompt = NULL,
  sections = NULL
)
}
\description{
Format and render contents as section.
}
