% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report.R
\name{rpt.summarize_drug_connectivity}
\alias{rpt.summarize_drug_connectivity}
\title{Summarize drug connectivity results}
\usage{
rpt.summarize_drug_connectivity(pgx, ct, model, db = 1, ntop = 10)
}
\description{
Summarize drug connectivity results
}
