% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA.LeadingEdge}
\alias{run.GSEA.LeadingEdge}
\title{Title}
\usage{
run.GSEA.LeadingEdge(
  rpt.path,
  gmt,
  xmx = 10,
  rpt.label = "gsea_leadingedge",
  gsea.program = GSEA.JAR,
  outdir = "."
)
}
\arguments{
\item{outdir}{}
}
\description{
Title
}
