% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA.preranked}
\alias{run.GSEA.preranked}
\title{Title}
\usage{
run.GSEA.preranked(
  rnk,
  gmt,
  output.dir = NULL,
  fdr = 0.25,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  rpt.label = "preranked",
  make.sets = TRUE,
  clean.files = TRUE,
  xmx = 10,
  scoring = "weighted",
  chip = "GENE_SYMBOL.chip",
  collapse = FALSE,
  do.leading.edge = FALSE,
  gsea.program = "/opt/GSEA/gsea-3.0.jar",
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\description{
Title
}
