% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA}
\alias{run.GSEA}
\title{Run Gene Set Enrichment Analysis (GSEA)}
\usage{
run.GSEA(
  X,
  y,
  gmt,
  output.dir = NULL,
  fdr = 0.25,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  permute = "phenotype",
  scoring = "weighted",
  do.leading.edge = FALSE,
  rpt.label = "gsea",
  sort.y = "ref.last",
  ref.type = c("0", "NT", "REF", "DMSO", "WT", "LOW"),
  make.sets = TRUE,
  clean.files = TRUE,
  xmx = 10,
  force.permute = FALSE,
  metric = c("Signal2Noise", "Diff_of_Classes", "Pearson"),
  gsea.program = "/opt/GSEA/gsea-3.0.jar",
  quiet = FALSE
)
}
\arguments{
\item{X}{The gene expression data matrix.}

\item{y}{The phenotype vector.}

\item{gmt}{The gene set collection in GMT format.}

\item{output.dir}{Directory where the output files will be saved. Default is NULL, which creates a temporary directory.}

\item{fdr}{False discovery rate (FDR) threshold for filtering results. Default is 0.25.}

\item{set.min}{Minimum size of gene sets to consider. Default is 15.}

\item{set.max}{Maximum size of gene sets to consider. Default is 500.}

\item{topgs}{Number of top gene sets to report. Default is 100.}

\item{nperm}{Number of permutations for statistical significance estimation. Default is 1000.}

\item{permute}{Permutation type, either "phenotype" or "gene_set". Default is "phenotype".}

\item{scoring}{The scoring scheme for ranking genes. Default is "weighted".}

\item{do.leading.edge}{Whether to compute the leading edge genes for each gene set. Default is FALSE.}

\item{rpt.label}{The label to use for the output files. Default is "gsea".}

\item{sort.y}{The sorting order of the phenotype classes. Default is "ref.last".}

\item{ref.type}{The reference class or classes in the phenotype vector. Default is c("0", "NT", "REF", "DMSO", "WT", "LOW").}

\item{make.sets}{Whether to create gene sets using the gene expression data. Default is TRUE.}

\item{clean.files}{Whether to remove old files in the output directory. Default is TRUE.}

\item{xmx}{Maximum memory allocation for GSEA in gigabytes. Default is 10.}

\item{force.permute}{Whether to force gene_set permutation when there are fewer than 7 samples. Default is FALSE.}

\item{metric}{The metric for ranking genes. Default is c("Signal2Noise", "Diff_of_Classes", "Pearson").}

\item{gsea.program}{Path to the GSEA program JAR file. Default is "/opt/GSEA/gsea-3.0.jar".}

\item{quiet}{Whether to suppress console output during the analysis. Default is FALSE.}
}
\value{
A data frame containing the GSEA results, including normalized enrichment score (NES), nominal p-value (NOM p-val), and false discovery rate (FDR) q-value.
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) on the given gene expression data.
}
