% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-grep2.R
\name{run_tximport_kallisto}
\alias{run_tximport_kallisto}
\title{Run tximport on Kallisto quantification}
\usage{
run_tximport_kallisto(
  srr_id,
  species = c("human", "mouse", "rat"),
  kallisto_dir
)
}
\arguments{
\item{srr_id}{Character vector of SRR IDs for samples}

\item{species}{Species name, used to get gene annotation data. Default is c("human", "mouse", "rat").}

\item{kallisto_dir}{Directory containing Kallisto output folders for each sample.}
}
\value{
List containing gene and transcript count matrices and tximport result objects.
}
\description{
Imports Kallisto transcript-level abundance estimates into R matrices at gene and transcript level using tximport.
}
\details{
This function takes a vector of SRR IDs and the path to the Kallisto output directory containing abundance estimates (abundance.tsv files) for each sample.
It imports the transcript counts into R using tximport, summarizing into gene-level counts based on gene annotation data for the specified species.

The output is a list containing the gene counts matrix, transcript counts matrix, and the tximport result objects.
}
