% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{SAMPLES}
\alias{SAMPLES}
\title{Example sample data from unknown GEO dataset}
\format{
## `SAMPLES`
data.frame
}
\source{
unknown
}
\usage{
SAMPLES
}
\description{
Example sample data from unknown GEO dataset
}
\keyword{datasets}
