% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-colors.R
\name{scale_color_omics_c}
\alias{scale_color_omics_c}
\title{Color scale constructor for continuous omics color palettes}
\usage{
scale_color_omics_c(palette = "brand_blue", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omics_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE
or FALSE}
}
\description{
Color scale constructor for continuous omics color palettes
}
\examples{
library(ggplot2)
ggplot(mpg, aes(hwy, cty, color = displ)) +
  geom_point(size = 4) +
  scale_color_omics_c()
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Petal.Width)) +
  geom_point(size = 4) +
  scale_color_omics_c("turq_orange_GREY", reverse = TRUE)
}
