% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-colors.R
\name{scale_color_omics_d}
\alias{scale_color_omics_d}
\title{Color scale constructor for discrete omics colors}
\usage{
scale_color_omics_d(palette = "default", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in omics_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE
or FALSE}
}
\description{
Color scale constructor for discrete omics colors
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, cty, color = class)) +
  geom_point(size = 4) +
  scale_color_omics_d()
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_omics_d("mUTED", reverse = TRUE)

}
