% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-singlecell.R
\name{seurat.transferLabels}
\alias{seurat.transferLabels}
\title{Transfer labels from reference to query (Seurat objects)}
\usage{
seurat.transferLabels(ref, query, labels, sct = TRUE)
}
\description{
Transfer labels from reference to query (Seurat objects)
}
