% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-singlecell.R
\name{seurat2pgx}
\alias{seurat2pgx}
\title{Convert Seurat to PGX}
\usage{
seurat2pgx(obj, do.cluster = FALSE, organism)
}
\arguments{
\item{obj}{Seurat object to convert}

\item{do.cluster}{Logical indicating whether to cluster samples. Default is FALSE.}
}
\value{
PGX object
}
\description{
Converts a Seurat single-cell RNA-seq object into a PGX object
}
\details{
This function takes a Seurat object containing single-cell RNA-seq data and converts it into a PGX object.
The count matrix, normalized expression matrix, and sample metadata are extracted from the Seurat object.
Gene annotations are added using the gene symbols.
If do.cluster=TRUE, dimensionality reduction and clustering of samples is performed.
Any existing tsne/umap embeddings and cluster assignments are copied over from the Seurat object.
}
