% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R, R/pgx-ensembl.R
\name{showProbeTypes}
\alias{showProbeTypes}
\title{Show some probe types for selected organism}
\usage{
showProbeTypes(organism, keytypes = NULL, use.ah = NULL, n = 10)

showProbeTypes(organism, keytypes = NULL, use.ah = NULL, n = 10)
}
\description{
Show some probe types for selected organism

Show some probe types for selected organism
}
