% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{sigdb.getConnectivityContrasts}
\alias{sigdb.getConnectivityContrasts}
\title{Get Connectivity Contrasts from a SIGDB File}
\usage{
sigdb.getConnectivityContrasts(sigdb, path = NULL)
}
\arguments{
\item{sigdb}{A character string specifying the name of the SIGDB file.}

\item{path}{An optional character string specifying the path to the directory containing the SIGDB file (default is NULL).}
}
\value{
A character vector containing the connectivity contrasts from the specified SIGDB file, or NULL if the file does not exist.
}
\description{
This function reads the connectivity contrasts from a specified SIGDB file and returns them as a character vector.
}
\details{
The function first checks if a path to the SIGDB file is provided.
If a path is provided, it constructs the full path to the SIGDB file by concatenating the path and the file name.
The function then checks if the specified SIGDB file exists.
If the file exists, it uses the `h5read` function from the rhdf5 package to read the connectivity contrasts from the "data/colnames" dataset in the SIGDB file.
}
