% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-signature.R
\name{sigdb.getConnectivityMatrix}
\alias{sigdb.getConnectivityMatrix}
\alias{sigdb.getEnrichmentMatrix}
\alias{sigdb.getSignatureMatrix}
\alias{sigdb.removeDataset}
\title{Get elements from a sigDB}
\usage{
sigdb.getConnectivityMatrix(sigdb, select = NULL, genes = NULL, path = NULL)

sigdb.getEnrichmentMatrix(
  sigdb,
  select = NULL,
  path = NULL,
  which = c("gsea", "rankcor")[1]
)

sigdb.getSignatureMatrix(sigdb, path = NULL)

sigdb.removeDataset(h5.file, pgxname)
}
\arguments{
\item{sigdb}{Path to signature database HDF5 file}

\item{select}{Character vector of contrasts to return}

\item{genes}{Character vector of genes to return}

\item{path}{Optional path prefix for sigdb file}

\item{which}{Method to use for enrichment, either "gsea" or "rankcor"}
}
\value{
Matrix of connectivity scores
}
\description{
Get the connectivity matrix from a signature database HDF5 file.
}
\details{
This function extracts the connectivity matrix from a signature database
stored in HDF5 format. The HDF5 file should contain the matrix under /data/matrix.

You can filter the returned matrix by specifying a subset of contrasts to return
with the \code{select} parameter. Use the \code{genes} parameter to return only a
subset of genes.

If the sigdb file is not in the working directory, set the \code{path} parameter.
}
\section{Functions}{
\itemize{
\item \code{sigdb.getEnrichmentMatrix()}: This function reads the enrichment data stored in a signature database in HDF5 format.
It extracts either the gene set enrichment analysis (GSEA) results or the rank
correlation analysis results.

\item \code{sigdb.getSignatureMatrix()}: reads the up/down regulated gene signatures from a signature database HDF5 file.

\item \code{sigdb.removeDataset()}: removes a dataset and its associated data from a signature database HDF5 file.

}}
