% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{sMNN}
\alias{sMNN}
\title{Supervised MNN correction. Finds mutual neighbours with same label
between two datasets and returns corresponding correction vectors.}
\usage{
sMNN(X, batch, y, nv = 0.33, nn = 3, return.idx = FALSE)
}
\description{
Supervised MNN correction. Finds mutual neighbours with same label
between two datasets and returns corresponding correction vectors.
}
