% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{sparse_cor_matrix}
\alias{sparse_cor_matrix}
\title{Calculate sparse correlation matrix handling missing values}
\usage{
sparse_cor_matrix(G, rank_matrix)
}
\arguments{
\item{G}{Sparse matrix containing gene sets}

\item{rank_matrix}{Matrix of ranked values}
}
\value{
Correlation matrix between G and rank_matrix
}
\description{
Calculate sparse correlation matrix handling missing values
}
\details{
If rank_matrix has no missing values, calculates correlation directly using corSparse.
Otherwise computes column-wise correlations only using non-missing values.
}
