% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{SPECIES_TABLE}
\alias{SPECIES_TABLE}
\title{Table for species ID conversin conversion from scientific name to Ensembl db ID}
\format{
## `SPECIES_TABLE`
data.frame with columns "dataset"	"description"	"version"	"species_name"
}
\source{
BiomaRt
}
\usage{
SPECIES_TABLE
}
\description{
Table for species ID conversin conversion from scientific name to Ensembl db ID
}
\keyword{datasets}
