% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{star.symbols}
\alias{star.symbols}
\title{Generate Star Symbols}
\usage{
star.symbols(n, pch = "★")
}
\arguments{
\item{n}{A numeric value specifying the number of star symbols to generate.}

\item{pch}{An optional character string specifying the symbol to use for the stars.
The default value is a black star.}
}
\value{
A character vector of length 1, containing the generated string of star symbols.
}
\description{
This function generates a string of star symbols.
}
\details{
This function takes a numeric value `n` and an optional character string `pch` as
input and generates a string of `n` star symbols using the specified `pch` symbol.
If `n` is 0, an empty string is returned.
The resulting string of star symbols is returned as a character vector of length 1.
}
\examples{
\dontrun{
# example code
star.symbols(3)
star.symbols(5, pch = "*")
}
}
