% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{stratifyContrasts}
\alias{stratifyContrasts}
\title{Stratify contrasts matrix within strata variable}
\usage{
stratifyContrasts(contrasts, strata, mingrp = 2)
}
\description{
Stratify contrasts matrix within strata variable
}
