% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{strip_postfix}
\alias{strip_postfix}
\title{non-greedy removal of numerical postfix. Postfix is defined as
numerical substring after . (dot)}
\usage{
strip_postfix(s)
}
\description{
non-greedy removal of numerical postfix. Postfix is defined as
numerical substring after . (dot)
}
