% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{strip_prefix}
\alias{strip_prefix}
\title{non-greedy removal of prefixes. Prefix is defined as any
character substring (no spaces, no special chars) before the
matching colon character :.}
\usage{
strip_prefix(s)
}
\description{
non-greedy removal of prefixes. Prefix is defined as any
character substring (no spaces, no special chars) before the
matching colon character :.
}
