% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{strwrap2}
\alias{strwrap2}
\title{Wrap Strings to a Specified Width}
\usage{
strwrap2(str, n)
}
\arguments{
\item{str}{A character vector of strings to be wrapped.}

\item{n}{A numeric value specifying the maximum width of each line in characters.}
}
\value{
A character vector of the same length as `str`, containing the wrapped strings.
}
\description{
This function wraps a character vector of strings to a specified width.
}
\details{
This function takes a character vector of strings `str` and a numeric value `n` as input and wraps each string in `str` to a maximum width of `n` characters. 
The resulting wrapped strings are returned as a character vector, where each element contains one or more lines separated by newline characters.
}
\examples{
\dontrun{
# example code
str <- c("The quick brown fox jumps over the lazy dog")
strwrap2(str, 10)
}

}
