% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_treps}
\alias{sum_treps}
\title{Sum up tech replicates of a linear count matrix 
This function combines (sums up) features of technical replicates
It does so in the linear space.}
\usage{
sum_treps(counts, trep_var = "")
}
\arguments{
\item{counts}{Numeric matrix}

\item{trep_var}{Character vector indicating tech rep info for each sample}
}
\value{
counts matrix with technical replicates combined
}
\description{
Sum up tech replicates of a linear count matrix 
This function combines (sums up) features of technical replicates
It does so in the linear space.
}
