% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{tagged.hamming}
\alias{tagged.hamming}
\title{Calculate tagged Hamming distance}
\usage{
tagged.hamming(aa, bb, align = TRUE)
}
\arguments{
\item{aa}{Character vector. The first sequence.}

\item{bb}{Character vector. The second sequence.}

\item{align}{Logical. Whether to align the sequences before calculating Hamming distance. Default is TRUE.}
}
\value{
Named numeric vector of Hamming distances for each matched tag.
}
\description{
Calculate the Hamming distance between two sequences, taking into account tags.
}
\details{
This function calculates the Hamming distance between two character vector 
sequences \code{aa} and \code{bb}. It first parses any tags in the sequences using \code{parse.tags}. 
It then calculates the Hamming distance between the sequences indicated by any matching tags (e.g. 
\code{cdr1}, \code{cdr2}). For tagged sequences, it will align the sequences before calculating Hamming 
distance if \code{align=TRUE}. For non-tagged sequences, it calculates simple Hamming distance.
}
