% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{tcosine_similarity}
\alias{tcosine_similarity}
\alias{cosine_similarity}
\alias{tcosine.sparse}
\title{Calculate cosine similarity}
\usage{
tcosine_similarity(X, Y = NULL, method = NULL)

cosine_similarity(X, Y = NULL, method = NULL)

tcosine.sparse(
  X,
  k = 100,
  th = 0.01,
  block = 100,
  ties.method = "random",
  gpu = FALSE
)
}
\arguments{
\item{X}{Numeric matrix}

\item{Y}{Optional second numeric matrix to compare against X}

\item{method}{Optional specification for method to calculate cosine similarity}

\item{k:}{number of nearest neighbors (default 100)}

\item{th:}{similarity threshold (default 0.01)}

\item{block:}{block size for processing (default 100)}

\item{ties.method:}{method for breaking ties (default "random")}

\item{gpu:}{use GPU acceleration (default FALSE)}
}
\value{
Matrix of cosine similarity values
}
\description{
Calculates the cosine similarity between two matrices.
}
\details{
This function takes two numeric matrices \code{X} and \code{Y} and calculates the cosine similarity between them.
It can handle sparse matrices and missing values.

Cosine similarity is defined as the cosine of the angle between two vectors. 
It is calculated as the dot product of the vectors divided by the L2 norms of the vectors.

If only \code{X} is provided, the cosine similarity between columns of \code{X} is calculated.
If both \code{X} and \code{Y} are provided, the cosine similarity between columns of \code{X} and \code{Y} is calculated.

The \code{method} parameter allows specifying alternate methods to calculate the cosine similarity.
By default the cosine similarity formula is used directly.
}
\section{Functions}{
\itemize{
\item \code{cosine_similarity()}: calculates the cosine similarity between two matrices X and Y.

\item \code{tcosine.sparse()}: calculates a sparse cosine similarity matrix for the data matrix X

}}
