% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{tiledb.addDataset.create}
\alias{tiledb.addDataset.create}
\title{Create new TileDB database with single file (internal)}
\usage{
tiledb.addDataset.create(tiledb_path, pgx_file, verbose = TRUE)
}
\description{
Create new TileDB database with single file (internal)
}
\keyword{internal}
