% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{tiledb.addDataset}
\alias{tiledb.addDataset}
\title{Add a dataset to existing TileDB database}
\usage{
tiledb.addDataset(tiledb_path, pgx_file, overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{tiledb_path}{Path to the TileDB database}

\item{pgx_file}{Path to .pgx file to add}

\item{overwrite}{If TRUE and dataset already exists, remove old data first. Default TRUE.}

\item{verbose}{Logical, whether to print progress messages. Default TRUE.}
}
\value{
Invisibly returns updated metadata
}
\description{
Adds a single PGX file to an existing TileDB database.
If the dataset already exists, it can be overwritten. If the TileDB
database doesn't exist, it will be created.
}
\details{
This function enables incremental updates to the TileDB database
without rebuilding from scratch. It:
\itemize{
  \item Loads the pgx file and extracts counts/phenotypes
  \item Appends new data to the TileDB sparse array
  \item Updates the metadata RDS file
}
}
