% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{tiledb.extractPhenotypes}
\alias{tiledb.extractPhenotypes}
\title{Extract phenotype data from PGX object}
\usage{
tiledb.extractPhenotypes(pgx, file_prefix)
}
\arguments{
\item{pgx}{PGX object with samples data.frame}

\item{file_prefix}{Dataset name prefix for sample IDs}
}
\value{
List with phenotype_data (list of data.frames) and phenotypes (column names)
}
\description{
Extracts phenotype data in long format
}
\keyword{internal}
