% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{tiledb.prepareData}
\alias{tiledb.prepareData}
\title{Extract counts and z-scores from PGX object}
\usage{
tiledb.prepareData(pgx)
}
\arguments{
\item{pgx}{PGX object with counts and X matrices}
}
\value{
List with counts, zscores, genes, and samples
}
\description{
Prepares counts matrix and computes z-scores from log-scaled X matrix
}
\keyword{internal}
