% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{tiledb.removeDataset.internal}
\alias{tiledb.removeDataset.internal}
\title{Remove dataset from TileDB (internal helper)}
\usage{
tiledb.removeDataset.internal(
  tiledb_path,
  dataset_name,
  metadata,
  verbose = TRUE
)
}
\description{
Remove dataset from TileDB (internal helper)
}
\keyword{internal}
