% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-tiledb.R
\name{tiledb.writeData}
\alias{tiledb.writeData}
\title{Write counts and z-scores to TileDB array}
\usage{
tiledb.writeData(arr, data, file_prefix)
}
\arguments{
\item{arr}{TileDB array object}

\item{data}{List from tiledb.prepareData()}

\item{file_prefix}{Dataset name prefix for sample IDs}
}
\description{
Writes data for all samples to TileDB sparse array
}
\keyword{internal}
