% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-singlecell.R
\name{transferLabels}
\alias{transferLabels}
\title{Transfer labels from reference to query (matrices)}
\usage{
transferLabels(ref.mat, query.mat, labels)
}
\description{
Transfer labels from reference to query (matrices)
}
