% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{trimsame}
\alias{trimsame}
\alias{trimsame.ends}
\alias{trimsame0}
\title{Trim Common Prefix from Strings}
\usage{
trimsame(s, split = " ", ends = TRUE, summarize = FALSE)

trimsame.ends(s, split = " ", summarize = FALSE)

trimsame0(s, split = " ", summarize = FALSE, rev = FALSE)
}
\arguments{
\item{s}{A character vector of strings to be trimmed.}

\item{split}{An optional character string specifying the delimiter used to split the strings into words.
The default value is a space (" ").}

\item{summarize}{An optional logical value indicating whether to summarize the common prefix using the first letter of each word.
The default value is `FALSE`.}

\item{rev}{An optional logical value indicating whether to reverse the order of the words in the summarized prefix.
The default value is `FALSE`.}
}
\value{
A character vector of the same length as `s`, containing the trimmed strings.
}
\description{
This function trims the common prefix from a character vector of strings.
}
\details{
The function takes a character vector of strings `s` as input and searches for a common prefix among the strings. 
The common prefix is defined as the longest sequence of characters that is shared by all strings and ends with the specified delimiter. 
If a common prefix is found, it is trimmed from all strings. 
If the `summarize` parameter is `TRUE`, the common prefix is summarized using the first letter of each word and appended to the beginning of each string. 
If the `rev` parameter is `TRUE`, the order of the words in the summarized prefix is reversed.
}
\section{Functions}{
\itemize{
\item \code{trimsame()}: trimsame is a function that trims common prefixes and/or 
suffixes from a character vector by applying trimsame0 forwards and/or backwards.

\item \code{trimsame.ends()}: trimsame.ends is a function that trims common prefixes and 
suffixes from a character vector by applying trimsame0 forwards and backwards.

}}
