% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-math.r
\name{uscale}
\alias{uscale}
\title{Rescale values to 0-1 range}
\usage{
uscale(x, symm = FALSE)
}
\arguments{
\item{x}{A numeric vector to rescale.}

\item{symm}{Logical indicating if rescaled values should be symmetrized around 0. Default is FALSE.}
}
\value{
A numeric vector with rescaled values.
}
\description{
Rescales a numeric vector to a 0-1 range.
}
\details{
Rescale values to 0-1 range

This function takes a numeric vector \code{x} and rescales the values to lie between 0 and 1.
It subtracts the minimum value and divides by the range.
This transforms the values to a 0-1 range while maintaining relative differences.

If \code{symm=TRUE}, the rescaled values are further transformed to be symmetrized around 0
by subtracting 0.5 and multiplying by 2.

The rescaled vector is returned.
}
\examples{
\dontrun{
x <- rnorm(100)
x_scaled <- uscale(x)
}
}
