% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{validate_contrasts}
\alias{validate_contrasts}
\title{Validate contrasts data}
\usage{
validate_contrasts(data)
}
\arguments{
\item{data}{matrix.}
}
\value{
boolean. true if data is valid
}
\description{
Contrasts data is valid if:
 - no duplicate rows
 - no empty rows
 - no duplicate cols
 - contains only cols with "_vs_" in names
}
