% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{visplot.PCSF}
\alias{visplot.PCSF}
\title{Plot interactive visnetwork graph from PCSF results object}
\usage{
visplot.PCSF(
  net,
  style = 0,
  edge_width = 5,
  node_size = 40,
  node_label_cex = 30,
  Steiner_node_color = "lightblue",
  Terminal_node_color = "lightgreen",
  Terminal_node_legend = "Terminal",
  Steiner_node_legend = "Steiner",
  layout = "layout_with_kk",
  physics = TRUE,
  layoutMatrix = NULL,
  width = 1800,
  height = 1800,
  invert.weight = FALSE,
  extra_node_colors = list(),
  edge_color = "lightgrey",
  ...
)
}
\value{
visNetwork plot object
}
\description{
Plot interactive visnetwork PCSF graph from result of
  pgx.computePCSF. To be called from plotPCSF.
}
