% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.calculateSignificanceScore}
\alias{wgcna.calculateSignificanceScore}
\title{Calculate compound significance scores per gene}
\usage{
wgcna.calculateSignificanceScore(
  wgcna,
  collapse = TRUE,
  sort.by = "score",
  digits = 4,
  annot = NULL,
  rownames = NULL,
  annot.cols = c("feature", "symbol", "gene_title")
)
}
\arguments{
\item{wgcna}{WGCNA result object with stats.}
}
\value{
Data frame of compound significance scores.
}
\description{
Calculate compound significance scores per gene
}
