% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.computeConsensusGeneStats}
\alias{wgcna.computeConsensusGeneStats}
\title{Compute gene statistics with original datExpr but with consensus
colors/labels for each layers. A separate function
wgcna.getConsensusGeneStats() extracts clean tables from this
results object.}
\usage{
wgcna.computeConsensusGeneStats(cons)
}
\description{
Compute gene statistics with original datExpr but with consensus
colors/labels for each layers. A separate function
wgcna.getConsensusGeneStats() extracts clean tables from this
results object.
}
\seealso{
WGCNAplus::computeConsensusGeneStats
}
