% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.computeConsensusModuleEnrichment}
\alias{wgcna.computeConsensusModuleEnrichment}
\title{Compute consensus enrichment by calculating overlapping enriched
terms.}
\usage{
wgcna.computeConsensusModuleEnrichment(
  cons,
  GMT,
  annot,
  methods = c("fisher", "gsetcor", "xcor"),
  min.genes = 3,
  ntop = 400
)
}
\description{
Compute consensus enrichment by calculating overlapping enriched
terms.
}
\seealso{
WGCNAplus::computeConsensusModuleEnrichment
}
