% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.computeDistinctMatrix}
\alias{wgcna.computeDistinctMatrix}
\title{Compute consensus matrix from list of matrices. The consensus
matrix checks for consistent sign and minimal threshold for each
matrix. Optionally filters on consistent p-value.}
\usage{
wgcna.computeDistinctMatrix(matlist, psig = 0.05, min.diff = 0.3)
}
\description{
Compute consensus matrix from list of matrices. The consensus
matrix checks for consistent sign and minimal threshold for each
matrix. Optionally filters on consistent p-value.
}
