% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.computeModules}
\alias{wgcna.computeModules}
\title{Reimplementation for WGCNA::blockwiseModules(). This returns exact
same object as WGCNA::blockwiseModules() but is faster and allows
different clustering linkage methods (ward, complete).}
\usage{
wgcna.computeModules(
  datExpr,
  power = 6,
  networkType = "signed",
  TOM = NULL,
  TOMType = "signed",
  calcMethod = "fast",
  lowrank = 20,
  clustMethod = "average",
  cutMethod = "hybrid",
  deepSplit = 2,
  treeCut = 0.99,
  treeCutCeiling = 1,
  minModuleSize = 20,
  minModuleSize2 = minModuleSize,
  mergeCutHeight = 0.15,
  minKMEtoStay = 0.3,
  numericLabels = FALSE,
  maxBlockSize = 9999,
  returnTOM = FALSE,
  verbose = 1
)
}
\description{
Reimplementation for WGCNA::blockwiseModules(). This returns exact
same object as WGCNA::blockwiseModules() but is faster and allows
different clustering linkage methods (ward, complete).
}
\seealso{
WGCNAplus::computeModules
}
