% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.create_lasagna_model}
\alias{wgcna.create_lasagna_model}
\title{Create lasagna graph for wgcna object}
\usage{
wgcna.create_lasagna_model(wgcna, layers = NULL)
}
\description{
Create lasagna graph for wgcna object
}
