% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.filterColors}
\alias{wgcna.filterColors}
\title{Filter color vector by minimum KME and mergeCutHeight. Set color of
features with KME smaller than minKME to grey (or 0) group. Merge
similar modules with (module) correlation larger than
(1-mergeCutHeight) together.}
\usage{
wgcna.filterColors(
  X,
  colors,
  minKME = 0.3,
  mergeCutHeight = 0.15,
  minmodsize = 20,
  ntop = -1
)
}
\description{
Filter color vector by minimum KME and mergeCutHeight. Set color of
features with KME smaller than minKME to grey (or 0) group. Merge
similar modules with (module) correlation larger than
(1-mergeCutHeight) together.
}
