% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.getGeneStats}
\alias{wgcna.getGeneStats}
\title{Compute per-gene significance table for a module/trait}
\usage{
wgcna.getGeneStats(
  wgcna,
  trait,
  module = NULL,
  plot = TRUE,
  stats = NULL,
  labels = NULL,
  showlogFC = TRUE,
  col = NULL,
  main = NULL
)
}
\description{
Compute per-gene significance table for a module/trait
}
\seealso{
WGCNAplus::getGeneStats
}
