% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.getModuleCrossGenes}
\alias{wgcna.getModuleCrossGenes}
\title{Find genes most correlated with each module eigengene across layers}
\usage{
wgcna.getModuleCrossGenes(
  wgcna,
  ref = NULL,
  ngenes = 100,
  multi = TRUE,
  modules = NULL
)
}
\description{
Find genes most correlated with each module eigengene across layers
}
\seealso{
WGCNAplus::getModuleCrossGenes
}
