% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-old-wgcnareport.R
\name{wgcna.graph2dot}
\alias{wgcna.graph2dot}
\title{Create DOT string object from graph object for sending to
LLM. Clean unneeded attributes.}
\usage{
wgcna.graph2dot(graph)
}
\description{
Create DOT string object from graph object for sending to
LLM. Clean unneeded attributes.
}
