% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.init}
\alias{wgcna.init}
\title{Init/update function}
\usage{
wgcna.init(
  wgcna,
  llm = NULL,
  img_model = NULL,
  annot = NULL,
  sv.tom = 40,
  progress = NULL
)
}
\description{
Init/update function
}
