% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.matchColors}
\alias{wgcna.matchColors}
\title{Match module colors of a wgcna result to a reference set of colors}
\usage{
wgcna.matchColors(wgcna, refcolors)
}
\description{
Match module colors of a wgcna result to a reference set of colors
}
\seealso{
WGCNAplus::matchColors
}
