% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.merge_block_dendrograms}
\alias{wgcna.merge_block_dendrograms}
\title{Merge per-block dendrograms into a single dendrogram}
\usage{
wgcna.merge_block_dendrograms(net, X, method = 1)
}
\description{
Merge per-block dendrograms into a single dendrogram
}
\seealso{
WGCNAplus::merge_block_dendrograms
}
