% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.mergeCloseModules}
\alias{wgcna.mergeCloseModules}
\title{Merge WGCNA modules whose eigengenes are closer than a cut height}
\usage{
wgcna.mergeCloseModules(datExpr, colors, cutHeight, MEs = NULL)
}
\description{
Merge WGCNA modules whose eigengenes are closer than a cut height
}
\seealso{
WGCNAplus::mergeCloseModules
}
