% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.mergeME}
\alias{wgcna.mergeME}
\title{Merge multiple ME matrices into one. Allow different dimensions.}
\usage{
wgcna.mergeME(mlist, me2 = NULL, prefix = FALSE)
}
\description{
WGCNAplus::mergeME() has no explicit handling for \code{mlist = NULL,
me2 = <matrix>}; that edge case is preserved here as an early return.
}
\seealso{
WGCNAplus::mergeME
}
