% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.pickSoftThreshold}
\alias{wgcna.pickSoftThreshold}
\title{Better (?) method to pick soft threshold (aka power).}
\usage{
wgcna.pickSoftThreshold(
  datExpr,
  sft = NULL,
  rcut = 0.85,
  method = c("sft", "iqr")[1],
  nmax = -1,
  powers = NULL,
  verbose = 1
)
}
\description{
Better (?) method to pick soft threshold (aka power).
}
\seealso{
WGCNAplus::pickSoftThreshold
}
