% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R, R/pgx-wgcna.R
\name{wgcna.plotMDS}
\alias{wgcna.plotMDS}
\title{Plot Multi-dimensional scaling (MDS) of centered data matrix.}
\usage{
wgcna.plotMDS(wgcna, main = NULL, scale = FALSE)

wgcna.plotMDS(wgcna, main = NULL, scale = FALSE)
}
\description{
Plot Multi-dimensional scaling (MDS) of centered data matrix.

Plot Multi-dimensional scaling (MDS) of centered data matrix.
}
\seealso{
WGCNAplus::plotMDS
}
