% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.plotMMvsGS}
\alias{wgcna.plotMMvsGS}
\title{Plot membership correlation vs gene signficance (correlation with
trait) to discover biomarkers/driver genes.}
\usage{
wgcna.plotMMvsGS(
  wgcna,
  module,
  trait,
  abs = TRUE,
  par = TRUE,
  plotlib = "base"
)
}
\description{
Plot membership correlation vs gene signficance (correlation with
trait) to discover biomarkers/driver genes.
}
