% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R
\name{wgcna.plotModuleHubGenes}
\alias{wgcna.plotModuleHubGenes}
\title{Plot module hub genes as graph in circle layout. This shows the
main connectivity structure for the modules. It can be used for
(edge) preservation analysis for comparing different group of
samples..}
\usage{
wgcna.plotModuleHubGenes(wgcna, modules = NULL, alpha = 0.5, setpar = TRUE)
}
\description{
Plot module hub genes as graph in circle layout. This shows the
main connectivity structure for the modules. It can be used for
(edge) preservation analysis for comparing different group of
samples..
}
\seealso{
WGCNAplus::plotModuleHubGenes
}
