% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.plotModuleTraitHeatmap}
\alias{wgcna.plotModuleTraitHeatmap}
\title{Plot membership correlation vs gene signficance (correlation with
trait) to discover biomarkers/driver genes.}
\usage{
wgcna.plotModuleTraitHeatmap(
  wgcna,
  setpar = TRUE,
  cluster = FALSE,
  multi = FALSE,
  main = NULL,
  justdata = FALSE,
  transpose = FALSE,
  colorlabel = TRUE,
  show = c("both", "traits", "contrasts")[1],
  nmax = -1,
  tmax = -1,
  text = TRUE,
  pstar = TRUE
)
}
\description{
Plot membership correlation vs gene signficance (correlation with
trait) to discover biomarkers/driver genes.
}
