% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R, R/pgx-wgcna.R
\name{wgcna.plotTopModules_multi}
\alias{wgcna.plotTopModules_multi}
\title{Plot top modules most correlated with trait for multi expression
data.}
\usage{
wgcna.plotTopModules_multi(
  multi,
  trait,
  nmax = 16,
  collapse = FALSE,
  plotlib = "base",
  setpar = TRUE
)

wgcna.plotTopModules_multi(
  multi,
  trait,
  nmax = 16,
  collapse = FALSE,
  plotlib = "base",
  setpar = TRUE
)
}
\description{
Plot top modules most correlated with trait for multi expression
data.

Plot top modules most correlated with trait for multi expression
data.
}
\seealso{
WGCNAplus::plotModuleScores
}
