% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.scaleTOMs}
\alias{wgcna.scaleTOMs}
\title{Scale a list of TOM matrices so that the quantiles (default p=0.95)
are equal after scaling with respect to the first TOM matrix.}
\usage{
wgcna.scaleTOMs(TOMs, scaleP = 0.95)
}
\description{
Scale a list of TOM matrices so that the quantiles (default p=0.95)
are equal after scaling with respect to the first TOM matrix.
}
