% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.tomclust}
\alias{wgcna.tomclust}
\title{Wrapper to hclust from matrix using default WGCNA parameters.}
\usage{
wgcna.tomclust(X, power = 6)
}
\description{
Wrapper to hclust from matrix using default WGCNA parameters.
}
