% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wmfcna.compute}
\alias{wmfcna.compute}
\title{WMFCNA: weighted multi-omics feature correlation
analysis... :). Essentially a WGCNA redo but redone-simpler-faster.}
\usage{
wmfcna.compute(
  X,
  y = NULL,
  k = 20,
  power = 1,
  vex = 1,
  ewidth = 1,
  min.cor = 0.7
)
}
\description{
WMFCNA: weighted multi-omics feature correlation
analysis... :). Essentially a WGCNA redo but redone-simpler-faster.
}
