% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{write.cls}
\alias{write.cls}
\title{Write data to a CLS file}
\usage{
write.cls(y, file, name = "")
}
\arguments{
\item{y}{The vector of class labels or clustering assignments.}

\item{file}{The file path to save the CLS file.}

\item{name}{(Optional) The name of the CLS file.}
}
\value{
None
}
\description{
This function writes data to a CLS (CLustering Solutions) file format.
The CLS format is commonly used to store class labels or clustering assignments.
}
