% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{write.gct}
\alias{write.gct}
\title{Write data to a GCT file}
\usage{
write.gct(X, file)
}
\arguments{
\item{X}{The data matrix or data frame to be written to the GCT file.
Rows represent genes and columns represent samples.}

\item{file}{The file path to save the GCT file.}
}
\value{
None
}
\description{
This function writes data to a GCT (Gene Cluster Text) file format.
The GCT format is commonly used to store gene expression data in a tabular format.
}
