% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-image.R
\name{ai.create_image_openai}
\alias{ai.create_image_openai}
\title{Generate image with OpenAI's dallE. Note the limitation of the
prompt of about 1000 characters.}
\usage{
ai.create_image_openai(
  prompt,
  model = NULL,
  size = c("512x512", "1024x1024", "256x256")[1],
  format = c("file", "base64", "raw"),
  filename = "image.png",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  base_url = "https://api.openai.com/v1",
  user = NULL,
  organization = NULL
)
}
\description{
Generate image with OpenAI's dallE. Note the limitation of the
prompt of about 1000 characters.
}
