% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{alias2hugo}
\alias{alias2hugo}
\title{Convert any gene alias to official gene symbol}
\usage{
alias2hugo(s, org = NULL, na.orig = TRUE)
}
\description{
Convert any gene alias to official gene symbol
}
