% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{annotate_methylomics}
\alias{annotate_methylomics}
\title{Handle entire annotation for 450K+850K methylomics.}
\usage{
annotate_methylomics(
  organism = "Human",
  probes = probes,
  meth_type = "450K array"
)
}
\arguments{
\item{organism}{Strictly human (for now).}

\item{probes}{450K or 850K array methylation probes.}
}
\value{
Ann. dframe: feature;symbol;gene_name;gene_title;chr;source;position;uniprot;
}
\description{
Handle entire annotation for 450K+850K methylomics.
}
