% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{annotate_phospho_residue}
\alias{annotate_phospho_residue}
\title{Annotate phosphosite with residue symbol. Feature names must be of
form 'uniprot_position'.}
\usage{
annotate_phospho_residue(features, detect.only = FALSE)
}
\description{
Annotate phosphosite with residue symbol. Feature names must be of
form 'uniprot_position'.
}
