% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{check_probetype}
\alias{check_probetype}
\title{Check if probes can be detected by Orthogene or AnnotHub/OrgDb
annotation engines.}
\usage{
check_probetype(organism, probes, verbose = 1)
}
\description{
export
}
