% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{clean_gmt}
\alias{clean_gmt}
\title{Clean gmt file function}
\usage{
clean_gmt(gmt.all, gmt.db)
}
\arguments{
\item{gmt.all}{a list of gmt files}

\item{gmt.db}{a vector of gmt database names}
}
\description{
Clean gmt file function
}
