% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{clean_probe_names}
\alias{clean_probe_names}
\title{Cleanup probe names from postfixes or version numbers}
\usage{
clean_probe_names(probes, sep = "_.-")
}
\description{
Cleanup probe names from postfixes or version numbers
}
