% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{clustermap}
\alias{clustermap}
\title{Create an annotated clustered heatmap}
\usage{
clustermap(
  x,
  nc = 6,
  nr = 6,
  na = 4,
  q = 0.8,
  p = 2,
  method = "multidist",
  nca = 24,
  col.annot = NULL,
  row.annot = NULL,
  plot = TRUE,
  na.fill = FALSE,
  nrlab = nrow(x),
  nclab = ncol(x),
  labrow = rownames(x),
  labcol = colnames(x),
  ...
)
}
\arguments{
\item{x}{Numeric data matrix to visualize}

\item{nc}{Number of sample clusters}

\item{nr}{Number of gene clusters}

\item{na}{value}

\item{q}{value}

\item{p}{value}

\item{method}{Distance metric for clustering (pearson, euclidean, etc)}

\item{nca}{value}

\item{col.annot}{Annotations for samples}

\item{row.annot}{Annotations for genes}

\item{plot}{value}

\item{na.fill}{value}

\item{nrlab}{value}

\item{nclab}{value}

\item{labrow}{value}

\item{labcol}{value}

\item{...}{value}
}
\value{
A heatmap grob object if plot = FALSE; otherwise NULL
}
\description{
Create an annotated clustered heatmap
}
\details{
This function clusters the rows and columns of a data matrix,
calculates cluster means, and visualizes the result as an annotated heatmap.

It performs soft clustering, where each value is a mixture of the cluster mean
and original value. Side color bars are added from sample and gene annotations.
}
\examples{
\dontrun{
mat <- matrix(rnorm(200), 20, 10)
amat <- matrix(sample(letters, 200, replace=TRUE), 20, 10)
clustermap(mat, col.annot = amat)
}

}
\seealso{
\code{\link[gplots]{heatmap.2}} for the underlying heatmap functions used
}
