% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{clustermap}
\alias{clustermap}
\title{Title}
\usage{
clustermap(
  x,
  nc = 6,
  nr = 6,
  na = 4,
  q = 0.8,
  p = 2,
  method = "multidist",
  nca = 24,
  col.annot = NULL,
  row.annot = NULL,
  plot = TRUE,
  na.fill = FALSE,
  nrlab = nrow(x),
  nclab = ncol(x),
  labrow = rownames(x),
  labcol = colnames(x),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
