% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{collapse_by_humansymbol}
\alias{collapse_by_humansymbol}
\title{Rename features names of object to available human symbol by
human_ortholog or other 'human-like' uppercased annotation
columns. WARNING: does not necessarily keep original length.}
\usage{
collapse_by_humansymbol(obj, annot)
}
\description{
Rename features names of object to available human symbol by
human_ortholog or other 'human-like' uppercased annotation
columns. WARNING: does not necessarily keep original length.
}
