% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{collapse_by_humansymbol}
\alias{collapse_by_humansymbol}
\title{Collapse object rownames/names to human symbol. Warning this function
does not maintain the original dimensions/length of object.}
\usage{
collapse_by_humansymbol(obj, annot)
}
\description{
Collapse object rownames/names to human symbol. Warning this function
does not maintain the original dimensions/length of object.
}
