% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{combine_feature_names}
\alias{combine_feature_names}
\title{Create new feature name by concatenating some columns of input
annotation table. Make all feature names unique.}
\usage{
combine_feature_names(annot, target)
}
\arguments{
\item{annot}{some annotation dataframe}

\item{target}{vector of character. e.g. c("feature","_","symbol")

export}
}
\description{
Create new feature name by concatenating some columns of input
annotation table. Make all feature names unique.
}
