% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-extra.R
\name{compute_deconvolution}
\alias{compute_deconvolution}
\title{Compute Deconvolution}
\usage{
compute_deconvolution(pgx, rna.counts = pgx$counts, full = FALSE)
}
\arguments{
\item{pgx}{An object containing the input data for analysis.}

\item{rna.counts}{A matrix or data frame of RNA expression counts. Defaults to the counts in the input object.}

\item{full}{A logical value indicating whether to use the full set of reference matrices and methods (TRUE),
or a subset of faster methods and references (FALSE).}
}
\value{
An updated object with deconvolution results.
}
\description{
This function performs deconvolution analysis on the input RNA expression data using various reference matrices
and methods. It estimates the abundance of different cell types or tissue components present in the data.
}
\examples{
\dontrun{
probes <- rownames(playbase::COUNTS)
mart <- biomaRt::useMart(biomart = "ensembl", dataset = species)
annot_table <- ngs.getGeneAnnotation(rownames(counts),
                                   probe_type = NULL,
                                   mart = mart)
deconv <- compute_deconvolution(probes, annot_table)
}
}
