% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-extra.R
\name{compute_deconvolution}
\alias{compute_deconvolution}
\title{Title}
\usage{
compute_deconvolution(ngs, rna.counts = ngs$counts, full = FALSE)
}
\arguments{
\item{ngs}{value}

\item{rna.counts}{value}

\item{full}{value}
}
\description{
Title
}
