% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-extra.R
\name{compute_drugSensitivityEnrichment}
\alias{compute_drugSensitivityEnrichment}
\title{Compute Drug Sensitivity Enrichment}
\usage{
compute_drugSensitivityEnrichment(pgx, libx.dir = NULL)
}
\arguments{
\item{pgx}{An object containing the input data for analysis.}

\item{libx.dir}{The directory path where the drug sensitivity databases are located.}
}
\value{
An updated object with drug sensitivity enrichment results.
}
\description{
This function performs drug sensitivity enrichment analysis based on the input RNA expression data.
It uses drug sensitivity databases to compute enrichment scores and attach the results to the input object.
The drug sensitivity databases are located in the specified directory (libx.dir).
}
