% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-extra.R
\name{compute_extra}
\alias{compute_extra}
\title{Title}
\usage{
compute_extra(
  ngs,
  extra = c("meta.go", "deconv", "infer", "drugs", "connectivity", "wordcloud"),
  lib.dir,
  sigdb = NULL
)
}
\arguments{
\item{sigdb}{}
}
\description{
Title
}
