% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{compute_testGenesets}
\alias{compute_testGenesets}
\title{Title}
\usage{
compute_testGenesets(
  pgx,
  max.features = 1000,
  test.methods = c("gsva", "camera", "fgsea"),
  remove.outputs = TRUE
)
}
\arguments{
\item{remove.outputs}{}
}
\description{
Title
}
