% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{convert.gmt}
\alias{convert.gmt}
\title{Convert gene sets to GMT format}
\usage{
convert.gmt(genes, gs_name)
}
\arguments{
\item{genes}{A character vector of gene names.}

\item{gs_name}{A character vector of gene set names corresponding to the genes.}
}
\value{
A list of gene sets in GMT format, where each gene set is represented as a character vector of gene names.
}
\description{
This function converts a set of genes and their corresponding gene set names to the GMT (Gene Matrix Transposed) format.
The GMT format is commonly used to store gene sets or gene annotations.
}
