% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{convert_probe_to_chebi}
\alias{convert_probe_to_chebi}
\title{convert IDs to CHEBI using base R functions}
\usage{
convert_probe_to_chebi(probes, probe_type)
}
\description{
convert IDs to CHEBI using base R functions
}
