% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{convert_probetype}
\alias{convert_probetype}
\title{Convert probetype unsing annothub}
\usage{
convert_probetype(
  organism,
  probes,
  target_id,
  from_id = NULL,
  datatype = NULL,
  orgdb = NULL,
  verbose = TRUE
)
}
\description{
Convert probetype unsing annothub
}
