% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_core.R
\name{create_pgx}
\alias{create_pgx}
\title{Create a pgx object}
\usage{
create_pgx(
  counts,
  samples,
  contrasts,
  X = NULL,
  is.logx = NULL,
  batch.correct = TRUE,
  auto.scale = TRUE,
  filter.genes = TRUE,
  prune.samples = FALSE,
  only.known = TRUE,
  only.hugo = TRUE,
  convert.hugo = TRUE,
  do.cluster = TRUE,
  cluster.contrasts = FALSE,
  do.clustergenes = TRUE,
  only.proteincoding = TRUE
)
}
\arguments{
\item{counts}{dataframe. The counts of the data}

\item{samples}{dataframe. The samples of the data}

\item{contrasts}{dataframe. The contrasts of the data}
}
\value{
a list (pgx object)
}
\description{
This function creates a pgx object (the core object
used in playbase) from an input of counts, samples, and
contrasts.
}
\details{
The pgx object is a list containing the following elements:
 - name
 - date
 - datatype
 - description
 - samples
 - counts
 - contrasts
 - X
 - total_counts
 - counts_multiplier
 - genes
 - and others that are dynamic / undocumented:
   - tsne2d
   - tsne3d
   - cluster
   - cluster.genes

This function relies on the following internal functions:
 - .probe_to_symbol (probe2symbol)
 - .get_gene_annotation (ngs.getGeneAnnotation)
 - .compute_cellcycle_gender (compute.cellcycle.gender)
 - .cluster_samples2  (pgx.clusterSamples2)
 - .find_louvain_clusters (pgx.findLouvainClusters)
 - .make_direct_contrasts (makeDirectContrasts)
 - .contrast_as_labels (contrastAsLabels)
 - .log_cpm (logCPM)
 - .cluster_genes (pgx.clusterGenes)

This functoin performs the following actions in order on the data:
 - unknown
}
\examples{
counts <- read_counts(example_file("counts.csv"))
samples <- read_samples(example_file("samples.csv"))
contrasts <- read_contrasts(example_file("contrasts.csv"))
## takes too long
# my_pgx <- create_pgx(counts, samples, contrasts)
}
