% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{createSparseGenesetMatrix}
\alias{createSparseGenesetMatrix}
\title{Build sparse gene set matrix}
\usage{
createSparseGenesetMatrix(
  gmt.all,
  min.geneset.size = 15,
  max.geneset.size = 500
)
}
\arguments{
\item{gmt.all}{a list of gmt files}

\item{gmt.db}{a vector of gmt database names}
}
\description{
Build sparse gene set matrix
}
