% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{detect_probetype.ANNOTHUB}
\alias{detect_probetype.ANNOTHUB}
\title{Detect probe type from probe set}
\usage{
detect_probetype.ANNOTHUB(organism, probes, ah = NULL, nprobe = 100)
}
\description{
Detect probe type from probe set
}
