% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{detect_probetype}
\alias{detect_probetype}
\title{Detect probe type from probe set}
\usage{
detect_probetype(
  organism,
  probes,
  orgdb = NULL,
  nprobe = 1000,
  use.ah = NULL,
  datatype = NULL,
  probe_type = NULL
)
}
\description{
Detect probe type from probe set
}
