% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{detect_probetype}
\alias{detect_probetype}
\title{Detect probe type from probe set}
\usage{
detect_probetype(organism, probes, orgdb = NULL, nprobe = 100, use.ah = NULL)
}
\description{
Detect probe type from probe set
}
