% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-utils.R
\name{.clean_symbols}
\alias{.clean_symbols}
\title{Cleanup symbols names from postfixes and prefixes. Take only first
symbol. This is mostly used for symbol lookup tables that need one
clean symbol.}
\usage{
.clean_symbols(symbols)
}
\description{
Cleanup symbols names from postfixes and prefixes. Take only first
symbol. This is mostly used for symbol lookup tables that need one
clean symbol.
}
