% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{.getSpeciesTable.ANNOTHUB}
\alias{.getSpeciesTable.ANNOTHUB}
\title{Get species table in AnnotationHub/OrgDB}
\usage{
.getSpeciesTable.ANNOTHUB(ah = NULL)
}
\description{
Get species table in AnnotationHub/OrgDB
}
