% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{.query_orthologs}
\alias{.query_orthologs}
\title{Single method lookup of genes/features to human orthologs. For
multi-methods search use .convert_orthologs()}
\usage{
.query_orthologs(
  genes,
  species,
  method = c("homologene", "gprofiler", "babelgene", "gprofiler2"),
  target_species = "Human",
  batch_size = ORTHOLOG_BATCH_SIZE,
  verbose = 1
)
}
\description{
Single method lookup of genes/features to human orthologs. For
multi-methods search use .convert_orthologs()
}
