% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R, R/pgx-pheno.R
\name{expandAnnotationMatrix}
\alias{expandAnnotationMatrix}
\title{Expand annotation matrix}
\usage{
expandAnnotationMatrix(A)

expandAnnotationMatrix(A)
}
\arguments{
\item{A}{Data frame containing the annotation variables.}
}
\value{
An expanded annotation matrix with dummy variables.

An expanded annotation matrix with dummy variables.
}
\description{
Expands a phenotype annotation matrix into dummy variables.

Expands a phenotype annotation matrix into dummy variables.
}
\details{
This function takes an annotation data frame and expands any categorical variables into
dummy variables using model.matrix.

For each column, it determines if the variable is numeric or categorical.
Numeric variables are ranked. Categorical variables are expanded into dummy variables.

This function takes an annotation data frame and expands any categorical variables into
dummy variables using model.matrix.

For each column, it determines if the variable is numeric or categorical.
Numeric variables are ranked. Categorical variables are expanded into dummy variables.
}
