% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{expandPhenoMatrix}
\alias{expandPhenoMatrix}
\title{Expand phenotype matrix}
\usage{
expandPhenoMatrix(M, drop.ref = TRUE, keep.numeric = FALSE)
}
\arguments{
\item{drop.ref}{Logical indicating whether to drop the reference level for each factor.}

\item{pheno}{Data frame containing the phenotype variables.}
}
\value{
An expanded phenotype matrix with dummy variables suitable for regression modeling.
}
\description{
Expands a phenotype data matrix into dummy variables while optionally
dropping the reference level.
}
\details{
This function takes a phenotype data matrix and expands any categorical variables into
dummy variables

For each column, it determines if the variable is numeric or categorical. Numeric variables are
dichotomized. Categorical variables are expanded into dummy variables using \code{model.matrix}.
}
