% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{filterProbes}
\alias{filterProbes}
\title{Filter probes from gene expression data}
\usage{
filterProbes(annot, genes)
}
\arguments{
\item{annot}{Gene annotation data.frame with probes as row names.}

\item{genes}{Character vector of genes to filter for.}
}
\value{
Character vector of matching probe names.
}
\description{
Filter probes from gene expression data
}
\details{
This function filters a  data.frame to return only probes matching the input vector.
It checks the rownames, symbol, and ortholog name for matches.
}
\examples{
\dontrun{
annot <- pgx$genes
genes <- c("FOO", "BAR")
probes <- filterProbes(annot, genes)
}
}
