% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbase-data.R
\docType{data}
\name{GENES_TABLE}
\alias{GENES_TABLE}
\title{Table for species ID conversin conversion from scientific name to Ensembl db ID}
\format{
## `SPECIES_TABLE`
data.frame with columns "dataset"	"description"	"version"	"species_name"

A data frame with 6826 rows and 10 columns:
\describe{
  \item{symbol}{Description of column1}
  \item{gene_title}{Description of column2}
  \item{chr}{Description of column2}
  \item{pos}{Description of column2}
  \item{tx_len}{Description of column2}
  \item{map}{Description of column2}
  \item{gene_name}{Description of column2}
  \item{feature}{Description of column2}
}
}
\source{
BiomaRt

Proteome profiles of activated vs resting human naive T cells at different times (Geiger et al., Cell 2016).
}
\usage{
GENES_TABLE

GENES_TABLE
}
\description{
A dataset containing the gene annotation table for the example dataset
}
\keyword{datasets}
