% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getExampleFeatures}
\alias{getExampleFeatures}
\title{Return n example features (symbols) for given organism}
\usage{
getExampleFeatures(organism, n = 20, db = c("gprofiler", "orgdb"))
}
\description{
Return n example features (symbols) for given organism
}
