% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getGeneAnnotation.GPROFILER}
\alias{getGeneAnnotation.GPROFILER}
\title{Annotate using gprofiler. This seems much faster than
AnnotHub. There are about 700 species supported. Online connection
to server is needed but we are already using remote AnnotHub and
gprofiler for ortholog matching. The advantage is that probe type
detection is not needed because orthogene seems to detect is
automatically.}
\usage{
getGeneAnnotation.GPROFILER(organism, probes, verbose = TRUE)
}
\description{
Annotate using gprofiler. This seems much faster than
AnnotHub. There are about 700 species supported. Online connection
to server is needed but we are already using remote AnnotHub and
gprofiler for ortholog matching. The advantage is that probe type
detection is not needed because orthogene seems to detect is
automatically.
}
