% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getGeneAnnotation}
\alias{getGeneAnnotation}
\title{Get gene annotation data using annothub or orthogene.}
\usage{
getGeneAnnotation(
  organism,
  probes,
  is.phospho = FALSE,
  use.ah = NULL,
  verbose = TRUE,
  methods = c("annothub", "gprofiler")
)
}
\description{
Get gene annotation data using annothub or orthogene.
}
