% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{getGseaOutput}
\alias{getGseaOutput}
\title{Title}
\usage{
getGseaOutput(path, i = 1, raster.png = FALSE)
}
\arguments{
\item{path}{value}

\item{i}{value}

\item{raster.png}{value}
}
\description{
Title
}
