% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{getGseaOutput}
\alias{getGseaOutput}
\title{Retrieve GSEA Output}
\usage{
getGseaOutput(
  path = "../analysis_v1b/output_GSEA/Th17_mut_2h_VS_mut_ut",
  i = 1,
  raster.png = FALSE
)
}
\arguments{
\item{path}{The path to the GSEA output directory.}

\item{i}{The index of the GSEA output folder to retrieve. Default is 1.}

\item{raster.png}{Logical value indicating whether to convert PNG images to raster format using grid::rasterGrob. Default is FALSE.}
}
\value{
A list containing the retrieved GSEA output files and data.
}
\description{
This function retrieves the output files and data generated from a GSEA analysis.
}
