% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{getGSETS_playbase.SAVE}
\alias{getGSETS_playbase.SAVE}
\title{Note: this function needs to be refactored since lib.dir will not exist}
\usage{
getGSETS_playbase.SAVE(
  pattern,
  lib.dir,
  custom_families_file = "custom-families.gmt"
)
}
\description{
Note: this function needs to be refactored since lib.dir will not exist
}
