% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{getHumanOrtholog}
\alias{getHumanOrtholog}
\title{Get human ortholog from given symbols of organism by using
  orthogene package. This package needs internet connection.}
\usage{
getHumanOrtholog(organism, symbols)
}
\description{
Get human ortholog from given symbols of organism by using
  orthogene package. This package needs internet connection.
}
