% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getMultiOmicsProbeAnnotation}
\alias{getMultiOmicsProbeAnnotation}
\title{Annotate multi-omics probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.}
\usage{
getMultiOmicsProbeAnnotation(organism, probes)
}
\description{
Annotate multi-omics probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.
}
