% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot-go.R
\name{getOrganismGO}
\alias{getOrganismGO}
\title{Get GO gene sets for organism directly from
AnnotationHub/OrgDB. Restrict to genes as background.}
\usage{
getOrganismGO(
  organism,
  features = NULL,
  minsize = 3,
  batch_size = 2000,
  db = c("annothub", "gprofiler"),
  include_iea = TRUE,
  symbol.annot = NULL
)
}
\description{
Get GO gene sets for organism directly from
AnnotationHub/OrgDB. Restrict to genes as background.
}
