% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{getOrganismGO}
\alias{getOrganismGO}
\title{Get GO gene sets for organism directly from
AnnotationHub/OrgDB. Restrict to genes as background.}
\usage{
getOrganismGO(organism, genes, ah = NULL)
}
\description{
export
}
