% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getOrganismGO}
\alias{getOrganismGO}
\title{Get GO gene sets for organism directly from
AnnotationHub/OrgDB. Restrict to genes as background.}
\usage{
getOrganismGO(organism, use.ah = NULL, orgdb = NULL)
}
\description{
Get GO gene sets for organism directly from
AnnotationHub/OrgDB. Restrict to genes as background.
}
