% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{getPlaydataGMT}
\alias{getPlaydataGMT}
\title{Get merged multiomics geneset x feature matrix. Combines geneset
and metabolite sets.}
\usage{
getPlaydataGMT()
}
\description{
Get merged multiomics geneset x feature matrix. Combines geneset
and metabolite sets.
}
