% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{getProbeAnnotation}
\alias{getProbeAnnotation}
\title{Convert multi-omics probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.}
\usage{
getProbeAnnotation(organism, probes, datatype, annot_table = NULL)
}
\description{
Convert multi-omics probetype. Probe names *must  be prefixed with
data type unless classical transcriptomics/proteomics.
}
