% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gmt2mat}
\alias{gmt2mat}
\title{Title}
\usage{
gmt2mat(
  gmt,
  max.genes = -1,
  ntop = -1,
  sparse = TRUE,
  bg = NULL,
  use.multicore = TRUE
)
}
\arguments{
\item{use.multicore}{}
}
\description{
Title
}
