% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.barplot}
\alias{gsea.barplot}
\title{Title}
\usage{
gsea.barplot(
  scores,
  names = NULL,
  xlab = "score",
  xlim = NULL,
  cex.text = 1,
  main = "enrichment",
  n = 16
)
}
\arguments{
\item{n}{}
}
\description{
Title
}
