% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.barplot}
\alias{gsea.barplot}
\title{Title}
\usage{
gsea.barplot(
  scores,
  names = NULL,
  xlab = "score",
  xlim = NULL,
  cex.text = 1,
  main = "enrichment",
  n = 16
)
}
\arguments{
\item{scores}{value}

\item{names}{value}

\item{xlab}{value}

\item{xlim}{value}

\item{cex.text}{value}

\item{main}{value}

\item{n}{value}
}
\description{
Title
}
