% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.enplot}
\alias{gsea.enplot}
\title{Title}
\usage{
gsea.enplot(
  rnk,
  gset,
  names = NULL,
  main = NULL,
  decreasing = TRUE,
  cex = 1,
  cex.main = 0.9,
  len.main = 40,
  lab.line = c(0.8, 2),
  cex.lab = 0.8,
  main.line = 0.3,
  xlab = "Rank in ordered dataset",
  res = 1200,
  ylab = "Rank metric"
)
}
\arguments{
\item{rnk}{value}

\item{gset}{value}

\item{names}{value}

\item{main}{value}

\item{decreasing}{value}

\item{cex}{value}

\item{cex.main}{value}

\item{len.main}{value}

\item{lab.line}{value}

\item{cex.lab}{value}

\item{main.line}{value}

\item{xlab}{value}

\item{res}{value}

\item{ylab}{value}
}
\description{
Title
}
