% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.enplot.UPDN}
\alias{gsea.enplot.UPDN}
\title{GSEA Enrichment Plot (Up/Down Genes)}
\usage{
gsea.enplot.UPDN(
  rnk,
  gset.up,
  gset.dn,
  names = NULL,
  main = NULL,
  decreasing = TRUE,
  cex.main = 0.9,
  len.main = 40,
  sum.trace = TRUE,
  res = 1200,
  xlab = "Rank in ordered dataset",
  ylab = "Ranked list metric"
)
}
\arguments{
\item{rnk}{A numeric vector representing the ranked list.}

\item{gset.up}{A character vector specifying the gene set for up-regulated genes.}

\item{gset.dn}{A character vector specifying the gene set for down-regulated genes.}

\item{names}{A character vector of names for the ranked list.}

\item{main}{The main title of the plot.}

\item{decreasing}{Logical indicating whether the ranked list should be sorted in decreasing order.}

\item{cex.main}{The size of the main title text.}

\item{len.main}{The maximum number of characters for the main title before line breaks are applied.}

\item{sum.trace}{Logical indicating whether to plot the sum of the traces for up-regulated and down-regulated genes.}

\item{res}{The resolution of the plot.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}
}
\value{
a plot
}
\description{
Generates an enrichment plot for GSEA analysis with separate traces for up-regulated and down-regulated genes.
}
