% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.enplot.UPDN}
\alias{gsea.enplot.UPDN}
\title{Title}
\usage{
gsea.enplot.UPDN(
  rnk,
  gset.up,
  gset.dn,
  names = NULL,
  main = NULL,
  decreasing = TRUE,
  cex.main = 0.9,
  len.main = 40,
  sum.trace = TRUE,
  res = 1200,
  xlab = "Rank in ordered dataset",
  ylab = "Ranked list metric"
)
}
\arguments{
\item{rnk}{value}

\item{gset.up}{value}

\item{gset.dn}{value}

\item{names}{value}

\item{main}{value}

\item{decreasing}{value}

\item{cex.main}{value}

\item{len.main}{value}

\item{sum.trace}{value}

\item{res}{value}

\item{xlab}{value}

\item{ylab}{value}
}
\description{
Title
}
