% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.fitAllContrasts}
\alias{gsea.fitAllContrasts}
\title{Perform GSEA analysis for all contrasts}
\usage{
gsea.fitAllContrasts(
  X,
  gmt,
  design,
  contr.matrix,
  output_dir,
  set.min = 15,
  set.max = 500,
  fdr = 0.25,
  skip.done = TRUE
)
}
\arguments{
\item{X}{The expression matrix.}

\item{gmt}{The gene set collection in GMT format.}

\item{design}{The design matrix.}

\item{contr.matrix}{The contrast matrix.}

\item{output_dir}{The directory to save the GSEA results.}

\item{set.min}{The minimum size of gene sets to consider in the analysis.}

\item{set.max}{The maximum size of gene sets to consider in the analysis.}

\item{fdr}{The false discovery rate (FDR) threshold.}

\item{skip.done}{Logical value indicating whether to skip the contrasts that have already been analyzed and have existing results in the output directory.}
}
\value{
A list of GSEA analysis results for each contrast.
}
\description{
This function performs GSEA (Gene Set Enrichment Analysis) analysis for all the contrasts specified in the design matrix.
It iterates over the contrasts, runs GSEA for each contrast, and saves the results in the specified output directory.
}
