% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.ftplot}
\alias{gsea.ftplot}
\title{Title}
\usage{
gsea.ftplot(
  x,
  rft = "var",
  cft = "var",
  rmax = nrow(x),
  cmax = ncol(x),
  main = "",
  cex.main = 1.5,
  xnames = NULL,
  ynames = NULL,
  cft.lab = "col metric",
  rft.lab = "row metric",
  cexRow = 0.65,
  cexCol = 0.65,
  scale = "rows",
  rsort = "hclust",
  csort = "hclust",
  p = 1,
  mar = c(8, 4, 4, 2),
  layout.reset = TRUE,
  sort.decreasing = FALSE,
  col = gplots::bluered(64),
  legend = NULL,
  cex.legend = 1,
  layout.widths = c(0.15, 0.15, 1),
  layout.heights = c(0.15, 0.15, 1)
)
}
\arguments{
\item{x}{value}

\item{rft}{value}

\item{cft}{value}

\item{rmax}{value}

\item{cmax}{value}

\item{main}{value}

\item{cex.main}{value}

\item{xnames}{value}

\item{ynames}{value}

\item{cft.lab}{value}

\item{rft.lab}{value}

\item{cexRow}{value}

\item{cexCol}{value}

\item{scale}{value}

\item{rsort}{value}

\item{csort}{value}

\item{p}{value}

\item{mar}{value}

\item{layout.reset}{value}

\item{sort.decreasing}{value}

\item{col}{value}

\item{legend}{value}

\item{cex.legend}{value}

\item{layout.widths}{value}

\item{layout.heights}{value}
}
\description{
Title
}
