% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.LeadingEdgeAnalysis}
\alias{gsea.LeadingEdgeAnalysis}
\title{Title}
\usage{
gsea.LeadingEdgeAnalysis(
  output.dir,
  ntop = 100,
  gsea.program = "/opt/GSEA/gsea-3.0.jar",
  xmx = 10
)
}
\arguments{
\item{xmx}{}
}
\description{
Title
}
