% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.radarplot}
\alias{gsea.radarplot}
\title{Title}
\usage{
gsea.radarplot(
  values,
  names = NULL,
  mar = c(2, 5, 3, 5) * 2,
  offset = 0.1,
  clust = TRUE,
  main = "",
  cex = 0.8,
  col = c("steelblue2", "tomato")
)
}
\arguments{
\item{col}{}
}
\description{
Title
}
