% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gsea.radarplot}
\alias{gsea.radarplot}
\title{gsea.radarplot}
\usage{
gsea.radarplot(
  values,
  names = NULL,
  mar = c(2, 5, 3, 5) * 2,
  offset = 0.1,
  clust = TRUE,
  main = "",
  cex = 0.8,
  col = c("steelblue2", "tomato")
)
}
\arguments{
\item{values}{A numeric vector of values.}

\item{names}{A character vector of names for the values (optional).}

\item{mar}{A numeric vector of margins (default: c(2, 5, 3, 5) * 2).}

\item{offset}{A numeric value specifying the offset (default: 0.1).}

\item{clust}{A logical value indicating whether clustering should be performed (default: TRUE).}

\item{main}{A character string specifying the plot title (default: "").}

\item{cex}{A numeric value specifying the character expansion factor (default: 0.8).}

\item{col}{A character vector of colors for the radar plot (default: c("steelblue2", "tomato")).}
}
\value{
a plot
}
\description{
A function to generate a radar plot for visualization of values.
}
