% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gseaLeadingEdgeHeatmap}
\alias{gseaLeadingEdgeHeatmap}
\title{Title}
\usage{
gseaLeadingEdgeHeatmap(
  gsea,
  maxrow = 60,
  maxcol = 60,
  gsets = NULL,
  render = "gx.heatmap",
  info.text = TRUE
)
}
\arguments{
\item{gsea}{value}

\item{maxrow}{value}

\item{maxcol}{value}

\item{gsets}{value}

\item{render}{value}

\item{info.text}{value}
}
\description{
Title
}
