% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{gseaPlotEnplots}
\alias{gseaPlotEnplots}
\title{Title}
\usage{
gseaPlotEnplots(gsea, gsets = NULL, ncol = 5)
}
\arguments{
\item{gsea}{value}

\item{gsets}{value}

\item{ncol}{value}
}
\description{
Title
}
