% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-fisher.r
\name{gset.fisher}
\alias{gset.fisher}
\title{Title}
\usage{
gset.fisher(
  genes,
  genesets,
  background = NULL,
  fdr = 0.05,
  mc = TRUE,
  sort.by = "zratio",
  nmin = 3,
  min.genes = 15,
  max.genes = 500,
  method = "fast.fisher",
  check.background = TRUE,
  common.genes = TRUE,
  verbose = 1
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
