% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-fisher.r
\name{gset.fisher2}
\alias{gset.fisher2}
\title{Title}
\usage{
gset.fisher2(
  genes.up,
  genes.dn,
  genesets,
  background = NULL,
  fdr = 0.05,
  mc = TRUE,
  sort.by = "zratio",
  nmin = 3,
  verbose = 1,
  min.genes = 15,
  max.genes = 500,
  method = "fast.fisher",
  check.background = TRUE,
  common.genes = TRUE
)
}
\arguments{
\item{common.genes}{}
}
\description{
Title
}
