% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-meta.r
\name{gset.fitContrastsWithAllMethods}
\alias{gset.fitContrastsWithAllMethods}
\title{Fit Contrasts with All Gene Set Enrichment Methods}
\usage{
gset.fitContrastsWithAllMethods(
  gmt,
  X,
  impX,
  Y,
  G,
  design,
  contr.matrix,
  methods,
  mc.threads = 1,
  mc.cores = NULL,
  batch.correct = TRUE
)
}
\arguments{
\item{gmt}{The gene set matrix.}

\item{X}{The gene expression matrix.}

\item{Y}{The phenotype data matrix.}

\item{G}{The gene annotation matrix.}

\item{design}{The experimental design matrix.}

\item{contr.matrix}{The contrast matrix.}

\item{methods}{A character vector specifying the gene set enrichment methods to use.}

\item{mc.threads}{The number of threads to use for parallel computing.}

\item{mc.cores}{The number of CPU cores to use for parallel computing.}

\item{batch.correct}{Logical indicating whether to correct for batch effects.}
}
\value{
A list containing the results of the gene set enrichment analyses for each method.
}
\description{
This function fits contrasts using multiple gene set enrichment methods, such as fisher, ssgsea, gsva, spearman, camera, fry, gsea.permPH, gsea.permGS, gseaPR, and fgsea.
}
