% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-meta.r
\name{gset.fitContrastsWithAllMethods}
\alias{gset.fitContrastsWithAllMethods}
\title{Title}
\usage{
gset.fitContrastsWithAllMethods(
  gmt,
  X,
  Y,
  G,
  design,
  contr.matrix,
  methods,
  mc.threads = 1,
  mc.cores = NULL,
  batch.correct = TRUE
)
}
\arguments{
\item{gmt}{value}

\item{X}{value}

\item{Y}{value}

\item{G}{value}

\item{design}{value}

\item{contr.matrix}{value}

\item{methods}{value}

\item{mc.threads}{value}

\item{mc.cores}{value}

\item{batch.correct}{value}
}
\description{
Title
}
