% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-meta.r
\name{gset.fitContrastsWithLIMMA}
\alias{gset.fitContrastsWithLIMMA}
\title{Fit Contrasts with LIMMA}
\usage{
gset.fitContrastsWithLIMMA(
  gsetX,
  contr.matrix,
  design,
  trend = TRUE,
  conform.output = FALSE
)
}
\arguments{
\item{gsetX}{A gene expression dataset.}

\item{contr.matrix}{Contrast matrix specifying the contrasts of interest.}

\item{design}{Design matrix specifying the experimental design. Default is \code{NULL}.}

\item{trend}{Logical indicating whether to use trend estimation in the empirical Bayes smoothing. Default is \code{TRUE}.}

\item{conform.output}{Logical indicating whether to conform the output table columns. Default is \code{FALSE}.}
}
\value{
A list of tables containing the results of statistical analysis for each contrast.
}
\description{
This function fits contrasts using the LIMMA package and performs statistical analysis on gene expression data.
}
