% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-meta.r
\name{gset.fitContrastsWithLIMMA}
\alias{gset.fitContrastsWithLIMMA}
\title{Title}
\usage{
gset.fitContrastsWithLIMMA(
  gsetX,
  contr.matrix,
  design,
  trend = TRUE,
  conform.output = FALSE
)
}
\arguments{
\item{conform.output}{}
}
\description{
Title
}
