% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-plot.r
\name{gx.barplot}
\alias{gx.barplot}
\title{Title}
\usage{
gx.barplot(
  x,
  main = "",
  cex.main = 1.2,
  cex.names = 0.85,
  cex.legend = 0.9,
  srt = 0,
  xlab = "",
  ylab = "",
  group = NULL,
  group.names = NULL,
  bar.names = NULL,
  voff = 2,
  legend = TRUE
)
}
\arguments{
\item{x}{value}

\item{main}{value}

\item{cex.main}{value}

\item{cex.names}{value}

\item{cex.legend}{value}

\item{srt}{value}

\item{xlab}{value}

\item{ylab}{value}

\item{group}{value}

\item{group.names}{value}

\item{bar.names}{value}

\item{voff}{value}

\item{legend}{value}
}
\description{
Title
}
