% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{gx.heatmap}
\alias{gx.heatmap}
\title{Title}
\usage{
gx.heatmap(
  gx,
  values = NULL,
  clust.method = "ward.D2",
  dist.method = "pearson",
  col.dist.method = "euclidean",
  plot.method = "heatmap.2",
  col = colorRampPalette(c("royalblue3", "grey90", "indianred3"))(64),
  softmax = FALSE,
  scale = "row",
  verbose = 1,
  symm = FALSE,
  col.annot = NULL,
  row.annot = NULL,
  annot.ht = 1,
  nmax = 1000,
  cmax = NULL,
  show_colnames = TRUE,
  indent.names = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
