% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-limma.r
\name{gx.limma.paired}
\alias{gx.limma.paired}
\title{Title}
\usage{
gx.limma.paired(
  X,
  pheno,
  pair,
  fdr = 0.05,
  lfc = 0.2,
  ref = REF.CLASS,
  compute.means = FALSE,
  trend = FALSE
)
}
\arguments{
\item{trend}{}
}
\description{
Title
}
