% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-limma.r
\name{gx.limma}
\alias{gx.limma}
\title{Differential expression analysis using limma}
\usage{
gx.limma(
  X,
  pheno,
  B = NULL,
  remove.na = TRUE,
  fdr = 0.05,
  compute.means = TRUE,
  lfc = 0.2,
  max.na = 0.2,
  sort.by = "FC",
  ref = c("ctrl", "ctr", "control", "ct", "dmso", "nt", "0", "0h", "0hr", "non", "no",
    "not", "neg", "negative", "ref", "veh", "vehicle", "wt", "wildtype", "untreated",
    "normal", "false", "healthy"),
  trend = FALSE,
  robust = FALSE,
  method = 1,
  f.test = FALSE,
  verbose = 1
)
}
\arguments{
\item{X}{Gene expression matrix with genes in rows and samples in columns.}

\item{pheno}{Phenotype vector or factor for samples.}

\item{B}{Optional batch covariate matrix.}

\item{remove.na}{Logical for removing samples with missing phenotype.}

\item{fdr}{FDR threshold for identifying differentially expressed genes.}

\item{compute.means}{Logical for computing group means.}

\item{lfc}{Log fold change cutoff.}

\item{max.na}{Max proportion of missing values allowed for a gene.}

\item{ref}{Character vector of possible reference levels.}

\item{trend}{Logical for fitting a trend model.}

\item{f.test}{Logical. Run f-test when phenotype has >2 levels.}

\item{verbose}{Verbosity level.}
}
\value{
Data frame with limma results.
}
\description{
Performs differential expression analysis on a gene expression matrix using limma.
Handles single sample case by duplicating sample. Auto-detects reference level.
}
\details{
Differential expression analysis using limma

Performs differential expression analysis on the gene expression matrix \code{X} using limma.
It handles filtering, model design matrices, and output formatting.
The phenotype vector \code{pheno} is used to define the linear model.
Batch effects can be accounted for by providing a \code{B} matrix.
The function auto-detects the reference level or the user can provide possible values in \code{ref}.
It returns a data frame containing the limma analysis results.
}
