% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-limma.r
\name{gx.limma}
\alias{gx.limma}
\title{Title}
\usage{
gx.limma(
  X,
  pheno,
  B = NULL,
  remove.na = TRUE,
  fdr = 0.05,
  compute.means = TRUE,
  lfc = 0.2,
  max.na = 0.2,
  ref = c("ctrl", "ctr", "control", "dmso", "nt", "0", "0h", "0hr", "non", "no", "not",
    "neg", "negative", "ref", "veh", "vehicle", "wt", "wildtype", "untreated", "normal",
    "false", "healthy"),
  trend = FALSE,
  verbose = 1
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
