% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-limma.r
\name{gx.limma.two.factorial}
\alias{gx.limma.two.factorial}
\title{Title}
\usage{
gx.limma.two.factorial(
  X,
  factors,
  fdr = 0.05,
  lfc = 0.2,
  trend = FALSE,
  ref = REF.CLASS,
  compute.means = TRUE
)
}
\arguments{
\item{compute.means}{}
}
\description{
Title
}
