% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-limma.r
\name{gx.limma.two.factorial}
\alias{gx.limma.two.factorial}
\title{Title}
\usage{
gx.limma.two.factorial(
  X,
  factors,
  fdr = 0.05,
  lfc = 0.2,
  trend = FALSE,
  ref = c("ctrl", "ctr", "control", "dmso", "nt", "0", "0h", "0hr", "non", "no", "not",
    "neg", "negative", "ref", "veh", "vehicle", "wt", "wildtype", "untreated", "normal",
    "false", "healthy"),
  compute.means = TRUE
)
}
\arguments{
\item{compute.means}{}
}
\description{
Title
}
