% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-limma.r
\name{gx.limmaF}
\alias{gx.limmaF}
\title{Title}
\usage{
gx.limmaF(
  X,
  pheno,
  B = NULL,
  fdr = 0.05,
  compute.means = TRUE,
  lfc = 0.2,
  max.na = 0.2,
  ref = REF.CLASS,
  trend = FALSE,
  verbose = 1
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
