% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-combat.r
\name{gx.nnmcorrect}
\alias{gx.nnmcorrect}
\title{Title}
\usage{
gx.nnmcorrect(
  X,
  y,
  use.design = TRUE,
  dist.method = "cor",
  center.x = TRUE,
  center.m = TRUE,
  sdtop = 1000,
  replace = FALSE
)
}
\arguments{
\item{X}{value}

\item{y}{value}

\item{use.design}{value}

\item{dist.method}{value}

\item{center.x}{value}

\item{center.m}{value}

\item{sdtop}{value}

\item{replace}{value}
}
\description{
Title
}
