% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{gx.splitmap}
\alias{gx.splitmap}
\title{Title}
\usage{
gx.splitmap(
  gx,
  split = 5,
  splitx = NULL,
  clust.method = "ward.D2",
  dist.method = "euclidean",
  col.dist.method = "euclidean",
  plot.method = "heatmap.2",
  scale = "row",
  softmax = 0,
  order.groups = "clust",
  symm.scale = FALSE,
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  sort_columns = NULL,
  col.annot = NULL,
  row.annot = NULL,
  annot.ht = 3,
  nmax = 1000,
  cmax = NULL,
  main = " ",
  verbose = 1,
  denoise = 0,
  cexRow = 1,
  cexCol = 1,
  mar = c(5, 5, 5, 5),
  rownames_width = 25,
  rowlab.maxlen = 20,
  title_cex = 1.2,
  column_title_rot = 0,
  column_names_rot = 90,
  show_legend = TRUE,
  show_key = TRUE,
  zlim = NULL,
  show_rownames = nmax,
  lab.len = 80,
  key.offset = c(0.05, 1.01),
  show_colnames = NULL,
  use.nclust = FALSE
)
}
\arguments{
\item{use.nclust}{}
}
\description{
Title
}
