% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-heatmap.r
\name{gx.splitmap}
\alias{gx.splitmap}
\title{Create a split heatmap of gene expression data using
ComplexHeatmap.}
\usage{
gx.splitmap(
  gx,
  split = 5,
  splitx = NULL,
  clust.method = "ward.D2",
  dist.method = "euclidean",
  col.dist.method = "euclidean",
  plot.method = "heatmap.2",
  scale = "row",
  softmax = 0,
  order.groups = "clust",
  symm.scale = FALSE,
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  sort_columns = NULL,
  col.annot = NULL,
  row.annot = NULL,
  annot.ht = 3,
  annot.cex = 1,
  nmax = 1000,
  cmax = NULL,
  main = " ",
  verbose = 1,
  denoise = 0,
  cexRow = 1,
  cexCol = 1,
  na_col = "green",
  na_text = NULL,
  mar = c(5, 5, 5, 5),
  rownames_width = 25,
  rowlab.maxlen = 20,
  title_cex = 1.2,
  column_title_rot = 0,
  column_names_rot = 90,
  show_legend = TRUE,
  show_key = TRUE,
  zlim = NULL,
  show_rownames = nmax,
  lab.len = 80,
  key.offset = c(0.05, 1.01),
  show_colnames = NULL,
  use.nclust = FALSE,
  data = FALSE
)
}
\arguments{
\item{gx}{Gene expression matrix with genes in rows and samples in columns.}

\item{split}{Vector defining sample groups, or number of groups for clustering.}

\item{splitx}{External group assignments for samples.}

\item{clust.method}{Hierarchical clustering method for genes.}

\item{dist.method}{Distance metric for gene clustering.}

\item{col.dist.method}{Distance for sample clustering.}

\item{plot.method}{Heatmap plotting method (heatmap.2 or ComplexHeatmap).}

\item{scale}{Scaling method for the data matrix.}

\item{softmax}{value}

\item{order.groups}{value}

\item{symm.scale}{value}

\item{cluster_rows}{value}

\item{cluster_columns}{value}

\item{sort_columns}{value}

\item{col.annot}{Annotations for samples.}

\item{row.annot}{Annotations for genes.}

\item{annot.ht}{value}

\item{nmax}{Maximum number of genes to include.}

\item{cmax}{value}

\item{main}{Title for the heatmap.}

\item{verbose}{value}

\item{denoise}{value}

\item{cexRow}{value}

\item{cexCol}{value}

\item{mar}{value}

\item{rownames_width}{value}

\item{rowlab.maxlen}{value}

\item{title_cex}{value}

\item{column_title_rot}{value}

\item{column_names_rot}{value}

\item{show_legend}{value}

\item{show_key}{value}

\item{zlim}{value}

\item{show_rownames}{value}

\item{lab.len}{value}

\item{key.offset}{value}

\item{show_colnames}{value}

\item{use.nclust}{value}
}
\value{
A heatmap grob object
}
\description{
Visualize a gene expression matrix as a heatmap, with samples split into groups.
}
\details{
This function subsets the input gene expression matrix to the top variable
 genes, clusters genes and samples, and visualizes the result as a heatmap with
samples split into groups defined by \code{split} or \code{splitx}.

Many options are provided for clustering, scaling, annotations, and heatmap display
 parameters. By default it uses Euclidean distance and Ward clustering.
}
\examples{
\dontrun{
set.seed(123)
gx <- matrix(rnorm(250)^2, nrow = 25, ncol = 10)
rownames(gx) <- sample(LETTERS, 25)
colnames(gx) <- sample(letters, 10)
# Create a grouping variable

p <- gx.splitmap(
  gx,
  scale = "row",
  cluster_rows = FALSE,
  show_key = FALSE,
  cluster_columns = FALSE
)
}
}
