% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-volcano.r
\name{gx.volcanoPlot.LIMMA}
\alias{gx.volcanoPlot.LIMMA}
\title{Volcano plot from LIMMA results}
\usage{
gx.volcanoPlot.LIMMA(
  tab,
  render = "scatterD3",
  n = 1000,
  highlight = NULL,
  p.sig = 0.05,
  cex = 1,
  lab.cex = 1,
  nlab = 15,
  xlim = NULL,
  ylim = NULL,
  use.fdr = FALSE,
  use.rpkm = FALSE,
  ma.plot = FALSE,
  cex.main = 1.2,
  main = "",
  cex.axis = 1,
  axes = TRUE
)
}
\arguments{
\item{tab}{Data frame containing statistics from LIMMA analysis, with 
columns for log2 fold change, p-values, and adjusted p-values.}

\item{render}{Plotting method (either base R graphics or scatterD3).}

\item{n}{Maximum number of points to plot.}

\item{highlight}{Vector of genes to highlight.}

\item{p.sig}{P-value cutoff for significance.}

\item{cex}{Point size.}

\item{lab.cex}{Label size for highlighted genes.}

\item{nlab}{Number of top genes to label.}

\item{xlim}{Limits of the x-axis.}

\item{ylim}{Limits of the y-axis.}

\item{use.fdr}{Use FDR adjusted p-values.}

\item{use.rpkm}{Use RPKM/TPM values for x-axis.}

\item{ma.plot}{Overlay MA plot lines?}

\item{cex.main}{Title size.}

\item{main}{Plot title.}

\item{cex.axis}{Axis label size.}

\item{axes}{Display axes?}
}
\value{
A volcano plot is generated, no value is returned.
}
\description{
Generate a volcano plot from a LIMMA differential 
expression analysis result table.
}
\details{
This function takes a LIMMA result table and generates a volcano plot 
to visualize significance versus log2 fold change. Points exceeding the specified 
p-value and fold change thresholds are highlighted. The top most significant 
genes can be labeled. Both raw and FDR corrected p-values can be used.
}
\examples{
\dontrun{
n <- 100
tab <- data.frame(logFC=rnorm(n), 
                 P.Value=runif(n), 
                 adj.P.Val = p.adjust(runif(n)), 
                 AveExpr = runif(n))
z<- gx.volcanoPlot.LIMMA(tab)
z
}
}
