% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-volcano.r
\name{gx.volcanoPlot.LIMMA}
\alias{gx.volcanoPlot.LIMMA}
\title{Title}
\usage{
gx.volcanoPlot.LIMMA(
  tab,
  render = "scatterD3",
  n = 1000,
  highlight = NULL,
  p.sig = 0.05,
  cex = 1,
  lab.cex = 1,
  nlab = 15,
  xlim = NULL,
  ylim = NULL,
  use.fdr = FALSE,
  use.rpkm = FALSE,
  ma.plot = FALSE,
  cex.main = 1.2,
  main = "",
  cex.axis = 1,
  axes = TRUE
)
}
\arguments{
\item{axes}{}
}
\description{
Title
}
