% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-volcano.r
\name{gx.volcanoPlot.XY}
\alias{gx.volcanoPlot.XY}
\title{Volcano Plot}
\usage{
gx.volcanoPlot.XY(
  x,
  pv,
  gene,
  ma_plot = FALSE,
  ma = NULL,
  p.sig = 0.05,
  lfc = 1,
  render = "scatterD3",
  n = 1000,
  highlight = NULL,
  main = "",
  cex = 1,
  lab.cex = 1,
  nlab = 10,
  xlim = NULL,
  ylim = NULL,
  use.fdr = FALSE,
  xlab = "effect size",
  ylab = "significance (-log10p)",
  use.rpkm = FALSE,
  maxchar = 40,
  hi.col = "#1e60bb",
  cex.main = 1.2,
  axes = TRUE,
  cex.axis = 1
)
}
\arguments{
\item{x}{Numeric vector of effect sizes (fold changes, differences, etc).}

\item{pv}{Numeric vector of p-values.}

\item{gene}{Character vector of gene IDs corresponding to x and pv.}

\item{ma_plot}{Logical, if TRUE creates an MA-plot instead.}

\item{ma}{Matrix of log-fold changes and mean average values for MA-plot.}

\item{p.sig}{Significance threshold p-value.}

\item{lfc}{Log2 fold change threshold.}

\item{render}{Plot rendering method (base, ggplot, scatterD3).}

\item{n}{Maximum number of points to plot.}

\item{highlight}{Vector of genes to highlight.}

\item{main}{Title for the plot.}

\item{cex}{Point size.}

\item{lab.cex}{Label size for highlighted genes.}

\item{nlab}{Number of top genes to label.}

\item{xlim}{Limits of the x-axis.}

\item{ylim}{Limits of the y-axis.}

\item{use.fdr}{Use FDR instead of p-values.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{use.rpkm}{Use RPKM/FPKM values instead of raw counts.}

\item{maxchar}{Maximum characters for gene labels.}

\item{hi.col}{Color for highlighted points.}

\item{cex.main}{Size of main title.}

\item{axes}{Display axes?}

\item{cex.axis}{Size of axis labels.}
}
\value{
A volcano plot is generated, no value is returned.
}
\description{
Generate a volcano plot to visualize significance versus magnitude of change.
}
\details{
This function takes a vector of p-values and effect sizes (fold changes, differences, etc)
and generates a volcano plot to visualize significance versus magnitude of change. Points exceeding
both p-value and effect size thresholds are highlighted. The top most significant genes can be
labeled. Both raw and FDR corrected p-values can be used.
}
