% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr-graph.r
\name{hclust_graph}
\alias{hclust_graph}
\title{Hierarchical Clustering of Graph}
\usage{
hclust_graph(g, k = NULL, mc.cores = 2)
}
\arguments{
\item{g}{An igraph graph object to cluster}

\item{k}{The number of hierarchical levels. If NULL iterates until convergence.}

\item{mc.cores}{Number of cores for parallel processing. Default 2.}
}
\value{
A matrix with hierarchical clustering assignments for nodes.
}
\description{
Performs hierarchical clustering on a graph using iterative Louvain clustering.
}
\details{
Hierarchical Clustering of Graph

This function takes an igraph graph object \code{g} and performs hierarchical clustering to detect communities.
It uses iterative Louvain clustering, optimizing modularity at each level of the hierarchy.

At each iteration, it runs Louvain clustering on the communities from the previous level.
This splits up the communities into smaller sub-communities in a hierarchical fashion.

The number of levels \code{k} can be specified, otherwise it iterates until convergence.
Parallel processing with \code{mc.cores} is used to speed up the computations.

The algorithm returns a matrix containing the hierarchical clustering assignments for each node.
The columns represent the clustering at each level of the hierarchy.
}
