% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{justGCEA}
\alias{justGCEA}
\title{Title}
\usage{
justGCEA(
  X,
  sets,
  concepts,
  Y = NULL,
  fdr = 1,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  rpt.label = "analysis",
  output.dir = NULL,
  xmx = 10,
  scoring = "weighted",
  clean.files = TRUE
)
}
\arguments{
\item{X}{value}

\item{sets}{value}

\item{concepts}{value}

\item{Y}{value}

\item{fdr}{value}

\item{set.min}{value}

\item{set.max}{value}

\item{topgs}{value}

\item{nperm}{value}

\item{rpt.label}{value}

\item{output.dir}{value}

\item{xmx}{value}

\item{scoring}{value}

\item{clean.files}{value}
}
\description{
Title
}
