% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{justGSEA}
\alias{justGSEA}
\title{Title}
\usage{
justGSEA(
  X,
  gmt,
  Y = NULL,
  fdr = 1,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  rpt.label = "analysis",
  output.dir = NULL,
  xmx = 10,
  scoring = "weighted",
  sort.y = "ref.last",
  ref.type = c("0", "nt", "ref", "dmso", "wt", "dsmo"),
  clean.files = TRUE
)
}
\arguments{
\item{X}{value}

\item{gmt}{value}

\item{Y}{value}

\item{fdr}{value}

\item{set.min}{value}

\item{set.max}{value}

\item{topgs}{value}

\item{nperm}{value}

\item{rpt.label}{value}

\item{output.dir}{value}

\item{xmx}{value}

\item{scoring}{value}

\item{sort.y}{value}

\item{ref.type}{value}

\item{clean.files}{value}
}
\description{
Title
}
