% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{justGSEA}
\alias{justGSEA}
\title{Title}
\usage{
justGSEA(
  X,
  gmt,
  Y = NULL,
  fdr = 1,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  rpt.label = "analysis",
  output.dir = NULL,
  xmx = 10,
  scoring = "weighted",
  sort.y = "ref.last",
  ref.type = c("0", "nt", "ref", "dmso", "wt", "dsmo"),
  clean.files = TRUE
)
}
\arguments{
\item{clean.files}{}
}
\description{
Title
}
