% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{makeClusterContrasts}
\alias{makeClusterContrasts}
\title{Title}
\usage{
makeClusterContrasts(
  clusters,
  min.freq = 0.01,
  full = FALSE,
  by.sample = FALSE
)
}
\arguments{
\item{by.sample}{}
}
\description{
Title
}
