% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{map_probes}
\alias{map_probes}
\title{Collapse object rownames/names to human symbol. Warning this function
does not maintain the original dimensions/length of object.}
\usage{
map_probes(annot, genes, column = NULL, ignore.case = FALSE)
}
\description{
Collapse object rownames/names to human symbol. Warning this function
does not maintain the original dimensions/length of object.
}
