% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{map_probes}
\alias{map_probes}
\title{Map any feature identifier to other feature column. Warning does
not retain original match order. Default mapping to rownames.}
\usage{
map_probes(
  annot,
  genes,
  column = NULL,
  ignore.case = FALSE,
  target = "rownames"
)
}
\description{
Map any feature identifier to other feature column. Warning does
not retain original match order. Default mapping to rownames.
}
