% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-utils.R
\name{markdownToPDF}
\alias{markdownToPDF}
\title{Export markdown to PDF}
\usage{
markdownToPDF(
  text,
  file,
  tmpdir = NULL,
  engine = "pdflatex",
  font = "mathpazo",
  documentclass = "report",
  quiet = TRUE
)
}
\description{
Export markdown to PDF
}
