% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{merge_sparse_matrix}
\alias{merge_sparse_matrix}
\title{Merge two sparse matrices}
\usage{
merge_sparse_matrix(m1, m2)
}
\arguments{
\item{m1}{matrix 1 in Matrix format}

\item{m2}{matrix 2 in Matrix format}
}
\description{
Merge two sparse matrices
}
