% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{merge_sparse_matrix}
\alias{merge_sparse_matrix}
\title{Merge Sparse Matrix}
\usage{
merge_sparse_matrix(m1, m2)
}
\arguments{
\item{m1}{The first sparse matrix.}

\item{m2}{The second sparse matrix.}
}
\value{
A merged sparse matrix with combined columns and matching gene order.
}
\description{
Merges two sparse matrices by combining their columns and ensuring the same gene order.
}
