% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{mergeCpG}
\alias{mergeCpG}
\title{mergeCpG}
\usage{
mergeCpG(data, genes = NULL, collapse.by = "gene")
}
\arguments{
\item{genes}{Annotation matrix: pgx$genes. if NULL, automatically retrieved.}

\item{collapse.by}{collapse.by variable. Can be later flexibly user-defined.}

\item{X}{Matrix of methyation values. Can be M or Beta. Beta values will be used.}
}
\value{
List of 2 elements: [[1]] collapsed CpG matrix; [[2]] updated annotation dataframe.
}
\description{
Collapse CpG into genes. Use average beta value.
}
