% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-predict.R
\name{mixHivePlot}
\alias{mixHivePlot}
\alias{mixPlotLoadings}
\title{Create Hive Plot for mixed network}
\usage{
mixHivePlot(res, ngs, ct, showloops = FALSE, numlab = 6, cex = 1)

mixPlotLoadings(res, showloops = FALSE, cex = 1)
}
\arguments{
\item{res}{Input data}

\item{ngs}{NGS object containing expression data}

\item{ct}{Contrast name or column in \code{ngs$samples} to use for node coloring}

\item{showloops}{Whether to show looping edges}

\item{numlab}{Number of node labels to show}

\item{cex}{Label size scaling factor}
}
\value{
Hive plot object
}
\description{
Creates a Hive Plot to visualize a mixed gene-gene set network.
}
\details{
It extracts the network graph and formats it into a Hive Plot layout.

Node importance scores are extracted from the NGS object based on the specified \code{ct} contrast.
These scores are used to determine node sizes in the plot.

The number of node labels can be reduced by setting \code{numlab} to avoid overplotting.
}
\section{Functions}{
\itemize{
\item \code{mixPlotLoadings()}: function generates a hive plot visualization of variable loadings 
from a lmer model result object.

}}
