% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-mofa.R
\name{mofa.compute_enrichment}
\alias{mofa.compute_enrichment}
\title{Compute enrichment for MOFA factors.}
\usage{
mofa.compute_enrichment(W, G = NULL, filter = NULL, ntop = 1000)
}
\description{
Compute enrichment for MOFA factors.
}
