% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-metabolomics.R
\name{mx.convert_probe}
\alias{mx.convert_probe}
\title{convert IDs to CHEBI using base R functions}
\usage{
mx.convert_probe(probes, probe_type = NULL, target_id = "ChEBI")
}
\description{
convert IDs to CHEBI using base R functions
}
