% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-cook.r
\name{ngs.cookForDESEQ2}
\alias{ngs.cookForDESEQ2}
\title{Title}
\usage{
ngs.cookForDESEQ2(
  counts,
  samples,
  genes,
  remove.batch = TRUE,
  test = "Wald",
  prior.cpm = 0,
  filter = TRUE
)
}
\arguments{
\item{counts}{value}

\item{samples}{value}

\item{genes}{value}

\item{remove.batch}{value}

\item{test}{value}

\item{prior.cpm}{value}

\item{filter}{value}
}
\description{
Title
}
