% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-cook.r
\name{ngs.cookForEDGER}
\alias{ngs.cookForEDGER}
\title{Title}
\usage{
ngs.cookForEDGER(
  counts,
  samples = NULL,
  genes = NULL,
  normalization = "none",
  filter = TRUE,
  prior.cpm = 0,
  remove.batch = TRUE
)
}
\arguments{
\item{counts}{value}

\item{samples}{value}

\item{genes}{value}

\item{normalization}{value}

\item{filter}{value}

\item{prior.cpm}{value}

\item{remove.batch}{value}
}
\description{
Title
}
