% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitConstrastsWithDESEQ2}
\alias{ngs.fitConstrastsWithDESEQ2}
\title{Title}
\usage{
ngs.fitConstrastsWithDESEQ2(
  counts,
  group,
  contr.matrix,
  design,
  X = NULL,
  genes = NULL,
  test = "Wald",
  prune.samples = FALSE,
  conform.output = FALSE
)
}
\arguments{
\item{conform.output}{}
}
\description{
Title
}
