% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithAllMethods}
\alias{ngs.fitContrastsWithAllMethods}
\title{Title}
\usage{
ngs.fitContrastsWithAllMethods(
  counts,
  X = NULL,
  samples,
  design,
  contr.matrix,
  genes = NULL,
  prior.cpm = 1,
  prune.samples = FALSE,
  conform.output = TRUE,
  do.filter = TRUE,
  cpm.scale = 1e+06,
  remove.batch = TRUE,
  methods = c("ttest", "ttest.welch", "voom.limma", "trend.limma", "notrend.limma",
    "deseq2.wald", "deseq2.lrt", "edger.qlf", "edger.lrt"),
  correct.AveExpr = TRUE,
  custom = NULL,
  custom.name = NULL
)
}
\arguments{
\item{counts}{value}

\item{X}{value}

\item{samples}{value}

\item{design}{value}

\item{contr.matrix}{value}

\item{genes}{value}

\item{prior.cpm}{value}

\item{prune.samples}{value}

\item{conform.output}{value}

\item{do.filter}{value}

\item{cpm.scale}{value}

\item{remove.batch}{value}

\item{methods}{value}

\item{correct.AveExpr}{value}

\item{custom}{value}

\item{custom.name}{value}
}
\description{
Title
}
