% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-fit.r
\name{ngs.fitContrastsWithLIMMA}
\alias{ngs.fitContrastsWithLIMMA}
\alias{ngs.fitContrastsWithLIMMA.timeseries}
\title{ngs.fitContrastsWithLIMMA}
\usage{
ngs.fitContrastsWithLIMMA(
  X,
  contr.matrix,
  design,
  method = c("voom", "limma"),
  trend = TRUE,
  robust = TRUE,
  prune.samples = FALSE,
  conform.output = FALSE,
  plot = FALSE,
  timeseries = NULL
)

ngs.fitContrastsWithLIMMA.timeseries(
  X,
  y,
  timeseries,
  trend = TRUE,
  use.spline = NULL
)
}
\arguments{
\item{X}{log2- and normalized expression matrix}
}
\value{
limma top tables
}
\description{
ngs.fitContrastsWithAllMethods Fits contrasts using LIMMA
}
\section{Functions}{
\itemize{
\item \code{ngs.fitContrastsWithLIMMA.timeseries()}: Fits contrasts using LIMMA with no design. For time-series analysis.

}}
