% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{ngs.getGeneAnnotation}
\alias{ngs.getGeneAnnotation}
\title{Get gene annotation data}
\usage{
ngs.getGeneAnnotation(probes, probe_type = NULL, mart = NULL, verbose = TRUE)
}
\arguments{
\item{probes}{Character vector of gene/transcript identifiers to retrieve annotation for.}

\item{probe_type}{Character specifying the type of input identifiers. If NULL,
it will be automatically detected. Options are "ensembl_gene_id", "ensembl_transcript_id", etc.}

\item{mart}{BioMart object specifying the database to query.}

\item{verbose}{Logical indicating whether to print status messages.}
}
\value{
Data frame with gene annotation data for the input identifiers. Columns are:
\itemize{
  \item \code{gene_name}: Gene name
  \item \code{hgnc_symbol}: Gene symbol
  \item \code{gene_title}: Gene description
  \item \code{gene_biotype}: Gene biotype
  \item \code{chr}: Chromosome
  \item \code{pos}: Transcript start position
  \item \code{tx_len}: Transcript length
  \item \code{map}: Chromosome band
}
}
\description{
Retrieves gene annotation information from BioMart for a set of input
gene/transcript identifiers.
}
\details{
This function queries BioMart to retrieve key gene annotation data for
a set of input gene/transcript identifiers. It can detect the identifier
type automatically if not provided.
}
\examples{
\dontrun{
probes <- c("ENSG00000142192", "ENST00000288602")
mart <- biomaRt::useMart("ensembl")
result <- ngs.getGeneAnnotation(probes, mart)
head(result)
}
}
