% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-salmon.r
\name{ngs.rawSalmon}
\alias{ngs.rawSalmon}
\title{Title}
\usage{
ngs.rawSalmon(sampleTable, gencode, txi = NULL)
}
\arguments{
\item{sampleTable}{value}

\item{gencode}{value}

\item{txi}{value}
}
\description{
Title
}
