% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-salmon.r
\name{ngs.tximportSalmon}
\alias{ngs.tximportSalmon}
\title{Title}
\usage{
ngs.tximportSalmon(
  sf.files,
  count.type = "lengthScaledTPM",
  organism = "Hsapiens",
  txOut = FALSE
)
}
\arguments{
\item{sf.files}{value}

\item{count.type}{value}

\item{organism}{value}

\item{txOut}{value}
}
\description{
Title
}
