% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute2-genesets.R
\name{normalize_matrix_by_row}
\alias{normalize_matrix_by_row}
\title{Normalize Matrix by Row}
\usage{
normalize_matrix_by_row(G)
}
\arguments{
\item{G}{The matrix to be normalized.}
}
\value{
The normalized matrix.
}
\description{
Normalizes a matrix by dividing each row by the sum of its elements.
}
