% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-check.R
\name{pgx.checkPGX}
\alias{pgx.checkPGX}
\title{Check input files for pgx.computePGX}
\usage{
pgx.checkPGX(df, type = c("SAMPLES", "COUNTS", "EXPRESSION", "CONTRASTS"))
}
\arguments{
\item{df}{data.frame. The data frame corresponding to the input file as
described in Omics Playground documentation}

\item{type}{character. One of "SAMPLES", "COUNTS", "EXPRESSION", "CONTRASTS" - The type
of data file to check.}
}
\value{
a list with two elements: `checks` which contains the status of the checks, and
 `PASS` which contains the overall status of the check.
}
\description{
Check input files for pgx.computePGX
}
