% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterGenes}
\alias{pgx.clusterGenes}
\title{Title}
\usage{
pgx.clusterGenes(
  pgx,
  methods = c("pca", "tsne", "umap"),
  dims = c(2, 3),
  reduce.pca = 50,
  perplexity = 30,
  level = "gene",
  rank.tf = FALSE,
  center.rows = TRUE,
  scale.rows = FALSE,
  X = NULL,
  umap.pkg = "uwot"
)
}
\arguments{
\item{pgx}{value}

\item{methods}{value}

\item{dims}{value}

\item{reduce.pca}{value}

\item{perplexity}{value}

\item{level}{value}

\item{rank.tf}{value}

\item{center.rows}{value}

\item{scale.rows}{value}

\item{X}{value}

\item{umap.pkg}{value}
}
\description{
Title
}
