% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterGenes}
\alias{pgx.clusterGenes}
\title{Cluster genes}
\usage{
pgx.clusterGenes(
  pgx,
  methods = c("pca", "tsne", "umap"),
  dims = c(2, 3),
  reduce.pca = 50,
  perplexity = 30,
  level = "gene",
  rank.tf = FALSE,
  center.rows = TRUE,
  scale.rows = FALSE,
  X = NULL,
  umap.pkg = "uwot"
)
}
\arguments{
\item{pgx}{A PGX object containing gene expression data}

\item{dims}{Dimensions for dimensionality reduction (by default 2,and 3)}

\item{reduce.pca}{Number of PCs to reduce to for PCA}

\item{perplexity}{Perplexity parameter for tSNE}

\item{level}{Data level to cluster ("gene", "exon", "junction")}

\item{rank.tf}{Rank transform expression first? Logical}

\item{center.rows}{Center expression rows? Logical}

\item{scale.rows}{Scale expression rows? Logical}

\item{X}{Expression matrix to use instead of pgx$X}

\item{umap.pkg}{UMAP package to use ("uwot" or "umap")}

\item{method}{Character vector of clustering methods to apply (choose between "pca", "tsne", and "umap")}
}
\value{
The updated PGX object with gene clusters added
}
\description{
Clusters the genes in a PGX object using dimensionality reduction methods like PCA, tSNE and UMAP.
}
\details{
This function takes a PGX object and performs dimensionality reduction on the gene expression data,
followed by clustering in the reduced space. Methods like PCA, tSNE and UMAP are used to reduce the dimensions.

The gene expression matrix is extracted from the PGX object, optionally transformed, then embedded.
The resulting coordinates are clustered using k-means to assign genes to clusters.

The gene clusters are added to the PGX object for downstream use.
}
