% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterMatrix}
\alias{pgx.clusterMatrix}
\title{Cluster large matrices}
\usage{
pgx.clusterMatrix(
  X,
  methods = c("pca", "tsne", "umap"),
  dims = c(2, 3),
  perplexity = 30,
  reduce.sd = 1000,
  reduce.pca = 50,
  center.features = TRUE,
  scale.features = FALSE,
  find.clusters = FALSE,
  umap.pkg = "uwot"
)
}
\arguments{
\item{X}{Numeric matrix of data to cluster}

\item{methods}{Character vector of methods to use (pca, tsne, umap)}

\item{dims}{Output dimensions for dimensionality reduction}

\item{perplexity}{Perplexity parameter for tSNE and UMAP}

\item{reduce.sd}{Perform PCA if original dimensions > reduce.sd}

\item{reduce.pca}{Number of PCA components for initial reduction}

\item{center.features}{Center the rows before clustering?}

\item{scale.features}{Scale the rows before clustering?}

\item{find.clusters}{Detect optimal number of clusters?}

\item{umap.pkg}{UMAP package to use (umap or uwot)}
}
\value{
List of cluster assignments
}
\description{
Performs dimensionality reduction and clustering on large data matrices
}
\details{
Cluster large matrices


This function takes a large data matrix and reduces the dimensions using
PCA and tSNE/UMAP before clustering the rows. It is designed to handle
matrices that are too large to embed directly with tSNE or UMAP.

PCA is first used to reduce the dimensions if the original size is over
reduce.sd rows. tSNE or UMAP is then applied to further reduce the
dimensions for clustering. k-means clustering is performed on the
reduced data.

Multiple parameters allow customization of the methods, perplexity,
dimensions, etc. By default it tries to detect the optimal number of
clusters using the elbow method.
}
