% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterSamples}
\alias{pgx.clusterSamples}
\title{Title}
\usage{
pgx.clusterSamples(
  pgx,
  X = NULL,
  skipifexists = FALSE,
  perplexity = 30,
  ntop = 1000,
  npca = 50,
  prefix = "C",
  kclust = 1,
  dims = c(2, 3),
  find.clusters = TRUE,
  clust.detect = c("louvain", "hclust"),
  row.center = TRUE,
  row.scale = FALSE,
  method = c("tsne", "umap", "pca")
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
