% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterSamples}
\alias{pgx.clusterSamples}
\title{Cluster samples in a PGX object}
\usage{
pgx.clusterSamples(
  pgx,
  X = NULL,
  skipifexists = FALSE,
  perplexity = 30,
  ntop = 1000,
  npca = 50,
  prefix = "C",
  kclust = 1,
  dims = c(2, 3),
  find.clusters = TRUE,
  clust.detect = c("louvain", "hclust"),
  row.center = TRUE,
  row.scale = FALSE,
  method = c("tsne", "umap", "pca")
)
}
\arguments{
\item{pgx}{A PGX object}

\item{perplexity}{Perplexity parameter for tSNE and UMAP}

\item{ntop}{Number of top features to use for PCA}

\item{npca}{Number of PCA components to reduce to}

\item{prefix}{Prefix for cluster variable names}

\item{kclust}{Number of clusters for k-means}

\item{dims}{Dimensions for dimensionality reduction (by default 2,3)}

\item{find.clusters}{Logical to find optimal number of clusters}

\item{method}{Character vector of clustering methods to apply (choose between "pca", "tsne", and "umap")}
}
\value{
The updated PGX object with cluster assignments in pgx$cluster$sample
}
\description{
Performs dimensionality reduction and clustering on the samples in a PGX object.
}
\details{
This function takes a PGX object and performs PCA, tSNE, and/or UMAP on the samples.
It clusters the samples in the reduced space and stores the cluster assignments in pgx$cluster$sample.

PCA is performed first to reduce dimensions. tSNE and UMAP are then applied for further reduction.
k-means clustering is performed on the reduced data to group samples into clusters.

Multiple parameters control the methods, perplexity, number of clusters, etc.
By default it tries to detect the optimal number of clusters using the elbow method.
}
