% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.clusterSamples2}
\alias{pgx.clusterSamples2}
\title{Cluster samples in a PGX object}
\usage{
pgx.clusterSamples2(
  pgx,
  methods = c("pca", "tsne", "umap"),
  dims = c(2, 3),
  reduce.sd = 1000,
  reduce.pca = 50,
  perplexity = 30,
  center.rows = TRUE,
  scale.rows = FALSE,
  X = NULL,
  umap.pkg = "uwot",
  replace.orig = TRUE
)
}
\arguments{
\item{pgx}{A PGX object}

\item{dims}{Dimensions for dimensionality reduction (by default 2,and 3)}

\item{reduce.sd}{Perform dimensionality reduction if original dims > reduce.sd}

\item{reduce.pca}{For PCA, take first reduce.pca components}

\item{perplexity}{Perplexity parameter for tSNE and UMAP}

\item{center.rows}{Center rows before clustering?}

\item{scale.rows}{Scale rows to unit variance before clustering?}

\item{X}{Use this matrix instead of pgx$X}

\item{umap.pkg}{UMAP package to use (umap or uwot)}

\item{replace.orig}{Replace original pgx$X with reduced data?}

\item{method}{Character vector of clustering methods to apply (choose between "pca", "tsne", and "umap")}
}
\value{
The updated PGX object with cluster assignments in pgx$cluster$sample
}
\description{
Performs dimensionality reduction and clustering on the samples in a PGX object.
}
\details{
This function takes a PGX object and performs PCA, tSNE, and/or UMAP on the
samples. It clusters the samples in the reduced space and stores the cluster
assignments in the pgx$cluster$sample slot.

Dimensionality reduction is performed on pgx$X or a provided matrix X. If the
original number of dimensions is large, PCA is used to reduce to reduce.pca dimensions
first. tSNE and UMAP are then applied to further reduce to 2 or 3 dimensions.

Multiple parameter options are available to control the clustering, including
choice of methods, perplexity, centering/scaling samples first, etc.

By default the original pgx$X matrix is replaced with the reduced data. Set
replace.orig=FALSE to retain the original.
}
