% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{pgx.computePGX}
\alias{pgx.computePGX}
\title{Title}
\usage{
pgx.computePGX(
  pgx,
  max.genes = 19999,
  max.genesets = 9999,
  gx.methods = c("ttest.welch", "trend.limma", "edger.qlf"),
  gset.methods = c("fisher", "gsva", "fgsea"),
  do.cluster = TRUE,
  use.design = TRUE,
  prune.samples = FALSE,
  extra.methods = c("meta.go", "deconv", "infer", "drugs", "wordcloud"),
  lib.dir,
  progress = NULL
)
}
\arguments{
\item{progress}{}
}
\description{
Title
}
