% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{pgx.computePGX}
\alias{pgx.computePGX}
\title{Compute PGX}
\usage{
pgx.computePGX(
  pgx,
  max.genes = 19999,
  max.genesets = 5000,
  gx.methods = c("ttest.welch", "trend.limma", "edger.qlf"),
  gset.methods = c("fisher", "gsva", "fgsea"),
  custom.geneset = c(gmt = NULL, info = NULL),
  do.cluster = TRUE,
  use.design = TRUE,
  prune.samples = FALSE,
  extra.methods = c("meta.go", "infer", "deconv", "drugs", "wordcloud", "wgcna")[c(1, 2)],
  libx.dir = NULL,
  progress = NULL
)
}
\arguments{
\item{pgx}{A pgx object containing the input data}

\item{max.genes}{Maximum number of genes to test. Default is 19999.}

\item{max.genesets}{Maximum number of gene sets to test. Default is 5000.}

\item{gx.methods}{Methods for differential expression analysis at the gene level. Default is c("ttest.welch", "trend.limma", "edger.qlf").}

\item{gset.methods}{Methods for differential analysis at the gene set level. Default is c("fisher", "gsva", "fgsea").}

\item{custom.geneset}{Custom gene sets to test, as a named list with gmt and info elements.}

\item{do.cluster}{Whether to run sample clustering. Default is TRUE.}

\item{use.design}{Whether to use model design matrix for testing. Default is TRUE.}

\item{prune.samples}{Whether to remove samples without valid contrasts. Default is FALSE.}

\item{extra.methods}{Additional analysis methods to run. Default is c("meta.go", "infer", "deconv", "drugs", "wordcloud", "wgcna")[c(1, 2)].}

\item{libx.dir}{Directory containing custom analysis modules.}

\item{progress}{A progress object for tracking status.}
}
\value{
An updated pgx object containing analysis results.
}
\description{
Main function to populate pgx with results. The function computes the analysis on a pgx object
}
