% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{pgx.createFromFiles}
\alias{pgx.createFromFiles}
\title{Create a pgx object}
\usage{
pgx.createFromFiles(
  count.file,
  samples.file,
  contrasts.file = NULL,
  gxmethods = "trend.limma,edger.qlf,edger.lrt",
  gsetmethods = "fisher,gsva,fgsea",
  extra = "meta.go,deconv,infer,drugs,wordcloud"
)
}
\arguments{
\item{samples.file}{a playbase::SAMPLES file}

\item{contrasts.file}{a playbase::CONTRASTS file}

\item{counts.file}{a playbase::COUNTS file}
}
\value{
list. represents a pgx object
}
\description{
This function creates a pgx object from files, which is the core object in the
OmicsPlayground.
}
