% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{pgx.createPGX}
\alias{pgx.createPGX}
\title{Create a pgx object}
\usage{
pgx.createPGX(
  counts,
  samples,
  contrasts,
  X = NULL,
  is.logx = NULL,
  batch.correct = TRUE,
  auto.scale = TRUE,
  filter.genes = TRUE,
  prune.samples = FALSE,
  only.known = TRUE,
  only.hugo = TRUE,
  convert.hugo = TRUE,
  do.cluster = TRUE,
  cluster.contrasts = FALSE,
  do.clustergenes = TRUE,
  only.proteincoding = TRUE
)
}
\arguments{
\item{counts}{dataframe.}

\item{samples}{dataframe.}

\item{contrasts}{dataframe}

\item{X}{dataframe.}

\item{is.logx}{boolean.}

\item{batch.correct}{boolean.}

\item{auto.scale}{boolean.}

\item{filter.genes}{boolean.}

\item{prune.samples}{boolean.}

\item{only.known}{boolean.}

\item{only.hugo}{boolean.}

\item{convert.hugo}{boolean.}

\item{do.cluster}{boolean.}

\item{cluster.contrasts}{boolean.}

\item{do.clustergenes}{boolean.}

\item{only.proteincoding}{boolean.}
}
\value{
list. represents a pgx object
}
\description{
This function creates a pgx object, which is the core object in the
OmicsPlayground.
}
\examples{
x <- 1
}
