% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-compute.R
\name{pgx.createPGX}
\alias{pgx.createPGX}
\title{Create a PGX object
This function creates a pgx object, which is the core object in the
OmicsPlayground.}
\usage{
pgx.createPGX(
  counts,
  samples,
  contrasts,
  X = NULL,
  is.logx = NULL,
  batch.correct = TRUE,
  species = "Human",
  auto.scale = TRUE,
  filter.genes = TRUE,
  prune.samples = FALSE,
  only.known = TRUE,
  only.hugo = TRUE,
  convert.hugo = TRUE,
  do.cluster = TRUE,
  cluster.contrasts = FALSE,
  do.clustergenes = TRUE,
  only.proteincoding = TRUE
)
}
\arguments{
\item{counts}{Matrix of count data with genes as rows and samples as columns.}

\item{samples}{Data frame containing sample information.}

\item{contrasts}{Data frame defining sample contrasts.}

\item{X}{(Optional) Matrix of normalized expression data. If NULL, will be calculated from counts.}

\item{is.logx}{Logical indicating if count matrix is already log-transformed. If NULL, guessed automatically.}

\item{batch.correct}{Logical indicating whether to perform batch correction. Default is TRUE.}

\item{species}{Default "hsapiens_gene_ensembl", it indicates the species used
for the gene annotation table and the probe to symbol conversion.}

\item{auto.scale}{Logical indicating whether to automatically scale/center genes. Default is TRUE.}

\item{filter.genes}{Logical indicating whether to filter lowly expressed genes. Default is TRUE.}

\item{prune.samples}{Logical indicating whether to remove samples without contrasts. Default is FALSE.}

\item{only.known}{Logical indicating whether to keep only known genes. Default is TRUE.}

\item{only.hugo}{Logical indicating whether to convert symbols to HUGO names. Default is TRUE.}

\item{convert.hugo}{Logical indicating whether to convert symbols to HUGO names. Default is TRUE.}

\item{do.cluster}{Logical indicating whether to run sample clustering. Default is TRUE.}

\item{cluster.contrasts}{Logical indicating whether to cluster contrasts. Default is FALSE.}

\item{do.clustergenes}{Logical indicating whether to cluster genes. Default is TRUE.}

\item{only.proteincoding}{Logical indicating whether to keep only protein-coding genes. Default is TRUE.}
}
\value{
List. PGX object containing input data and parameters.
}
\description{
Create a PGX object
This function creates a pgx object, which is the core object in the
OmicsPlayground.
}
