% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R, R/pgx-pheno.R
\name{pgx.discretizePhenotypeMatrix}
\alias{pgx.discretizePhenotypeMatrix}
\title{Discretize phenotype matrix}
\usage{
pgx.discretizePhenotypeMatrix(
  df,
  min.ncat = 2,
  max.ncat = 20,
  remove.dup = FALSE
)

pgx.discretizePhenotypeMatrix(
  df,
  min.ncat = 2,
  max.ncat = 20,
  remove.dup = FALSE
)
}
\arguments{
\item{df}{Data frame containing phenotype data}

\item{min.ncat}{Minimum number of categories for discretization. Default is 2.}

\item{max.ncat}{Maximum number of categories for discretization. Default is 20.}

\item{remove.dup}{Logical indicating whether to remove duplicate factor levels. Default is FALSE.}
}
\value{
Discretized phenotype data frame

Discretized phenotype data frame
}
\description{
Discretize phenotype matrix

Discretize phenotype matrix
}
\details{
This function discretizes the phenotype data frame \code{df} into categorical variables.
Numerical variables are converted into "high" and "low" values based on the median.
Categorical variables are kept as factors.

The number of categories is controlled by \code{min.ncat} and \code{max.ncat} parameters.
Duplicate factor levels can be removed by setting \code{remove.dup = TRUE}.

This function discretizes the phenotype data frame \code{df} into categorical variables.
Numerical variables are converted into "high" and "low" values based on the median.
Categorical variables are kept as factors.

The number of categories is controlled by \code{min.ncat} and \code{max.ncat} parameters.
Duplicate factor levels can be removed by setting \code{remove.dup = TRUE}.
}
