% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.FindClusters}
\alias{pgx.FindClusters}
\title{Find optimal sample clusters}
\usage{
pgx.FindClusters(
  X,
  method = c("kmeans", "hclust", "louvain", "meta"),
  top.sd = 1000,
  npca = 50
)
}
\arguments{
\item{X}{Numeric matrix of expression data}

\item{method}{Clustering methods to evaluate (kmeans, hclust, louvain, meta)}

\item{top.sd}{Number of top variable features to use}

\item{npca}{Number of PCA components for dimensionality reduction}
}
\value{
Optimal number of clusters
}
\description{
Finds the optimal number of clusters in the samples using multiple methods.
}
\details{
Find optimal number of sample clusters

This function takes a gene expression matrix and evaluates multiple clustering algorithms
(kmeans, hierarchical, Louvain) across a range of cluster numbers to determine the optimal number
of clusters.

The data is first reduced to the top variable features and PCA dimensions. Clustering is then
performed for a range of k clusters. The optimal k is determined by the minimum average silhouette
width across methods.
}
