% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cluster.R
\name{pgx.findLouvainClusters}
\alias{pgx.findLouvainClusters}
\title{Find Louvain clusters}
\usage{
pgx.findLouvainClusters(
  X,
  graph.method = "dist",
  level = 1,
  prefix = "C",
  knn = 100,
  gamma = 1,
  small.zero = 0.01
)
}
\arguments{
\item{X}{A data matrix, data frame, or object that can be coerced to a matrix.}

\item{graph.method}{Method for building graph from data. Either "dist" or "snn".}

\item{level}{Number of levels to build hierarchy.}

\item{prefix}{Prefix for cluster labels.}

\item{gamma}{Exponent for distance matrix.}

\item{small.zero}{Proportion of smallest clusters to reassign to 0.}
}
\value{
Vector of cluster assignments
}
\description{
Find clusters in a dataset using the Louvain algorithm.
}
\details{
Find Louvain clusters

Builds a graph from the input data and runs the Louvain algorithm to find clusters.
The number of levels in the hierarchy can be specified. Small clusters can optionally be
reassigned to cluster 0.
}
