% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.getContrastGroups}
\alias{pgx.getContrastGroups}
\title{Title}
\usage{
pgx.getContrastGroups(pgx, contrast, as.factor = TRUE)
}
\arguments{
\item{pgx}{value}

\item{contrast}{value}

\item{as.factor}{value}
}
\description{
Title
}
