% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-api.R
\name{pgx.getFamilies}
\alias{pgx.getFamilies}
\title{Title}
\usage{
pgx.getFamilies(pgx, nmin = 10, extended = FALSE)
}
\arguments{
\item{pgx}{value}

\item{nmin}{value}

\item{extended}{value}
}
\description{
Title
}
