% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-getgeo.R
\name{pgx.getGEOseries}
\alias{pgx.getGEOseries}
\title{pgx.getGEOseries}
\usage{
pgx.getGEOseries(accession, archs.h5 = "human_matrix.h5", get.info = TRUE)
}
\arguments{
\item{accession}{GEO accession}

\item{archs.h5}{Path to ARCHS4 HDF5 file containing GEO data}
}
\value{
List containing processed counts, sample metadata, and genes
}
\description{
Download and process GEO dataset
}
\details{
Downloads GEO accession ID data. First checks if the data is
in ARCHS4. If not, it tries to get from GEO. Ultimately, it checks if
it is in recount. Counts and sample matrices are aligned.
}
