% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.makeAutoContrasts}
\alias{pgx.makeAutoContrasts}
\title{Title}
\usage{
pgx.makeAutoContrasts(
  df,
  mingrp = 3,
  slen = 20,
  ref = NULL,
  fix.degenerate = FALSE,
  skip.hidden = TRUE
)
}
\arguments{
\item{skip.hidden}{}
}
\description{
Title
}
