% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.makeAutoContrastsStratified}
\alias{pgx.makeAutoContrastsStratified}
\title{Title}
\usage{
pgx.makeAutoContrastsStratified(
  df,
  strata.var,
  mingrp = 3,
  slen = 20,
  ref = NULL,
  fix.degenerate = FALSE,
  skip.hidden = TRUE
)
}
\arguments{
\item{df}{value}

\item{strata.var}{value}

\item{mingrp}{value}

\item{slen}{value}

\item{ref}{value}

\item{fix.degenerate}{value}

\item{skip.hidden}{value}
}
\description{
Title
}
