% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.makeStratifiedContrasts}
\alias{pgx.makeStratifiedContrasts}
\title{Title}
\usage{
pgx.makeStratifiedContrasts(Y, strata, ref)
}
\arguments{
\item{Y}{value}

\item{strata}{value}

\item{ref}{value}
}
\description{
Title
}
