% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-contrasts.R
\name{pgx.makeStratifiedContrastsDF}
\alias{pgx.makeStratifiedContrastsDF}
\title{Create stratified contrasts design matrix}
\usage{
pgx.makeStratifiedContrastsDF(data, vars, strata, ref)
}
\arguments{
\item{data}{Data frame containing variables and strata column}

\item{vars}{Character vector of variable names}

\item{strata}{Name of stratification variable column}

\item{ref}{Reference level for strata}
}
\value{
A design matrix for stratified contrasts
}
\description{
Generates a design matrix for stratified contrasts from a data frame.
}
\details{
This function takes a data frame containing variables of interest and a stratification
variable. It constructs a model matrix encoding contrasts between a reference level
and other levels of the stratification variable. The contrasts are computed separately
within each stratum.

The result is a design matrix with columns for the variables of interest, followed by
columns encoding the stratified contrasts. This matrix can be used as a design matrix in
differential expression analysis.
}
