% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-correct.R
\name{pgx.PC_correlation}
\alias{pgx.PC_correlation}
\title{Correlate principal components with phenotypes}
\usage{
pgx.PC_correlation(X, pheno, nv = 3, stat = "F", plot = TRUE, main = NULL)
}
\arguments{
\item{X}{Expression matrix}

\item{pheno}{Data frame of sample phenotypes}

\item{nv}{Number of principal components to use}

\item{stat}{Statistic for categorical phenotypes ("F" or "t").}

\item{plot}{Logical, whether to generate a PCA plot.}

\item{main}{Title for PCA plot.}
}
\value{
Named vector of correlation coefficients
}
\description{
Calculates the correlation between principal components of the expression matrix X
and sample phenotypes in pheno.
}
\details{
This function calculates the top nv principal components of the expression matrix X
using irlba::irlba. It then correlates each PC with each phenotype in the pheno data frame.

For categorical phenotypes, it converts to a factor and calculates the correlation with
the model matrix. For numeric phenotypes it calculates the standard correlation coefficient.

It returns a named vector of correlation coefficients, with names corresponding to
the phenotypes.
}
