% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-cna.R
\name{pgx.plotCNAHeatmap}
\alias{pgx.plotCNAHeatmap}
\title{Title}
\usage{
pgx.plotCNAHeatmap(
  ngs,
  res,
  annot = NA,
  pca.filter = -1,
  lwd = 1,
  downsample = 10,
  order.by = "clust",
  clip = 0,
  lab.cex = 0.6
)
}
\arguments{
\item{lab.cex}{}
}
\description{
Title
}
