% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{pgx.splitHeatmapFromMatrix}
\alias{pgx.splitHeatmapFromMatrix}
\title{Split heatmap from matrix}
\usage{
pgx.splitHeatmapFromMatrix(
  X,
  annot,
  idx = NULL,
  splitx = NULL,
  xtips = NULL,
  ytips = NULL,
  row_clust = TRUE,
  row_annot_width = 0.03,
  scale = "row.center",
  colors = NULL,
  lmar = 60,
  rowcex = 1,
  colcex = 1,
  show_legend = TRUE
)
}
\arguments{
\item{X}{Numeric data matrix}

\item{annot}{Data frame with row and column annotations}

\item{idx}{Rows to highlight}

\item{splitx}{Columns to split heatmap}

\item{xtips}{Custom column tooltips}

\item{ytips}{Custom row tooltips}

\item{row_clust}{Cluster rows? Default is TRUE.}

\item{row_annot_width}{Width for row annotations. Default is 0.03.}

\item{scale}{Scaling for data. Default is "row.center".}

\item{colors}{Vector of colors to use. Default is RColorBrewer Set1.}

\item{lmar}{Label margin parameter. Default is 60.}

\item{rowcex}{Row text size scaling. Default is 1.}

\item{colcex}{Column text size scaling. Default is 1.}

\item{show_legend}{Show color legend? Default is TRUE.}
}
\value{
A plotly split heatmap object
}
\description{
Split heatmap from matrix
}
\details{
This function generates an interactive split heatmap from a data matrix.
Rows and/or columns can be split according to the input annotations.
Data is scaled and clustered using hierarchical clustering by default.
Various graphical parameters like colors, labels, text sizes can be adjusted.
}
