% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-drugs.R
\name{pgx.update_drugs_results}
\alias{pgx.update_drugs_results}
\title{Update older versions of pgx$drugs with MOA, report and infographic}
\usage{
pgx.update_drugs_results(pgx, model, img_model)
}
\description{
Update older versions of pgx$drugs with MOA, report and infographic
}
