% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report.R
\name{pgx.update_infographics}
\alias{pgx.update_infographics}
\title{Update infographics. Create if missing.}
\usage{
pgx.update_infographics(
  pgx,
  llm_model,
  img_model,
  force = FALSE,
  progress = NULL
)
}
\description{
Update infographics. Create if missing.
}
