% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report.R
\name{pgx.update_reports}
\alias{pgx.update_reports}
\title{Update all reports in pgx object. Create if missing.}
\usage{
pgx.update_reports(
  pgx,
  llm_model,
  img_model = NULL,
  verbose = 1,
  force = FALSE,
  select = c("wgcna", "mofa", "cmap", "summary")
)
}
\description{
Update all reports in pgx object. Create if missing.
}
