% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgxinfo.needUpdate}
\alias{pgxinfo.needUpdate}
\title{Checks if any of metadata files (info, allFC, sigdb) need
update. All pgx files in the datafolder should be included.}
\usage{
pgxinfo.needUpdate(pgx.dir, verbose = TRUE)
}
\description{
Checks if any of metadata files (info, allFC, sigdb) need
update. All pgx files in the datafolder should be included.
}
