% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-files.R
\name{pgxinfo.read}
\alias{pgxinfo.read}
\title{Read dataset information file (aka PGX info)}
\usage{
pgxinfo.read(pgx.dir, file = "datasets-info.csv", match = TRUE)
}
\description{
Read dataset information file (aka PGX info)
}
