% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{playbase_example}
\alias{playbase_example}
\title{Get path to omp example dataset(s)}
\usage{
playbase_example(file = NULL)
}
\arguments{
\item{file}{string. Name of file. If `NULL`, the example files will
be listed.}
}
\description{
`playbase` comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access. This function was
taken from tidyverse/readr.
}
\examples{
playbase_example()
playbase_example("counts.csv")
}
