% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-plotting.R
\name{plotMethylIdeogram}
\alias{plotMethylIdeogram}
\title{plotMethylIdeogram}
\usage{
plotMethylIdeogram(beta_matrix, annot, organism = NULL, chromosomes = NULL)
}
\arguments{
\item{beta_matrix}{Matrix of beta values. Features in rows. Samples in columns.}

\item{annot}{Annotation dataframe of CpGs. Must have chromosomes and positions.}

\item{chroms}{Chromosomes to plot. Chromosomes with fewer than 10 CpG probes are removed.}
}
\description{
Creates ideograms for the selected chromosomes and regions. Plot beta values
}
