% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-pcsf.R
\name{plotPCSF}
\alias{plotPCSF}
\title{Plot PCSF graph from results object}
\usage{
plotPCSF(
  pcsf,
  highlightby = c("centrality", "prize")[1],
  plot = c("visnet", "igraph"),
  node_cex = 30,
  label_cex = 30,
  nlabel = -1
)
}
\description{
Plot PCSF graph (visnetwork or igraph) from result of
  pgx.computePCSF
}
