% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{probe2symbol}
\alias{probe2symbol}
\title{Map probe identifiers to gene symbols}
\usage{
probe2symbol(probes, type = NULL, keep.na = FALSE)
}
\arguments{
\item{probes}{Character vector of probe identifiers}

\item{type}{Character specifying the type of identifier in probes. If NULL,
the function will try to detect the identifier type automatically.}

\item{keep.na}{Logical indicating whether to keep NA symbols (TRUE) or replace
them with the original probe identifiers (FALSE).}
}
\value{
Character vector of mapped gene symbols.
}
\description{
Map probe identifiers to gene symbols
}
\examples{
\dontrun{
counts <- playbase::COUNTS
subset_genes <- round(seq(1, nrow(counts), length.out = 10))
probes <- rownames(playbase::COUNTS)[subset_genes]
symbols <- playbase::probe2symbol(probes)
}

}
