% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_read.R
\name{read_counts}
\alias{read_counts}
\title{Read counts data from file}
\usage{
read_counts(file, convert_names = TRUE)
}
\arguments{
\item{file}{string. path to file}
}
\value{
matrix. the file with the data
}
\description{
Read counts data from file
}
\examples{
counts <- read_counts(playbase::playbase_example('counts.csv'))
}
