% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-read.R
\name{read_counts}
\alias{read_counts}
\title{Read counts data from file}
\usage{
read_counts(file, convert_names = FALSE)
}
\arguments{
\item{file}{string. path to file}

\item{convert_names}{boolean.}
}
\description{
Read counts data from file
}
\details{
This function reads a count matrix using \code{playbase::read.as_matrix()},
validates it with \code{validate_counts()}, and optionally converts row names from IDs to
gene symbols using \code{playbase::probe2symbol()}.

It removes rows with NA, blank or invalid symbols, and collapses any duplicate symbols by
summing counts across rows.
}
\examples{
\dontrun{
counts <- read_counts(playbase::example_file("counts.csv"))
}
}
