% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_read.R
\name{read_expression}
\alias{read_expression}
\title{Read expression data from file}
\usage{
read_expression(file, convert_names = TRUE)
}
\arguments{
\item{file}{Path to input expression data file. Should be matrix with genes as rows and samples as columns.}

\item{convert_names}{Logical indicating whether to convert row names from IDs to gene symbols. Default is TRUE.}
}
\value{
matrix. the file with the data
}
\description{
Reads an expression data matrix from file and performs validation and preprocessing.
}
\details{
This function reads an expression matrix using playbase::read.as_matrix(),
squares the values, validates with validate_counts(), and optionally converts row names
from IDs to gene symbols using playbase::probe2symbol().

It removes rows with NA, blank or invalid symbols, and collapses duplicate symbols by
summing expression across rows.
}
\examples{
counts <- read_expression(playbase::example_file("counts.csv"))
}
