% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-preprocess.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{betaToM}
\alias{detectOutlierSamples}
\alias{getPrior}
\alias{imputeMissing}
\alias{imputeMissing.mox}
\alias{is.multiomics}
\alias{log1s}
\alias{logCPM}
\alias{mToBeta}
\alias{maxMedianNormalization}
\alias{maxSumNormalization}
\alias{mofa.get_prefix}
\alias{nmfImpute}
\alias{nmfImpute2}
\alias{normalizeExpression}
\alias{normalizeMethylation}
\alias{normalizeMultiOmics}
\alias{normalizeRLE}
\alias{normalizeTMM}
\alias{perseusImpute}
\alias{pgx.countNormalization}
\alias{plotOutlierScores}
\alias{referenceNormalization}
\alias{svdImpute2}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{playbase.preprocess}{\code{\link[playbase.preprocess]{betaToM}}, \code{\link[playbase.preprocess]{detectOutlierSamples}}, \code{\link[playbase.preprocess]{getPrior}}, \code{\link[playbase.preprocess]{imputeMissing}}, \code{\link[playbase.preprocess]{imputeMissing.mox}}, \code{\link[playbase.preprocess]{is.multiomics}}, \code{\link[playbase.preprocess]{log1s}}, \code{\link[playbase.preprocess]{logCPM}}, \code{\link[playbase.preprocess]{maxMedianNormalization}}, \code{\link[playbase.preprocess]{maxSumNormalization}}, \code{\link[playbase.preprocess]{mofa.get_prefix}}, \code{\link[playbase.preprocess]{mToBeta}}, \code{\link[playbase.preprocess]{nmfImpute}}, \code{\link[playbase.preprocess]{nmfImpute2}}, \code{\link[playbase.preprocess]{normalizeExpression}}, \code{\link[playbase.preprocess]{normalizeMethylation}}, \code{\link[playbase.preprocess]{normalizeMultiOmics}}, \code{\link[playbase.preprocess]{normalizeRLE}}, \code{\link[playbase.preprocess]{normalizeTMM}}, \code{\link[playbase.preprocess]{perseusImpute}}, \code{\link[playbase.preprocess]{pgx.countNormalization}}, \code{\link[playbase.preprocess]{plotOutlierScores}}, \code{\link[playbase.preprocess]{referenceNormalization}}, \code{\link[playbase.preprocess]{svdImpute2}}}
}}

