% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{rho2bluered}
\alias{rho2bluered}
\title{Converts correlation values [-1;1] to blue-white-red colors. Good
for creating color labels for labeledHeatmaps that expect colors.
NOTE: use WGCNA::numbers2colors???}
\usage{
rho2bluered(R, a = 1, f = 0.95)
}
\description{
Converts correlation values [-1;1] to blue-white-red colors. Good
for creating color labels for labeledHeatmaps that expect colors.
NOTE: use WGCNA::numbers2colors???
}
