% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-report.R
\name{rpt.create_full_report}
\alias{rpt.create_full_report}
\title{This is a full report of all analyses of the experiment and can
serve as data source for LLM questions.}
\usage{
rpt.create_full_report(
  pgx,
  ntop = 20,
  llm = NULL,
  sections = NULL,
  collate = FALSE
)
}
\description{
This is a full report of all analyses of the experiment and can
serve as data source for LLM questions.
}
