% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA.LeadingEdge}
\alias{run.GSEA.LeadingEdge}
\title{Title}
\usage{
run.GSEA.LeadingEdge(
  rpt.path,
  gmt,
  xmx = 10,
  rpt.label = "gsea_leadingedge",
  gsea.program = "/opt/GSEA/gsea-3.0.jar",
  outdir = "."
)
}
\arguments{
\item{rpt.path}{value}

\item{gmt}{value}

\item{xmx}{value}

\item{rpt.label}{value}

\item{gsea.program}{value}

\item{outdir}{value}
}
\description{
Title
}
