% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA.LeadingEdge}
\alias{run.GSEA.LeadingEdge}
\title{Run LeadingEdge analysis for GSEA results}
\usage{
run.GSEA.LeadingEdge(
  rpt.path,
  gmt,
  xmx = 10,
  rpt.label = "gsea_leadingedge",
  gsea.program = "/opt/GSEA/gsea-3.0.jar",
  outdir = "."
)
}
\arguments{
\item{rpt.path}{The path to the GSEA result folder.}

\item{gmt}{A character vector specifying the gene sets to analyze.}

\item{xmx}{The maximum memory allocation for the GSEA program in gigabytes (default: 10).}

\item{rpt.label}{The label for the GSEA report (default: "gsea_leadingedge").}

\item{gsea.program}{The path to the GSEA program JAR file (default: "/opt/GSEA/gsea-3.0.jar").}

\item{outdir}{The output directory where the leading edge analysis results will be stored (default: ".").}
}
\value{
The function runs the leading edge analysis and does not return any value.
}
\description{
This function runs the LeadingEdge analysis for Gene Set Enrichment Analysis (GSEA) results.
It analyzes the leading edge genes of enriched gene sets to identify key genes and their
contribution to the enrichment signal.
}
