% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA.preranked}
\alias{run.GSEA.preranked}
\title{Title}
\usage{
run.GSEA.preranked(
  rnk,
  gmt,
  output.dir = NULL,
  fdr = 0.25,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  rpt.label = "preranked",
  make.sets = TRUE,
  clean.files = TRUE,
  xmx = 10,
  scoring = "weighted",
  chip = "GENE_SYMBOL.chip",
  collapse = FALSE,
  do.leading.edge = FALSE,
  gsea.program = "/opt/GSEA/gsea-3.0.jar",
  quiet = FALSE
)
}
\arguments{
\item{rnk}{value}

\item{gmt}{value}

\item{output.dir}{value}

\item{fdr}{value}

\item{set.min}{value}

\item{set.max}{value}

\item{topgs}{value}

\item{nperm}{value}

\item{rpt.label}{value}

\item{make.sets}{value}

\item{clean.files}{value}

\item{xmx}{value}

\item{scoring}{value}

\item{chip}{value}

\item{collapse}{value}

\item{do.leading.edge}{value}

\item{gsea.program}{value}

\item{quiet}{value}
}
\description{
Title
}
