% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA}
\alias{run.GSEA}
\title{Title}
\usage{
run.GSEA(
  X,
  y,
  gmt,
  output.dir = NULL,
  fdr = 0.25,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  permute = "phenotype",
  scoring = "weighted",
  do.leading.edge = FALSE,
  rpt.label = "gsea",
  sort.y = "ref.last",
  ref.type = c("0", "NT", "REF", "DMSO", "WT", "LOW"),
  make.sets = TRUE,
  clean.files = TRUE,
  xmx = 10,
  force.permute = FALSE,
  metric = c("Signal2Noise", "Diff_of_Classes", "Pearson"),
  gsea.program = "/opt/GSEA/gsea-3.0.jar",
  quiet = FALSE
)
}
\arguments{
\item{X}{value}

\item{y}{value}

\item{gmt}{value}

\item{output.dir}{value}

\item{fdr}{value}

\item{set.min}{value}

\item{set.max}{value}

\item{topgs}{value}

\item{nperm}{value}

\item{permute}{value}

\item{scoring}{value}

\item{do.leading.edge}{value}

\item{rpt.label}{value}

\item{sort.y}{value}

\item{ref.type}{value}

\item{make.sets}{value}

\item{clean.files}{value}

\item{xmx}{value}

\item{force.permute}{value}

\item{metric}{value}

\item{gsea.program}{value}

\item{quiet}{value}
}
\description{
Title
}
