% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gset-gsea.r
\name{run.GSEA}
\alias{run.GSEA}
\title{Title}
\usage{
run.GSEA(
  X,
  y,
  gmt,
  output.dir = NULL,
  fdr = 0.25,
  set.min = 15,
  set.max = 500,
  topgs = 100,
  nperm = 1000,
  permute = "phenotype",
  scoring = "weighted",
  do.leading.edge = FALSE,
  rpt.label = "gsea",
  sort.y = "ref.last",
  ref.type = c("0", "NT", "REF", "DMSO", "WT", "LOW"),
  make.sets = TRUE,
  clean.files = TRUE,
  xmx = 10,
  force.permute = FALSE,
  metric = c("Signal2Noise", "Diff_of_Classes", "Pearson"),
  gsea.program = GSEA.JAR,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\description{
Title
}
