% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-ensembl.R
\name{showProbeTypes}
\alias{showProbeTypes}
\title{Show some probe types for selected organism}
\usage{
showProbeTypes(organism, keytypes = NULL, ah = NULL, nprobe = 10)
}
\description{
Show some probe types for selected organism
}
