% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-annot.R
\name{strip_postfix}
\alias{strip_postfix}
\title{non-greedy removal of numerical postfix. Postfix is defined as (1)
last numerical substring after - (minus), or (2) any substring
after special separators [_.].}
\usage{
strip_postfix(s)
}
\description{
non-greedy removal of numerical postfix. Postfix is defined as (1)
last numerical substring after - (minus), or (2) any substring
after special separators [_.].
}
