% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gx-util.r
\name{symbol2hugo}
\alias{symbol2hugo}
\title{Title}
\usage{
symbol2hugo(
  genes,
  remove.non.hugo = TRUE,
  silent = FALSE,
  take.only.first = FALSE,
  split.char = ";",
  unknown = "unknown_gene"
)
}
\arguments{
\item{unknown}{}
}
\description{
Title
}
