% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.computeConsensusMatrix}
\alias{wgcna.computeConsensusMatrix}
\title{Compute consensus matrix from list of matrices. The consensus
matrix checks for consistent sign and minimal threshold for each
matrix. Optionally filters on consistent p-value.}
\usage{
wgcna.computeConsensusMatrix(matlist, ydim, psig = 0.05, consfun = "min")
}
\arguments{
\item{ydim}{original dimension of data}
}
\description{
Compute consensus matrix from list of matrices. The consensus
matrix checks for consistent sign and minimal threshold for each
matrix. Optionally filters on consistent p-value.
}
