% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.computeGeneStats}
\alias{wgcna.computeGeneStats}
\title{Compute general feature statistics after WGCNA results.}
\usage{
wgcna.computeGeneStats(net, datExpr, datTraits, TOM)
}
\description{
Compute general feature statistics after WGCNA results.
}
