% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.create_infographic}
\alias{wgcna.create_infographic}
\title{Given the output from wgcna.create_report() this function creates
an infographic by calling Gemini3. The genAI model is given the
report and asked to adhere to the included or external given
diagram (in DOT format).}
\usage{
wgcna.create_infographic(
  report,
  diagram = NULL,
  prompt = NULL,
  model = "gemini-3-pro-image-preview",
  add.fallback = FALSE,
  filename = "infographic.png"
)
}
\description{
Given the output from wgcna.create_report() this function creates
an infographic by calling Gemini3. The genAI model is given the
report and asked to adhere to the included or external given
diagram (in DOT format).
}
