% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.create_report}
\alias{wgcna.create_report}
\title{Create report}
\usage{
wgcna.create_report(
  wgcna,
  ai_model,
  graph = NULL,
  annot = NULL,
  multi = FALSE,
  ntop = 100,
  topratio = 0.85,
  psig = 0.05,
  format = "markdown",
  verbose = 1,
  progress = NULL
)
}
\description{
Create report
}
