% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-report.R
\name{wgcna.getConsensusTopGenesAndSets}
\alias{wgcna.getConsensusTopGenesAndSets}
\title{Top genes/sets/phenotypes for a consensus WGCNA object.}
\usage{
wgcna.getConsensusTopGenesAndSets(
  cons,
  annot = NULL,
  module = NULL,
  ntop = 40,
  level = c("gene", "geneset")[1],
  rename = "symbol"
)
}
\description{
Top genes/sets/phenotypes for a consensus WGCNA object.
}
