% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-report.R
\name{wgcna.getMultiTopGenesAndSets}
\alias{wgcna.getMultiTopGenesAndSets}
\title{Top genes/sets/phenotypes for a multi-omics WGCNA object (layered).}
\usage{
wgcna.getMultiTopGenesAndSets(
  multi_wgcna,
  annot = NULL,
  module = NULL,
  psig = 0.05,
  ntop = 40,
  level = NULL,
  rename = "symbol"
)
}
\description{
Top genes/sets/phenotypes for a multi-omics WGCNA object (layered).
}
