% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.labels2colors}
\alias{wgcna.labels2colors}
\title{Converts WGCNA labels (numeric or color) to colors.}
\usage{
wgcna.labels2colors(colors, ...)
}
\description{
Converts WGCNA labels (numeric or color) to colors.
}
