% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.mergeME}
\alias{wgcna.mergeME}
\title{Merge multiple ME matrices into one. Allow different dimensions.}
\usage{
wgcna.mergeME(mlist, me2 = NULL, prefix = FALSE)
}
\description{
Merge multiple ME matrices into one. Allow different dimensions.
}
