% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R, R/pgx-wgcna.R
\name{wgcna.plotDendroAndTraitCorrelation_cons}
\alias{wgcna.plotDendroAndTraitCorrelation_cons}
\title{wgcna.plotDendroAndTraits for Consensus output}
\usage{
wgcna.plotDendroAndTraitCorrelation_cons(
  cons,
  show.traits = TRUE,
  show.contrasts = TRUE,
  traits = NULL,
  main = NULL,
  rm.na = TRUE,
  use.tree = 0,
  marAll = c(0.2, 8, 2, 0.2),
  setLayout = TRUE,
  ...
)

wgcna.plotDendroAndTraitCorrelation_cons(
  cons,
  show.traits = TRUE,
  show.contrasts = TRUE,
  traits = NULL,
  main = NULL,
  rm.na = TRUE,
  use.tree = 0,
  marAll = c(0.2, 8, 2, 0.2),
  setLayout = TRUE,
  ...
)
}
\description{
wgcna.plotDendroAndTraits for Consensus output

wgcna.plotDendroAndTraits for Consensus output
}
