% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.plotEigenGeneAdjacencyHeatmap}
\alias{wgcna.plotEigenGeneAdjacencyHeatmap}
\title{Plot the adjacency correlation heatmap matrix of eigengenes with or
without traits. This can show how traits cluster together with the
eigengenes.}
\usage{
wgcna.plotEigenGeneAdjacencyHeatmap(
  res,
  add_traits = TRUE,
  marx = 1,
  main = NULL
)
}
\description{
Plot the adjacency correlation heatmap matrix of eigengenes with or
without traits. This can show how traits cluster together with the
eigengenes.
}
