% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.plotEigenGeneAdjacencyHeatmap}
\alias{wgcna.plotEigenGeneAdjacencyHeatmap}
\title{Plot the adjacency correlation heatmap matrix of eigengenes with or
without traits. This can show how traits cluster together with the
eigengenes.}
\usage{
wgcna.plotEigenGeneAdjacencyHeatmap(
  wgcna,
  add_traits = TRUE,
  marx = 1,
  main = NULL,
  multi = FALSE,
  colorlabel = TRUE,
  text = FALSE,
  pstar = TRUE,
  setMargins = TRUE,
  mar1 = c(5.5, 5, 1.6, 1),
  mar2 = c(8, 10, 4, 2),
  plotDendro = TRUE,
  plotHeatmap = TRUE,
  dendro.horiz = TRUE,
  dendro.width = 0.3,
  dendro.labels = TRUE,
  method = 2,
  nmax = -1,
  mask.intra = FALSE,
  justdata = FALSE
)
}
\description{
Plot the adjacency correlation heatmap matrix of eigengenes with or
without traits. This can show how traits cluster together with the
eigengenes.
}
