% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R, R/pgx-wgcna.R
\name{wgcna.plotEigenGeneClusterDendrogram}
\alias{wgcna.plotEigenGeneClusterDendrogram}
\title{Plot cluster dendrogram with eigengenes and traits.}
\usage{
wgcna.plotEigenGeneClusterDendrogram(
  wgcna = NULL,
  ME = NULL,
  add_traits = TRUE,
  horiz = FALSE,
  setMargins = TRUE,
  method = "wgcna",
  showlabels = TRUE,
  plot = TRUE,
  multi = FALSE,
  main = NULL
)

wgcna.plotEigenGeneClusterDendrogram(
  wgcna = NULL,
  ME = NULL,
  add_traits = TRUE,
  horiz = FALSE,
  setMargins = TRUE,
  method = "wgcna",
  showlabels = TRUE,
  plot = TRUE,
  multi = FALSE,
  main = NULL
)
}
\description{
Plot cluster dendrogram with eigengenes and traits.

Plot cluster dendrogram with eigengenes and traits.
}
\seealso{
WGCNAplus::plotEigenGeneClusterDendrogram
}
