% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna.R
\name{wgcna.plotEigenGeneClusterDendrogram}
\alias{wgcna.plotEigenGeneClusterDendrogram}
\title{Plot cluster dendrogram with eigengenes and traits.}
\usage{
wgcna.plotEigenGeneClusterDendrogram(wgcna, add_traits = TRUE, main = NULL)
}
\description{
Plot cluster dendrogram with eigengenes and traits.
}
