% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R
\name{wgcna.plotModuleScores}
\alias{wgcna.plotModuleScores}
\title{Plot top modules most correlated with trait for multi expression
data.}
\usage{
wgcna.plotModuleScores(
  res,
  trait,
  multi = FALSE,
  nmax = 16,
  collapse.trait = FALSE,
  plotlib = "base",
  setpar = TRUE
)
}
\description{
Plot top modules most correlated with trait for multi expression
data.
}
