% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-wgcna-plotting.R, R/pgx-wgcna.R
\name{wgcna.plotModuleSignificance}
\alias{wgcna.plotModuleSignificance}
\title{Plot module significance.}
\usage{
wgcna.plotModuleSignificance(wgcna, trait, main = NULL, abs = FALSE)

wgcna.plotModuleSignificance(wgcna, trait, main = NULL, abs = FALSE)
}
\description{
Plot module significance.

Plot module significance.
}
\seealso{
WGCNAplus::plotModuleSignificance
}
